Download timeseries data from waterinfo.be

Using the ts_id codes and by providing a given date period, download the corresponding time series from the waterinfo.be website

Usage

get_timeseries_tsid(
  ts_id,
  period = NULL,
  from = NULL,
  to = NULL,
  datasource = 1,
  token = NULL
)

Format

A data.frame with 3 variables:

Timestamp

Datetime of the measurement.

Value

Measured value.

Quality Code

Quality code of the measurement, dependent on the data source used:

VMM Quality Code Interpretation (datasource 1)

10/110 - Excellent

30/100/130 - Good

50/150 - Moderate

70/170 - Poor

80/180 - Estimated

90/190 - Suspect

220 - Default

-1 - Missing

HIC Quality Code Interpretation (datasource 2)

40 - Good

80 - Estimated

120 - Suspect

200 - Unchecked

60 - Complete

160 - Incomplete

-1 - Missing

Aggregated timeseries

40 - Good

100 - Estimated

120 - Suspect

200 - Unchecked

-1 - Missing

The URL of the specific request is provided as a comment attribute to the returned data.frame. Use comment(df) to get the request URL.

Arguments

ts_id

waterinfo.be database ts_id, defining a timeserie variable and frequency it is defined.

period

input string according to format required by waterinfo: De period string is provided as P#Y#M#DT#H#M#S, with P defines `Period`, each # is an integer value and the codes define the number of... Y - years M - months D - days T required if information about sub-day resolution is present H - hours D - days M - minutes S - seconds Instead of D (days), the usage of W - weeks is possible as well Examples of valid period strings: P3D, P1Y, P1DT12H, PT6H, P1Y6M3DT4H20M30S.

from

date of datestring as start of the time series

to

date of datestring as end of the time series

datasource

int [0-4] defines the `meetnet` of which the measurement station is part of. VMM based stations are net '1', MOW-HIC is net '2'

token

token to use with the call (optional, can be retrieved via get_token)

Value

data.frame with the timestamps, values and quality code

Examples

get_timeseries_tsid("35055042", from = "2017-01-01", to = "2017-01-02")
#>              Timestamp Value Quality Code
#> 1  2016-12-31 23:00:00  0.00          130
#> 2  2016-12-31 23:15:00  0.00          130
#> 3  2016-12-31 23:30:00  0.00          130
#> 4  2016-12-31 23:45:00  0.00          130
#> 5  2017-01-01 00:00:00  0.00          130
#> 6  2017-01-01 00:15:00  0.00          130
#> 7  2017-01-01 00:30:00  0.00          130
#> 8  2017-01-01 00:45:00  0.00          130
#> 9  2017-01-01 01:00:00  0.00          130
#> 10 2017-01-01 01:15:00  0.00          130
#> 11 2017-01-01 01:30:00  0.00          130
#> 12 2017-01-01 01:45:00  0.00          130
#> 13 2017-01-01 02:00:00  0.00          130
#> 14 2017-01-01 02:15:00  0.00          130
#> 15 2017-01-01 02:30:00  0.00          130
#> 16 2017-01-01 02:45:00  0.00          130
#> 17 2017-01-01 03:00:00  0.00          130
#> 18 2017-01-01 03:15:00  0.00          130
#> 19 2017-01-01 03:30:00  0.00          130
#> 20 2017-01-01 03:45:00  0.00          130
#> 21 2017-01-01 04:00:00  0.00          130
#> 22 2017-01-01 04:15:00  0.00          130
#> 23 2017-01-01 04:30:00  0.00          130
#> 24 2017-01-01 04:45:00  0.00          130
#> 25 2017-01-01 05:00:00  0.00          130
#> 26 2017-01-01 05:15:00  0.00          130
#> 27 2017-01-01 05:30:00  0.00          130
#> 28 2017-01-01 05:45:00  0.00          130
#> 29 2017-01-01 06:00:00  0.00          130
#> 30 2017-01-01 06:15:00  0.00          130
#> 31 2017-01-01 06:30:00  0.00          130
#> 32 2017-01-01 06:45:00  0.00          130
#> 33 2017-01-01 07:00:00  0.00          130
#> 34 2017-01-01 07:15:00  0.00          130
#> 35 2017-01-01 07:30:00  0.00          130
#> 36 2017-01-01 07:45:00  0.00          130
#> 37 2017-01-01 08:00:00  0.00          130
#> 38 2017-01-01 08:15:00  0.00          130
#> 39 2017-01-01 08:30:00  0.00          130
#> 40 2017-01-01 08:45:00  0.00          130
#> 41 2017-01-01 09:00:00  0.00          130
#> 42 2017-01-01 09:15:00  0.00          130
#> 43 2017-01-01 09:30:00  0.00          130
#> 44 2017-01-01 09:45:00  0.00          130
#> 45 2017-01-01 10:00:00  0.00          130
#> 46 2017-01-01 10:15:00  0.00          130
#> 47 2017-01-01 10:30:00  0.00          130
#> 48 2017-01-01 10:45:00  0.00          130
#> 49 2017-01-01 11:00:00  0.00          130
#> 50 2017-01-01 11:15:00  0.00          130
#> 51 2017-01-01 11:30:00  0.00          130
#> 52 2017-01-01 11:45:00  0.00          130
#> 53 2017-01-01 12:00:00  0.00          130
#> 54 2017-01-01 12:15:00  0.00          130
#> 55 2017-01-01 12:30:00  0.00          130
#> 56 2017-01-01 12:45:00  0.00          130
#> 57 2017-01-01 13:00:00  0.00          130
#> 58 2017-01-01 13:15:00  0.00          130
#> 59 2017-01-01 13:30:00  0.00          130
#> 60 2017-01-01 13:45:00  0.00          130
#> 61 2017-01-01 14:00:00  0.00          130
#> 62 2017-01-01 14:15:00  0.00          130
#> 63 2017-01-01 14:30:00  0.00          130
#> 64 2017-01-01 14:45:00  0.00          130
#> 65 2017-01-01 15:00:00  0.00          130
#> 66 2017-01-01 15:15:00  0.00          130
#> 67 2017-01-01 15:30:00  0.00          130
#> 68 2017-01-01 15:45:00  0.00          130
#> 69 2017-01-01 16:00:00  0.00          130
#> 70 2017-01-01 16:15:00  0.00          130
#> 71 2017-01-01 16:30:00  0.00          130
#> 72 2017-01-01 16:45:00  0.00          130
#> 73 2017-01-01 17:00:00  0.00          130
#> 74 2017-01-01 17:15:00  0.00          130
#> 75 2017-01-01 17:30:00  0.00          130
#> 76 2017-01-01 17:45:00  0.03          130
#> 77 2017-01-01 18:00:00  0.00          130
#> 78 2017-01-01 18:15:00  0.04          130
#> 79 2017-01-01 18:30:00  0.00          130
#> 80 2017-01-01 18:45:00  0.03          130
#> 81 2017-01-01 19:00:00  0.00          130
#> 82 2017-01-01 19:15:00  0.03          130
#> 83 2017-01-01 19:30:00  0.03          130
#> 84 2017-01-01 19:45:00  0.00          130
#> 85 2017-01-01 20:00:00  0.03          130
#> 86 2017-01-01 20:15:00  0.00          130
#> 87 2017-01-01 20:30:00  0.00          130
#> 88 2017-01-01 20:45:00  0.00          130
#> 89 2017-01-01 21:00:00  0.00          130
#> 90 2017-01-01 21:15:00  0.03          130
#> 91 2017-01-01 21:30:00  0.00          130
#> 92 2017-01-01 21:45:00  0.00          130
#> 93 2017-01-01 22:00:00  0.00          130
#> 94 2017-01-01 22:15:00  0.11          130
#> 95 2017-01-01 22:30:00  0.82          130
#> 96 2017-01-01 22:45:00  0.69          130
#> 97 2017-01-01 23:00:00  0.34          130
get_timeseries_tsid("5156042", period = "P3D")
#>               Timestamp  Value Quality Code
#> 1   2026-08-27 08:00:00 17.449          110
#> 2   2026-08-27 08:15:00 17.449          110
#> 3   2026-08-27 08:30:00 17.449          110
#> 4   2026-08-27 08:45:00 17.449          110
#> 5   2026-08-27 09:00:00 17.449          110
#> 6   2026-08-27 09:15:00 17.450          110
#> 7   2026-08-27 09:30:00 17.450          110
#> 8   2026-08-27 09:45:00 17.450          110
#> 9   2026-08-27 10:00:00 17.450          110
#> 10  2026-08-27 10:15:00 17.450          110
#> 11  2026-08-27 10:30:00 17.449          110
#> 12  2026-08-27 10:45:00 17.449          110
#> 13  2026-08-27 11:00:00 17.449          110
#> 14  2026-08-27 11:15:00 17.449          110
#> 15  2026-08-27 11:30:00 17.449          110
#> 16  2026-08-27 11:45:00 17.449          110
#> 17  2026-08-27 12:00:00 17.449          110
#> 18  2026-08-27 12:15:00 17.449          110
#> 19  2026-08-27 12:30:00 17.449          110
#> 20  2026-08-27 12:45:00 17.449          110
#> 21  2026-08-27 13:00:00 17.449          110
#> 22  2026-08-27 13:15:00 17.449          110
#> 23  2026-08-27 13:30:00 17.448          110
#> 24  2026-08-27 13:45:00 17.448          110
#> 25  2026-08-27 14:00:00 17.448          110
#> 26  2026-08-27 14:15:00 17.448          110
#> 27  2026-08-27 14:30:00 17.448          110
#> 28  2026-08-27 14:45:00 17.448          110
#> 29  2026-08-27 15:00:00 17.448          110
#> 30  2026-08-27 15:15:00 17.447          110
#> 31  2026-08-27 15:30:00 17.447          110
#> 32  2026-08-27 15:45:00 17.447          110
#> 33  2026-08-27 16:00:00 17.447          110
#> 34  2026-08-27 16:15:00 17.446          110
#> 35  2026-08-27 16:30:00 17.446          110
#> 36  2026-08-27 16:45:00 17.446          110
#> 37  2026-08-27 17:00:00 17.446          110
#> 38  2026-08-27 17:15:00 17.446          110
#> 39  2026-08-27 17:30:00 17.446          110
#> 40  2026-08-27 17:45:00 17.445          110
#> 41  2026-08-27 18:00:00 17.445          110
#> 42  2026-08-27 18:15:00 17.445          110
#> 43  2026-08-27 18:30:00 17.445          110
#> 44  2026-08-27 18:45:00 17.445          110
#> 45  2026-08-27 19:00:00 17.444          110
#> 46  2026-08-27 19:15:00 17.444          110
#> 47  2026-08-27 19:30:00 17.444          110
#> 48  2026-08-27 19:45:00 17.443          110
#> 49  2026-08-27 20:00:00 17.444          110
#> 50  2026-08-27 20:15:00 17.452          110
#> 51  2026-08-27 20:30:00 17.457          110
#> 52  2026-08-27 20:45:00 17.460          110
#> 53  2026-08-27 21:00:00 17.461          110
#> 54  2026-08-27 21:15:00 17.461          110
#> 55  2026-08-27 21:30:00 17.460          110
#> 56  2026-08-27 21:45:00 17.459          110
#> 57  2026-08-27 22:00:00 17.458          110
#> 58  2026-08-27 22:15:00 17.458          110
#> 59  2026-08-27 22:30:00 17.457          110
#> 60  2026-08-27 22:45:00 17.457          110
#> 61  2026-08-27 23:00:00 17.457          110
#> 62  2026-08-27 23:15:00 17.456          110
#> 63  2026-08-27 23:30:00 17.456          110
#> 64  2026-08-27 23:45:00 17.455          110
#> 65  2026-08-28 00:00:00 17.455          110
#> 66  2026-08-28 00:15:00 17.455          110
#> 67  2026-08-28 00:30:00 17.456          110
#> 68  2026-08-28 00:45:00 17.456          110
#> 69  2026-08-28 01:00:00 17.456          110
#> 70  2026-08-28 01:15:00 17.456          110
#> 71  2026-08-28 01:30:00 17.456          110
#> 72  2026-08-28 01:45:00 17.456          110
#> 73  2026-08-28 02:00:00 17.456          110
#> 74  2026-08-28 02:15:00 17.456          110
#> 75  2026-08-28 02:30:00 17.456          110
#> 76  2026-08-28 02:45:00 17.456          110
#> 77  2026-08-28 03:00:00 17.456          110
#> 78  2026-08-28 03:15:00 17.456          110
#> 79  2026-08-28 03:30:00 17.456          110
#> 80  2026-08-28 03:45:00 17.457          110
#> 81  2026-08-28 04:00:00 17.456          110
#> 82  2026-08-28 04:15:00 17.456          110
#> 83  2026-08-28 04:30:00 17.456          110
#> 84  2026-08-28 04:45:00 17.456          110
#> 85  2026-08-28 05:00:00 17.455          110
#> 86  2026-08-28 05:15:00 17.455          110
#> 87  2026-08-28 05:30:00 17.454          110
#> 88  2026-08-28 05:45:00 17.454          110
#> 89  2026-08-28 06:00:00 17.453          110
#> 90  2026-08-28 06:15:00 17.453          110
#> 91  2026-08-28 06:30:00 17.452          110
#> 92  2026-08-28 06:45:00 17.452          110
#> 93  2026-08-28 07:00:00 17.452          110
#> 94  2026-08-28 07:15:00 17.452          110
#> 95  2026-08-28 07:30:00 17.451          110
#> 96  2026-08-28 07:45:00 17.450          110
#> 97  2026-08-28 08:00:00 17.450          110
#> 98  2026-08-28 08:15:00 17.450          110
#> 99  2026-08-28 08:30:00 17.450          110
#> 100 2026-08-28 08:45:00 17.450          110
#> 101 2026-08-28 09:00:00 17.449          110
#> 102 2026-08-28 09:15:00 17.449          110
#> 103 2026-08-28 09:30:00 17.450          110
#> 104 2026-08-28 09:45:00 17.450          110
#> 105 2026-08-28 10:00:00 17.450          110
#> 106 2026-08-28 10:15:00 17.450          110
#> 107 2026-08-28 10:30:00 17.451          110
#> 108 2026-08-28 10:45:00 17.452          110
#> 109 2026-08-28 11:00:00 17.453          110
#> 110 2026-08-28 11:15:00 17.453          110
#> 111 2026-08-28 11:30:00 17.455          110
#> 112 2026-08-28 11:45:00 17.458          110
#> 113 2026-08-28 12:00:00 17.462          110
#> 114 2026-08-28 12:15:00 17.466          110
#> 115 2026-08-28 12:30:00 17.471          110
#> 116 2026-08-28 12:45:00 17.476          110
#> 117 2026-08-28 13:00:00 17.481          110
#> 118 2026-08-28 13:15:00 17.487          110
#> 119 2026-08-28 13:30:00 17.493          110
#> 120 2026-08-28 13:45:00 17.500          110
#> 121 2026-08-28 14:00:00 17.505          110
#> 122 2026-08-28 14:15:00 17.518          110
#> 123 2026-08-28 14:30:00 17.528          110
#> 124 2026-08-28 14:45:00 17.536          110
#> 125 2026-08-28 15:00:00 17.541          110
#> 126 2026-08-28 15:15:00 17.544          110
#> 127 2026-08-28 15:30:00 17.546          110
#> 128 2026-08-28 15:45:00 17.548          110
#> 129 2026-08-28 16:00:00 17.549          110
#> 130 2026-08-28 16:15:00 17.549          110
#> 131 2026-08-28 16:30:00 17.550          110
#> 132 2026-08-28 16:45:00 17.555          110
#> 133 2026-08-28 17:00:00 17.562          110
#> 134 2026-08-28 17:15:00 17.566          110
#> 135 2026-08-28 17:30:00 17.569          110
#> 136 2026-08-28 17:45:00 17.571          110
#> 137 2026-08-28 18:00:00 17.571          110
#> 138 2026-08-28 18:15:00 17.570          110
#> 139 2026-08-28 18:30:00 17.568          110
#> 140 2026-08-28 18:45:00 17.565          110
#> 141 2026-08-28 19:00:00 17.562          110
#> 142 2026-08-28 19:15:00 17.558          110
#> 143 2026-08-28 19:30:00 17.555          110
#> 144 2026-08-28 19:45:00 17.553          110
#> 145 2026-08-28 20:00:00 17.551          110
#> 146 2026-08-28 20:15:00 17.549          110
#> 147 2026-08-28 20:30:00 17.549          110
#> 148 2026-08-28 20:45:00 17.548          110
#> 149 2026-08-28 21:00:00 17.547          110
#> 150 2026-08-28 21:15:00 17.547          110
#> 151 2026-08-28 21:30:00 17.547          110
#> 152 2026-08-28 21:45:00 17.546          110
#> 153 2026-08-28 22:00:00 17.547          110
#> 154 2026-08-28 22:15:00 17.547          110
#> 155 2026-08-28 22:30:00 17.547          110
#> 156 2026-08-28 22:45:00 17.547          110
#> 157 2026-08-28 23:00:00 17.547          110
#> 158 2026-08-28 23:15:00 17.547          110
#> 159 2026-08-28 23:30:00 17.548          110
#> 160 2026-08-28 23:45:00 17.548          110
#> 161 2026-08-29 00:00:00 17.548          110
#> 162 2026-08-29 00:15:00 17.548          110
#> 163 2026-08-29 00:30:00 17.549          110
#> 164 2026-08-29 00:45:00 17.549          110
#> 165 2026-08-29 01:00:00 17.550          110
#> 166 2026-08-29 01:15:00 17.550          110
#> 167 2026-08-29 01:30:00 17.551          110
#> 168 2026-08-29 01:45:00 17.551          110
#> 169 2026-08-29 02:00:00 17.552          110
#> 170 2026-08-29 02:15:00 17.552          110
#> 171 2026-08-29 02:30:00 17.553          110
#> 172 2026-08-29 02:45:00 17.554          110
#> 173 2026-08-29 03:00:00 17.555          110
#> 174 2026-08-29 03:15:00 17.556          110
#> 175 2026-08-29 03:30:00 17.557          110
#> 176 2026-08-29 03:45:00 17.558          110
#> 177 2026-08-29 04:00:00 17.559          110
#> 178 2026-08-29 04:15:00 17.560          110
#> 179 2026-08-29 04:30:00 17.561          110
#> 180 2026-08-29 04:45:00 17.562          110
#> 181 2026-08-29 05:00:00 17.563          110
#> 182 2026-08-29 05:15:00 17.564          110
#> 183 2026-08-29 05:30:00 17.566          110
#> 184 2026-08-29 05:45:00 17.567          110
#> 185 2026-08-29 06:00:00 17.568          110
#> 186 2026-08-29 06:15:00 17.569          110
#> 187 2026-08-29 06:30:00 17.571          110
#> 188 2026-08-29 06:45:00 17.572          110
#> 189 2026-08-29 07:00:00 17.573          110
#> 190 2026-08-29 07:15:00 17.574          110
#> 191 2026-08-29 07:30:00 17.575          110
#> 192 2026-08-29 07:45:00 17.576          110
#> 193 2026-08-29 08:00:00 17.577          110
#> 194 2026-08-29 08:15:00 17.578          110
#> 195 2026-08-29 08:30:00 17.579          110
#> 196 2026-08-29 08:45:00 17.580          110
#> 197 2026-08-29 09:00:00 17.580          110
#> 198 2026-08-29 09:15:00 17.581          110
#> 199 2026-08-29 09:30:00 17.581          110
#> 200 2026-08-29 09:45:00 17.582          110
#> 201 2026-08-29 10:00:00 17.583          110
#> 202 2026-08-29 10:15:00 17.584          110
#> 203 2026-08-29 10:30:00 17.585          110
#> 204 2026-08-29 10:45:00 17.585          110
#> 205 2026-08-29 11:00:00 17.586          110
#> 206 2026-08-29 11:15:00 17.586          110
#> 207 2026-08-29 11:30:00 17.587          110
#> 208 2026-08-29 11:45:00 17.588          110
#> 209 2026-08-29 12:00:00 17.588          110
#> 210 2026-08-29 12:15:00 17.588          110
#> 211 2026-08-29 12:30:00 17.588          110
#> 212 2026-08-29 12:45:00 17.589          110
#> 213 2026-08-29 13:00:00 17.589          110
#> 214 2026-08-29 13:15:00 17.590          110
#> 215 2026-08-29 13:30:00 17.590          110
#> 216 2026-08-29 13:45:00 17.591          110
#> 217 2026-08-29 14:00:00 17.591          110
#> 218 2026-08-29 14:15:00 17.592          110
#> 219 2026-08-29 14:30:00 17.592          110
#> 220 2026-08-29 14:45:00 17.592          110
#> 221 2026-08-29 15:00:00 17.592          110
#> 222 2026-08-29 15:15:00 17.592          110
#> 223 2026-08-29 15:30:00 17.592          110
#> 224 2026-08-29 15:45:00 17.592          110
#> 225 2026-08-29 16:00:00 17.592          110
#> 226 2026-08-29 16:15:00 17.592          110
#> 227 2026-08-29 16:30:00 17.592          110
#> 228 2026-08-29 16:45:00 17.592          110
#> 229 2026-08-29 17:00:00 17.592          110
#> 230 2026-08-29 17:15:00 17.592          110
#> 231 2026-08-29 17:30:00 17.591          110
#> 232 2026-08-29 17:45:00 17.591          110
#> 233 2026-08-29 18:00:00 17.591          110
#> 234 2026-08-29 18:15:00 17.591          110
#> 235 2026-08-29 18:30:00 17.590          110
#> 236 2026-08-29 18:45:00 17.590          110
#> 237 2026-08-29 19:00:00 17.590          110
#> 238 2026-08-29 19:15:00 17.590          110
#> 239 2026-08-29 19:30:00 17.590          110
#> 240 2026-08-29 19:45:00 17.589          110
#> 241 2026-08-29 20:00:00 17.589          110
#> 242 2026-08-29 20:15:00 17.588          110
#> 243 2026-08-29 20:30:00 17.588          110
#> 244 2026-08-29 20:45:00 17.587          110
#> 245 2026-08-29 21:00:00 17.587          110
#> 246 2026-08-29 21:15:00 17.586          110
#> 247 2026-08-29 21:30:00 17.585          110
#> 248 2026-08-29 21:45:00 17.585          110
#> 249 2026-08-29 22:00:00 17.585          110
#> 250 2026-08-29 22:15:00 17.584          110
#> 251 2026-08-29 22:30:00 17.582          110
#> 252 2026-08-29 22:45:00 17.582          110
#> 253 2026-08-29 23:00:00 17.582          110
#> 254 2026-08-29 23:15:00 17.581          110
#> 255 2026-08-29 23:30:00 17.580          110
#> 256 2026-08-29 23:45:00 17.580          110
#> 257 2026-08-30 00:00:00 17.579          110
#> 258 2026-08-30 00:15:00 17.578          110
#> 259 2026-08-30 00:30:00 17.577          110
#> 260 2026-08-30 00:45:00 17.577          110
#> 261 2026-08-30 01:00:00 17.576          110
#> 262 2026-08-30 01:15:00 17.574          110
#> 263 2026-08-30 01:30:00 17.573          110
#> 264 2026-08-30 01:45:00 17.572          110
#> 265 2026-08-30 02:00:00 17.572          110
#> 266 2026-08-30 02:15:00 17.571          110
#> 267 2026-08-30 02:30:00 17.570          110
#> 268 2026-08-30 02:45:00 17.569          110
#> 269 2026-08-30 03:00:00 17.568          110
#> 270 2026-08-30 03:15:00 17.567          110
#> 271 2026-08-30 03:30:00 17.567          110
#> 272 2026-08-30 03:45:00 17.566          110
#> 273 2026-08-30 04:00:00 17.565          110
#> 274 2026-08-30 04:15:00 17.565          110
#> 275 2026-08-30 04:30:00 17.564          110
#> 276 2026-08-30 04:45:00 17.563          110
#> 277 2026-08-30 05:00:00 17.563          110
#> 278 2026-08-30 05:15:00 17.562          110
#> 279 2026-08-30 05:30:00 17.561          110
#> 280 2026-08-30 05:45:00 17.561          110
#> 281 2026-08-30 06:00:00 17.560          110
#> 282 2026-08-30 06:15:00 17.560          110
#> 283 2026-08-30 06:30:00 17.559          110
#> 284 2026-08-30 06:45:00 17.559          110
#> 285 2026-08-30 07:00:00 17.558          110
#> 286 2026-08-30 07:15:00 17.557          110
#> 287 2026-08-30 07:30:00 17.557          110
#> 288 2026-08-30 07:45:00 17.557          110
get_timeseries_tsid("55419010", from = "2017-06-01", to = "2017-06-03",
                    datasource = 4)
#>               Timestamp Value Quality Code
#> 1   2017-05-31 22:00:00  3.97           11
#> 2   2017-05-31 22:10:00  3.80           11
#> 3   2017-05-31 22:20:00  3.63           11
#> 4   2017-05-31 22:30:00  3.48           11
#> 5   2017-05-31 22:40:00  3.32           11
#> 6   2017-05-31 22:50:00  3.17           11
#> 7   2017-05-31 23:00:00  3.02           11
#> 8   2017-05-31 23:10:00  2.87           11
#> 9   2017-05-31 23:20:00  2.74           11
#> 10  2017-05-31 23:30:00  2.60           11
#> 11  2017-05-31 23:40:00  2.48           11
#> 12  2017-05-31 23:50:00  2.35           11
#> 13  2017-06-01 00:00:00  2.23           11
#> 14  2017-06-01 00:10:00  2.11           11
#> 15  2017-06-01 00:20:00  1.99           11
#> 16  2017-06-01 00:30:00  1.87           11
#> 17  2017-06-01 00:40:00  1.75           11
#> 18  2017-06-01 00:50:00  1.64           11
#> 19  2017-06-01 01:00:00  1.53           11
#> 20  2017-06-01 01:10:00  1.42           11
#> 21  2017-06-01 01:20:00  1.31           11
#> 22  2017-06-01 01:30:00  1.20           11
#> 23  2017-06-01 01:40:00  1.09           11
#> 24  2017-06-01 01:50:00  0.99           11
#> 25  2017-06-01 02:00:00  0.90           11
#> 26  2017-06-01 02:10:00  0.80           11
#> 27  2017-06-01 02:20:00  0.71           11
#> 28  2017-06-01 02:30:00  0.62           11
#> 29  2017-06-01 02:40:00  0.53           11
#> 30  2017-06-01 02:50:00  0.44           11
#> 31  2017-06-01 03:00:00  0.36           11
#> 32  2017-06-01 03:10:00  0.28           11
#> 33  2017-06-01 03:20:00  0.20           11
#> 34  2017-06-01 03:30:00  0.13           11
#> 35  2017-06-01 03:40:00  0.07           11
#> 36  2017-06-01 03:50:00  0.01           11
#> 37  2017-06-01 04:00:00 -0.02           11
#> 38  2017-06-01 04:10:00  0.08           11
#> 39  2017-06-01 04:20:00  0.38           11
#> 40  2017-06-01 04:30:00  0.69           11
#> 41  2017-06-01 04:40:00  0.94           11
#> 42  2017-06-01 04:50:00  1.14           11
#> 43  2017-06-01 05:00:00  1.31           11
#> 44  2017-06-01 05:10:00  1.47           11
#> 45  2017-06-01 05:20:00  1.63           11
#> 46  2017-06-01 05:30:00  1.79           11
#> 47  2017-06-01 05:40:00  1.95           11
#> 48  2017-06-01 05:50:00  2.11           11
#> 49  2017-06-01 06:00:00  2.26           11
#> 50  2017-06-01 06:10:00  2.43           11
#> 51  2017-06-01 06:20:00  2.58           11
#> 52  2017-06-01 06:30:00  2.75           11
#> 53  2017-06-01 06:40:00  2.92           11
#> 54  2017-06-01 06:50:00  3.13           11
#> 55  2017-06-01 07:00:00  3.36           11
#> 56  2017-06-01 07:10:00  3.62           11
#> 57  2017-06-01 07:20:00  3.91           11
#> 58  2017-06-01 07:30:00  4.19           11
#> 59  2017-06-01 07:40:00  4.47           11
#> 60  2017-06-01 07:50:00  4.71           11
#> 61  2017-06-01 08:00:00  4.92           11
#> 62  2017-06-01 08:10:00  5.09           11
#> 63  2017-06-01 08:20:00  5.21           11
#> 64  2017-06-01 08:30:00  5.31           11
#> 65  2017-06-01 08:40:00  5.37           11
#> 66  2017-06-01 08:50:00  5.39           11
#> 67  2017-06-01 09:00:00  5.39           11
#> 68  2017-06-01 09:10:00  5.36           11
#> 69  2017-06-01 09:20:00  5.27           11
#> 70  2017-06-01 09:30:00  5.14           11
#> 71  2017-06-01 09:40:00  4.96           11
#> 72  2017-06-01 09:50:00  4.76           11
#> 73  2017-06-01 10:00:00  4.56           11
#> 74  2017-06-01 10:10:00  4.37           11
#> 75  2017-06-01 10:20:00  4.18           11
#> 76  2017-06-01 10:30:00  4.01           11
#> 77  2017-06-01 10:40:00  3.83           11
#> 78  2017-06-01 10:50:00  3.66           11
#> 79  2017-06-01 11:00:00  3.49           11
#> 80  2017-06-01 11:10:00  3.34           11
#> 81  2017-06-01 11:20:00  3.18           11
#> 82  2017-06-01 11:30:00  3.03           11
#> 83  2017-06-01 11:40:00  2.88           11
#> 84  2017-06-01 11:50:00  2.73           11
#> 85  2017-06-01 12:00:00  2.60           11
#> 86  2017-06-01 12:10:00  2.46           11
#> 87  2017-06-01 12:20:00  2.33           11
#> 88  2017-06-01 12:30:00  2.20           11
#> 89  2017-06-01 12:40:00  2.07           11
#> 90  2017-06-01 12:50:00  1.95           11
#> 91  2017-06-01 13:00:00  1.84           11
#> 92  2017-06-01 13:10:00  1.72           11
#> 93  2017-06-01 13:20:00  1.61           11
#> 94  2017-06-01 13:30:00  1.50           11
#> 95  2017-06-01 13:40:00  1.39           11
#> 96  2017-06-01 13:50:00  1.29           11
#> 97  2017-06-01 14:00:00  1.19           11
#> 98  2017-06-01 14:10:00  1.09           11
#> 99  2017-06-01 14:20:00  0.99           11
#> 100 2017-06-01 14:30:00  0.89           11
#> 101 2017-06-01 14:40:00  0.80           11
#> 102 2017-06-01 14:50:00  0.72           11
#> 103 2017-06-01 15:00:00  0.63           11
#> 104 2017-06-01 15:10:00  0.55           11
#> 105 2017-06-01 15:20:00  0.47           11
#> 106 2017-06-01 15:30:00  0.40           11
#> 107 2017-06-01 15:40:00  0.33           11
#> 108 2017-06-01 15:50:00  0.27           11
#> 109 2017-06-01 16:00:00  0.23           11
#> 110 2017-06-01 16:10:00  0.22           11
#> 111 2017-06-01 16:20:00  0.33           11
#> 112 2017-06-01 16:30:00  0.66           11
#> 113 2017-06-01 16:40:00  0.93           11
#> 114 2017-06-01 16:50:00  1.16           11
#> 115 2017-06-01 17:00:00  1.35           11
#> 116 2017-06-01 17:10:00  1.51           11
#> 117 2017-06-01 17:20:00  1.65           11
#> 118 2017-06-01 17:30:00  1.79           11
#> 119 2017-06-01 17:40:00  1.95           11
#> 120 2017-06-01 17:50:00  2.10           11
#> 121 2017-06-01 18:00:00  2.24           11
#> 122 2017-06-01 18:10:00  2.37           11
#> 123 2017-06-01 18:20:00  2.51           11
#> 124 2017-06-01 18:30:00  2.64           11
#> 125 2017-06-01 18:40:00  2.77           11
#> 126 2017-06-01 18:50:00  2.90           11
#> 127 2017-06-01 19:00:00  3.04           11
#> 128 2017-06-01 19:10:00  3.18           11
#> 129 2017-06-01 19:20:00  3.33           11
#> 130 2017-06-01 19:30:00  3.49           11
#> 131 2017-06-01 19:40:00  3.66           11
#> 132 2017-06-01 19:50:00  3.87           11
#> 133 2017-06-01 20:00:00  4.08           11
#> 134 2017-06-01 20:10:00  4.31           11
#> 135 2017-06-01 20:20:00  4.52           11
#> 136 2017-06-01 20:30:00  4.72           11
#> 137 2017-06-01 20:40:00  4.89           11
#> 138 2017-06-01 20:50:00  5.02           11
#> 139 2017-06-01 21:00:00  5.11           11
#> 140 2017-06-01 21:10:00  5.17           11
#> 141 2017-06-01 21:20:00  5.20           11
#> 142 2017-06-01 21:30:00  5.19           11
#> 143 2017-06-01 21:40:00  5.12           11
#> 144 2017-06-01 21:50:00  5.00           11
#> 145 2017-06-01 22:00:00  4.82           11
#> 146 2017-06-01 22:10:00  4.62           11
#> 147 2017-06-01 22:20:00  4.41           11
#> 148 2017-06-01 22:30:00  4.21           11
#> 149 2017-06-01 22:40:00  4.03           11
#> 150 2017-06-01 22:50:00  3.86           11
#> 151 2017-06-01 23:00:00  3.70           15
#> 152 2017-06-01 23:10:00  3.54           11
#> 153 2017-06-01 23:20:00  3.38           11
#> 154 2017-06-01 23:30:00  3.23           11
#> 155 2017-06-01 23:40:00  3.08           11
#> 156 2017-06-01 23:50:00  2.93           11
#> 157 2017-06-02 00:00:00  2.79           11
#> 158 2017-06-02 00:10:00  2.65           11
#> 159 2017-06-02 00:20:00  2.52           11
#> 160 2017-06-02 00:30:00  2.39           11
#> 161 2017-06-02 00:40:00  2.27           11
#> 162 2017-06-02 00:50:00  2.14           11
#> 163 2017-06-02 01:00:00  2.02           11
#> 164 2017-06-02 01:10:00  1.91           11
#> 165 2017-06-02 01:20:00  1.79           11
#> 166 2017-06-02 01:30:00  1.68           11
#> 167 2017-06-02 01:40:00  1.57           11
#> 168 2017-06-02 01:50:00  1.46           11
#> 169 2017-06-02 02:00:00  1.35           11
#> 170 2017-06-02 02:10:00  1.25           11
#> 171 2017-06-02 02:20:00  1.15           11
#> 172 2017-06-02 02:30:00  1.05           11
#> 173 2017-06-02 02:40:00  0.95           11
#> 174 2017-06-02 02:50:00  0.86           11
#> 175 2017-06-02 03:00:00  0.77           11
#> 176 2017-06-02 03:10:00  0.68           11
#> 177 2017-06-02 03:20:00  0.59           11
#> 178 2017-06-02 03:30:00  0.51           11
#> 179 2017-06-02 03:40:00  0.43           11
#> 180 2017-06-02 03:50:00  0.35           11
#> 181 2017-06-02 04:00:00  0.28           11
#> 182 2017-06-02 04:10:00  0.21           11
#> 183 2017-06-02 04:20:00  0.15           11
#> 184 2017-06-02 04:30:00  0.09           11
#> 185 2017-06-02 04:40:00  0.04           11
#> 186 2017-06-02 04:50:00  0.02           11
#> 187 2017-06-02 05:00:00  0.11           11
#> 188 2017-06-02 05:10:00  0.37           11
#> 189 2017-06-02 05:20:00  0.66           11
#> 190 2017-06-02 05:30:00  0.90           11
#> 191 2017-06-02 05:40:00  1.10           11
#> 192 2017-06-02 05:50:00  1.27           11
#> 193 2017-06-02 06:00:00  1.43           11
#> 194 2017-06-02 06:10:00  1.60           11
#> 195 2017-06-02 06:20:00  1.77           11
#> 196 2017-06-02 06:30:00  1.94           11
#> 197 2017-06-02 06:40:00  2.10           11
#> 198 2017-06-02 06:50:00  2.26           11
#> 199 2017-06-02 07:00:00  2.42           11
#> 200 2017-06-02 07:10:00  2.59           11
#> 201 2017-06-02 07:20:00  2.76           11
#> 202 2017-06-02 07:30:00  2.95           11
#> 203 2017-06-02 07:40:00  3.15           11
#> 204 2017-06-02 07:50:00  3.36           11
#> 205 2017-06-02 08:00:00  3.59           11
#> 206 2017-06-02 08:10:00  3.83           11
#> 207 2017-06-02 08:20:00  4.08           11
#> 208 2017-06-02 08:30:00  4.33           11
#> 209 2017-06-02 08:40:00  4.57           11
#> 210 2017-06-02 08:50:00  4.79           11
#> 211 2017-06-02 09:00:00  4.97           11
#> 212 2017-06-02 09:10:00  5.11           11
#> 213 2017-06-02 09:20:00  5.23           11
#> 214 2017-06-02 09:30:00  5.31           11
#> 215 2017-06-02 09:40:00  5.38           11
#> 216 2017-06-02 09:50:00  5.41           11
#> 217 2017-06-02 10:00:00  5.41           11
#> 218 2017-06-02 10:10:00  5.39           11
#> 219 2017-06-02 10:20:00  5.31           11
#> 220 2017-06-02 10:30:00  5.19           11
#> 221 2017-06-02 10:40:00  5.03           11
#> 222 2017-06-02 10:50:00  4.84           11
#> 223 2017-06-02 11:00:00  4.66           11
#> 224 2017-06-02 11:10:00  4.48           11
#> 225 2017-06-02 11:20:00  4.30           11
#> 226 2017-06-02 11:30:00  4.11           11
#> 227 2017-06-02 11:40:00  3.95           11
#> 228 2017-06-02 11:50:00  3.78           11
#> 229 2017-06-02 12:00:00  3.63           11
#> 230 2017-06-02 12:10:00  3.46           11
#> 231 2017-06-02 12:20:00  3.30           11
#> 232 2017-06-02 12:30:00  3.15           11
#> 233 2017-06-02 12:40:00  3.00           11
#> 234 2017-06-02 12:50:00  2.86           11
#> 235 2017-06-02 13:00:00  2.72           11
#> 236 2017-06-02 13:10:00  2.59           11
#> 237 2017-06-02 13:20:00  2.46           11
#> 238 2017-06-02 13:30:00  2.33           11
#> 239 2017-06-02 13:40:00  2.20           11
#> 240 2017-06-02 13:50:00  2.09           11
#> 241 2017-06-02 14:00:00  1.97           11
#> 242 2017-06-02 14:10:00  1.86           11
#> 243 2017-06-02 14:20:00  1.75           11
#> 244 2017-06-02 14:30:00  1.64           11
#> 245 2017-06-02 14:40:00  1.54           11
#> 246 2017-06-02 14:50:00  1.44           11
#> 247 2017-06-02 15:00:00  1.33           11
#> 248 2017-06-02 15:10:00  1.23           11
#> 249 2017-06-02 15:20:00  1.14           11
#> 250 2017-06-02 15:30:00  1.04           11
#> 251 2017-06-02 15:40:00  0.95           11
#> 252 2017-06-02 15:50:00  0.87           11
#> 253 2017-06-02 16:00:00  0.78           11
#> 254 2017-06-02 16:10:00  0.71           11
#> 255 2017-06-02 16:20:00  0.63           11
#> 256 2017-06-02 16:30:00  0.57           11
#> 257 2017-06-02 16:40:00  0.51           11
#> 258 2017-06-02 16:50:00  0.46           11
#> 259 2017-06-02 17:00:00  0.45           11
#> 260 2017-06-02 17:10:00  0.53           11
#> 261 2017-06-02 17:20:00  0.75           11
#> 262 2017-06-02 17:30:00  1.04           11
#> 263 2017-06-02 17:40:00  1.26           11
#> 264 2017-06-02 17:50:00  1.46           11
#> 265 2017-06-02 18:00:00  1.62           11
#> 266 2017-06-02 18:10:00  1.77           11
#> 267 2017-06-02 18:20:00  1.91           11
#> 268 2017-06-02 18:30:00  2.07           11
#> 269 2017-06-02 18:40:00  2.21           11
#> 270 2017-06-02 18:50:00  2.35           11
#> 271 2017-06-02 19:00:00  2.48           11
#> 272 2017-06-02 19:10:00  2.62           11
#> 273 2017-06-02 19:20:00  2.74           11
#> 274 2017-06-02 19:30:00  2.87           11
#> 275 2017-06-02 19:40:00  2.99           11
#> 276 2017-06-02 19:50:00  3.11           11
#> 277 2017-06-02 20:00:00  3.24           11
#> 278 2017-06-02 20:10:00  3.37           11
#> 279 2017-06-02 20:20:00  3.52           11
#> 280 2017-06-02 20:30:00  3.67           11
#> 281 2017-06-02 20:40:00  3.84           11
#> 282 2017-06-02 20:50:00  4.01           11
#> 283 2017-06-02 21:00:00  4.20           11
#> 284 2017-06-02 21:10:00  4.39           11
#> 285 2017-06-02 21:20:00  4.58           11
#> 286 2017-06-02 21:30:00  4.75           11
#> 287 2017-06-02 21:40:00  4.91           11
#> 288 2017-06-02 21:50:00  5.03           11
#> 289 2017-06-02 22:00:00  5.13           11