Get list of stations for a variable

For a given timeseriesgroup (variable), provide a list of measurement stations providing data. An overview of the variables is provided by the function supported_variables.

get_stations(variable_name = NULL, frequency = "15min", token = NULL)

Arguments

variable_name
char valid nam of available variable as timeseriesgroup

frequency
char valid frequency for the given variable, for most variables, the 15min frequency is available

token
token to use with the call (optional, can be retrieved via get_token)

Format

A data.frame with 10 variables:

ts_id

Unique timeseries identifier to access time series data corresponding to a combination of the station, measured variable and frequency.

station_latitude

Latitude coordinates of the station (WGS84)

station_longitude

Longitude coordinates of the station (WGS84)

station_id

Identifier of the station as used in the waterinfo backend

station_no

Station ID as provided on the waterinfo.be website.

station_name

Official name of the measurement station.

stationparameter_name

Station specific variable description.

parametertype_name

Measured variable description.

ts_unitsymbol

Unit of the variable.

dataprovider

Data provider of the time series data.

Value

data.frame with an overview of the available stations for the requested variable

Details

For the moment, this only works for measurement stations of VMM (meetnet 1), and stations from other measurement data sources are not included in the list

See also

Examples

get_stations('irradiance')

#>      ts_id station_latitude station_longitude station_id station_no
#> 1 78845042         51.27226          3.728299      12207   ME03_017
#> 2 78879042         50.86149          3.411318      12209   ME05_019
#> 3 78913042         50.73795          5.141976      12211   ME09_012
#> 4 78947042         51.20300          5.439589      12213   ME11_002
#> 5 78896042         50.88663          4.094898      12210   ME07_006
#> 6 78862042         51.24379          4.266912      12208   ME04_001
#> 7 78930042         51.16224          4.845708      12212   ME10_011
#>              station_name stationparameter_name parametertype_name
#> 1            Boekhoute_ME                   Rad                 Rn
#> 2              Waregem_ME                   Rad                 Rn
#> 3 Niel-bij-St.-Truiden_ME                   Rad                 Rn
#> 4             Overpelt_ME                   Rad                 Rn
#> 5           Liedekerke_ME                   Rad                 Rn
#> 6              Melsele_ME                   Rad                 Rn
#> 7            Herentals_ME                   Rad                 Rn
#>   ts_unitsymbol dataprovider
#> 1          W/m²          VMM
#> 2          W/m²          VMM
#> 3          W/m²          VMM
#> 4          W/m²          VMM
#> 5          W/m²          VMM
#> 6          W/m²          VMM
#> 7          W/m²          VMM

get_stations('soil_saturation')

#>       ts_id station_latitude station_longitude station_id station_no
#> 1  95857042         51.27226          3.728299      12207   ME03_017
#> 2  95841042         51.27226          3.728299      12207   ME03_017
#> 3  95921042         50.88663          4.094898      12210   ME07_006
#> 4  95891042         50.86149          3.411318      12209   ME05_019
#> 5 201506042         50.73795          5.141976      12211   ME09_012
#> 6  95876042         50.86149          3.411318      12209   ME05_019
#> 7  95906042         50.88663          4.094898      12210   ME07_006
#>              station_name    stationparameter_name parametertype_name
#> 1            Boekhoute_ME Bodemverzadiging(0-70cm)    Soil Saturation
#> 2            Boekhoute_ME Bodemverzadiging(0-10cm)    Soil Saturation
#> 3           Liedekerke_ME Bodemverzadiging(0-70cm)    Soil Saturation
#> 4              Waregem_ME Bodemverzadiging(0-70cm)    Soil Saturation
#> 5 Niel-bij-St.-Truiden_ME Bodemverzadiging(0-10cm)    Soil Saturation
#> 6              Waregem_ME Bodemverzadiging(0-10cm)    Soil Saturation
#> 7           Liedekerke_ME Bodemverzadiging(0-10cm)    Soil Saturation
#>   ts_unitsymbol dataprovider
#> 1             %          VMM
#> 2             %          VMM
#> 3             %          VMM
#> 4             %          VMM
#> 5             %          VMM
#> 6             %          VMM
#> 7             %          VMM