Download timeseries data from waterinfo.be

Using the ts_id codes and by providing a given date period, download the corresponding time series from the waterinfo.be website

get_timeseries_tsid(ts_id, period = NULL, from = NULL, to = NULL,
  datasource = 1, token = NULL)

Arguments

ts_id
waterinfo.be database ts_id, defining a timeserie variable and frequency it is defined.

period
input string according to format required by waterinfo: De period string is provided as P#Y#M#DT#H#M#S, with P defines `Period`, each # is an integer value and the codes define the number of... Y - years M - months D - days T required if information about sub-day resolution is present H - hours D - days M - minutes S - seconds Instead of D (days), the usage of W - weeks is possible as well Examples of valid period strings: P3D, P1Y, P1DT12H, PT6H, P1Y6M3DT4H20M30S.

from
date of datestring as start of the time series

to
date of datestring as end of the time series

datasource
int [0-4] defines the `meetnet` of which the measurement station is part of. VMM based stations are net '1', MOW-HIC is net '2'

token
token to use with the call (optional, can be retrieved via get_token)

Format

A data.frame with 3 variables:

Timestamp

Datetime of the measurement.

Value

Measured value.

Quality Code

Quality code of the measurement, dependent on the data source used:

VMM Quality Code Interpretation (datasource 1)

10/110 - Excellent

30/100/130 - Good

50/150 - Moderate

70/170 - Poor

80/180 - Estimated

90/190 - Suspect

220 - Default

-1 - Missing

HIC Quality Code Interpretation (datasource 2)

40 - Good

80 - Estimated

120 - Suspect

200 - Unchecked

60 - Complete

160 - Incomplete

-1 - Missing

Aggregated timeseries

40 - Good

100 - Estimated

120 - Suspect

200 - Unchecked

-1 - Missing

The URL of the specific request is provided as a comment attribute to the returned data.frame. Use comment(df) to get the request URL.

Value

data.frame with the timestamps, values and quality code

Examples

get_timeseries_tsid("35055042", from = "2017-01-01", to = "2017-01-02")

#>              Timestamp Value Quality Code
#> 1  2016-12-31 23:00:00  0.00          130
#> 2  2016-12-31 23:15:00  0.00          130
#> 3  2016-12-31 23:30:00  0.00          130
#> 4  2016-12-31 23:45:00  0.00          130
#> 5  2017-01-01 00:00:00  0.00          130
#> 6  2017-01-01 00:15:00  0.00          130
#> 7  2017-01-01 00:30:00  0.00          130
#> 8  2017-01-01 00:45:00  0.00          130
#> 9  2017-01-01 01:00:00  0.00          130
#> 10 2017-01-01 01:15:00  0.00          130
#> 11 2017-01-01 01:30:00  0.00          130
#> 12 2017-01-01 01:45:00  0.00          130
#> 13 2017-01-01 02:00:00  0.00          130
#> 14 2017-01-01 02:15:00  0.00          130
#> 15 2017-01-01 02:30:00  0.00          130
#> 16 2017-01-01 02:45:00  0.00          130
#> 17 2017-01-01 03:00:00  0.00          130
#> 18 2017-01-01 03:15:00  0.00          130
#> 19 2017-01-01 03:30:00  0.00          130
#> 20 2017-01-01 03:45:00  0.00          130
#> 21 2017-01-01 04:00:00  0.00          130
#> 22 2017-01-01 04:15:00  0.00          130
#> 23 2017-01-01 04:30:00  0.00          130
#> 24 2017-01-01 04:45:00  0.00          130
#> 25 2017-01-01 05:00:00  0.00          130
#> 26 2017-01-01 05:15:00  0.00          130
#> 27 2017-01-01 05:30:00  0.00          130
#> 28 2017-01-01 05:45:00  0.00          130
#> 29 2017-01-01 06:00:00  0.00          130
#> 30 2017-01-01 06:15:00  0.00          130
#> 31 2017-01-01 06:30:00  0.00          130
#> 32 2017-01-01 06:45:00  0.00          130
#> 33 2017-01-01 07:00:00  0.00          130
#> 34 2017-01-01 07:15:00  0.00          130
#> 35 2017-01-01 07:30:00  0.00          130
#> 36 2017-01-01 07:45:00  0.00          130
#> 37 2017-01-01 08:00:00  0.00          130
#> 38 2017-01-01 08:15:00  0.00          130
#> 39 2017-01-01 08:30:00  0.00          130
#> 40 2017-01-01 08:45:00  0.00          130
#> 41 2017-01-01 09:00:00  0.00          130
#> 42 2017-01-01 09:15:00  0.00          130
#> 43 2017-01-01 09:30:00  0.00          130
#> 44 2017-01-01 09:45:00  0.00          130
#> 45 2017-01-01 10:00:00  0.00          130
#> 46 2017-01-01 10:15:00  0.00          130
#> 47 2017-01-01 10:30:00  0.00          130
#> 48 2017-01-01 10:45:00  0.00          130
#> 49 2017-01-01 11:00:00  0.00          130
#> 50 2017-01-01 11:15:00  0.00          130
#> 51 2017-01-01 11:30:00  0.00          130
#> 52 2017-01-01 11:45:00  0.00          130
#> 53 2017-01-01 12:00:00  0.00          130
#> 54 2017-01-01 12:15:00  0.00          130
#> 55 2017-01-01 12:30:00  0.00          130
#> 56 2017-01-01 12:45:00  0.00          130
#> 57 2017-01-01 13:00:00  0.00          130
#> 58 2017-01-01 13:15:00  0.00          130
#> 59 2017-01-01 13:30:00  0.00          130
#> 60 2017-01-01 13:45:00  0.00          130
#> 61 2017-01-01 14:00:00  0.00          130
#> 62 2017-01-01 14:15:00  0.00          130
#> 63 2017-01-01 14:30:00  0.00          130
#> 64 2017-01-01 14:45:00  0.00          130
#> 65 2017-01-01 15:00:00  0.00          130
#> 66 2017-01-01 15:15:00  0.00          130
#> 67 2017-01-01 15:30:00  0.00          130
#> 68 2017-01-01 15:45:00  0.00          130
#> 69 2017-01-01 16:00:00  0.00          130
#> 70 2017-01-01 16:15:00  0.00          130
#> 71 2017-01-01 16:30:00  0.00          130
#> 72 2017-01-01 16:45:00  0.00          130
#> 73 2017-01-01 17:00:00  0.00          130
#> 74 2017-01-01 17:15:00  0.00          130
#> 75 2017-01-01 17:30:00  0.00          130
#> 76 2017-01-01 17:45:00  0.03          130
#> 77 2017-01-01 18:00:00  0.00          130
#> 78 2017-01-01 18:15:00  0.04          130
#> 79 2017-01-01 18:30:00  0.00          130
#> 80 2017-01-01 18:45:00  0.03          130
#> 81 2017-01-01 19:00:00  0.00          130
#> 82 2017-01-01 19:15:00  0.03          130
#> 83 2017-01-01 19:30:00  0.03          130
#> 84 2017-01-01 19:45:00  0.00          130
#> 85 2017-01-01 20:00:00  0.03          130
#> 86 2017-01-01 20:15:00  0.00          130
#> 87 2017-01-01 20:30:00  0.00          130
#> 88 2017-01-01 20:45:00  0.00          130
#> 89 2017-01-01 21:00:00  0.00          130
#> 90 2017-01-01 21:15:00  0.03          130
#> 91 2017-01-01 21:30:00  0.00          130
#> 92 2017-01-01 21:45:00  0.00          130
#> 93 2017-01-01 22:00:00  0.00          130
#> 94 2017-01-01 22:15:00  0.11          130
#> 95 2017-01-01 22:30:00  0.82          130
#> 96 2017-01-01 22:45:00  0.69          130
#> 97 2017-01-01 23:00:00  0.34          130

get_timeseries_tsid("5156042", period = "P3D")

#>               Timestamp  Value Quality Code
#> 1   2021-06-28 18:00:00 18.039          110
#> 2   2021-06-28 18:15:00 18.039          110
#> 3   2021-06-28 18:30:00 18.039          110
#> 4   2021-06-28 18:45:00 18.039          110
#> 5   2021-06-28 19:00:00 18.039          110
#> 6   2021-06-28 19:15:00 18.039          110
#> 7   2021-06-28 19:30:00 18.039          110
#> 8   2021-06-28 19:45:00 18.039          110
#> 9   2021-06-28 20:00:00 18.039          110
#> 10  2021-06-28 20:15:00 18.039          110
#> 11  2021-06-28 20:30:00 18.039          110
#> 12  2021-06-28 20:45:00 18.040          110
#> 13  2021-06-28 21:00:00 18.040          110
#> 14  2021-06-28 21:15:00 18.040          110
#> 15  2021-06-28 21:30:00 18.040          110
#> 16  2021-06-28 21:45:00 18.040          110
#> 17  2021-06-28 22:00:00 18.041          110
#> 18  2021-06-28 22:15:00 18.041          110
#> 19  2021-06-28 22:30:00 18.041          110
#> 20  2021-06-28 22:45:00 18.041          110
#> 21  2021-06-28 23:00:00 18.042          110
#> 22  2021-06-28 23:15:00 18.042          110
#> 23  2021-06-28 23:30:00 18.042          110
#> 24  2021-06-28 23:45:00 18.042          110
#> 25  2021-06-29 00:00:00 18.043          110
#> 26  2021-06-29 00:15:00 18.043          110
#> 27  2021-06-29 00:30:00 18.043          110
#> 28  2021-06-29 00:45:00 18.044          110
#> 29  2021-06-29 01:00:00 18.044          110
#> 30  2021-06-29 01:15:00 18.044          110
#> 31  2021-06-29 01:30:00 18.044          110
#> 32  2021-06-29 01:45:00 18.045          110
#> 33  2021-06-29 02:00:00 18.045          110
#> 34  2021-06-29 02:15:00 18.046          110
#> 35  2021-06-29 02:30:00 18.046          110
#> 36  2021-06-29 02:45:00 18.046          110
#> 37  2021-06-29 03:00:00 18.046          110
#> 38  2021-06-29 03:15:00 18.046          110
#> 39  2021-06-29 03:30:00 18.046          110
#> 40  2021-06-29 03:45:00 18.047          110
#> 41  2021-06-29 04:00:00 18.047          110
#> 42  2021-06-29 04:15:00 18.047          110
#> 43  2021-06-29 04:30:00 18.046          110
#> 44  2021-06-29 04:45:00 18.047          110
#> 45  2021-06-29 05:00:00 18.047          110
#> 46  2021-06-29 05:15:00 18.047          110
#> 47  2021-06-29 05:30:00 18.047          110
#> 48  2021-06-29 05:45:00 18.048          110
#> 49  2021-06-29 06:00:00 18.048          110
#> 50  2021-06-29 06:15:00 18.048          110
#> 51  2021-06-29 06:30:00 18.048          110
#> 52  2021-06-29 06:45:00 18.048          110
#> 53  2021-06-29 07:00:00 18.048          110
#> 54  2021-06-29 07:15:00 18.049          110
#> 55  2021-06-29 07:30:00 18.049          110
#> 56  2021-06-29 07:45:00 18.048          110
#> 57  2021-06-29 08:00:00 18.048          110
#> 58  2021-06-29 08:15:00 18.049          110
#> 59  2021-06-29 08:30:00 18.048          110
#> 60  2021-06-29 08:45:00 18.048          110
#> 61  2021-06-29 09:00:00 18.048          110
#> 62  2021-06-29 09:15:00 18.048          110
#> 63  2021-06-29 09:30:00 18.048          110
#> 64  2021-06-29 09:45:00 18.048          110
#> 65  2021-06-29 10:00:00 18.048          110
#> 66  2021-06-29 10:15:00 18.048          110
#> 67  2021-06-29 10:30:00 18.048          110
#> 68  2021-06-29 10:45:00 18.048          110
#> 69  2021-06-29 11:00:00 18.048          110
#> 70  2021-06-29 11:15:00 18.048          110
#> 71  2021-06-29 11:30:00 18.048          110
#> 72  2021-06-29 11:45:00 18.048          110
#> 73  2021-06-29 12:00:00 18.047          110
#> 74  2021-06-29 12:15:00 18.047          110
#> 75  2021-06-29 12:30:00 18.047          110
#> 76  2021-06-29 12:45:00 18.047          110
#> 77  2021-06-29 13:00:00 18.047          110
#> 78  2021-06-29 13:15:00 18.047          110
#> 79  2021-06-29 13:30:00 18.047          110
#> 80  2021-06-29 13:45:00 18.046          110
#> 81  2021-06-29 14:00:00 18.047          110
#> 82  2021-06-29 14:15:00 18.046          110
#> 83  2021-06-29 14:30:00 18.046          110
#> 84  2021-06-29 14:45:00 18.046          110
#> 85  2021-06-29 15:00:00 18.046          110
#> 86  2021-06-29 15:15:00 18.046          110
#> 87  2021-06-29 15:30:00 18.046          110
#> 88  2021-06-29 15:45:00 18.047          110
#> 89  2021-06-29 16:00:00 18.046          110
#> 90  2021-06-29 16:15:00 18.046          110
#> 91  2021-06-29 16:30:00 18.046          110
#> 92  2021-06-29 16:45:00 18.046          110
#> 93  2021-06-29 17:00:00 18.046          110
#> 94  2021-06-29 17:15:00 18.046          110
#> 95  2021-06-29 17:30:00 18.046          110
#> 96  2021-06-29 17:45:00 18.046          110
#> 97  2021-06-29 18:00:00 18.045          110
#> 98  2021-06-29 18:15:00 18.045          110
#> 99  2021-06-29 18:30:00 18.045          110
#> 100 2021-06-29 18:45:00 18.045          110
#> 101 2021-06-29 19:00:00 18.044          110
#> 102 2021-06-29 19:15:00 18.044          110
#> 103 2021-06-29 19:30:00 18.044          110
#> 104 2021-06-29 19:45:00 18.044          110
#> 105 2021-06-29 20:00:00 18.044          110
#> 106 2021-06-29 20:15:00 18.044          110
#> 107 2021-06-29 20:30:00 18.044          110
#> 108 2021-06-29 20:45:00 18.044          110
#> 109 2021-06-29 21:00:00 18.044          110
#> 110 2021-06-29 21:15:00 18.045          110
#> 111 2021-06-29 21:30:00 18.045          110
#> 112 2021-06-29 21:45:00 18.045          110
#> 113 2021-06-29 22:00:00 18.045          110
#> 114 2021-06-29 22:15:00 18.045          110
#> 115 2021-06-29 22:30:00 18.045          110
#> 116 2021-06-29 22:45:00 18.045          110
#> 117 2021-06-29 23:00:00 18.045          110
#> 118 2021-06-29 23:15:00 18.045          110
#> 119 2021-06-29 23:30:00 18.045          110
#> 120 2021-06-29 23:45:00 18.045          110
#> 121 2021-06-30 00:00:00 18.046          110
#> 122 2021-06-30 00:15:00 18.046          110
#> 123 2021-06-30 00:30:00 18.046          110
#> 124 2021-06-30 00:45:00 18.047          110
#> 125 2021-06-30 01:00:00 18.047          110
#> 126 2021-06-30 01:15:00 18.047          110
#> 127 2021-06-30 01:30:00 18.047          110
#> 128 2021-06-30 01:45:00 18.048          110
#> 129 2021-06-30 02:00:00 18.049          110
#> 130 2021-06-30 02:15:00 18.050          110
#> 131 2021-06-30 02:30:00 18.051          110
#> 132 2021-06-30 02:45:00 18.053          110
#> 133 2021-06-30 03:00:00 18.054          110
#> 134 2021-06-30 03:15:00 18.055          110
#> 135 2021-06-30 03:30:00 18.056          110
#> 136 2021-06-30 03:45:00 18.057          110
#> 137 2021-06-30 04:00:00 18.057          110
#> 138 2021-06-30 04:15:00 18.058          110
#> 139 2021-06-30 04:30:00 18.058          110
#> 140 2021-06-30 04:45:00 18.060          110
#> 141 2021-06-30 05:00:00 18.061          110
#> 142 2021-06-30 05:15:00 18.062          110
#> 143 2021-06-30 05:30:00 18.063          110
#> 144 2021-06-30 05:45:00 18.064          110
#> 145 2021-06-30 06:00:00 18.066          110
#> 146 2021-06-30 06:15:00 18.072          110
#> 147 2021-06-30 06:30:00 18.085          110
#> 148 2021-06-30 06:45:00 18.101          110
#> 149 2021-06-30 07:00:00 18.115          110
#> 150 2021-06-30 07:15:00 18.127          110
#> 151 2021-06-30 07:30:00 18.135          110
#> 152 2021-06-30 07:45:00 18.141          110
#> 153 2021-06-30 08:00:00 18.144          110
#> 154 2021-06-30 08:15:00 18.146          110
#> 155 2021-06-30 08:30:00 18.146          110
#> 156 2021-06-30 08:45:00 18.145          110
#> 157 2021-06-30 09:00:00 18.144          110
#> 158 2021-06-30 09:15:00 18.143          110
#> 159 2021-06-30 09:30:00 18.143          110
#> 160 2021-06-30 09:45:00 18.142          110
#> 161 2021-06-30 10:00:00 18.142          110
#> 162 2021-06-30 10:15:00 18.142          110
#> 163 2021-06-30 10:30:00 18.142          110
#> 164 2021-06-30 10:45:00 18.142          110
#> 165 2021-06-30 11:00:00 18.142          110
#> 166 2021-06-30 11:15:00 18.142          110
#> 167 2021-06-30 11:30:00 18.141          110
#> 168 2021-06-30 11:45:00 18.140          110
#> 169 2021-06-30 12:00:00 18.139          110
#> 170 2021-06-30 12:15:00 18.139          110
#> 171 2021-06-30 12:30:00 18.138          110
#> 172 2021-06-30 12:45:00 18.138          110
#> 173 2021-06-30 13:00:00 18.139          110
#> 174 2021-06-30 13:15:00 18.139          110
#> 175 2021-06-30 13:30:00 18.140          110
#> 176 2021-06-30 13:45:00 18.142          110
#> 177 2021-06-30 14:00:00 18.143          110
#> 178 2021-06-30 14:15:00 18.144          110
#> 179 2021-06-30 14:30:00 18.146          110
#> 180 2021-06-30 14:45:00 18.148          110
#> 181 2021-06-30 15:00:00 18.149          110
#> 182 2021-06-30 15:15:00 18.151          110
#> 183 2021-06-30 15:30:00 18.153          110
#> 184 2021-06-30 15:45:00 18.154          110
#> 185 2021-06-30 16:00:00 18.156          110
#> 186 2021-06-30 16:15:00 18.157          110
#> 187 2021-06-30 16:30:00 18.159          110
#> 188 2021-06-30 16:45:00 18.161          110
#> 189 2021-06-30 17:00:00 18.162          110
#> 190 2021-06-30 17:15:00 18.164          110
#> 191 2021-06-30 17:30:00 18.165          110
#> 192 2021-06-30 17:45:00 18.166          110
#> 193 2021-06-30 18:00:00 18.167          110
#> 194 2021-06-30 18:15:00 18.169          110
#> 195 2021-06-30 18:30:00 18.169          110
#> 196 2021-06-30 18:45:00 18.171          110
#> 197 2021-06-30 19:00:00 18.172          110
#> 198 2021-06-30 19:15:00 18.172          110
#> 199 2021-06-30 19:30:00 18.173          110
#> 200 2021-06-30 19:45:00 18.174          110
#> 201 2021-06-30 20:00:00 18.175          110
#> 202 2021-06-30 20:15:00 18.176          110
#> 203 2021-06-30 20:30:00 18.177          110
#> 204 2021-06-30 20:45:00 18.178          110
#> 205 2021-06-30 21:00:00 18.179          110
#> 206 2021-06-30 21:15:00 18.180          110
#> 207 2021-06-30 21:30:00 18.181          110
#> 208 2021-06-30 21:45:00 18.183          110
#> 209 2021-06-30 22:00:00 18.183          110
#> 210 2021-06-30 22:15:00 18.185          110
#> 211 2021-06-30 22:30:00 18.185          110
#> 212 2021-06-30 22:45:00 18.186          110
#> 213 2021-06-30 23:00:00 18.187          110
#> 214 2021-06-30 23:15:00 18.188          110
#> 215 2021-06-30 23:30:00 18.189          110
#> 216 2021-06-30 23:45:00 18.189          110
#> 217 2021-07-01 00:00:00 18.190          110
#> 218 2021-07-01 00:15:00 18.191          110
#> 219 2021-07-01 00:30:00 18.192          110
#> 220 2021-07-01 00:45:00 18.193          110
#> 221 2021-07-01 01:00:00 18.194          110
#> 222 2021-07-01 01:15:00 18.194          110
#> 223 2021-07-01 01:30:00 18.195          110
#> 224 2021-07-01 01:45:00 18.196          110
#> 225 2021-07-01 02:00:00 18.197          110
#> 226 2021-07-01 02:15:00 18.197          110
#> 227 2021-07-01 02:30:00 18.198          110
#> 228 2021-07-01 02:45:00 18.200          110
#> 229 2021-07-01 03:00:00 18.200          110
#> 230 2021-07-01 03:15:00 18.201          110
#> 231 2021-07-01 03:30:00 18.202          110
#> 232 2021-07-01 03:45:00 18.203          110
#> 233 2021-07-01 04:00:00 18.204          110
#> 234 2021-07-01 04:15:00 18.205          110
#> 235 2021-07-01 04:30:00 18.206          110
#> 236 2021-07-01 04:45:00 18.207          110
#> 237 2021-07-01 05:00:00 18.208          110
#> 238 2021-07-01 05:15:00 18.208          110
#> 239 2021-07-01 05:30:00 18.209          110
#> 240 2021-07-01 05:45:00 18.210          110
#> 241 2021-07-01 06:00:00 18.211          110
#> 242 2021-07-01 06:15:00 18.212          110
#> 243 2021-07-01 06:30:00 18.213          110
#> 244 2021-07-01 06:45:00 18.214          110
#> 245 2021-07-01 07:00:00 18.215          110
#> 246 2021-07-01 07:15:00 18.216          110
#> 247 2021-07-01 07:30:00 18.217          110
#> 248 2021-07-01 07:45:00 18.218          110
#> 249 2021-07-01 08:00:00 18.219          110
#> 250 2021-07-01 08:15:00 18.220          110
#> 251 2021-07-01 08:30:00 18.221          110
#> 252 2021-07-01 08:45:00 18.222          110
#> 253 2021-07-01 09:00:00 18.223          110
#> 254 2021-07-01 09:15:00 18.224          110
#> 255 2021-07-01 09:30:00 18.225          110
#> 256 2021-07-01 09:45:00 18.226          110
#> 257 2021-07-01 10:00:00 18.227          110
#> 258 2021-07-01 10:15:00 18.228          110
#> 259 2021-07-01 10:30:00 18.229          110
#> 260 2021-07-01 10:45:00 18.230          110
#> 261 2021-07-01 11:00:00 18.231          110
#> 262 2021-07-01 11:15:00 18.231          110
#> 263 2021-07-01 11:30:00 18.233          110
#> 264 2021-07-01 11:45:00 18.234          110
#> 265 2021-07-01 12:00:00 18.234          110
#> 266 2021-07-01 12:15:00 18.235          110
#> 267 2021-07-01 12:30:00 18.236          110
#> 268 2021-07-01 12:45:00 18.236          110
#> 269 2021-07-01 13:00:00 18.237          110
#> 270 2021-07-01 13:15:00 18.238          110
#> 271 2021-07-01 13:30:00 18.238          110
#> 272 2021-07-01 13:45:00 18.239          110
#> 273 2021-07-01 14:00:00 18.240          110
#> 274 2021-07-01 14:15:00 18.241          110
#> 275 2021-07-01 14:30:00 18.241          110
#> 276 2021-07-01 14:45:00 18.242          110
#> 277 2021-07-01 15:00:00 18.242          110
#> 278 2021-07-01 15:15:00 18.243          110
#> 279 2021-07-01 15:30:00 18.244          110
#> 280 2021-07-01 15:45:00 18.245          110
#> 281 2021-07-01 16:00:00 18.245          110
#> 282 2021-07-01 16:15:00 18.246          110
#> 283 2021-07-01 16:30:00 18.246          110
#> 284 2021-07-01 16:45:00 18.247          110
#> 285 2021-07-01 17:00:00 18.248          110
#> 286 2021-07-01 17:15:00 18.248          110
#> 287 2021-07-01 17:30:00 18.249          110

get_timeseries_tsid("55419010", from = "2017-06-01", to = "2017-06-03",
                    datasource = 4)

#>               Timestamp Value Quality Code
#> 1   2017-05-31 22:00:00  3.97           11
#> 2   2017-05-31 22:10:00  3.80           11
#> 3   2017-05-31 22:20:00  3.63           11
#> 4   2017-05-31 22:30:00  3.48           11
#> 5   2017-05-31 22:40:00  3.32           11
#> 6   2017-05-31 22:50:00  3.17           11
#> 7   2017-05-31 23:00:00  3.02           11
#> 8   2017-05-31 23:10:00  2.87           11
#> 9   2017-05-31 23:20:00  2.74           11
#> 10  2017-05-31 23:30:00  2.60           11
#> 11  2017-05-31 23:40:00  2.48           11
#> 12  2017-05-31 23:50:00  2.35           11
#> 13  2017-06-01 00:00:00  2.23           11
#> 14  2017-06-01 00:10:00  2.11           11
#> 15  2017-06-01 00:20:00  1.99           11
#> 16  2017-06-01 00:30:00  1.87           11
#> 17  2017-06-01 00:40:00  1.75           11
#> 18  2017-06-01 00:50:00  1.64           11
#> 19  2017-06-01 01:00:00  1.53           11
#> 20  2017-06-01 01:10:00  1.42           11
#> 21  2017-06-01 01:20:00  1.31           11
#> 22  2017-06-01 01:30:00  1.20           11
#> 23  2017-06-01 01:40:00  1.09           11
#> 24  2017-06-01 01:50:00  0.99           11
#> 25  2017-06-01 02:00:00  0.90           11
#> 26  2017-06-01 02:10:00  0.80           11
#> 27  2017-06-01 02:20:00  0.71           11
#> 28  2017-06-01 02:30:00  0.62           11
#> 29  2017-06-01 02:40:00  0.53           11
#> 30  2017-06-01 02:50:00  0.44           11
#> 31  2017-06-01 03:00:00  0.36           11
#> 32  2017-06-01 03:10:00  0.28           11
#> 33  2017-06-01 03:20:00  0.20           11
#> 34  2017-06-01 03:30:00  0.13           11
#> 35  2017-06-01 03:40:00  0.07           11
#> 36  2017-06-01 03:50:00  0.01           11
#> 37  2017-06-01 04:00:00 -0.02           11
#> 38  2017-06-01 04:10:00  0.08           11
#> 39  2017-06-01 04:20:00  0.38           11
#> 40  2017-06-01 04:30:00  0.69           11
#> 41  2017-06-01 04:40:00  0.94           11
#> 42  2017-06-01 04:50:00  1.14           11
#> 43  2017-06-01 05:00:00  1.31           11
#> 44  2017-06-01 05:10:00  1.47           11
#> 45  2017-06-01 05:20:00  1.63           11
#> 46  2017-06-01 05:30:00  1.79           11
#> 47  2017-06-01 05:40:00  1.95           11
#> 48  2017-06-01 05:50:00  2.11           11
#> 49  2017-06-01 06:00:00  2.26           11
#> 50  2017-06-01 06:10:00  2.43           11
#> 51  2017-06-01 06:20:00  2.58           11
#> 52  2017-06-01 06:30:00  2.75           11
#> 53  2017-06-01 06:40:00  2.92           11
#> 54  2017-06-01 06:50:00  3.13           11
#> 55  2017-06-01 07:00:00  3.36           11
#> 56  2017-06-01 07:10:00  3.62           11
#> 57  2017-06-01 07:20:00  3.91           11
#> 58  2017-06-01 07:30:00  4.19           11
#> 59  2017-06-01 07:40:00  4.47           11
#> 60  2017-06-01 07:50:00  4.71           11
#> 61  2017-06-01 08:00:00  4.92           11
#> 62  2017-06-01 08:10:00  5.09           11
#> 63  2017-06-01 08:20:00  5.21           11
#> 64  2017-06-01 08:30:00  5.31           11
#> 65  2017-06-01 08:40:00  5.37           11
#> 66  2017-06-01 08:50:00  5.39           11
#> 67  2017-06-01 09:00:00  5.39           11
#> 68  2017-06-01 09:10:00  5.36           11
#> 69  2017-06-01 09:20:00  5.27           11
#> 70  2017-06-01 09:30:00  5.14           11
#> 71  2017-06-01 09:40:00  4.96           11
#> 72  2017-06-01 09:50:00  4.76           11
#> 73  2017-06-01 10:00:00  4.56           11
#> 74  2017-06-01 10:10:00  4.37           11
#> 75  2017-06-01 10:20:00  4.18           11
#> 76  2017-06-01 10:30:00  4.01           11
#> 77  2017-06-01 10:40:00  3.83           11
#> 78  2017-06-01 10:50:00  3.66           11
#> 79  2017-06-01 11:00:00  3.49           11
#> 80  2017-06-01 11:10:00  3.34           11
#> 81  2017-06-01 11:20:00  3.18           11
#> 82  2017-06-01 11:30:00  3.03           11
#> 83  2017-06-01 11:40:00  2.88           11
#> 84  2017-06-01 11:50:00  2.73           11
#> 85  2017-06-01 12:00:00  2.60           11
#> 86  2017-06-01 12:10:00  2.46           11
#> 87  2017-06-01 12:20:00  2.33           11
#> 88  2017-06-01 12:30:00  2.20           11
#> 89  2017-06-01 12:40:00  2.07           11
#> 90  2017-06-01 12:50:00  1.95           11
#> 91  2017-06-01 13:00:00  1.84           11
#> 92  2017-06-01 13:10:00  1.72           11
#> 93  2017-06-01 13:20:00  1.61           11
#> 94  2017-06-01 13:30:00  1.50           11
#> 95  2017-06-01 13:40:00  1.39           11
#> 96  2017-06-01 13:50:00  1.29           11
#> 97  2017-06-01 14:00:00  1.19           11
#> 98  2017-06-01 14:10:00  1.09           11
#> 99  2017-06-01 14:20:00  0.99           11
#> 100 2017-06-01 14:30:00  0.89           11
#> 101 2017-06-01 14:40:00  0.80           11
#> 102 2017-06-01 14:50:00  0.72           11
#> 103 2017-06-01 15:00:00  0.63           11
#> 104 2017-06-01 15:10:00  0.55           11
#> 105 2017-06-01 15:20:00  0.47           11
#> 106 2017-06-01 15:30:00  0.40           11
#> 107 2017-06-01 15:40:00  0.33           11
#> 108 2017-06-01 15:50:00  0.27           11
#> 109 2017-06-01 16:00:00  0.23           11
#> 110 2017-06-01 16:10:00  0.22           11
#> 111 2017-06-01 16:20:00  0.33           11
#> 112 2017-06-01 16:30:00  0.66           11
#> 113 2017-06-01 16:40:00  0.93           11
#> 114 2017-06-01 16:50:00  1.16           11
#> 115 2017-06-01 17:00:00  1.35           11
#> 116 2017-06-01 17:10:00  1.51           11
#> 117 2017-06-01 17:20:00  1.65           11
#> 118 2017-06-01 17:30:00  1.79           11
#> 119 2017-06-01 17:40:00  1.95           11
#> 120 2017-06-01 17:50:00  2.10           11
#> 121 2017-06-01 18:00:00  2.24           11
#> 122 2017-06-01 18:10:00  2.37           11
#> 123 2017-06-01 18:20:00  2.51           11
#> 124 2017-06-01 18:30:00  2.64           11
#> 125 2017-06-01 18:40:00  2.77           11
#> 126 2017-06-01 18:50:00  2.90           11
#> 127 2017-06-01 19:00:00  3.04           11
#> 128 2017-06-01 19:10:00  3.18           11
#> 129 2017-06-01 19:20:00  3.33           11
#> 130 2017-06-01 19:30:00  3.49           11
#> 131 2017-06-01 19:40:00  3.66           11
#> 132 2017-06-01 19:50:00  3.87           11
#> 133 2017-06-01 20:00:00  4.08           11
#> 134 2017-06-01 20:10:00  4.31           11
#> 135 2017-06-01 20:20:00  4.52           11
#> 136 2017-06-01 20:30:00  4.72           11
#> 137 2017-06-01 20:40:00  4.89           11
#> 138 2017-06-01 20:50:00  5.02           11
#> 139 2017-06-01 21:00:00  5.11           11
#> 140 2017-06-01 21:10:00  5.17           11
#> 141 2017-06-01 21:20:00  5.20           11
#> 142 2017-06-01 21:30:00  5.19           11
#> 143 2017-06-01 21:40:00  5.12           11
#> 144 2017-06-01 21:50:00  5.00           11
#> 145 2017-06-01 22:00:00  4.82           11
#> 146 2017-06-01 22:10:00  4.62           11
#> 147 2017-06-01 22:20:00  4.41           11
#> 148 2017-06-01 22:30:00  4.21           11
#> 149 2017-06-01 22:40:00  4.03           11
#> 150 2017-06-01 22:50:00  3.86           11
#> 151 2017-06-01 23:00:00  3.70           15
#> 152 2017-06-01 23:10:00  3.54           11
#> 153 2017-06-01 23:20:00  3.38           11
#> 154 2017-06-01 23:30:00  3.23           11
#> 155 2017-06-01 23:40:00  3.08           11
#> 156 2017-06-01 23:50:00  2.93           11
#> 157 2017-06-02 00:00:00  2.79           11
#> 158 2017-06-02 00:10:00  2.65           11
#> 159 2017-06-02 00:20:00  2.52           11
#> 160 2017-06-02 00:30:00  2.39           11
#> 161 2017-06-02 00:40:00  2.27           11
#> 162 2017-06-02 00:50:00  2.14           11
#> 163 2017-06-02 01:00:00  2.02           11
#> 164 2017-06-02 01:10:00  1.91           11
#> 165 2017-06-02 01:20:00  1.79           11
#> 166 2017-06-02 01:30:00  1.68           11
#> 167 2017-06-02 01:40:00  1.57           11
#> 168 2017-06-02 01:50:00  1.46           11
#> 169 2017-06-02 02:00:00  1.35           11
#> 170 2017-06-02 02:10:00  1.25           11
#> 171 2017-06-02 02:20:00  1.15           11
#> 172 2017-06-02 02:30:00  1.05           11
#> 173 2017-06-02 02:40:00  0.95           11
#> 174 2017-06-02 02:50:00  0.86           11
#> 175 2017-06-02 03:00:00  0.77           11
#> 176 2017-06-02 03:10:00  0.68           11
#> 177 2017-06-02 03:20:00  0.59           11
#> 178 2017-06-02 03:30:00  0.51           11
#> 179 2017-06-02 03:40:00  0.43           11
#> 180 2017-06-02 03:50:00  0.35           11
#> 181 2017-06-02 04:00:00  0.28           11
#> 182 2017-06-02 04:10:00  0.21           11
#> 183 2017-06-02 04:20:00  0.15           11
#> 184 2017-06-02 04:30:00  0.09           11
#> 185 2017-06-02 04:40:00  0.04           11
#> 186 2017-06-02 04:50:00  0.02           11
#> 187 2017-06-02 05:00:00  0.11           11
#> 188 2017-06-02 05:10:00  0.37           11
#> 189 2017-06-02 05:20:00  0.66           11
#> 190 2017-06-02 05:30:00  0.90           11
#> 191 2017-06-02 05:40:00  1.10           11
#> 192 2017-06-02 05:50:00  1.27           11
#> 193 2017-06-02 06:00:00  1.43           11
#> 194 2017-06-02 06:10:00  1.60           11
#> 195 2017-06-02 06:20:00  1.77           11
#> 196 2017-06-02 06:30:00  1.94           11
#> 197 2017-06-02 06:40:00  2.10           11
#> 198 2017-06-02 06:50:00  2.26           11
#> 199 2017-06-02 07:00:00  2.42           11
#> 200 2017-06-02 07:10:00  2.59           11
#> 201 2017-06-02 07:20:00  2.76           11
#> 202 2017-06-02 07:30:00  2.95           11
#> 203 2017-06-02 07:40:00  3.15           11
#> 204 2017-06-02 07:50:00  3.36           11
#> 205 2017-06-02 08:00:00  3.59           11
#> 206 2017-06-02 08:10:00  3.83           11
#> 207 2017-06-02 08:20:00  4.08           11
#> 208 2017-06-02 08:30:00  4.33           11
#> 209 2017-06-02 08:40:00  4.57           11
#> 210 2017-06-02 08:50:00  4.79           11
#> 211 2017-06-02 09:00:00  4.97           11
#> 212 2017-06-02 09:10:00  5.11           11
#> 213 2017-06-02 09:20:00  5.23           11
#> 214 2017-06-02 09:30:00  5.31           11
#> 215 2017-06-02 09:40:00  5.38           11
#> 216 2017-06-02 09:50:00  5.41           11
#> 217 2017-06-02 10:00:00  5.41           11
#> 218 2017-06-02 10:10:00  5.39           11
#> 219 2017-06-02 10:20:00  5.31           11
#> 220 2017-06-02 10:30:00  5.19           11
#> 221 2017-06-02 10:40:00  5.03           11
#> 222 2017-06-02 10:50:00  4.84           11
#> 223 2017-06-02 11:00:00  4.66           11
#> 224 2017-06-02 11:10:00  4.48           11
#> 225 2017-06-02 11:20:00  4.30           11
#> 226 2017-06-02 11:30:00  4.11           11
#> 227 2017-06-02 11:40:00  3.95           11
#> 228 2017-06-02 11:50:00  3.78           11
#> 229 2017-06-02 12:00:00  3.63           11
#> 230 2017-06-02 12:10:00  3.46           11
#> 231 2017-06-02 12:20:00  3.30           11
#> 232 2017-06-02 12:30:00  3.15           11
#> 233 2017-06-02 12:40:00  3.00           11
#> 234 2017-06-02 12:50:00  2.86           11
#> 235 2017-06-02 13:00:00  2.72           11
#> 236 2017-06-02 13:10:00  2.59           11
#> 237 2017-06-02 13:20:00  2.46           11
#> 238 2017-06-02 13:30:00  2.33           11
#> 239 2017-06-02 13:40:00  2.20           11
#> 240 2017-06-02 13:50:00  2.09           11
#> 241 2017-06-02 14:00:00  1.97           11
#> 242 2017-06-02 14:10:00  1.86           11
#> 243 2017-06-02 14:20:00  1.75           11
#> 244 2017-06-02 14:30:00  1.64           11
#> 245 2017-06-02 14:40:00  1.54           11
#> 246 2017-06-02 14:50:00  1.44           11
#> 247 2017-06-02 15:00:00  1.33           11
#> 248 2017-06-02 15:10:00  1.23           11
#> 249 2017-06-02 15:20:00  1.14           11
#> 250 2017-06-02 15:30:00  1.04           11
#> 251 2017-06-02 15:40:00  0.95           11
#> 252 2017-06-02 15:50:00  0.87           11
#> 253 2017-06-02 16:00:00  0.78           11
#> 254 2017-06-02 16:10:00  0.71           11
#> 255 2017-06-02 16:20:00  0.63           11
#> 256 2017-06-02 16:30:00  0.57           11
#> 257 2017-06-02 16:40:00  0.51           11
#> 258 2017-06-02 16:50:00  0.46           11
#> 259 2017-06-02 17:00:00  0.45           11
#> 260 2017-06-02 17:10:00  0.53           11
#> 261 2017-06-02 17:20:00  0.75           11
#> 262 2017-06-02 17:30:00  1.04           11
#> 263 2017-06-02 17:40:00  1.26           11
#> 264 2017-06-02 17:50:00  1.46           11
#> 265 2017-06-02 18:00:00  1.62           11
#> 266 2017-06-02 18:10:00  1.77           11
#> 267 2017-06-02 18:20:00  1.91           11
#> 268 2017-06-02 18:30:00  2.07           11
#> 269 2017-06-02 18:40:00  2.21           11
#> 270 2017-06-02 18:50:00  2.35           11
#> 271 2017-06-02 19:00:00  2.48           11
#> 272 2017-06-02 19:10:00  2.62           11
#> 273 2017-06-02 19:20:00  2.74           11
#> 274 2017-06-02 19:30:00  2.87           11
#> 275 2017-06-02 19:40:00  2.99           11
#> 276 2017-06-02 19:50:00  3.11           11
#> 277 2017-06-02 20:00:00  3.24           11
#> 278 2017-06-02 20:10:00  3.37           11
#> 279 2017-06-02 20:20:00  3.52           11
#> 280 2017-06-02 20:30:00  3.67           11
#> 281 2017-06-02 20:40:00  3.84           11
#> 282 2017-06-02 20:50:00  4.01           11
#> 283 2017-06-02 21:00:00  4.20           11
#> 284 2017-06-02 21:10:00  4.39           11
#> 285 2017-06-02 21:20:00  4.58           11
#> 286 2017-06-02 21:30:00  4.75           11
#> 287 2017-06-02 21:40:00  4.91           11
#> 288 2017-06-02 21:50:00  5.03           11
#> 289 2017-06-02 22:00:00  5.13           11