Download timeseries data from waterinfo.be

Using the ts_id codes and by providing a given date period, download the corresponding time series from the waterinfo.be website

get_timeseries_tsid(ts_id, period = NULL, from = NULL, to = NULL,
  datasource = 1, token = NULL)

Arguments

ts_id
waterinfo.be database ts_id, defining a timeserie variable and frequency it is defined.

period
input string according to format required by waterinfo: De period string is provided as P#Y#M#DT#H#M#S, with P defines `Period`, each # is an integer value and the codes define the number of... Y - years M - months D - days T required if information about sub-day resolution is present H - hours D - days M - minutes S - seconds Instead of D (days), the usage of W - weeks is possible as well Examples of valid period strings: P3D, P1Y, P1DT12H, PT6H, P1Y6M3DT4H20M30S.

from
date of datestring as start of the time series

to
date of datestring as end of the time series

datasource
int [0-3] defines the `meetnet` of which the measurement station is part of. VMM based stations are net '1', MOW-HIC is net '2'

token
token to use with the call (optional, can be retrieved via get_token)

Format

A data.frame with 3 variables:

Timestamp

Datetime of the measurement.

Value

Measured value.

Quality Code

Quality code of the measurement, dependent on the data source used:

VMM Quality Code Interpretation (datasource 1)

10/110 - Excellent

30/100/130 - Good

50/150 - Moderate

70/170 - Poor

80/180 - Estimated

90/190 - Suspect

220 - Default

-1 - Missing

HIC Quality Code Interpretation (datasource 2)

40 - Good

80 - Estimated

120 - Suspect

200 - Unchecked

60 - Complete

160 - Incomplete

-1 - Missing

Aggregated timeseries

40 - Good

100 - Estimated

120 - Suspect

200 - Unchecked

-1 - Missing

Value

data.frame with the timestamps, values and quality code

Examples

get_timeseries_tsid("35055042", from = "2017-01-01", to = "2017-01-02")

#>              Timestamp Value Quality Code
#> 1  2017-01-01 00:00:00  0.00          130
#> 2  2017-01-01 00:15:00  0.00          130
#> 3  2017-01-01 00:30:00  0.00          130
#> 4  2017-01-01 00:45:00  0.00          130
#> 5  2017-01-01 01:00:00  0.00          130
#> 6  2017-01-01 01:15:00  0.00          130
#> 7  2017-01-01 01:30:00  0.00          130
#> 8  2017-01-01 01:45:00  0.00          130
#> 9  2017-01-01 02:00:00  0.00          130
#> 10 2017-01-01 02:15:00  0.00          130
#> 11 2017-01-01 02:30:00  0.00          130
#> 12 2017-01-01 02:45:00  0.00          130
#> 13 2017-01-01 03:00:00  0.00          130
#> 14 2017-01-01 03:15:00  0.00          130
#> 15 2017-01-01 03:30:00  0.00          130
#> 16 2017-01-01 03:45:00  0.00          130
#> 17 2017-01-01 04:00:00  0.00          130
#> 18 2017-01-01 04:15:00  0.00          130
#> 19 2017-01-01 04:30:00  0.00          130
#> 20 2017-01-01 04:45:00  0.00          130
#> 21 2017-01-01 05:00:00  0.00          130
#> 22 2017-01-01 05:15:00  0.00          130
#> 23 2017-01-01 05:30:00  0.00          130
#> 24 2017-01-01 05:45:00  0.00          130
#> 25 2017-01-01 06:00:00  0.00          130
#> 26 2017-01-01 06:15:00  0.00          130
#> 27 2017-01-01 06:30:00  0.00          130
#> 28 2017-01-01 06:45:00  0.00          130
#> 29 2017-01-01 07:00:00  0.00          130
#> 30 2017-01-01 07:15:00  0.00          130
#> 31 2017-01-01 07:30:00  0.00          130
#> 32 2017-01-01 07:45:00  0.00          130
#> 33 2017-01-01 08:00:00  0.00          130
#> 34 2017-01-01 08:15:00  0.00          130
#> 35 2017-01-01 08:30:00  0.00          130
#> 36 2017-01-01 08:45:00  0.00          130
#> 37 2017-01-01 09:00:00  0.00          130
#> 38 2017-01-01 09:15:00  0.00          130
#> 39 2017-01-01 09:30:00  0.00          130
#> 40 2017-01-01 09:45:00  0.00          130
#> 41 2017-01-01 10:00:00  0.00          130
#> 42 2017-01-01 10:15:00  0.00          130
#> 43 2017-01-01 10:30:00  0.00          130
#> 44 2017-01-01 10:45:00  0.00          130
#> 45 2017-01-01 11:00:00  0.00          130
#> 46 2017-01-01 11:15:00  0.00          130
#> 47 2017-01-01 11:30:00  0.00          130
#> 48 2017-01-01 11:45:00  0.00          130
#> 49 2017-01-01 12:00:00  0.00          130
#> 50 2017-01-01 12:15:00  0.00          130
#> 51 2017-01-01 12:30:00  0.00          130
#> 52 2017-01-01 12:45:00  0.00          130
#> 53 2017-01-01 13:00:00  0.00          130
#> 54 2017-01-01 13:15:00  0.00          130
#> 55 2017-01-01 13:30:00  0.00          130
#> 56 2017-01-01 13:45:00  0.00          130
#> 57 2017-01-01 14:00:00  0.00          130
#> 58 2017-01-01 14:15:00  0.00          130
#> 59 2017-01-01 14:30:00  0.00          130
#> 60 2017-01-01 14:45:00  0.00          130
#> 61 2017-01-01 15:00:00  0.00          130
#> 62 2017-01-01 15:15:00  0.00          130
#> 63 2017-01-01 15:30:00  0.00          130
#> 64 2017-01-01 15:45:00  0.00          130
#> 65 2017-01-01 16:00:00  0.00          130
#> 66 2017-01-01 16:15:00  0.00          130
#> 67 2017-01-01 16:30:00  0.00          130
#> 68 2017-01-01 16:45:00  0.00          130
#> 69 2017-01-01 17:00:00  0.00          130
#> 70 2017-01-01 17:15:00  0.00          130
#> 71 2017-01-01 17:30:00  0.00          130
#> 72 2017-01-01 17:45:00  0.03          130
#> 73 2017-01-01 18:00:00  0.00          130
#> 74 2017-01-01 18:15:00  0.04          130
#> 75 2017-01-01 18:30:00  0.00          130
#> 76 2017-01-01 18:45:00  0.03          130
#> 77 2017-01-01 19:00:00  0.00          130
#> 78 2017-01-01 19:15:00  0.03          130
#> 79 2017-01-01 19:30:00  0.03          130
#> 80 2017-01-01 19:45:00  0.00          130
#> 81 2017-01-01 20:00:00  0.03          130
#> 82 2017-01-01 20:15:00  0.00          130
#> 83 2017-01-01 20:30:00  0.00          130
#> 84 2017-01-01 20:45:00  0.00          130
#> 85 2017-01-01 21:00:00  0.00          130
#> 86 2017-01-01 21:15:00  0.03          130
#> 87 2017-01-01 21:30:00  0.00          130
#> 88 2017-01-01 21:45:00  0.00          130
#> 89 2017-01-01 22:00:00  0.00          130
#> 90 2017-01-01 22:15:00  0.11          130
#> 91 2017-01-01 22:30:00  0.82          130
#> 92 2017-01-01 22:45:00  0.69          130
#> 93 2017-01-01 23:00:00  0.34          130
#> 94 2017-01-01 23:15:00  0.13          130
#> 95 2017-01-01 23:30:00  0.03          130
#> 96 2017-01-01 23:45:00  0.03          130
#> 97 2017-01-02 00:00:00  0.05          130

get_timeseries_tsid("5156042", period = "P3D")

#>               Timestamp Value Quality Code
#> 1   2018-09-22 22:45:00 0.132          110
#> 2   2018-09-22 23:00:00 0.132          110
#> 3   2018-09-22 23:15:00 0.132          110
#> 4   2018-09-22 23:30:00 0.132          110
#> 5   2018-09-22 23:45:00 0.133          110
#> 6   2018-09-23 00:00:00 0.133          110
#> 7   2018-09-23 00:15:00 0.133          110
#> 8   2018-09-23 00:30:00 0.133          110
#> 9   2018-09-23 00:45:00 0.133          110
#> 10  2018-09-23 01:00:00 0.134          110
#> 11  2018-09-23 01:15:00 0.134          110
#> 12  2018-09-23 01:30:00 0.135          110
#> 13  2018-09-23 01:45:00 0.135          110
#> 14  2018-09-23 02:00:00 0.136          110
#> 15  2018-09-23 02:15:00 0.136          110
#> 16  2018-09-23 02:30:00 0.137          110
#> 17  2018-09-23 02:45:00 0.137          110
#> 18  2018-09-23 03:00:00 0.138          110
#> 19  2018-09-23 03:15:00 0.139          110
#> 20  2018-09-23 03:30:00 0.139          110
#> 21  2018-09-23 03:45:00 0.140          110
#> 22  2018-09-23 04:00:00 0.141          110
#> 23  2018-09-23 04:15:00 0.142          110
#> 24  2018-09-23 04:30:00 0.142          110
#> 25  2018-09-23 04:45:00 0.144          110
#> 26  2018-09-23 05:00:00 0.145          110
#> 27  2018-09-23 05:15:00 0.146          110
#> 28  2018-09-23 05:30:00 0.147          110
#> 29  2018-09-23 05:45:00 0.149          110
#> 30  2018-09-23 06:00:00 0.151          110
#> 31  2018-09-23 06:15:00 0.153          110
#> 32  2018-09-23 06:30:00 0.155          110
#> 33  2018-09-23 06:45:00 0.158          110
#> 34  2018-09-23 07:00:00 0.160          110
#> 35  2018-09-23 07:15:00 0.162          110
#> 36  2018-09-23 07:30:00 0.165          110
#> 37  2018-09-23 07:45:00 0.168          110
#> 38  2018-09-23 08:00:00 0.171          110
#> 39  2018-09-23 08:15:00 0.174          110
#> 40  2018-09-23 08:30:00 0.177          110
#> 41  2018-09-23 08:45:00 0.180          110
#> 42  2018-09-23 09:00:00 0.183          110
#> 43  2018-09-23 09:15:00 0.186          110
#> 44  2018-09-23 09:30:00 0.189          110
#> 45  2018-09-23 09:45:00 0.192          110
#> 46  2018-09-23 10:00:00 0.195          110
#> 47  2018-09-23 10:15:00 0.200          110
#> 48  2018-09-23 10:30:00 0.205          110
#> 49  2018-09-23 10:45:00 0.212          110
#> 50  2018-09-23 11:00:00 0.217          110
#> 51  2018-09-23 11:15:00 0.221          110
#> 52  2018-09-23 11:30:00 0.227          110
#> 53  2018-09-23 11:45:00 0.233          110
#> 54  2018-09-23 12:00:00 0.237          110
#> 55  2018-09-23 12:15:00 0.244          110
#> 56  2018-09-23 12:30:00 0.264          110
#> 57  2018-09-23 12:45:00 0.283          110
#> 58  2018-09-23 13:00:00 0.297          110
#> 59  2018-09-23 13:15:00 0.305          110
#> 60  2018-09-23 13:30:00 0.310          110
#> 61  2018-09-23 13:45:00 0.313          110
#> 62  2018-09-23 14:00:00 0.313          110
#> 63  2018-09-23 14:15:00 0.313          110
#> 64  2018-09-23 14:30:00 0.312          110
#> 65  2018-09-23 14:45:00 0.311          110
#> 66  2018-09-23 15:00:00 0.309          110
#> 67  2018-09-23 15:15:00 0.307          110
#> 68  2018-09-23 15:30:00 0.305          110
#> 69  2018-09-23 15:45:00 0.304          110
#> 70  2018-09-23 16:00:00 0.303          110
#> 71  2018-09-23 16:15:00 0.303          110
#> 72  2018-09-23 16:30:00 0.305          110
#> 73  2018-09-23 16:45:00 0.306          110
#> 74  2018-09-23 17:00:00 0.307          110
#> 75  2018-09-23 17:15:00 0.308          110
#> 76  2018-09-23 17:30:00 0.310          110
#> 77  2018-09-23 17:45:00 0.309          110
#> 78  2018-09-23 18:00:00 0.309          110
#> 79  2018-09-23 18:15:00 0.310          110
#> 80  2018-09-23 18:30:00 0.310          110
#> 81  2018-09-23 18:45:00 0.309          110
#> 82  2018-09-23 19:00:00 0.309          110
#> 83  2018-09-23 19:15:00 0.309          110
#> 84  2018-09-23 19:30:00 0.308          110
#> 85  2018-09-23 19:45:00 0.308          110
#> 86  2018-09-23 20:00:00 0.309          110
#> 87  2018-09-23 20:15:00 0.310          110
#> 88  2018-09-23 20:30:00 0.311          110
#> 89  2018-09-23 20:45:00 0.314          110
#> 90  2018-09-23 21:00:00 0.317          110
#> 91  2018-09-23 21:15:00 0.320          110
#> 92  2018-09-23 21:30:00 0.323          110
#> 93  2018-09-23 21:45:00 0.326          110
#> 94  2018-09-23 22:00:00 0.329          110
#> 95  2018-09-23 22:15:00 0.331          110
#> 96  2018-09-23 22:30:00 0.334          110
#> 97  2018-09-23 22:45:00 0.337          110
#> 98  2018-09-23 23:00:00 0.339          110
#> 99  2018-09-23 23:15:00 0.341          110
#> 100 2018-09-23 23:30:00 0.344          110
#> 101 2018-09-23 23:45:00 0.346          110
#> 102 2018-09-24 00:00:00 0.347          110
#> 103 2018-09-24 00:15:00 0.350          110
#> 104 2018-09-24 00:30:00 0.352          110
#> 105 2018-09-24 00:45:00 0.353          110
#> 106 2018-09-24 01:00:00 0.355          110
#> 107 2018-09-24 01:15:00 0.356          110
#> 108 2018-09-24 01:30:00 0.358          110
#> 109 2018-09-24 01:45:00 0.359          110
#> 110 2018-09-24 02:00:00 0.361          110
#> 111 2018-09-24 02:15:00 0.362          110
#> 112 2018-09-24 02:30:00 0.362          110
#> 113 2018-09-24 02:45:00 0.363          110
#> 114 2018-09-24 03:00:00 0.364          110
#> 115 2018-09-24 03:15:00 0.364          110
#> 116 2018-09-24 03:30:00 0.364          110
#> 117 2018-09-24 03:45:00 0.365          110
#> 118 2018-09-24 04:00:00 0.365          110
#> 119 2018-09-24 04:15:00 0.365          110
#> 120 2018-09-24 04:30:00 0.364          110
#> 121 2018-09-24 04:45:00 0.364          110
#> 122 2018-09-24 05:00:00 0.363          110
#> 123 2018-09-24 05:15:00 0.362          110
#> 124 2018-09-24 05:30:00 0.361          110
#> 125 2018-09-24 05:45:00 0.361          110
#> 126 2018-09-24 06:00:00 0.360          110
#> 127 2018-09-24 06:15:00 0.358          110
#> 128 2018-09-24 06:30:00 0.357          110
#> 129 2018-09-24 06:45:00 0.356          110
#> 130 2018-09-24 07:00:00 0.354          110
#> 131 2018-09-24 07:15:00 0.352          110
#> 132 2018-09-24 07:30:00 0.350          110
#> 133 2018-09-24 07:45:00 0.348          110
#> 134 2018-09-24 08:00:00 0.347          110
#> 135 2018-09-24 08:15:00 0.345          110
#> 136 2018-09-24 08:30:00 0.343          110
#> 137 2018-09-24 08:45:00 0.341          110
#> 138 2018-09-24 09:00:00 0.338          110
#> 139 2018-09-24 09:15:00 0.336          110
#> 140 2018-09-24 09:30:00 0.334          110
#> 141 2018-09-24 09:45:00 0.332          110
#> 142 2018-09-24 10:00:00 0.330          110
#> 143 2018-09-24 10:15:00 0.328          110
#> 144 2018-09-24 10:30:00 0.326          110
#> 145 2018-09-24 10:45:00 0.324          110
#> 146 2018-09-24 11:00:00 0.322          110
#> 147 2018-09-24 11:15:00 0.321          110
#> 148 2018-09-24 11:30:00 0.320          110
#> 149 2018-09-24 11:45:00 0.318          110
#> 150 2018-09-24 12:00:00 0.316          110
#> 151 2018-09-24 12:15:00 0.315          110
#> 152 2018-09-24 12:30:00 0.313          110
#> 153 2018-09-24 12:45:00 0.311          110
#> 154 2018-09-24 13:00:00 0.310          110
#> 155 2018-09-24 13:15:00 0.308          110
#> 156 2018-09-24 13:30:00 0.307          110
#> 157 2018-09-24 13:45:00 0.305          110
#> 158 2018-09-24 14:00:00 0.303          110
#> 159 2018-09-24 14:15:00 0.301          110
#> 160 2018-09-24 14:30:00 0.300          110
#> 161 2018-09-24 14:45:00 0.299          110
#> 162 2018-09-24 15:00:00 0.297          110
#> 163 2018-09-24 15:15:00 0.295          110
#> 164 2018-09-24 15:30:00 0.293          110
#> 165 2018-09-24 15:45:00 0.292          110
#> 166 2018-09-24 16:00:00 0.291          110
#> 167 2018-09-24 16:15:00 0.289          110
#> 168 2018-09-24 16:30:00 0.289          110
#> 169 2018-09-24 16:45:00 0.288          110
#> 170 2018-09-24 17:00:00 0.287          110
#> 171 2018-09-24 17:15:00 0.286          110
#> 172 2018-09-24 17:30:00 0.286          110
#> 173 2018-09-24 17:45:00 0.285          110
#> 174 2018-09-24 18:00:00 0.284          110
#> 175 2018-09-24 18:15:00 0.284          110
#> 176 2018-09-24 18:30:00 0.284          110
#> 177 2018-09-24 18:45:00 0.283          110
#> 178 2018-09-24 19:00:00 0.283          110
#> 179 2018-09-24 19:15:00 0.283          110
#> 180 2018-09-24 19:30:00 0.283          110
#> 181 2018-09-24 19:45:00 0.283          110
#> 182 2018-09-24 20:00:00 0.283          110
#> 183 2018-09-24 20:15:00 0.283          110
#> 184 2018-09-24 20:30:00 0.283          110
#> 185 2018-09-24 20:45:00 0.283          110
#> 186 2018-09-24 21:00:00 0.283          110
#> 187 2018-09-24 21:15:00 0.283          110
#> 188 2018-09-24 21:30:00 0.283          110
#> 189 2018-09-24 21:45:00 0.284          110
#> 190 2018-09-24 22:00:00 0.284          110
#> 191 2018-09-24 22:15:00 0.284          110
#> 192 2018-09-24 22:30:00 0.284          110
#> 193 2018-09-24 22:45:00 0.284          110
#> 194 2018-09-24 23:00:00 0.284          110
#> 195 2018-09-24 23:15:00 0.285          110
#> 196 2018-09-24 23:30:00 0.285          110
#> 197 2018-09-24 23:45:00 0.285          110
#> 198 2018-09-25 00:00:00 0.285          110
#> 199 2018-09-25 00:15:00 0.285          110
#> 200 2018-09-25 00:30:00 0.286          110
#> 201 2018-09-25 00:45:00 0.285          110
#> 202 2018-09-25 01:00:00 0.285          110
#> 203 2018-09-25 07:45:00 0.277          110
#> 204 2018-09-25 08:00:00 0.276          110
#> 205 2018-09-25 08:15:00 0.276          110
#> 206 2018-09-25 08:30:00 0.275          110
#> 207 2018-09-25 08:45:00 0.274          110
#> 208 2018-09-25 09:00:00 0.273          110
#> 209 2018-09-25 09:15:00 0.272          110
#> 210 2018-09-25 09:30:00 0.272          110
#> 211 2018-09-25 09:45:00 0.271          110
#> 212 2018-09-25 10:00:00 0.270          110
#> 213 2018-09-25 10:15:00 0.269          110
#> 214 2018-09-25 10:30:00 0.268          110
#> 215 2018-09-25 10:45:00 0.268          110
#> 216 2018-09-25 11:00:00 0.267          110
#> 217 2018-09-25 11:15:00 0.267          110
#> 218 2018-09-25 11:30:00 0.266          110
#> 219 2018-09-25 11:45:00 0.265          110
#> 220 2018-09-25 12:00:00 0.264          110
#> 221 2018-09-25 12:15:00 0.263          110
#> 222 2018-09-25 12:30:00 0.262          110
#> 223 2018-09-25 12:45:00 0.261          110
#> 224 2018-09-25 13:00:00 0.260          110
#> 225 2018-09-25 13:15:00 0.259          110
#> 226 2018-09-25 13:30:00 0.259          110
#> 227 2018-09-25 13:45:00 0.258          110
#> 228 2018-09-25 14:00:00 0.258          110
#> 229 2018-09-25 14:15:00 0.257          110
#> 230 2018-09-25 14:30:00 0.256          110
#> 231 2018-09-25 14:45:00 0.256          110
#> 232 2018-09-25 15:00:00 0.255          110
#> 233 2018-09-25 15:15:00 0.255          110
#> 234 2018-09-25 15:30:00 0.255          110
#> 235 2018-09-25 15:45:00 0.254          110
#> 236 2018-09-25 16:00:00 0.254          110
#> 237 2018-09-25 16:15:00 0.253          110
#> 238 2018-09-25 16:30:00 0.253          110
#> 239 2018-09-25 16:45:00 0.253          110
#> 240 2018-09-25 17:00:00 0.252          110
#> 241 2018-09-25 17:15:00 0.252          110
#> 242 2018-09-25 17:30:00 0.252          110
#> 243 2018-09-25 17:45:00 0.252          110
#> 244 2018-09-25 18:00:00 0.251          110
#> 245 2018-09-25 18:15:00 0.251          110
#> 246 2018-09-25 18:30:00 0.251          110
#> 247 2018-09-25 18:45:00 0.251          110
#> 248 2018-09-25 19:00:00 0.251          110
#> 249 2018-09-25 19:15:00 0.251          110
#> 250 2018-09-25 19:30:00 0.251          110
#> 251 2018-09-25 19:45:00 0.251          110
#> 252 2018-09-25 20:00:00 0.251          110
#> 253 2018-09-25 20:15:00 0.251          110
#> 254 2018-09-25 20:30:00 0.251          110
#> 255 2018-09-25 20:45:00 0.250          110
#> 256 2018-09-25 21:00:00 0.250          110
#> 257 2018-09-25 21:15:00 0.250          110
#> 258 2018-09-25 21:30:00 0.250          110
#> 259 2018-09-25 21:45:00 0.249          110
#> 260 2018-09-25 22:00:00 0.249          110
#> 261 2018-09-25 22:15:00 0.249          110

get_timeseries_tsid("2813562", period = "P1D", datasource = 2)

#>                Timestamp Value Quality Code
#> 1    2018-09-24 22:31:00  1.44          200
#> 2    2018-09-24 22:32:00  1.44          200
#> 3    2018-09-24 22:33:00  1.43          200
#> 4    2018-09-24 22:34:00  1.42          200
#> 5    2018-09-24 22:35:00  1.41          200
#> 6    2018-09-24 22:36:00  1.40          200
#> 7    2018-09-24 22:37:00  1.39          200
#> 8    2018-09-24 22:38:00  1.39          200
#> 9    2018-09-24 22:39:00  1.38          200
#> 10   2018-09-24 22:40:00  1.37          200
#> 11   2018-09-24 22:41:00  1.36          200
#> 12   2018-09-24 22:42:00  1.35          200
#> 13   2018-09-24 22:43:00  1.34          200
#> 14   2018-09-24 22:44:00  1.34          200
#> 15   2018-09-24 22:45:00  1.33          200
#> 16   2018-09-24 22:46:00  1.32          200
#> 17   2018-09-24 22:47:00  1.31          200
#> 18   2018-09-24 22:48:00  1.31          200
#> 19   2018-09-24 22:49:00  1.30          200
#> 20   2018-09-24 22:50:00  1.29          200
#> 21   2018-09-24 22:51:00  1.28          200
#> 22   2018-09-24 22:52:00  1.27          200
#> 23   2018-09-24 22:53:00  1.26          200
#> 24   2018-09-24 22:54:00  1.26          200
#> 25   2018-09-24 22:55:00  1.25          200
#> 26   2018-09-24 22:56:00  1.24          200
#> 27   2018-09-24 22:57:00  1.24          200
#> 28   2018-09-24 22:58:00  1.22          200
#> 29   2018-09-24 22:59:00  1.22          200
#> 30   2018-09-24 23:00:00  1.21          200
#> 31   2018-09-24 23:01:00  1.20          200
#> 32   2018-09-24 23:02:00  1.20          200
#> 33   2018-09-24 23:03:00  1.19          200
#> 34   2018-09-24 23:04:00  1.18          200
#> 35   2018-09-24 23:05:00  1.17          200
#> 36   2018-09-24 23:06:00  1.17          200
#> 37   2018-09-24 23:07:00  1.16          200
#> 38   2018-09-24 23:08:00  1.15          200
#> 39   2018-09-24 23:09:00  1.14          200
#> 40   2018-09-24 23:10:00  1.14          200
#> 41   2018-09-24 23:11:00  1.13          200
#> 42   2018-09-24 23:12:00  1.12          200
#> 43   2018-09-24 23:13:00  1.12          200
#> 44   2018-09-24 23:14:00  1.11          200
#> 45   2018-09-24 23:15:00  1.10          200
#> 46   2018-09-24 23:16:00  1.09          200
#> 47   2018-09-24 23:17:00  1.08          200
#> 48   2018-09-24 23:18:00  1.08          200
#> 49   2018-09-24 23:19:00  1.07          200
#> 50   2018-09-24 23:20:00  1.06          200
#> 51   2018-09-24 23:21:00  1.06          200
#> 52   2018-09-24 23:22:00  1.05          200
#> 53   2018-09-24 23:23:00  1.04          200
#> 54   2018-09-24 23:24:00  1.03          200
#> 55   2018-09-24 23:25:00  1.03          200
#> 56   2018-09-24 23:26:00  1.02          200
#> 57   2018-09-24 23:27:00  1.01          200
#> 58   2018-09-24 23:28:00  1.01          200
#> 59   2018-09-24 23:29:00  1.00          200
#> 60   2018-09-24 23:30:00  1.00          200
#> 61   2018-09-24 23:31:00  0.99          200
#> 62   2018-09-24 23:32:00  0.98          200
#> 63   2018-09-24 23:33:00  0.97          200
#> 64   2018-09-24 23:34:00  0.97          200
#> 65   2018-09-24 23:35:00  0.96          200
#> 66   2018-09-24 23:36:00  0.95          200
#> 67   2018-09-24 23:37:00  0.94          200
#> 68   2018-09-24 23:38:00  0.94          200
#> 69   2018-09-24 23:39:00  0.93          200
#> 70   2018-09-24 23:40:00  0.92          200
#> 71   2018-09-24 23:41:00  0.92          200
#> 72   2018-09-24 23:42:00  0.91          200
#> 73   2018-09-24 23:43:00  0.91          200
#> 74   2018-09-24 23:44:00  0.90          200
#> 75   2018-09-24 23:45:00  0.89          200
#> 76   2018-09-24 23:46:00  0.89          200
#> 77   2018-09-24 23:47:00  0.88          200
#> 78   2018-09-24 23:48:00  0.88          200
#> 79   2018-09-24 23:49:00  0.87          200
#> 80   2018-09-24 23:50:00  0.86          200
#> 81   2018-09-24 23:51:00  0.86          200
#> 82   2018-09-24 23:52:00  0.85          200
#> 83   2018-09-24 23:53:00  0.85          200
#> 84   2018-09-24 23:54:00  0.84          200
#> 85   2018-09-24 23:55:00  0.83          200
#> 86   2018-09-24 23:56:00  0.83          200
#> 87   2018-09-24 23:57:00  0.83          200
#> 88   2018-09-24 23:58:00  0.82          200
#> 89   2018-09-24 23:59:00  0.83          200
#> 90   2018-09-25 00:00:00  0.83          200
#> 91   2018-09-25 00:01:00  0.84          200
#> 92   2018-09-25 00:02:00  0.88          200
#> 93   2018-09-25 00:03:00  0.92          200
#> 94   2018-09-25 00:04:00  0.94          200
#> 95   2018-09-25 00:05:00  0.97          200
#> 96   2018-09-25 00:06:00  1.03          200
#> 97   2018-09-25 00:07:00  1.06          200
#> 98   2018-09-25 00:08:00  1.09          200
#> 99   2018-09-25 00:09:00  1.12          200
#> 100  2018-09-25 00:10:00  1.16          200
#> 101  2018-09-25 00:11:00  1.19          200
#> 102  2018-09-25 00:12:00  1.23          200
#> 103  2018-09-25 00:13:00  1.27          200
#> 104  2018-09-25 00:14:00  1.30          200
#> 105  2018-09-25 00:15:00  1.31          200
#> 106  2018-09-25 00:16:00  1.33          200
#> 107  2018-09-25 00:17:00  1.36          200
#> 108  2018-09-25 00:18:00  1.39          200
#> 109  2018-09-25 00:19:00  1.41          200
#> 110  2018-09-25 00:20:00  1.44          200
#> 111  2018-09-25 00:21:00  1.46          200
#> 112  2018-09-25 00:22:00  1.48          200
#> 113  2018-09-25 00:23:00  1.50          200
#> 114  2018-09-25 00:24:00  1.52          200
#> 115  2018-09-25 00:25:00  1.54          200
#> 116  2018-09-25 00:26:00  1.56          200
#> 117  2018-09-25 00:27:00  1.58          200
#> 118  2018-09-25 00:28:00  1.60          200
#> 119  2018-09-25 00:29:00  1.62          200
#> 120  2018-09-25 00:30:00  1.63          200
#> 121  2018-09-25 00:31:00  1.65          200
#> 122  2018-09-25 00:32:00  1.66          200
#> 123  2018-09-25 00:33:00  1.68          200
#> 124  2018-09-25 00:34:00  1.70          200
#> 125  2018-09-25 00:35:00  1.72          200
#> 126  2018-09-25 00:36:00  1.73          200
#> 127  2018-09-25 00:37:00  1.74          200
#> 128  2018-09-25 00:38:00  1.76          200
#> 129  2018-09-25 00:39:00  1.77          200
#> 130  2018-09-25 00:40:00  1.79          200
#> 131  2018-09-25 00:41:00  1.80          200
#> 132  2018-09-25 00:42:00  1.82          200
#> 133  2018-09-25 00:43:00  1.83          200
#> 134  2018-09-25 00:44:00  1.84          200
#> 135  2018-09-25 00:45:00  1.85          200
#> 136  2018-09-25 00:46:00  1.87          200
#> 137  2018-09-25 00:47:00  1.88          200
#> 138  2018-09-25 00:48:00  1.90          200
#> 139  2018-09-25 00:49:00  1.92          200
#> 140  2018-09-25 00:50:00  1.93          200
#> 141  2018-09-25 00:51:00  1.94          200
#> 142  2018-09-25 00:52:00  1.95          200
#> 143  2018-09-25 00:53:00  1.96          200
#> 144  2018-09-25 00:54:00  1.97          200
#> 145  2018-09-25 00:55:00  1.98          200
#> 146  2018-09-25 00:56:00  1.99          200
#> 147  2018-09-25 00:57:00  2.00          200
#> 148  2018-09-25 00:58:00  2.01          200
#> 149  2018-09-25 00:59:00  2.03          200
#> 150  2018-09-25 01:00:00  2.04          200
#> 151  2018-09-25 01:01:00  2.05          200
#> 152  2018-09-25 01:02:00  2.06          200
#> 153  2018-09-25 01:03:00  2.07          200
#> 154  2018-09-25 01:04:00  2.09          200
#> 155  2018-09-25 01:05:00  2.10          200
#> 156  2018-09-25 01:06:00  2.11          200
#> 157  2018-09-25 01:07:00  2.12          200
#> 158  2018-09-25 01:08:00  2.13          200
#> 159  2018-09-25 01:09:00  2.15          200
#> 160  2018-09-25 01:10:00  2.16          200
#> 161  2018-09-25 01:11:00  2.17          200
#> 162  2018-09-25 01:12:00  2.19          200
#> 163  2018-09-25 01:13:00  2.20          200
#> 164  2018-09-25 01:14:00  2.21          200
#> 165  2018-09-25 01:15:00  2.22          200
#> 166  2018-09-25 01:16:00  2.24          200
#> 167  2018-09-25 01:17:00  2.25          200
#> 168  2018-09-25 01:18:00  2.26          200
#> 169  2018-09-25 01:19:00  2.27          200
#> 170  2018-09-25 01:20:00  2.28          200
#> 171  2018-09-25 01:21:00  2.30          200
#> 172  2018-09-25 01:22:00  2.31          200
#> 173  2018-09-25 01:23:00  2.32          200
#> 174  2018-09-25 01:24:00  2.33          200
#> 175  2018-09-25 01:25:00  2.34          200
#> 176  2018-09-25 01:26:00  2.35          200
#> 177  2018-09-25 01:27:00  2.36          200
#> 178  2018-09-25 01:28:00  2.37          200
#> 179  2018-09-25 01:29:00  2.38          200
#> 180  2018-09-25 01:30:00  2.39          200
#> 181  2018-09-25 01:31:00  2.41          200
#> 182  2018-09-25 01:32:00  2.42          200
#> 183  2018-09-25 01:33:00  2.43          200
#> 184  2018-09-25 01:34:00  2.44          200
#> 185  2018-09-25 01:35:00  2.46          200
#> 186  2018-09-25 01:36:00  2.47          200
#> 187  2018-09-25 01:37:00  2.48          200
#> 188  2018-09-25 01:38:00  2.49          200
#> 189  2018-09-25 01:39:00  2.50          200
#> 190  2018-09-25 01:40:00  2.51          200
#> 191  2018-09-25 01:41:00  2.52          200
#> 192  2018-09-25 01:42:00  2.54          200
#> 193  2018-09-25 01:43:00  2.55          200
#> 194  2018-09-25 01:44:00  2.56          200
#> 195  2018-09-25 01:45:00  2.57          200
#> 196  2018-09-25 01:46:00  2.59          200
#> 197  2018-09-25 01:47:00  2.60          200
#> 198  2018-09-25 01:48:00  2.61          200
#> 199  2018-09-25 01:49:00  2.62          200
#> 200  2018-09-25 01:50:00  2.63          200
#> 201  2018-09-25 01:51:00  2.64          200
#> 202  2018-09-25 01:52:00  2.66          200
#> 203  2018-09-25 01:53:00  2.68          200
#> 204  2018-09-25 01:54:00  2.69          200
#> 205  2018-09-25 01:55:00  2.70          200
#> 206  2018-09-25 01:56:00  2.71          200
#> 207  2018-09-25 01:57:00  2.72          200
#> 208  2018-09-25 01:58:00  2.73          200
#> 209  2018-09-25 01:59:00  2.75          200
#> 210  2018-09-25 02:00:00  2.76          200
#> 211  2018-09-25 02:01:00  2.77          200
#> 212  2018-09-25 02:02:00  2.79          200
#> 213  2018-09-25 02:03:00  2.80          200
#> 214  2018-09-25 02:04:00  2.81          200
#> 215  2018-09-25 02:05:00  2.82          200
#> 216  2018-09-25 02:06:00  2.83          200
#> 217  2018-09-25 02:07:00  2.85          200
#> 218  2018-09-25 02:08:00  2.86          200
#> 219  2018-09-25 02:09:00  2.87          200
#> 220  2018-09-25 02:10:00  2.88          200
#> 221  2018-09-25 02:11:00  2.90          200
#> 222  2018-09-25 02:12:00  2.91          200
#> 223  2018-09-25 02:13:00  2.92          200
#> 224  2018-09-25 02:14:00  2.94          200
#> 225  2018-09-25 02:15:00  2.96          200
#> 226  2018-09-25 02:16:00  2.97          200
#> 227  2018-09-25 02:17:00  2.98          200
#> 228  2018-09-25 02:18:00  2.99          200
#> 229  2018-09-25 02:19:00  3.01          200
#> 230  2018-09-25 02:20:00  3.02          200
#> 231  2018-09-25 02:21:00  3.04          200
#> 232  2018-09-25 02:22:00  3.05          200
#> 233  2018-09-25 02:23:00  3.07          200
#> 234  2018-09-25 02:24:00  3.08          200
#> 235  2018-09-25 02:25:00  3.10          200
#> 236  2018-09-25 02:26:00  3.12          200
#> 237  2018-09-25 02:27:00  3.13          200
#> 238  2018-09-25 02:28:00  3.15          200
#> 239  2018-09-25 02:29:00  3.17          200
#> 240  2018-09-25 02:30:00  3.19          200
#> 241  2018-09-25 02:31:00  3.20          200
#> 242  2018-09-25 02:32:00  3.22          200
#> 243  2018-09-25 02:33:00  3.24          200
#> 244  2018-09-25 02:34:00  3.26          200
#> 245  2018-09-25 02:35:00  3.28          200
#> 246  2018-09-25 02:36:00  3.30          200
#> 247  2018-09-25 02:37:00  3.31          200
#> 248  2018-09-25 02:38:00  3.33          200
#> 249  2018-09-25 02:39:00  3.35          200
#> 250  2018-09-25 02:40:00  3.37          200
#> 251  2018-09-25 02:41:00  3.39          200
#> 252  2018-09-25 02:42:00  3.41          200
#> 253  2018-09-25 02:43:00  3.43          200
#> 254  2018-09-25 02:44:00  3.46          200
#> 255  2018-09-25 02:45:00  3.48          200
#> 256  2018-09-25 02:46:00  3.50          200
#> 257  2018-09-25 02:47:00  3.52          200
#> 258  2018-09-25 02:48:00  3.54          200
#> 259  2018-09-25 02:49:00  3.57          200
#> 260  2018-09-25 02:50:00  3.59          200
#> 261  2018-09-25 02:51:00  3.61          200
#> 262  2018-09-25 02:52:00  3.64          200
#> 263  2018-09-25 02:53:00  3.67          200
#> 264  2018-09-25 02:54:00  3.70          200
#> 265  2018-09-25 02:55:00  3.73          200
#> 266  2018-09-25 02:56:00  3.75          200
#> 267  2018-09-25 02:57:00  3.78          200
#> 268  2018-09-25 02:58:00  3.80          200
#> 269  2018-09-25 02:59:00  3.83          200
#> 270  2018-09-25 03:00:00  3.86          200
#> 271  2018-09-25 03:01:00  3.88          200
#> 272  2018-09-25 03:02:00  3.91          200
#> 273  2018-09-25 03:03:00  3.94          200
#> 274  2018-09-25 03:04:00  3.97          200
#> 275  2018-09-25 03:05:00  4.00          200
#> 276  2018-09-25 03:06:00  4.02          200
#> 277  2018-09-25 03:07:00  4.06          200
#> 278  2018-09-25 03:08:00  4.09          200
#> 279  2018-09-25 03:09:00  4.11          200
#> 280  2018-09-25 03:10:00  4.14          200
#> 281  2018-09-25 03:11:00  4.18          200
#> 282  2018-09-25 03:12:00  4.21          200
#> 283  2018-09-25 03:13:00  4.24          200
#> 284  2018-09-25 03:14:00  4.26          200
#> 285  2018-09-25 03:15:00  4.30          200
#> 286  2018-09-25 03:16:00  4.32          200
#> 287  2018-09-25 03:17:00  4.36          200
#> 288  2018-09-25 03:18:00  4.39          200
#> 289  2018-09-25 03:19:00  4.42          200
#> 290  2018-09-25 03:20:00  4.45          200
#> 291  2018-09-25 03:21:00  4.47          200
#> 292  2018-09-25 03:22:00  4.50          200
#> 293  2018-09-25 03:23:00  4.52          200
#> 294  2018-09-25 03:24:00  4.54          200
#> 295  2018-09-25 03:25:00  4.57          200
#> 296  2018-09-25 03:26:00  4.59          200
#> 297  2018-09-25 03:27:00  4.61          200
#> 298  2018-09-25 03:28:00  4.64          200
#> 299  2018-09-25 03:29:00  4.67          200
#> 300  2018-09-25 03:30:00  4.69          200
#> 301  2018-09-25 03:31:00  4.72          200
#> 302  2018-09-25 03:32:00  4.74          200
#> 303  2018-09-25 03:33:00  4.75          200
#> 304  2018-09-25 03:34:00  4.78          200
#> 305  2018-09-25 03:35:00  4.80          200
#> 306  2018-09-25 03:36:00  4.82          200
#> 307  2018-09-25 03:37:00  4.84          200
#> 308  2018-09-25 03:38:00  4.87          200
#> 309  2018-09-25 03:39:00  4.88          200
#> 310  2018-09-25 03:40:00  4.90          200
#> 311  2018-09-25 03:41:00  4.92          200
#> 312  2018-09-25 03:42:00  4.94          200
#> 313  2018-09-25 03:43:00  4.96          200
#> 314  2018-09-25 03:44:00  4.98          200
#> 315  2018-09-25 03:45:00  5.00          200
#> 316  2018-09-25 03:46:00  5.01          200
#> 317  2018-09-25 03:47:00  5.03          200
#> 318  2018-09-25 03:48:00  5.05          200
#> 319  2018-09-25 03:49:00  5.06          200
#> 320  2018-09-25 03:50:00  5.07          200
#> 321  2018-09-25 03:51:00  5.09          200
#> 322  2018-09-25 03:52:00  5.11          200
#> 323  2018-09-25 03:53:00  5.12          200
#> 324  2018-09-25 03:54:00  5.14          200
#> 325  2018-09-25 03:55:00  5.15          200
#> 326  2018-09-25 03:56:00  5.16          200
#> 327  2018-09-25 03:57:00  5.18          200
#> 328  2018-09-25 03:58:00  5.20          200
#> 329  2018-09-25 03:59:00  5.21          200
#> 330  2018-09-25 04:00:00  5.22          200
#> 331  2018-09-25 04:01:00  5.23          200
#> 332  2018-09-25 04:02:00  5.25          200
#> 333  2018-09-25 04:03:00  5.25          200
#>  [ reached getOption("max.print") -- omitted 1082 rows ]