Download timeseries data from waterinfo.be

Using the ts_id codes and by providing a given date period, download the corresponding time series from the waterinfo.be website

get_timeseries_tsid(ts_id, period = NULL, from = NULL, to = NULL,
  datasource = 1, token = NULL)

Arguments

ts_id
waterinfo.be database ts_id, defining a timeserie variable and frequency it is defined.

period
input string according to format required by waterinfo: De period string is provided as P#Y#M#DT#H#M#S, with P defines `Period`, each # is an integer value and the codes define the number of... Y - years M - months D - days T required if information about sub-day resolution is present H - hours D - days M - minutes S - seconds Instead of D (days), the usage of W - weeks is possible as well Examples of valid period strings: P3D, P1Y, P1DT12H, PT6H, P1Y6M3DT4H20M30S.

from
date of datestring as start of the time series

to
date of datestring as end of the time series

datasource
int [0-3] defines the `meetnet` of which the measurement station is part of. VMM based stations are net '1', MOW-HIC is net '2'

token
token to use with the call (optional, can be retrieved via get_token)

Format

A data.frame with 3 variables:

Timestamp

Datetime of the measurement.

Value

Measured value.

Quality Code

Quality code of the measurement, dependent on the data source used:

VMM Quality Code Interpretation (datasource 1)

10/110 - Excellent

30/100/130 - Good

50/150 - Moderate

70/170 - Poor

80/180 - Estimated

90/190 - Suspect

220 - Default

-1 - Missing

HIC Quality Code Interpretation (datasource 2)

40 - Good

80 - Estimated

120 - Suspect

200 - Unchecked

60 - Complete

160 - Incomplete

-1 - Missing

Aggregated timeseries

40 - Good

100 - Estimated

120 - Suspect

200 - Unchecked

-1 - Missing

The URL of the specific request is provided as a comment attribute to the returned data.frame. Use comment(df) to get the request URL.

Value

data.frame with the timestamps, values and quality code

Examples

get_timeseries_tsid("35055042", from = "2017-01-01", to = "2017-01-02")

#>              Timestamp Value Quality Code
#> 1  2017-01-01 00:00:00  0.00          130
#> 2  2017-01-01 00:15:00  0.00          130
#> 3  2017-01-01 00:30:00  0.00          130
#> 4  2017-01-01 00:45:00  0.00          130
#> 5  2017-01-01 01:00:00  0.00          130
#> 6  2017-01-01 01:15:00  0.00          130
#> 7  2017-01-01 01:30:00  0.00          130
#> 8  2017-01-01 01:45:00  0.00          130
#> 9  2017-01-01 02:00:00  0.00          130
#> 10 2017-01-01 02:15:00  0.00          130
#> 11 2017-01-01 02:30:00  0.00          130
#> 12 2017-01-01 02:45:00  0.00          130
#> 13 2017-01-01 03:00:00  0.00          130
#> 14 2017-01-01 03:15:00  0.00          130
#> 15 2017-01-01 03:30:00  0.00          130
#> 16 2017-01-01 03:45:00  0.00          130
#> 17 2017-01-01 04:00:00  0.00          130
#> 18 2017-01-01 04:15:00  0.00          130
#> 19 2017-01-01 04:30:00  0.00          130
#> 20 2017-01-01 04:45:00  0.00          130
#> 21 2017-01-01 05:00:00  0.00          130
#> 22 2017-01-01 05:15:00  0.00          130
#> 23 2017-01-01 05:30:00  0.00          130
#> 24 2017-01-01 05:45:00  0.00          130
#> 25 2017-01-01 06:00:00  0.00          130
#> 26 2017-01-01 06:15:00  0.00          130
#> 27 2017-01-01 06:30:00  0.00          130
#> 28 2017-01-01 06:45:00  0.00          130
#> 29 2017-01-01 07:00:00  0.00          130
#> 30 2017-01-01 07:15:00  0.00          130
#> 31 2017-01-01 07:30:00  0.00          130
#> 32 2017-01-01 07:45:00  0.00          130
#> 33 2017-01-01 08:00:00  0.00          130
#> 34 2017-01-01 08:15:00  0.00          130
#> 35 2017-01-01 08:30:00  0.00          130
#> 36 2017-01-01 08:45:00  0.00          130
#> 37 2017-01-01 09:00:00  0.00          130
#> 38 2017-01-01 09:15:00  0.00          130
#> 39 2017-01-01 09:30:00  0.00          130
#> 40 2017-01-01 09:45:00  0.00          130
#> 41 2017-01-01 10:00:00  0.00          130
#> 42 2017-01-01 10:15:00  0.00          130
#> 43 2017-01-01 10:30:00  0.00          130
#> 44 2017-01-01 10:45:00  0.00          130
#> 45 2017-01-01 11:00:00  0.00          130
#> 46 2017-01-01 11:15:00  0.00          130
#> 47 2017-01-01 11:30:00  0.00          130
#> 48 2017-01-01 11:45:00  0.00          130
#> 49 2017-01-01 12:00:00  0.00          130
#> 50 2017-01-01 12:15:00  0.00          130
#> 51 2017-01-01 12:30:00  0.00          130
#> 52 2017-01-01 12:45:00  0.00          130
#> 53 2017-01-01 13:00:00  0.00          130
#> 54 2017-01-01 13:15:00  0.00          130
#> 55 2017-01-01 13:30:00  0.00          130
#> 56 2017-01-01 13:45:00  0.00          130
#> 57 2017-01-01 14:00:00  0.00          130
#> 58 2017-01-01 14:15:00  0.00          130
#> 59 2017-01-01 14:30:00  0.00          130
#> 60 2017-01-01 14:45:00  0.00          130
#> 61 2017-01-01 15:00:00  0.00          130
#> 62 2017-01-01 15:15:00  0.00          130
#> 63 2017-01-01 15:30:00  0.00          130
#> 64 2017-01-01 15:45:00  0.00          130
#> 65 2017-01-01 16:00:00  0.00          130
#> 66 2017-01-01 16:15:00  0.00          130
#> 67 2017-01-01 16:30:00  0.00          130
#> 68 2017-01-01 16:45:00  0.00          130
#> 69 2017-01-01 17:00:00  0.00          130
#> 70 2017-01-01 17:15:00  0.00          130
#> 71 2017-01-01 17:30:00  0.00          130
#> 72 2017-01-01 17:45:00  0.03          130
#> 73 2017-01-01 18:00:00  0.00          130
#> 74 2017-01-01 18:15:00  0.04          130
#> 75 2017-01-01 18:30:00  0.00          130
#> 76 2017-01-01 18:45:00  0.03          130
#> 77 2017-01-01 19:00:00  0.00          130
#> 78 2017-01-01 19:15:00  0.03          130
#> 79 2017-01-01 19:30:00  0.03          130
#> 80 2017-01-01 19:45:00  0.00          130
#> 81 2017-01-01 20:00:00  0.03          130
#> 82 2017-01-01 20:15:00  0.00          130
#> 83 2017-01-01 20:30:00  0.00          130
#> 84 2017-01-01 20:45:00  0.00          130
#> 85 2017-01-01 21:00:00  0.00          130
#> 86 2017-01-01 21:15:00  0.03          130
#> 87 2017-01-01 21:30:00  0.00          130
#> 88 2017-01-01 21:45:00  0.00          130
#> 89 2017-01-01 22:00:00  0.00          130
#> 90 2017-01-01 22:15:00  0.11          130
#> 91 2017-01-01 22:30:00  0.82          130
#> 92 2017-01-01 22:45:00  0.69          130
#> 93 2017-01-01 23:00:00  0.34          130
#> 94 2017-01-01 23:15:00  0.13          130
#> 95 2017-01-01 23:30:00  0.03          130
#> 96 2017-01-01 23:45:00  0.03          130
#> 97 2017-01-02 00:00:00  0.05          130

get_timeseries_tsid("5156042", period = "P3D")

#>               Timestamp Value Quality Code
#> 1   2019-05-24 19:00:00 0.551          110
#> 2   2019-05-24 19:15:00 0.553          110
#> 3   2019-05-24 19:30:00 0.556          110
#> 4   2019-05-24 19:45:00 0.558          110
#> 5   2019-05-24 20:00:00 0.560          110
#> 6   2019-05-24 20:15:00 0.562          110
#> 7   2019-05-24 20:30:00 0.563          110
#> 8   2019-05-24 20:45:00 0.565          110
#> 9   2019-05-24 21:00:00 0.566          110
#> 10  2019-05-24 21:15:00 0.567          110
#> 11  2019-05-24 21:30:00 0.568          110
#> 12  2019-05-24 21:45:00 0.569          110
#> 13  2019-05-24 22:00:00 0.570          110
#> 14  2019-05-24 22:15:00 0.571          110
#> 15  2019-05-24 22:30:00 0.573          110
#> 16  2019-05-24 22:45:00 0.574          110
#> 17  2019-05-24 23:00:00 0.575          110
#> 18  2019-05-24 23:15:00 0.576          110
#> 19  2019-05-24 23:30:00 0.577          110
#> 20  2019-05-24 23:45:00 0.577          110
#> 21  2019-05-25 00:00:00 0.578          110
#> 22  2019-05-25 00:15:00 0.578          110
#> 23  2019-05-25 00:30:00 0.579          110
#> 24  2019-05-25 00:45:00 0.579          110
#> 25  2019-05-25 01:00:00 0.580          110
#> 26  2019-05-25 01:15:00 0.580          110
#> 27  2019-05-25 01:30:00 0.579          110
#> 28  2019-05-25 01:45:00 0.580          110
#> 29  2019-05-25 02:00:00 0.579          110
#> 30  2019-05-25 02:15:00 0.579          110
#> 31  2019-05-25 02:30:00 0.579          110
#> 32  2019-05-25 02:45:00 0.579          110
#> 33  2019-05-25 03:00:00 0.578          110
#> 34  2019-05-25 03:15:00 0.578          110
#> 35  2019-05-25 03:30:00 0.577          110
#> 36  2019-05-25 03:45:00 0.577          110
#> 37  2019-05-25 04:00:00 0.577          110
#> 38  2019-05-25 04:15:00 0.576          110
#> 39  2019-05-25 04:30:00 0.576          110
#> 40  2019-05-25 04:45:00 0.575          110
#> 41  2019-05-25 05:00:00 0.574          110
#> 42  2019-05-25 05:15:00 0.574          110
#> 43  2019-05-25 05:30:00 0.573          110
#> 44  2019-05-25 05:45:00 0.572          110
#> 45  2019-05-25 06:00:00 0.571          110
#> 46  2019-05-25 06:15:00 0.570          110
#> 47  2019-05-25 06:30:00 0.569          110
#> 48  2019-05-25 06:45:00 0.568          110
#> 49  2019-05-25 07:00:00 0.567          110
#> 50  2019-05-25 07:15:00 0.566          110
#> 51  2019-05-25 07:30:00 0.565          110
#> 52  2019-05-25 07:45:00 0.564          110
#> 53  2019-05-25 08:00:00 0.563          110
#> 54  2019-05-25 08:15:00 0.562          110
#> 55  2019-05-25 08:30:00 0.561          110
#> 56  2019-05-25 08:45:00 0.560          110
#> 57  2019-05-25 09:00:00 0.559          110
#> 58  2019-05-25 09:15:00 0.557          110
#> 59  2019-05-25 09:30:00 0.557          110
#> 60  2019-05-25 09:45:00 0.556          110
#> 61  2019-05-25 10:00:00 0.554          110
#> 62  2019-05-25 10:15:00 0.554          110
#> 63  2019-05-25 10:30:00 0.553          110
#> 64  2019-05-25 10:45:00 0.552          110
#> 65  2019-05-25 11:00:00 0.552          110
#> 66  2019-05-25 11:15:00 0.552          110
#> 67  2019-05-25 11:30:00 0.552          110
#> 68  2019-05-25 11:45:00 0.552          110
#> 69  2019-05-25 12:00:00 0.553          110
#> 70  2019-05-25 12:15:00 0.554          110
#> 71  2019-05-25 12:30:00 0.555          110
#> 72  2019-05-25 12:45:00 0.556          110
#> 73  2019-05-25 13:00:00 0.557          110
#> 74  2019-05-25 13:15:00 0.558          110
#> 75  2019-05-25 13:30:00 0.558          110
#> 76  2019-05-25 13:45:00 0.559          110
#> 77  2019-05-25 14:00:00 0.560          110
#> 78  2019-05-25 14:15:00 0.560          110
#> 79  2019-05-25 14:30:00 0.561          110
#> 80  2019-05-25 14:45:00 0.561          110
#> 81  2019-05-25 15:00:00 0.561          110
#> 82  2019-05-25 15:15:00 0.561          110
#> 83  2019-05-25 15:30:00 0.561          110
#> 84  2019-05-25 15:45:00 0.561          110
#> 85  2019-05-25 16:00:00 0.561          110
#> 86  2019-05-25 16:15:00 0.560          110
#> 87  2019-05-25 16:30:00 0.560          110
#> 88  2019-05-25 16:45:00 0.560          110
#> 89  2019-05-25 17:00:00 0.560          110
#> 90  2019-05-25 17:15:00 0.560          110
#> 91  2019-05-25 17:30:00 0.560          110
#> 92  2019-05-25 17:45:00 0.559          110
#> 93  2019-05-25 18:00:00 0.559          110
#> 94  2019-05-25 18:15:00 0.559          110
#> 95  2019-05-25 18:30:00 0.559          110
#> 96  2019-05-25 18:45:00 0.558          110
#> 97  2019-05-25 19:00:00 0.557          110
#> 98  2019-05-25 19:15:00 0.557          110
#> 99  2019-05-25 19:30:00 0.557          110
#> 100 2019-05-25 19:45:00 0.557          110
#> 101 2019-05-25 20:00:00 0.556          110
#> 102 2019-05-25 20:15:00 0.556          110
#> 103 2019-05-25 20:30:00 0.555          110
#> 104 2019-05-25 20:45:00 0.554          110
#> 105 2019-05-25 21:00:00 0.554          110
#> 106 2019-05-25 21:15:00 0.553          110
#> 107 2019-05-25 21:30:00 0.552          110
#> 108 2019-05-25 21:45:00 0.552          110
#> 109 2019-05-25 22:00:00 0.551          110
#> 110 2019-05-25 22:15:00 0.550          110
#> 111 2019-05-25 22:30:00 0.549          110
#> 112 2019-05-25 22:45:00 0.549          110
#> 113 2019-05-25 23:00:00 0.548          110
#> 114 2019-05-25 23:15:00 0.547          110
#> 115 2019-05-25 23:30:00 0.546          110
#> 116 2019-05-25 23:45:00 0.545          110
#> 117 2019-05-26 00:00:00 0.545          110
#> 118 2019-05-26 00:15:00 0.544          110
#> 119 2019-05-26 00:30:00 0.543          110
#> 120 2019-05-26 00:45:00 0.542          110
#> 121 2019-05-26 01:00:00 0.542          110
#> 122 2019-05-26 01:15:00 0.541          110
#> 123 2019-05-26 01:30:00 0.540          110
#> 124 2019-05-26 01:45:00 0.539          110
#> 125 2019-05-26 02:00:00 0.538          110
#> 126 2019-05-26 02:15:00 0.538          110
#> 127 2019-05-26 02:30:00 0.537          110
#> 128 2019-05-26 02:45:00 0.536          110
#> 129 2019-05-26 03:00:00 0.536          110
#> 130 2019-05-26 03:15:00 0.535          110
#> 131 2019-05-26 03:30:00 0.534          110
#> 132 2019-05-26 03:45:00 0.533          110
#> 133 2019-05-26 04:00:00 0.532          110
#> 134 2019-05-26 04:15:00 0.531          110
#> 135 2019-05-26 04:30:00 0.530          110
#> 136 2019-05-26 04:45:00 0.530          110
#> 137 2019-05-26 05:00:00 0.529          110
#> 138 2019-05-26 05:15:00 0.528          110
#> 139 2019-05-26 05:30:00 0.527          110
#> 140 2019-05-26 05:45:00 0.526          110
#> 141 2019-05-26 06:00:00 0.525          110
#> 142 2019-05-26 06:15:00 0.524          110
#> 143 2019-05-26 06:30:00 0.523          110
#> 144 2019-05-26 06:45:00 0.522          110
#> 145 2019-05-26 07:00:00 0.521          110
#> 146 2019-05-26 07:15:00 0.520          110
#> 147 2019-05-26 07:30:00 0.519          110
#> 148 2019-05-26 07:45:00 0.518          110
#> 149 2019-05-26 08:00:00 0.518          110
#> 150 2019-05-26 08:15:00 0.517          110
#> 151 2019-05-26 08:30:00 0.516          110
#> 152 2019-05-26 08:45:00 0.515          110
#> 153 2019-05-26 09:00:00 0.514          110
#> 154 2019-05-26 09:15:00 0.513          110
#> 155 2019-05-26 09:30:00 0.512          110
#> 156 2019-05-26 09:45:00 0.511          110
#> 157 2019-05-26 10:00:00 0.510          110
#> 158 2019-05-26 10:15:00 0.509          110
#> 159 2019-05-26 10:30:00 0.509          110
#> 160 2019-05-26 10:45:00 0.508          110
#> 161 2019-05-26 11:00:00 0.507          110
#> 162 2019-05-26 11:15:00 0.506          110
#> 163 2019-05-26 11:30:00 0.505          110
#> 164 2019-05-26 11:45:00 0.505          110
#> 165 2019-05-26 12:00:00 0.504          110
#> 166 2019-05-26 12:15:00 0.503          110
#> 167 2019-05-26 12:30:00 0.503          110
#> 168 2019-05-26 12:45:00 0.502          110
#> 169 2019-05-26 13:00:00 0.501          110
#> 170 2019-05-26 13:15:00 0.500          110
#> 171 2019-05-26 13:30:00 0.499          110
#> 172 2019-05-26 13:45:00 0.499          110
#> 173 2019-05-26 14:00:00 0.498          110
#> 174 2019-05-26 14:15:00 0.497          110
#> 175 2019-05-26 14:30:00 0.497          110
#> 176 2019-05-26 14:45:00 0.498          110
#> 177 2019-05-26 15:00:00 0.497          110
#> 178 2019-05-26 15:15:00 0.497          110
#> 179 2019-05-26 15:30:00 0.496          110
#> 180 2019-05-26 15:45:00 0.495          110
#> 181 2019-05-26 16:00:00 0.494          110
#> 182 2019-05-26 16:15:00 0.495          110
#> 183 2019-05-26 16:30:00 0.494          110
#> 184 2019-05-26 16:45:00 0.493          110
#> 185 2019-05-26 17:00:00 0.494          110
#> 186 2019-05-26 17:15:00 0.493          110
#> 187 2019-05-26 17:30:00 0.493          110
#> 188 2019-05-26 17:45:00 0.493          110
#> 189 2019-05-26 18:00:00 0.492          110
#> 190 2019-05-26 18:15:00 0.492          110
#> 191 2019-05-26 18:30:00 0.492          110
#> 192 2019-05-26 18:45:00 0.492          110
#> 193 2019-05-26 19:00:00 0.492          110
#> 194 2019-05-26 19:15:00 0.492          110
#> 195 2019-05-26 19:30:00 0.491          110
#> 196 2019-05-26 19:45:00 0.491          110
#> 197 2019-05-26 20:00:00 0.491          110
#> 198 2019-05-26 20:15:00 0.491          110
#> 199 2019-05-26 20:30:00 0.491          110
#> 200 2019-05-26 20:45:00 0.490          110
#> 201 2019-05-26 21:00:00 0.489          110
#> 202 2019-05-26 21:15:00 0.489          110
#> 203 2019-05-26 21:30:00 0.488          110
#> 204 2019-05-26 21:45:00 0.487          110
#> 205 2019-05-26 22:00:00 0.488          110
#> 206 2019-05-26 22:15:00 0.487          110
#> 207 2019-05-26 22:30:00 0.487          110
#> 208 2019-05-26 22:45:00 0.487          110
#> 209 2019-05-26 23:00:00 0.487          110
#> 210 2019-05-26 23:15:00 0.487          110
#> 211 2019-05-26 23:30:00 0.487          110
#> 212 2019-05-26 23:45:00 0.487          110
#> 213 2019-05-27 00:00:00 0.488          110
#> 214 2019-05-27 00:15:00 0.488          110
#> 215 2019-05-27 00:30:00 0.488          110
#> 216 2019-05-27 00:45:00 0.489          110
#> 217 2019-05-27 01:00:00 0.489          110
#> 218 2019-05-27 01:15:00 0.489          110
#> 219 2019-05-27 01:30:00 0.489          110
#> 220 2019-05-27 01:45:00 0.489          110
#> 221 2019-05-27 02:00:00 0.490          110
#> 222 2019-05-27 02:15:00 0.490          110
#> 223 2019-05-27 02:30:00 0.490          110
#> 224 2019-05-27 02:45:00 0.490          110
#> 225 2019-05-27 03:00:00 0.490          110
#> 226 2019-05-27 03:15:00 0.490          110
#> 227 2019-05-27 03:30:00 0.489          110
#> 228 2019-05-27 03:45:00 0.489          110
#> 229 2019-05-27 04:00:00 0.489          110
#> 230 2019-05-27 04:15:00 0.489          110
#> 231 2019-05-27 04:30:00 0.488          110
#> 232 2019-05-27 04:45:00 0.488          110
#> 233 2019-05-27 05:00:00 0.488          110
#> 234 2019-05-27 05:15:00 0.488          110
#> 235 2019-05-27 05:30:00 0.487          110
#> 236 2019-05-27 05:45:00 0.487          110
#> 237 2019-05-27 06:00:00 0.487          110
#> 238 2019-05-27 06:15:00 0.487          110
#> 239 2019-05-27 06:30:00 0.487          110
#> 240 2019-05-27 06:45:00 0.486          110
#> 241 2019-05-27 07:00:00 0.486          110
#> 242 2019-05-27 07:15:00 0.487          110
#> 243 2019-05-27 07:30:00 0.487          110
#> 244 2019-05-27 07:45:00 0.486          110
#> 245 2019-05-27 08:00:00 0.486          110
#> 246 2019-05-27 08:15:00 0.486          110
#> 247 2019-05-27 08:30:00 0.485          110
#> 248 2019-05-27 08:45:00 0.485          110
#> 249 2019-05-27 09:00:00 0.485          110
#> 250 2019-05-27 09:15:00 0.485          110
#> 251 2019-05-27 09:30:00 0.485          110
#> 252 2019-05-27 09:45:00 0.484          180
#> 253 2019-05-27 10:00:00 0.482          110
#> 254 2019-05-27 10:15:00 0.486          110
#> 255 2019-05-27 10:30:00 0.487          110
#> 256 2019-05-27 10:45:00 0.488          110
#> 257 2019-05-27 11:00:00 0.489          110
#> 258 2019-05-27 11:15:00 0.489          110
#> 259 2019-05-27 11:30:00 0.489          110
#> 260 2019-05-27 11:45:00 0.490          110
#> 261 2019-05-27 12:00:00 0.490          110
#> 262 2019-05-27 12:15:00 0.490          110
#> 263 2019-05-27 12:30:00 0.490          110
#> 264 2019-05-27 12:45:00 0.490          110
#> 265 2019-05-27 13:00:00 0.490          110
#> 266 2019-05-27 13:15:00 0.490          110
#> 267 2019-05-27 13:30:00 0.491          110
#> 268 2019-05-27 13:45:00 0.491          110
#> 269 2019-05-27 14:00:00 0.490          110
#> 270 2019-05-27 14:15:00 0.490          110
#> 271 2019-05-27 14:30:00 0.489          110
#> 272 2019-05-27 14:45:00 0.488          110
#> 273 2019-05-27 15:00:00 0.487          110
#> 274 2019-05-27 15:15:00 0.486          110
#> 275 2019-05-27 15:30:00 0.486          110
#> 276 2019-05-27 15:45:00 0.485          110
#> 277 2019-05-27 16:00:00 0.485          110
#> 278 2019-05-27 16:15:00 0.485          110
#> 279 2019-05-27 16:30:00 0.485          110
#> 280 2019-05-27 16:45:00 0.485          110
#> 281 2019-05-27 17:00:00 0.485          110
#> 282 2019-05-27 17:15:00 0.485          110
#> 283 2019-05-27 17:30:00 0.485          110
#> 284 2019-05-27 17:45:00 0.485          110
#> 285 2019-05-27 18:00:00 0.485          110
#> 286 2019-05-27 18:15:00 0.486          110

get_timeseries_tsid("55419010", from = "2017-06-01", to = "2017-06-03",
                    datasource = 4)

#>               Timestamp Value Quality Code
#> 1   2017-06-01 00:00:00  2.23           11
#> 2   2017-06-01 00:10:00  2.11           11
#> 3   2017-06-01 00:20:00  1.99           11
#> 4   2017-06-01 00:30:00  1.87           11
#> 5   2017-06-01 00:40:00  1.75           11
#> 6   2017-06-01 00:50:00  1.64           11
#> 7   2017-06-01 01:00:00  1.53           11
#> 8   2017-06-01 01:10:00  1.42           11
#> 9   2017-06-01 01:20:00  1.31           11
#> 10  2017-06-01 01:30:00  1.20           11
#> 11  2017-06-01 01:40:00  1.09           11
#> 12  2017-06-01 01:50:00  0.99           11
#> 13  2017-06-01 02:00:00  0.90           11
#> 14  2017-06-01 02:10:00  0.80           11
#> 15  2017-06-01 02:20:00  0.71           11
#> 16  2017-06-01 02:30:00  0.62           11
#> 17  2017-06-01 02:40:00  0.53           11
#> 18  2017-06-01 02:50:00  0.44           11
#> 19  2017-06-01 03:00:00  0.36           11
#> 20  2017-06-01 03:10:00  0.28           11
#> 21  2017-06-01 03:20:00  0.20           11
#> 22  2017-06-01 03:30:00  0.13           11
#> 23  2017-06-01 03:40:00  0.07           11
#> 24  2017-06-01 03:50:00  0.01           11
#> 25  2017-06-01 04:00:00 -0.02           11
#> 26  2017-06-01 04:10:00  0.08           11
#> 27  2017-06-01 04:20:00  0.38           11
#> 28  2017-06-01 04:30:00  0.69           11
#> 29  2017-06-01 04:40:00  0.94           11
#> 30  2017-06-01 04:50:00  1.14           11
#> 31  2017-06-01 05:00:00  1.31           11
#> 32  2017-06-01 05:10:00  1.47           11
#> 33  2017-06-01 05:20:00  1.63           11
#> 34  2017-06-01 05:30:00  1.79           11
#> 35  2017-06-01 05:40:00  1.95           11
#> 36  2017-06-01 05:50:00  2.11           11
#> 37  2017-06-01 06:00:00  2.26           11
#> 38  2017-06-01 06:10:00  2.43           11
#> 39  2017-06-01 06:20:00  2.58           11
#> 40  2017-06-01 06:30:00  2.75           11
#> 41  2017-06-01 06:40:00  2.92           11
#> 42  2017-06-01 06:50:00  3.13           11
#> 43  2017-06-01 07:00:00  3.36           11
#> 44  2017-06-01 07:10:00  3.62           11
#> 45  2017-06-01 07:20:00  3.91           11
#> 46  2017-06-01 07:30:00  4.19           11
#> 47  2017-06-01 07:40:00  4.47           11
#> 48  2017-06-01 07:50:00  4.71           11
#> 49  2017-06-01 08:00:00  4.92           11
#> 50  2017-06-01 08:10:00  5.09           11
#> 51  2017-06-01 08:20:00  5.21           11
#> 52  2017-06-01 08:30:00  5.31           11
#> 53  2017-06-01 08:40:00  5.37           11
#> 54  2017-06-01 08:50:00  5.39           11
#> 55  2017-06-01 09:00:00  5.39           11
#> 56  2017-06-01 09:10:00  5.36           11
#> 57  2017-06-01 09:20:00  5.27           11
#> 58  2017-06-01 09:30:00  5.14           11
#> 59  2017-06-01 09:40:00  4.96           11
#> 60  2017-06-01 09:50:00  4.76           11
#> 61  2017-06-01 10:00:00  4.56           11
#> 62  2017-06-01 10:10:00  4.37           11
#> 63  2017-06-01 10:20:00  4.18           11
#> 64  2017-06-01 10:30:00  4.01           11
#> 65  2017-06-01 10:40:00  3.83           11
#> 66  2017-06-01 10:50:00  3.66           11
#> 67  2017-06-01 11:00:00  3.49           11
#> 68  2017-06-01 11:10:00  3.34           11
#> 69  2017-06-01 11:20:00  3.18           11
#> 70  2017-06-01 11:30:00  3.03           11
#> 71  2017-06-01 11:40:00  2.88           11
#> 72  2017-06-01 11:50:00  2.73           11
#> 73  2017-06-01 12:00:00  2.60           11
#> 74  2017-06-01 12:10:00  2.46           11
#> 75  2017-06-01 12:20:00  2.33           11
#> 76  2017-06-01 12:30:00  2.20           11
#> 77  2017-06-01 12:40:00  2.07           11
#> 78  2017-06-01 12:50:00  1.95           11
#> 79  2017-06-01 13:00:00  1.84           11
#> 80  2017-06-01 13:10:00  1.72           11
#> 81  2017-06-01 13:20:00  1.61           11
#> 82  2017-06-01 13:30:00  1.50           11
#> 83  2017-06-01 13:40:00  1.39           11
#> 84  2017-06-01 13:50:00  1.29           11
#> 85  2017-06-01 14:00:00  1.19           11
#> 86  2017-06-01 14:10:00  1.09           11
#> 87  2017-06-01 14:20:00  0.99           11
#> 88  2017-06-01 14:30:00  0.89           11
#> 89  2017-06-01 14:40:00  0.80           11
#> 90  2017-06-01 14:50:00  0.72           11
#> 91  2017-06-01 15:00:00  0.63           11
#> 92  2017-06-01 15:10:00  0.55           11
#> 93  2017-06-01 15:20:00  0.47           11
#> 94  2017-06-01 15:30:00  0.40           11
#> 95  2017-06-01 15:40:00  0.33           11
#> 96  2017-06-01 15:50:00  0.27           11
#> 97  2017-06-01 16:00:00  0.23           11
#> 98  2017-06-01 16:10:00  0.22           11
#> 99  2017-06-01 16:20:00  0.33           11
#> 100 2017-06-01 16:30:00  0.66           11
#> 101 2017-06-01 16:40:00  0.93           11
#> 102 2017-06-01 16:50:00  1.16           11
#> 103 2017-06-01 17:00:00  1.35           11
#> 104 2017-06-01 17:10:00  1.51           11
#> 105 2017-06-01 17:20:00  1.65           11
#> 106 2017-06-01 17:30:00  1.79           11
#> 107 2017-06-01 17:40:00  1.95           11
#> 108 2017-06-01 17:50:00  2.10           11
#> 109 2017-06-01 18:00:00  2.24           11
#> 110 2017-06-01 18:10:00  2.37           11
#> 111 2017-06-01 18:20:00  2.51           11
#> 112 2017-06-01 18:30:00  2.64           11
#> 113 2017-06-01 18:40:00  2.77           11
#> 114 2017-06-01 18:50:00  2.90           11
#> 115 2017-06-01 19:00:00  3.04           11
#> 116 2017-06-01 19:10:00  3.18           11
#> 117 2017-06-01 19:20:00  3.33           11
#> 118 2017-06-01 19:30:00  3.49           11
#> 119 2017-06-01 19:40:00  3.66           11
#> 120 2017-06-01 19:50:00  3.87           11
#> 121 2017-06-01 20:00:00  4.08           11
#> 122 2017-06-01 20:10:00  4.31           11
#> 123 2017-06-01 20:20:00  4.52           11
#> 124 2017-06-01 20:30:00  4.72           11
#> 125 2017-06-01 20:40:00  4.89           11
#> 126 2017-06-01 20:50:00  5.02           11
#> 127 2017-06-01 21:00:00  5.11           11
#> 128 2017-06-01 21:10:00  5.17           11
#> 129 2017-06-01 21:20:00  5.20           11
#> 130 2017-06-01 21:30:00  5.19           11
#> 131 2017-06-01 21:40:00  5.12           11
#> 132 2017-06-01 21:50:00  5.00           11
#> 133 2017-06-01 22:00:00  4.82           11
#> 134 2017-06-01 22:10:00  4.62           11
#> 135 2017-06-01 22:20:00  4.41           11
#> 136 2017-06-01 22:30:00  4.21           11
#> 137 2017-06-01 22:40:00  4.03           11
#> 138 2017-06-01 22:50:00  3.86           11
#> 139 2017-06-01 23:00:00  3.70           15
#> 140 2017-06-01 23:10:00  3.54           11
#> 141 2017-06-01 23:20:00  3.38           11
#> 142 2017-06-01 23:30:00  3.23           11
#> 143 2017-06-01 23:40:00  3.08           11
#> 144 2017-06-01 23:50:00  2.93           11
#> 145 2017-06-02 00:00:00  2.79           11
#> 146 2017-06-02 00:10:00  2.65           11
#> 147 2017-06-02 00:20:00  2.52           11
#> 148 2017-06-02 00:30:00  2.39           11
#> 149 2017-06-02 00:40:00  2.27           11
#> 150 2017-06-02 00:50:00  2.14           11
#> 151 2017-06-02 01:00:00  2.02           11
#> 152 2017-06-02 01:10:00  1.91           11
#> 153 2017-06-02 01:20:00  1.79           11
#> 154 2017-06-02 01:30:00  1.68           11
#> 155 2017-06-02 01:40:00  1.57           11
#> 156 2017-06-02 01:50:00  1.46           11
#> 157 2017-06-02 02:00:00  1.35           11
#> 158 2017-06-02 02:10:00  1.25           11
#> 159 2017-06-02 02:20:00  1.15           11
#> 160 2017-06-02 02:30:00  1.05           11
#> 161 2017-06-02 02:40:00  0.95           11
#> 162 2017-06-02 02:50:00  0.86           11
#> 163 2017-06-02 03:00:00  0.77           11
#> 164 2017-06-02 03:10:00  0.68           11
#> 165 2017-06-02 03:20:00  0.59           11
#> 166 2017-06-02 03:30:00  0.51           11
#> 167 2017-06-02 03:40:00  0.43           11
#> 168 2017-06-02 03:50:00  0.35           11
#> 169 2017-06-02 04:00:00  0.28           11
#> 170 2017-06-02 04:10:00  0.21           11
#> 171 2017-06-02 04:20:00  0.15           11
#> 172 2017-06-02 04:30:00  0.09           11
#> 173 2017-06-02 04:40:00  0.04           11
#> 174 2017-06-02 04:50:00  0.02           11
#> 175 2017-06-02 05:00:00  0.11           11
#> 176 2017-06-02 05:10:00  0.37           11
#> 177 2017-06-02 05:20:00  0.66           11
#> 178 2017-06-02 05:30:00  0.90           11
#> 179 2017-06-02 05:40:00  1.10           11
#> 180 2017-06-02 05:50:00  1.27           11
#> 181 2017-06-02 06:00:00  1.43           11
#> 182 2017-06-02 06:10:00  1.60           11
#> 183 2017-06-02 06:20:00  1.77           11
#> 184 2017-06-02 06:30:00  1.94           11
#> 185 2017-06-02 06:40:00  2.10           11
#> 186 2017-06-02 06:50:00  2.26           11
#> 187 2017-06-02 07:00:00  2.42           11
#> 188 2017-06-02 07:10:00  2.59           11
#> 189 2017-06-02 07:20:00  2.76           11
#> 190 2017-06-02 07:30:00  2.95           11
#> 191 2017-06-02 07:40:00  3.15           11
#> 192 2017-06-02 07:50:00  3.36           11
#> 193 2017-06-02 08:00:00  3.59           11
#> 194 2017-06-02 08:10:00  3.83           11
#> 195 2017-06-02 08:20:00  4.08           11
#> 196 2017-06-02 08:30:00  4.33           11
#> 197 2017-06-02 08:40:00  4.57           11
#> 198 2017-06-02 08:50:00  4.79           11
#> 199 2017-06-02 09:00:00  4.97           11
#> 200 2017-06-02 09:10:00  5.11           11
#> 201 2017-06-02 09:20:00  5.23           11
#> 202 2017-06-02 09:30:00  5.31           11
#> 203 2017-06-02 09:40:00  5.38           11
#> 204 2017-06-02 09:50:00  5.41           11
#> 205 2017-06-02 10:00:00  5.41           11
#> 206 2017-06-02 10:10:00  5.39           11
#> 207 2017-06-02 10:20:00  5.31           11
#> 208 2017-06-02 10:30:00  5.19           11
#> 209 2017-06-02 10:40:00  5.03           11
#> 210 2017-06-02 10:50:00  4.84           11
#> 211 2017-06-02 11:00:00  4.66           11
#> 212 2017-06-02 11:10:00  4.48           11
#> 213 2017-06-02 11:20:00  4.30           11
#> 214 2017-06-02 11:30:00  4.11           11
#> 215 2017-06-02 11:40:00  3.95           11
#> 216 2017-06-02 11:50:00  3.78           11
#> 217 2017-06-02 12:00:00  3.63           11
#> 218 2017-06-02 12:10:00  3.46           11
#> 219 2017-06-02 12:20:00  3.30           11
#> 220 2017-06-02 12:30:00  3.15           11
#> 221 2017-06-02 12:40:00  3.00           11
#> 222 2017-06-02 12:50:00  2.86           11
#> 223 2017-06-02 13:00:00  2.72           11
#> 224 2017-06-02 13:10:00  2.59           11
#> 225 2017-06-02 13:20:00  2.46           11
#> 226 2017-06-02 13:30:00  2.33           11
#> 227 2017-06-02 13:40:00  2.20           11
#> 228 2017-06-02 13:50:00  2.09           11
#> 229 2017-06-02 14:00:00  1.97           11
#> 230 2017-06-02 14:10:00  1.86           11
#> 231 2017-06-02 14:20:00  1.75           11
#> 232 2017-06-02 14:30:00  1.64           11
#> 233 2017-06-02 14:40:00  1.54           11
#> 234 2017-06-02 14:50:00  1.44           11
#> 235 2017-06-02 15:00:00  1.33           11
#> 236 2017-06-02 15:10:00  1.23           11
#> 237 2017-06-02 15:20:00  1.14           11
#> 238 2017-06-02 15:30:00  1.04           11
#> 239 2017-06-02 15:40:00  0.95           11
#> 240 2017-06-02 15:50:00  0.87           11
#> 241 2017-06-02 16:00:00  0.78           11
#> 242 2017-06-02 16:10:00  0.71           11
#> 243 2017-06-02 16:20:00  0.63           11
#> 244 2017-06-02 16:30:00  0.57           11
#> 245 2017-06-02 16:40:00  0.51           11
#> 246 2017-06-02 16:50:00  0.46           11
#> 247 2017-06-02 17:00:00  0.45           11
#> 248 2017-06-02 17:10:00  0.53           11
#> 249 2017-06-02 17:20:00  0.75           11
#> 250 2017-06-02 17:30:00  1.04           11
#> 251 2017-06-02 17:40:00  1.26           11
#> 252 2017-06-02 17:50:00  1.46           11
#> 253 2017-06-02 18:00:00  1.62           11
#> 254 2017-06-02 18:10:00  1.77           11
#> 255 2017-06-02 18:20:00  1.91           11
#> 256 2017-06-02 18:30:00  2.07           11
#> 257 2017-06-02 18:40:00  2.21           11
#> 258 2017-06-02 18:50:00  2.35           11
#> 259 2017-06-02 19:00:00  2.48           11
#> 260 2017-06-02 19:10:00  2.62           11
#> 261 2017-06-02 19:20:00  2.74           11
#> 262 2017-06-02 19:30:00  2.87           11
#> 263 2017-06-02 19:40:00  2.99           11
#> 264 2017-06-02 19:50:00  3.11           11
#> 265 2017-06-02 20:00:00  3.24           11
#> 266 2017-06-02 20:10:00  3.37           11
#> 267 2017-06-02 20:20:00  3.52           11
#> 268 2017-06-02 20:30:00  3.67           11
#> 269 2017-06-02 20:40:00  3.84           11
#> 270 2017-06-02 20:50:00  4.01           11
#> 271 2017-06-02 21:00:00  4.20           11
#> 272 2017-06-02 21:10:00  4.39           11
#> 273 2017-06-02 21:20:00  4.58           11
#> 274 2017-06-02 21:30:00  4.75           11
#> 275 2017-06-02 21:40:00  4.91           11
#> 276 2017-06-02 21:50:00  5.03           11
#> 277 2017-06-02 22:00:00  5.13           11
#> 278 2017-06-02 22:10:00  5.20           11
#> 279 2017-06-02 22:20:00  5.22           11
#> 280 2017-06-02 22:30:00  5.22           11
#> 281 2017-06-02 22:40:00  5.17           11
#> 282 2017-06-02 22:50:00  5.07           11
#> 283 2017-06-02 23:00:00  4.92           11
#> 284 2017-06-02 23:10:00  4.74           11
#> 285 2017-06-02 23:20:00  4.54           11
#> 286 2017-06-02 23:30:00  4.34           11
#> 287 2017-06-02 23:40:00  4.17           11
#> 288 2017-06-02 23:50:00  4.01           11
#> 289 2017-06-03 00:00:00  3.85           11