Download timeseries data from waterinfo.be

Using the ts_id codes and by providing a given date period, download the corresponding time series from the waterinfo.be website

get_timeseries_tsid(ts_id, period = NULL, from = NULL, to = NULL,
  datasource = 1, token = NULL)

Arguments

ts_id
waterinfo.be database ts_id, defining a timeserie variable and frequency it is defined.

period
input string according to format required by waterinfo: De period string is provided as P#Y#M#DT#H#M#S, with P defines `Period`, each # is an integer value and the codes define the number of... Y - years M - months D - days T required if information about sub-day resolution is present H - hours D - days M - minutes S - seconds Instead of D (days), the usage of W - weeks is possible as well Examples of valid period strings: P3D, P1Y, P1DT12H, PT6H, P1Y6M3DT4H20M30S.

from
date of datestring as start of the time series

to
date of datestring as end of the time series

datasource
int [0-3] defines the `meetnet` of which the measurement station is part of. VMM based stations are net '1', MOW-HIC is net '2'

token
token to use with the call (optional, can be retrieved via get_token)

Value

data.frame with the timestamps, values and quality code

Examples

get_timeseries_tsid("35055042", from="2017-01-01", to="2017-01-02")

#>              Timestamp Value Quality Code
#> 1  2017-01-01 00:00:00  0.00          130
#> 2  2017-01-01 00:15:00  0.00          130
#> 3  2017-01-01 00:30:00  0.00          130
#> 4  2017-01-01 00:45:00  0.00          130
#> 5  2017-01-01 01:00:00  0.00          130
#> 6  2017-01-01 01:15:00  0.00          130
#> 7  2017-01-01 01:30:00  0.00          130
#> 8  2017-01-01 01:45:00  0.00          130
#> 9  2017-01-01 02:00:00  0.00          130
#> 10 2017-01-01 02:15:00  0.00          130
#> 11 2017-01-01 02:30:00  0.00          130
#> 12 2017-01-01 02:45:00  0.00          130
#> 13 2017-01-01 03:00:00  0.00          130
#> 14 2017-01-01 03:15:00  0.00          130
#> 15 2017-01-01 03:30:00  0.00          130
#> 16 2017-01-01 03:45:00  0.00          130
#> 17 2017-01-01 04:00:00  0.00          130
#> 18 2017-01-01 04:15:00  0.00          130
#> 19 2017-01-01 04:30:00  0.00          130
#> 20 2017-01-01 04:45:00  0.00          130
#> 21 2017-01-01 05:00:00  0.00          130
#> 22 2017-01-01 05:15:00  0.00          130
#> 23 2017-01-01 05:30:00  0.00          130
#> 24 2017-01-01 05:45:00  0.00          130
#> 25 2017-01-01 06:00:00  0.00          130
#> 26 2017-01-01 06:15:00  0.00          130
#> 27 2017-01-01 06:30:00  0.00          130
#> 28 2017-01-01 06:45:00  0.00          130
#> 29 2017-01-01 07:00:00  0.00          130
#> 30 2017-01-01 07:15:00  0.00          130
#> 31 2017-01-01 07:30:00  0.00          130
#> 32 2017-01-01 07:45:00  0.00          130
#> 33 2017-01-01 08:00:00  0.00          130
#> 34 2017-01-01 08:15:00  0.00          130
#> 35 2017-01-01 08:30:00  0.00          130
#> 36 2017-01-01 08:45:00  0.00          130
#> 37 2017-01-01 09:00:00  0.00          130
#> 38 2017-01-01 09:15:00  0.00          130
#> 39 2017-01-01 09:30:00  0.00          130
#> 40 2017-01-01 09:45:00  0.00          130
#> 41 2017-01-01 10:00:00  0.00          130
#> 42 2017-01-01 10:15:00  0.00          130
#> 43 2017-01-01 10:30:00  0.00          130
#> 44 2017-01-01 10:45:00  0.00          130
#> 45 2017-01-01 11:00:00  0.00          130
#> 46 2017-01-01 11:15:00  0.00          130
#> 47 2017-01-01 11:30:00  0.00          130
#> 48 2017-01-01 11:45:00  0.00          130
#> 49 2017-01-01 12:00:00  0.00          130
#> 50 2017-01-01 12:15:00  0.00          130
#> 51 2017-01-01 12:30:00  0.00          130
#> 52 2017-01-01 12:45:00  0.00          130
#> 53 2017-01-01 13:00:00  0.00          130
#> 54 2017-01-01 13:15:00  0.00          130
#> 55 2017-01-01 13:30:00  0.00          130
#> 56 2017-01-01 13:45:00  0.00          130
#> 57 2017-01-01 14:00:00  0.00          130
#> 58 2017-01-01 14:15:00  0.00          130
#> 59 2017-01-01 14:30:00  0.00          130
#> 60 2017-01-01 14:45:00  0.00          130
#> 61 2017-01-01 15:00:00  0.00          130
#> 62 2017-01-01 15:15:00  0.00          130
#> 63 2017-01-01 15:30:00  0.00          130
#> 64 2017-01-01 15:45:00  0.00          130
#> 65 2017-01-01 16:00:00  0.00          130
#> 66 2017-01-01 16:15:00  0.00          130
#> 67 2017-01-01 16:30:00  0.00          130
#> 68 2017-01-01 16:45:00  0.00          130
#> 69 2017-01-01 17:00:00  0.00          130
#> 70 2017-01-01 17:15:00  0.00          130
#> 71 2017-01-01 17:30:00  0.00          130
#> 72 2017-01-01 17:45:00  0.03          130
#> 73 2017-01-01 18:00:00  0.00          130
#> 74 2017-01-01 18:15:00  0.04          130
#> 75 2017-01-01 18:30:00  0.00          130
#> 76 2017-01-01 18:45:00  0.03          130
#> 77 2017-01-01 19:00:00  0.00          130
#> 78 2017-01-01 19:15:00  0.03          130
#> 79 2017-01-01 19:30:00  0.03          130
#> 80 2017-01-01 19:45:00  0.00          130
#> 81 2017-01-01 20:00:00  0.03          130
#> 82 2017-01-01 20:15:00  0.00          130
#> 83 2017-01-01 20:30:00  0.00          130
#> 84 2017-01-01 20:45:00  0.00          130
#> 85 2017-01-01 21:00:00  0.00          130
#> 86 2017-01-01 21:15:00  0.03          130
#> 87 2017-01-01 21:30:00  0.00          130
#> 88 2017-01-01 21:45:00  0.00          130
#> 89 2017-01-01 22:00:00  0.00          130
#> 90 2017-01-01 22:15:00  0.11          130
#> 91 2017-01-01 22:30:00  0.82          130
#> 92 2017-01-01 22:45:00  0.69          130
#> 93 2017-01-01 23:00:00  0.34          130
#> 94 2017-01-01 23:15:00  0.13          130
#> 95 2017-01-01 23:30:00  0.03          130
#> 96 2017-01-01 23:45:00  0.03          130
#> 97 2017-01-02 00:00:00  0.05          130

get_timeseries_tsid("5156042", period = "P3D")

#>               Timestamp Value Quality Code
#> 1   2018-01-15 10:45:00 0.422          110
#> 2   2018-01-15 11:00:00 0.422          110
#> 3   2018-01-15 11:15:00 0.421          110
#> 4   2018-01-15 11:30:00 0.420          110
#> 5   2018-01-15 11:45:00 0.420          110
#> 6   2018-01-15 12:00:00 0.420          110
#> 7   2018-01-15 12:15:00 0.420          110
#> 8   2018-01-15 12:30:00 0.420          110
#> 9   2018-01-15 12:45:00 0.420          110
#> 10  2018-01-15 13:00:00 0.420          110
#> 11  2018-01-15 13:15:00 0.419          110
#> 12  2018-01-15 13:30:00 0.420          110
#> 13  2018-01-15 13:45:00 0.421          110
#> 14  2018-01-15 14:00:00 0.421          110
#> 15  2018-01-15 14:15:00 0.421          110
#> 16  2018-01-15 14:30:00 0.422          110
#> 17  2018-01-15 14:45:00 0.423          110
#> 18  2018-01-15 15:00:00 0.423          110
#> 19  2018-01-15 15:15:00 0.423          110
#> 20  2018-01-15 15:30:00 0.424          110
#> 21  2018-01-15 15:45:00 0.426          110
#> 22  2018-01-15 16:00:00 0.428          110
#> 23  2018-01-15 16:15:00 0.430          110
#> 24  2018-01-15 16:30:00 0.432          110
#> 25  2018-01-15 16:45:00 0.434          110
#> 26  2018-01-15 17:00:00 0.435          110
#> 27  2018-01-15 17:15:00 0.436          110
#> 28  2018-01-15 17:30:00 0.438          110
#> 29  2018-01-15 17:45:00 0.439          110
#> 30  2018-01-15 18:00:00 0.441          110
#> 31  2018-01-15 18:15:00 0.443          110
#> 32  2018-01-15 18:30:00 0.445          110
#> 33  2018-01-15 18:45:00 0.447          110
#> 34  2018-01-15 19:00:00 0.449          110
#> 35  2018-01-15 19:15:00 0.451          110
#> 36  2018-01-15 19:30:00 0.455          110
#> 37  2018-01-15 19:45:00 0.458          110
#> 38  2018-01-15 20:00:00 0.468          110
#> 39  2018-01-15 20:15:00 0.481          110
#> 40  2018-01-15 20:30:00 0.494          110
#> 41  2018-01-15 20:45:00 0.504          110
#> 42  2018-01-15 21:00:00 0.514          110
#> 43  2018-01-15 21:15:00 0.522          110
#> 44  2018-01-15 21:30:00 0.532          110
#> 45  2018-01-15 21:45:00 0.541          110
#> 46  2018-01-15 22:00:00 0.550          110
#> 47  2018-01-15 22:15:00 0.558          110
#> 48  2018-01-15 22:30:00 0.566          110
#> 49  2018-01-15 22:45:00 0.573          110
#> 50  2018-01-15 23:00:00 0.580          110
#> 51  2018-01-15 23:15:00 0.586          110
#> 52  2018-01-15 23:30:00 0.592          110
#> 53  2018-01-15 23:45:00 0.597          110
#> 54  2018-01-16 00:00:00 0.604          110
#> 55  2018-01-16 00:15:00 0.610          110
#> 56  2018-01-16 00:30:00 0.617          110
#> 57  2018-01-16 00:45:00 0.624          110
#> 58  2018-01-16 01:00:00 0.630          110
#> 59  2018-01-16 01:15:00 0.637          110
#> 60  2018-01-16 01:30:00 0.642          110
#> 61  2018-01-16 01:45:00 0.648          110
#> 62  2018-01-16 02:00:00 0.653          110
#> 63  2018-01-16 02:15:00 0.658          110
#> 64  2018-01-16 02:30:00 0.663          110
#> 65  2018-01-16 02:45:00 0.668          110
#> 66  2018-01-16 03:00:00 0.672          110
#> 67  2018-01-16 03:15:00 0.677          110
#> 68  2018-01-16 03:30:00 0.682          110
#> 69  2018-01-16 03:45:00 0.686          110
#> 70  2018-01-16 04:00:00 0.689          110
#> 71  2018-01-16 04:15:00 0.692          110
#> 72  2018-01-16 04:30:00 0.695          110
#> 73  2018-01-16 04:45:00 0.697          110
#> 74  2018-01-16 05:00:00 0.699          110
#> 75  2018-01-16 05:15:00 0.701          110
#> 76  2018-01-16 05:30:00 0.703          110
#> 77  2018-01-16 05:45:00 0.704          110
#> 78  2018-01-16 06:00:00 0.706          110
#> 79  2018-01-16 06:15:00 0.707          110
#> 80  2018-01-16 06:30:00 0.708          110
#> 81  2018-01-16 06:45:00 0.710          110
#> 82  2018-01-16 07:00:00 0.712          110
#> 83  2018-01-16 07:15:00 0.713          110
#> 84  2018-01-16 07:30:00 0.714          110
#> 85  2018-01-16 07:45:00 0.715          110
#> 86  2018-01-16 08:00:00 0.716          110
#> 87  2018-01-16 08:15:00 0.716          110
#> 88  2018-01-16 08:30:00 0.717          110
#> 89  2018-01-16 08:45:00 0.717          110
#> 90  2018-01-16 09:00:00 0.718          110
#> 91  2018-01-16 09:15:00 0.718          110
#> 92  2018-01-16 09:30:00 0.718          110
#> 93  2018-01-16 09:45:00 0.718          110
#> 94  2018-01-16 10:00:00 0.718          110
#> 95  2018-01-16 10:15:00 0.718          110
#> 96  2018-01-16 10:30:00 0.719          110
#> 97  2018-01-16 10:45:00 0.718          110
#> 98  2018-01-16 11:00:00 0.718          110
#> 99  2018-01-16 11:15:00 0.719          110
#> 100 2018-01-16 11:30:00 0.720          110
#> 101 2018-01-16 11:45:00 0.719          110
#> 102 2018-01-16 12:00:00 0.720          110
#> 103 2018-01-16 12:15:00 0.720          110
#> 104 2018-01-16 12:30:00 0.720          110
#> 105 2018-01-16 12:45:00 0.720          110
#> 106 2018-01-16 13:00:00 0.722          110
#> 107 2018-01-16 13:15:00 0.724          110
#> 108 2018-01-16 13:30:00 0.726          110
#> 109 2018-01-16 13:45:00 0.727          110
#> 110 2018-01-16 14:00:00 0.727          110
#> 111 2018-01-16 14:15:00 0.727          110
#> 112 2018-01-16 14:30:00 0.728          110
#> 113 2018-01-16 14:45:00 0.728          110
#> 114 2018-01-16 15:00:00 0.728          110
#> 115 2018-01-16 15:15:00 0.728          110
#> 116 2018-01-16 15:30:00 0.729          110
#> 117 2018-01-16 15:45:00 0.730          110
#> 118 2018-01-16 16:00:00 0.732          110
#> 119 2018-01-16 16:15:00 0.736          110
#> 120 2018-01-16 16:30:00 0.740          110
#> 121 2018-01-16 16:45:00 0.743          110
#> 122 2018-01-16 17:00:00 0.746          110
#> 123 2018-01-16 17:15:00 0.749          110
#> 124 2018-01-16 17:30:00 0.751          110
#> 125 2018-01-16 17:45:00 0.752          110
#> 126 2018-01-16 18:00:00 0.754          110
#> 127 2018-01-16 18:15:00 0.755          110
#> 128 2018-01-16 18:30:00 0.755          110
#> 129 2018-01-16 18:45:00 0.756          110
#> 130 2018-01-16 19:00:00 0.757          110
#> 131 2018-01-16 19:15:00 0.757          110
#> 132 2018-01-16 19:30:00 0.757          110
#> 133 2018-01-16 19:45:00 0.758          110
#> 134 2018-01-16 20:00:00 0.758          110
#> 135 2018-01-16 20:15:00 0.758          110
#> 136 2018-01-16 20:30:00 0.759          110
#> 137 2018-01-16 20:45:00 0.760          110
#> 138 2018-01-16 21:00:00 0.761          110
#> 139 2018-01-16 21:15:00 0.762          110
#> 140 2018-01-16 21:30:00 0.763          110
#> 141 2018-01-16 21:45:00 0.765          110
#> 142 2018-01-16 22:00:00 0.766          110
#> 143 2018-01-16 22:15:00 0.767          110
#> 144 2018-01-16 22:30:00 0.769          110
#> 145 2018-01-16 22:45:00 0.770          110
#> 146 2018-01-16 23:00:00 0.771          110
#> 147 2018-01-16 23:15:00 0.773          110
#> 148 2018-01-16 23:30:00 0.774          110
#> 149 2018-01-16 23:45:00 0.774          110
#> 150 2018-01-17 00:00:00 0.775          110
#> 151 2018-01-17 00:15:00 0.776          110
#> 152 2018-01-17 00:30:00 0.777          110
#> 153 2018-01-17 00:45:00 0.778          110
#> 154 2018-01-17 01:00:00 0.779          110
#> 155 2018-01-17 01:15:00 0.779          110
#> 156 2018-01-17 01:30:00 0.779          110
#> 157 2018-01-17 01:45:00 0.779          110
#> 158 2018-01-17 02:00:00 0.780          110
#> 159 2018-01-17 02:15:00 0.779          110
#> 160 2018-01-17 02:30:00 0.780          110
#> 161 2018-01-17 02:45:00 0.780          110
#> 162 2018-01-17 03:00:00 0.780          110
#> 163 2018-01-17 03:15:00 0.780          110
#> 164 2018-01-17 03:30:00 0.781          110
#> 165 2018-01-17 03:45:00 0.781          110
#> 166 2018-01-17 04:00:00 0.781          110
#> 167 2018-01-17 04:15:00 0.782          110
#> 168 2018-01-17 04:30:00 0.782          110
#> 169 2018-01-17 04:45:00 0.783          110
#> 170 2018-01-17 05:00:00 0.782          110
#> 171 2018-01-17 05:15:00 0.782          110
#> 172 2018-01-17 05:30:00 0.783          110
#> 173 2018-01-17 05:45:00 0.782          110
#> 174 2018-01-17 06:00:00 0.784          110
#> 175 2018-01-17 06:15:00 0.785          110
#> 176 2018-01-17 06:30:00 0.785          110
#> 177 2018-01-17 06:45:00 0.785          110
#> 178 2018-01-17 07:00:00 0.787          110
#> 179 2018-01-17 07:15:00 0.788          110
#> 180 2018-01-17 07:30:00 0.788          110
#> 181 2018-01-17 07:45:00 0.789          110
#> 182 2018-01-17 08:00:00 0.790          110
#> 183 2018-01-17 08:15:00 0.791          110
#> 184 2018-01-17 08:30:00 0.791          110
#> 185 2018-01-17 08:45:00 0.792          110
#> 186 2018-01-17 09:00:00 0.794          110
#> 187 2018-01-17 09:15:00 0.796          110
#> 188 2018-01-17 09:30:00 0.797          110
#> 189 2018-01-17 09:45:00 0.799          110
#> 190 2018-01-17 10:00:00 0.800          110
#> 191 2018-01-17 10:15:00 0.802          110
#> 192 2018-01-17 10:30:00 0.805          110
#> 193 2018-01-17 10:45:00 0.807          110
#> 194 2018-01-17 11:00:00 0.810          110
#> 195 2018-01-17 11:15:00 0.816          110
#> 196 2018-01-17 11:30:00 0.820          110
#> 197 2018-01-17 11:45:00 0.823          110
#> 198 2018-01-17 12:00:00 0.826          110
#> 199 2018-01-17 12:15:00 0.828          110
#> 200 2018-01-17 12:30:00 0.830          110
#> 201 2018-01-17 12:45:00 0.832          110
#> 202 2018-01-17 13:00:00 0.834          110
#> 203 2018-01-17 13:15:00 0.836          110
#> 204 2018-01-17 13:30:00 0.838          110
#> 205 2018-01-17 13:45:00 0.838          110
#> 206 2018-01-17 14:00:00 0.838          110
#> 207 2018-01-17 14:15:00 0.838          110
#> 208 2018-01-17 14:30:00 0.838          110
#> 209 2018-01-17 14:45:00 0.838          110
#> 210 2018-01-17 15:00:00 0.838          110
#> 211 2018-01-17 15:15:00 0.839          110
#> 212 2018-01-17 15:30:00 0.840          110
#> 213 2018-01-17 15:45:00 0.841          110
#> 214 2018-01-17 16:00:00 0.842          110
#> 215 2018-01-17 16:15:00 0.842          110
#> 216 2018-01-17 16:30:00 0.842          110
#> 217 2018-01-17 16:45:00 0.842          110
#> 218 2018-01-17 17:00:00 0.842          110
#> 219 2018-01-17 17:15:00 0.843          110
#> 220 2018-01-17 17:30:00 0.844          110
#> 221 2018-01-17 17:45:00 0.845          110
#> 222 2018-01-17 18:00:00 0.846          110
#> 223 2018-01-17 18:15:00 0.849          110
#> 224 2018-01-17 18:30:00 0.850          110
#> 225 2018-01-17 18:45:00 0.850          110
#> 226 2018-01-17 19:00:00 0.850          110
#> 227 2018-01-17 19:15:00 0.850          110
#> 228 2018-01-17 19:30:00 0.850          110
#> 229 2018-01-17 19:45:00 0.850          110
#> 230 2018-01-17 20:00:00 0.850          110
#> 231 2018-01-17 20:15:00 0.850          110
#> 232 2018-01-17 20:30:00 0.851          110
#> 233 2018-01-17 20:45:00 0.851          110
#> 234 2018-01-17 21:00:00 0.852          110
#> 235 2018-01-17 21:15:00 0.853          110
#> 236 2018-01-17 21:30:00 0.853          110
#> 237 2018-01-17 21:45:00 0.853          110
#> 238 2018-01-17 22:00:00 0.852          110
#> 239 2018-01-17 22:15:00 0.851          110
#> 240 2018-01-17 22:30:00 0.851          110
#> 241 2018-01-17 22:45:00 0.850          110
#> 242 2018-01-17 23:00:00 0.850          110
#> 243 2018-01-17 23:15:00 0.849          110
#> 244 2018-01-17 23:30:00 0.849          110
#> 245 2018-01-17 23:45:00 0.850          110
#> 246 2018-01-18 00:00:00 0.850          110
#> 247 2018-01-18 00:15:00 0.850          110
#> 248 2018-01-18 00:30:00 0.849          110
#> 249 2018-01-18 00:45:00 0.848          110
#> 250 2018-01-18 01:00:00 0.847          110
#> 251 2018-01-18 01:15:00 0.846          110
#> 252 2018-01-18 01:30:00 0.845          110
#> 253 2018-01-18 01:45:00 0.844          110
#> 254 2018-01-18 02:00:00 0.845          110
#> 255 2018-01-18 02:15:00 0.844          110
#> 256 2018-01-18 02:30:00 0.845          110
#> 257 2018-01-18 02:45:00 0.845          110
#> 258 2018-01-18 03:00:00 0.845          110
#> 259 2018-01-18 03:15:00 0.845          110
#> 260 2018-01-18 03:30:00 0.845          110
#> 261 2018-01-18 03:45:00 0.845          110
#> 262 2018-01-18 04:00:00 0.846          110
#> 263 2018-01-18 04:15:00 0.847          110
#> 264 2018-01-18 04:30:00 0.848          110
#> 265 2018-01-18 04:45:00 0.849          110
#> 266 2018-01-18 05:00:00 0.852          110
#> 267 2018-01-18 05:15:00 0.855          110
#> 268 2018-01-18 05:30:00 0.857          110
#> 269 2018-01-18 05:45:00 0.858          110
#> 270 2018-01-18 06:00:00 0.863          110
#> 271 2018-01-18 06:15:00 0.868          110
#> 272 2018-01-18 06:30:00 0.875          110
#> 273 2018-01-18 06:45:00 0.881          110
#> 274 2018-01-18 07:00:00 0.887          110

get_timeseries_tsid("2813562", period = "P1D", datasource = 2)

#>                Timestamp Value Quality Code
#> 1    2018-01-17 10:45:00  2.27          200
#> 2    2018-01-17 10:46:00  2.26          200
#> 3    2018-01-17 10:47:00  2.26          200
#> 4    2018-01-17 10:48:00  2.25          200
#> 5    2018-01-17 10:49:00  2.24          200
#> 6    2018-01-17 10:50:00  2.24          200
#> 7    2018-01-17 10:51:00  2.23          200
#> 8    2018-01-17 10:52:00  2.22          200
#> 9    2018-01-17 10:53:00  2.21          200
#> 10   2018-01-17 10:54:00  2.20          200
#> 11   2018-01-17 10:55:00  2.20          200
#> 12   2018-01-17 10:56:00  2.19          200
#> 13   2018-01-17 10:57:00  2.19          200
#> 14   2018-01-17 10:58:00  2.18          200
#> 15   2018-01-17 10:59:00  2.17          200
#> 16   2018-01-17 11:00:00  2.16          200
#> 17   2018-01-17 11:01:00  2.16          200
#> 18   2018-01-17 11:02:00  2.15          200
#> 19   2018-01-17 11:03:00  2.15          200
#> 20   2018-01-17 11:04:00  2.14          200
#> 21   2018-01-17 11:05:00  2.14          200
#> 22   2018-01-17 11:06:00  2.12          200
#> 23   2018-01-17 11:07:00  2.12          200
#> 24   2018-01-17 11:08:00  2.11          200
#> 25   2018-01-17 11:09:00  2.11          200
#> 26   2018-01-17 11:10:00  2.10          200
#> 27   2018-01-17 11:11:00  2.09          200
#> 28   2018-01-17 11:12:00  2.09          200
#> 29   2018-01-17 11:13:00  2.08          200
#> 30   2018-01-17 11:14:00  2.07          200
#> 31   2018-01-17 11:15:00  2.07          200
#> 32   2018-01-17 11:16:00  2.06          200
#> 33   2018-01-17 11:17:00  2.05          200
#> 34   2018-01-17 11:18:00  2.05          200
#> 35   2018-01-17 11:19:00  2.05          200
#> 36   2018-01-17 11:20:00  2.04          200
#> 37   2018-01-17 11:21:00  2.03          200
#> 38   2018-01-17 11:22:00  2.01          200
#> 39   2018-01-17 11:23:00  2.02          200
#> 40   2018-01-17 11:24:00  2.02          200
#> 41   2018-01-17 11:25:00  2.01          200
#> 42   2018-01-17 11:26:00  1.99          200
#> 43   2018-01-17 11:27:00  1.99          200
#> 44   2018-01-17 11:28:00  1.99          200
#> 45   2018-01-17 11:29:00  1.98          200
#> 46   2018-01-17 11:30:00  1.97          200
#> 47   2018-01-17 11:31:00  1.97          200
#> 48   2018-01-17 11:32:00  1.96          200
#> 49   2018-01-17 11:33:00  1.95          200
#> 50   2018-01-17 11:34:00  1.95          200
#> 51   2018-01-17 11:35:00  1.94          200
#> 52   2018-01-17 11:36:00  1.94          200
#> 53   2018-01-17 11:37:00  1.92          200
#> 54   2018-01-17 11:38:00  1.92          200
#> 55   2018-01-17 11:39:00  1.92          200
#> 56   2018-01-17 11:40:00  1.92          200
#> 57   2018-01-17 11:41:00  1.91          200
#> 58   2018-01-17 11:42:00  1.91          200
#> 59   2018-01-17 11:43:00  1.90          200
#> 60   2018-01-17 11:44:00  1.90          200
#> 61   2018-01-17 11:45:00  1.89          200
#> 62   2018-01-17 11:46:00  1.89          200
#> 63   2018-01-17 11:47:00  1.89          200
#> 64   2018-01-17 11:48:00  1.88          200
#> 65   2018-01-17 11:49:00  1.88          200
#> 66   2018-01-17 11:50:00  1.88          200
#> 67   2018-01-17 11:51:00  1.87          200
#> 68   2018-01-17 11:52:00  1.87          200
#> 69   2018-01-17 11:53:00  1.86          200
#> 70   2018-01-17 11:54:00  1.87          200
#> 71   2018-01-17 11:55:00  1.86          200
#> 72   2018-01-17 11:56:00  1.86          200
#> 73   2018-01-17 11:57:00  1.86          200
#> 74   2018-01-17 11:58:00  1.87          200
#> 75   2018-01-17 11:59:00  1.87          200
#> 76   2018-01-17 12:00:00  1.88          200
#> 77   2018-01-17 12:01:00  1.88          200
#> 78   2018-01-17 12:02:00  1.91          200
#> 79   2018-01-17 12:03:00  1.93          200
#> 80   2018-01-17 12:04:00  1.95          200
#> 81   2018-01-17 12:05:00  1.96          200
#> 82   2018-01-17 12:06:00  2.00          200
#> 83   2018-01-17 12:07:00  2.04          200
#> 84   2018-01-17 12:08:00  2.06          200
#> 85   2018-01-17 12:09:00  2.10          200
#> 86   2018-01-17 12:10:00  2.13          200
#> 87   2018-01-17 12:11:00  2.17          200
#> 88   2018-01-17 12:12:00  2.19          200
#> 89   2018-01-17 12:13:00  2.22          200
#> 90   2018-01-17 12:14:00  2.29          200
#> 91   2018-01-17 12:15:00  2.34          200
#> 92   2018-01-17 12:16:00  2.37          200
#> 93   2018-01-17 12:17:00  2.40          200
#> 94   2018-01-17 12:18:00  2.41          200
#> 95   2018-01-17 12:19:00  2.46          200
#> 96   2018-01-17 12:20:00  2.46          200
#> 97   2018-01-17 12:21:00  2.50          200
#> 98   2018-01-17 12:22:00  2.51          200
#> 99   2018-01-17 12:23:00  2.55          200
#> 100  2018-01-17 12:24:00  2.57          200
#> 101  2018-01-17 12:25:00  2.58          200
#> 102  2018-01-17 12:26:00  2.64          200
#> 103  2018-01-17 12:27:00  2.64          200
#> 104  2018-01-17 12:28:00  2.68          200
#> 105  2018-01-17 12:29:00  2.67          200
#> 106  2018-01-17 12:30:00  2.72          200
#> 107  2018-01-17 12:31:00  2.73          200
#> 108  2018-01-17 12:32:00  2.75          200
#> 109  2018-01-17 12:33:00  2.76          200
#> 110  2018-01-17 12:34:00  2.80          200
#> 111  2018-01-17 12:35:00  2.78          200
#> 112  2018-01-17 12:36:00  2.82          200
#> 113  2018-01-17 12:37:00  2.85          200
#> 114  2018-01-17 12:38:00  2.85          200
#> 115  2018-01-17 12:39:00  2.86          200
#> 116  2018-01-17 12:40:00  2.89          200
#> 117  2018-01-17 12:41:00  2.90          200
#> 118  2018-01-17 12:42:00  2.91          200
#> 119  2018-01-17 12:43:00  2.93          200
#> 120  2018-01-17 12:44:00  2.96          200
#> 121  2018-01-17 12:45:00  2.96          200
#> 122  2018-01-17 12:46:00  2.96          200
#> 123  2018-01-17 12:47:00  2.99          200
#> 124  2018-01-17 12:48:00  3.00          200
#> 125  2018-01-17 12:49:00  3.03          200
#> 126  2018-01-17 12:50:00  3.04          200
#> 127  2018-01-17 12:51:00  3.06          200
#> 128  2018-01-17 12:52:00  3.08          200
#> 129  2018-01-17 12:53:00  3.09          200
#> 130  2018-01-17 12:54:00  3.10          200
#> 131  2018-01-17 12:55:00  3.12          200
#> 132  2018-01-17 12:56:00  3.13          200
#> 133  2018-01-17 12:57:00  3.15          200
#> 134  2018-01-17 12:58:00  3.17          200
#> 135  2018-01-17 12:59:00  3.16          200
#> 136  2018-01-17 13:00:00  3.19          200
#> 137  2018-01-17 13:01:00  3.21          200
#> 138  2018-01-17 13:02:00  3.22          200
#> 139  2018-01-17 13:03:00  3.23          200
#> 140  2018-01-17 13:04:00  3.23          200
#> 141  2018-01-17 13:05:00  3.26          200
#> 142  2018-01-17 13:06:00  3.28          200
#> 143  2018-01-17 13:07:00  3.29          200
#> 144  2018-01-17 13:08:00  3.30          200
#> 145  2018-01-17 13:09:00  3.32          200
#> 146  2018-01-17 13:10:00  3.34          200
#> 147  2018-01-17 13:11:00  3.36          200
#> 148  2018-01-17 13:12:00  3.37          200
#> 149  2018-01-17 13:13:00  3.39          200
#> 150  2018-01-17 13:14:00  3.39          200
#> 151  2018-01-17 13:15:00  3.42          200
#> 152  2018-01-17 13:16:00  3.43          200
#> 153  2018-01-17 13:17:00  3.44          200
#> 154  2018-01-17 13:18:00  3.44          200
#> 155  2018-01-17 13:19:00  3.46          200
#> 156  2018-01-17 13:20:00  3.48          200
#> 157  2018-01-17 13:21:00  3.50          200
#> 158  2018-01-17 13:22:00  3.50          200
#> 159  2018-01-17 13:23:00  3.51          200
#> 160  2018-01-17 13:24:00  3.54          200
#> 161  2018-01-17 13:25:00  3.56          200
#> 162  2018-01-17 13:26:00  3.56          200
#> 163  2018-01-17 13:27:00  3.58          200
#> 164  2018-01-17 13:28:00  3.59          200
#> 165  2018-01-17 13:29:00  3.61          200
#> 166  2018-01-17 13:30:00  3.62          200
#> 167  2018-01-17 13:31:00  3.64          200
#> 168  2018-01-17 13:32:00  3.64          200
#> 169  2018-01-17 13:33:00  3.66          200
#> 170  2018-01-17 13:34:00  3.67          200
#> 171  2018-01-17 13:35:00  3.68          200
#> 172  2018-01-17 13:36:00  3.71          200
#> 173  2018-01-17 13:37:00  3.72          200
#> 174  2018-01-17 13:38:00  3.73          200
#> 175  2018-01-17 13:39:00  3.75          200
#> 176  2018-01-17 13:40:00  3.76          200
#> 177  2018-01-17 13:41:00  3.79          200
#> 178  2018-01-17 13:42:00  3.79          200
#> 179  2018-01-17 13:43:00  3.81          200
#> 180  2018-01-17 13:44:00  3.83          200
#> 181  2018-01-17 13:45:00  3.83          200
#> 182  2018-01-17 13:46:00  3.84          200
#> 183  2018-01-17 13:47:00  3.86          200
#> 184  2018-01-17 13:48:00  3.87          200
#> 185  2018-01-17 13:49:00  3.89          200
#> 186  2018-01-17 13:50:00  3.90          200
#> 187  2018-01-17 13:51:00  3.91          200
#> 188  2018-01-17 13:52:00  3.93          200
#> 189  2018-01-17 13:53:00  3.95          200
#> 190  2018-01-17 13:54:00  3.96          200
#> 191  2018-01-17 13:55:00  3.97          200
#> 192  2018-01-17 13:56:00  3.96          200
#> 193  2018-01-17 13:57:00  3.99          200
#> 194  2018-01-17 13:58:00  4.01          200
#> 195  2018-01-17 13:59:00  4.03          200
#> 196  2018-01-17 14:00:00  4.04          200
#> 197  2018-01-17 14:01:00  4.05          200
#> 198  2018-01-17 14:02:00  4.06          200
#> 199  2018-01-17 14:03:00  4.06          200
#> 200  2018-01-17 14:04:00  4.08          200
#> 201  2018-01-17 14:05:00  4.09          200
#> 202  2018-01-17 14:06:00  4.11          200
#> 203  2018-01-17 14:07:00  4.10          200
#> 204  2018-01-17 14:08:00  4.14          200
#> 205  2018-01-17 14:09:00  4.15          200
#> 206  2018-01-17 14:10:00  4.17          200
#> 207  2018-01-17 14:11:00  4.18          200
#> 208  2018-01-17 14:12:00  4.19          200
#> 209  2018-01-17 14:13:00  4.21          200
#> 210  2018-01-17 14:14:00  4.23          200
#> 211  2018-01-17 14:15:00  4.24          200
#> 212  2018-01-17 14:16:00  4.26          200
#> 213  2018-01-17 14:17:00  4.29          200
#> 214  2018-01-17 14:18:00  4.30          200
#> 215  2018-01-17 14:19:00  4.31          200
#> 216  2018-01-17 14:20:00  4.30          200
#> 217  2018-01-17 14:21:00  4.35          200
#> 218  2018-01-17 14:22:00  4.35          200
#> 219  2018-01-17 14:23:00  4.37          200
#> 220  2018-01-17 14:24:00  4.39          200
#> 221  2018-01-17 14:25:00  4.41          200
#> 222  2018-01-17 14:26:00  4.43          200
#> 223  2018-01-17 14:27:00  4.45          200
#> 224  2018-01-17 14:28:00  4.46          200
#> 225  2018-01-17 14:29:00  4.48          200
#> 226  2018-01-17 14:30:00  4.50          200
#> 227  2018-01-17 14:31:00  4.51          200
#> 228  2018-01-17 14:32:00  4.52          200
#> 229  2018-01-17 14:33:00  4.54          200
#> 230  2018-01-17 14:34:00  4.57          200
#> 231  2018-01-17 14:35:00  4.58          200
#> 232  2018-01-17 14:36:00  4.60          200
#> 233  2018-01-17 14:37:00  4.61          200
#> 234  2018-01-17 14:38:00  4.63          200
#> 235  2018-01-17 14:39:00  4.65          200
#> 236  2018-01-17 14:40:00  4.67          200
#> 237  2018-01-17 14:41:00  4.68          200
#> 238  2018-01-17 14:42:00  4.70          200
#> 239  2018-01-17 14:43:00  4.71          200
#> 240  2018-01-17 14:44:00  4.73          200
#> 241  2018-01-17 14:45:00  4.75          200
#> 242  2018-01-17 14:46:00  4.77          200
#> 243  2018-01-17 14:47:00  4.78          200
#> 244  2018-01-17 14:48:00  4.81          200
#> 245  2018-01-17 14:49:00  4.83          200
#> 246  2018-01-17 14:50:00  4.85          200
#> 247  2018-01-17 14:51:00  4.86          200
#> 248  2018-01-17 14:52:00  4.88          200
#> 249  2018-01-17 14:53:00  4.91          200
#> 250  2018-01-17 14:54:00  4.92          200
#> 251  2018-01-17 14:55:00  4.95          200
#> 252  2018-01-17 14:56:00  4.98          200
#> 253  2018-01-17 14:57:00  5.00          200
#> 254  2018-01-17 14:58:00  5.02          200
#> 255  2018-01-17 14:59:00  5.05          200
#> 256  2018-01-17 15:00:00  5.06          200
#> 257  2018-01-17 15:01:00  5.10          200
#> 258  2018-01-17 15:02:00  5.12          200
#> 259  2018-01-17 15:03:00  5.15          200
#> 260  2018-01-17 15:04:00  5.17          200
#> 261  2018-01-17 15:05:00  5.20          200
#> 262  2018-01-17 15:06:00  5.23          200
#> 263  2018-01-17 15:07:00  5.26          200
#> 264  2018-01-17 15:08:00  5.28          200
#> 265  2018-01-17 15:09:00  5.30          200
#> 266  2018-01-17 15:10:00  5.33          200
#> 267  2018-01-17 15:11:00  5.36          200
#> 268  2018-01-17 15:12:00  5.38          200
#> 269  2018-01-17 15:13:00  5.40          200
#> 270  2018-01-17 15:14:00  5.42          200
#> 271  2018-01-17 15:15:00  5.45          200
#> 272  2018-01-17 15:16:00  5.47          200
#> 273  2018-01-17 15:17:00  5.49          200
#> 274  2018-01-17 15:18:00  5.51          200
#> 275  2018-01-17 15:19:00  5.52          200
#> 276  2018-01-17 15:20:00  5.54          200
#> 277  2018-01-17 15:21:00  5.56          200
#> 278  2018-01-17 15:22:00  5.58          200
#> 279  2018-01-17 15:23:00  5.60          200
#> 280  2018-01-17 15:24:00  5.62          200
#> 281  2018-01-17 15:25:00  5.64          200
#> 282  2018-01-17 15:26:00  5.64          200
#> 283  2018-01-17 15:27:00  5.66          200
#> 284  2018-01-17 15:28:00  5.68          200
#> 285  2018-01-17 15:29:00  5.69          200
#> 286  2018-01-17 15:30:00  5.71          200
#> 287  2018-01-17 15:31:00  5.71          200
#> 288  2018-01-17 15:32:00  5.72          200
#> 289  2018-01-17 15:33:00  5.74          200
#> 290  2018-01-17 15:34:00  5.75          200
#> 291  2018-01-17 15:35:00  5.76          200
#> 292  2018-01-17 15:36:00  5.78          200
#> 293  2018-01-17 15:37:00  5.79          200
#> 294  2018-01-17 15:38:00  5.80          200
#> 295  2018-01-17 15:39:00  5.82          200
#> 296  2018-01-17 15:40:00  5.83          200
#> 297  2018-01-17 15:41:00  5.84          200
#> 298  2018-01-17 15:42:00  5.85          200
#> 299  2018-01-17 15:43:00  5.86          200
#> 300  2018-01-17 15:44:00  5.87          200
#> 301  2018-01-17 15:45:00  5.88          200
#> 302  2018-01-17 15:46:00  5.90          200
#> 303  2018-01-17 15:47:00  5.90          200
#> 304  2018-01-17 15:48:00  5.91          200
#> 305  2018-01-17 15:49:00  5.92          200
#> 306  2018-01-17 15:50:00  5.93          200
#> 307  2018-01-17 15:51:00  5.94          200
#> 308  2018-01-17 15:52:00  5.95          200
#> 309  2018-01-17 15:53:00  5.97          200
#> 310  2018-01-17 15:54:00  5.97          200
#> 311  2018-01-17 15:55:00  5.97          200
#> 312  2018-01-17 15:56:00  5.98          200
#> 313  2018-01-17 15:57:00  5.99          200
#> 314  2018-01-17 15:58:00  6.01          200
#> 315  2018-01-17 15:59:00  6.02          200
#> 316  2018-01-17 16:00:00  6.03          200
#> 317  2018-01-17 16:01:00  6.05          200
#> 318  2018-01-17 16:02:00  6.05          200
#> 319  2018-01-17 16:03:00  6.06          200
#> 320  2018-01-17 16:04:00  6.07          200
#> 321  2018-01-17 16:05:00  6.08          200
#> 322  2018-01-17 16:06:00  6.08          200
#> 323  2018-01-17 16:07:00  6.09          200
#> 324  2018-01-17 16:08:00  6.10          200
#> 325  2018-01-17 16:09:00  6.11          200
#> 326  2018-01-17 16:10:00  6.12          200
#> 327  2018-01-17 16:11:00  6.13          200
#> 328  2018-01-17 16:12:00  6.14          200
#> 329  2018-01-17 16:13:00  6.14          200
#> 330  2018-01-17 16:14:00  6.15          200
#> 331  2018-01-17 16:15:00  6.16          200
#> 332  2018-01-17 16:16:00  6.17          200
#> 333  2018-01-17 16:17:00  6.18          200
#>  [ reached getOption("max.print") -- omitted 1083 rows ]