Download timeseries data from waterinfo.be

Using the ts_id codes and by providing a given date period, download the corresponding time series from the waterinfo.be website

get_timeseries_tsid(ts_id, period = NULL, from = NULL, to = NULL,
  datasource = 1, token = NULL)

Arguments

ts_id
waterinfo.be database ts_id, defining a timeserie variable and frequency it is defined.

period
input string according to format required by waterinfo: De period string is provided as P#Y#M#DT#H#M#S, with P defines `Period`, each # is an integer value and the codes define the number of... Y - years M - months D - days T required if information about sub-day resolution is present H - hours D - days M - minutes S - seconds Instead of D (days), the usage of W - weeks is possible as well Examples of valid period strings: P3D, P1Y, P1DT12H, PT6H, P1Y6M3DT4H20M30S.

from
date of datestring as start of the time series

to
date of datestring as end of the time series

datasource
int [0-3] defines the `meetnet` of which the measurement station is part of. VMM based stations are net '1', MOW-HIC is net '2'

token
token to use with the call (optional, can be retrieved via get_token)

Value

data.frame with the timestamps, values and quality code

Examples

get_timeseries_tsid("35055042", from = "2017-01-01", to = "2017-01-02")

#>              Timestamp Value Quality Code
#> 1  2017-01-01 00:00:00  0.00          130
#> 2  2017-01-01 00:15:00  0.00          130
#> 3  2017-01-01 00:30:00  0.00          130
#> 4  2017-01-01 00:45:00  0.00          130
#> 5  2017-01-01 01:00:00  0.00          130
#> 6  2017-01-01 01:15:00  0.00          130
#> 7  2017-01-01 01:30:00  0.00          130
#> 8  2017-01-01 01:45:00  0.00          130
#> 9  2017-01-01 02:00:00  0.00          130
#> 10 2017-01-01 02:15:00  0.00          130
#> 11 2017-01-01 02:30:00  0.00          130
#> 12 2017-01-01 02:45:00  0.00          130
#> 13 2017-01-01 03:00:00  0.00          130
#> 14 2017-01-01 03:15:00  0.00          130
#> 15 2017-01-01 03:30:00  0.00          130
#> 16 2017-01-01 03:45:00  0.00          130
#> 17 2017-01-01 04:00:00  0.00          130
#> 18 2017-01-01 04:15:00  0.00          130
#> 19 2017-01-01 04:30:00  0.00          130
#> 20 2017-01-01 04:45:00  0.00          130
#> 21 2017-01-01 05:00:00  0.00          130
#> 22 2017-01-01 05:15:00  0.00          130
#> 23 2017-01-01 05:30:00  0.00          130
#> 24 2017-01-01 05:45:00  0.00          130
#> 25 2017-01-01 06:00:00  0.00          130
#> 26 2017-01-01 06:15:00  0.00          130
#> 27 2017-01-01 06:30:00  0.00          130
#> 28 2017-01-01 06:45:00  0.00          130
#> 29 2017-01-01 07:00:00  0.00          130
#> 30 2017-01-01 07:15:00  0.00          130
#> 31 2017-01-01 07:30:00  0.00          130
#> 32 2017-01-01 07:45:00  0.00          130
#> 33 2017-01-01 08:00:00  0.00          130
#> 34 2017-01-01 08:15:00  0.00          130
#> 35 2017-01-01 08:30:00  0.00          130
#> 36 2017-01-01 08:45:00  0.00          130
#> 37 2017-01-01 09:00:00  0.00          130
#> 38 2017-01-01 09:15:00  0.00          130
#> 39 2017-01-01 09:30:00  0.00          130
#> 40 2017-01-01 09:45:00  0.00          130
#> 41 2017-01-01 10:00:00  0.00          130
#> 42 2017-01-01 10:15:00  0.00          130
#> 43 2017-01-01 10:30:00  0.00          130
#> 44 2017-01-01 10:45:00  0.00          130
#> 45 2017-01-01 11:00:00  0.00          130
#> 46 2017-01-01 11:15:00  0.00          130
#> 47 2017-01-01 11:30:00  0.00          130
#> 48 2017-01-01 11:45:00  0.00          130
#> 49 2017-01-01 12:00:00  0.00          130
#> 50 2017-01-01 12:15:00  0.00          130
#> 51 2017-01-01 12:30:00  0.00          130
#> 52 2017-01-01 12:45:00  0.00          130
#> 53 2017-01-01 13:00:00  0.00          130
#> 54 2017-01-01 13:15:00  0.00          130
#> 55 2017-01-01 13:30:00  0.00          130
#> 56 2017-01-01 13:45:00  0.00          130
#> 57 2017-01-01 14:00:00  0.00          130
#> 58 2017-01-01 14:15:00  0.00          130
#> 59 2017-01-01 14:30:00  0.00          130
#> 60 2017-01-01 14:45:00  0.00          130
#> 61 2017-01-01 15:00:00  0.00          130
#> 62 2017-01-01 15:15:00  0.00          130
#> 63 2017-01-01 15:30:00  0.00          130
#> 64 2017-01-01 15:45:00  0.00          130
#> 65 2017-01-01 16:00:00  0.00          130
#> 66 2017-01-01 16:15:00  0.00          130
#> 67 2017-01-01 16:30:00  0.00          130
#> 68 2017-01-01 16:45:00  0.00          130
#> 69 2017-01-01 17:00:00  0.00          130
#> 70 2017-01-01 17:15:00  0.00          130
#> 71 2017-01-01 17:30:00  0.00          130
#> 72 2017-01-01 17:45:00  0.03          130
#> 73 2017-01-01 18:00:00  0.00          130
#> 74 2017-01-01 18:15:00  0.04          130
#> 75 2017-01-01 18:30:00  0.00          130
#> 76 2017-01-01 18:45:00  0.03          130
#> 77 2017-01-01 19:00:00  0.00          130
#> 78 2017-01-01 19:15:00  0.03          130
#> 79 2017-01-01 19:30:00  0.03          130
#> 80 2017-01-01 19:45:00  0.00          130
#> 81 2017-01-01 20:00:00  0.03          130
#> 82 2017-01-01 20:15:00  0.00          130
#> 83 2017-01-01 20:30:00  0.00          130
#> 84 2017-01-01 20:45:00  0.00          130
#> 85 2017-01-01 21:00:00  0.00          130
#> 86 2017-01-01 21:15:00  0.03          130
#> 87 2017-01-01 21:30:00  0.00          130
#> 88 2017-01-01 21:45:00  0.00          130
#> 89 2017-01-01 22:00:00  0.00          130
#> 90 2017-01-01 22:15:00  0.11          130
#> 91 2017-01-01 22:30:00  0.82          130
#> 92 2017-01-01 22:45:00  0.69          130
#> 93 2017-01-01 23:00:00  0.34          130
#> 94 2017-01-01 23:15:00  0.13          130
#> 95 2017-01-01 23:30:00  0.03          130
#> 96 2017-01-01 23:45:00  0.03          130
#> 97 2017-01-02 00:00:00  0.05          130

get_timeseries_tsid("5156042", period = "P3D")

#>               Timestamp Value Quality Code
#> 1   2018-04-30 16:45:00 1.012          110
#> 2   2018-04-30 17:00:00 1.011          110
#> 3   2018-04-30 17:15:00 1.010          110
#> 4   2018-04-30 17:30:00 1.009          110
#> 5   2018-04-30 17:45:00 1.008          110
#> 6   2018-04-30 18:00:00 1.008          110
#> 7   2018-04-30 18:15:00 1.008          110
#> 8   2018-04-30 18:30:00 1.009          110
#> 9   2018-04-30 18:45:00 1.010          110
#> 10  2018-04-30 19:00:00 1.011          110
#> 11  2018-04-30 19:15:00 1.012          110
#> 12  2018-04-30 19:30:00 1.013          110
#> 13  2018-04-30 19:45:00 1.014          110
#> 14  2018-04-30 20:00:00 1.015          110
#> 15  2018-04-30 20:15:00 1.016          110
#> 16  2018-04-30 20:30:00 1.016          110
#> 17  2018-04-30 20:45:00 1.017          110
#> 18  2018-04-30 21:00:00 1.019          110
#> 19  2018-04-30 21:15:00 1.020          110
#> 20  2018-04-30 21:30:00 1.021          110
#> 21  2018-04-30 21:45:00 1.022          110
#> 22  2018-04-30 22:00:00 1.023          110
#> 23  2018-04-30 22:15:00 1.024          110
#> 24  2018-04-30 22:30:00 1.026          110
#> 25  2018-04-30 22:45:00 1.028          110
#> 26  2018-04-30 23:00:00 1.030          110
#> 27  2018-04-30 23:15:00 1.033          110
#> 28  2018-04-30 23:30:00 1.036          110
#> 29  2018-04-30 23:45:00 1.040          110
#> 30  2018-05-01 00:00:00 1.043          110
#> 31  2018-05-01 00:15:00 1.047          110
#> 32  2018-05-01 00:30:00 1.051          110
#> 33  2018-05-01 00:45:00 1.054          110
#> 34  2018-05-01 01:00:00 1.057          110
#> 35  2018-05-01 01:15:00 1.061          110
#> 36  2018-05-01 01:30:00 1.064          110
#> 37  2018-05-01 01:45:00 1.066          110
#> 38  2018-05-01 02:00:00 1.069          110
#> 39  2018-05-01 02:15:00 1.072          110
#> 40  2018-05-01 02:30:00 1.074          110
#> 41  2018-05-01 02:45:00 1.075          110
#> 42  2018-05-01 03:00:00 1.077          110
#> 43  2018-05-01 03:15:00 1.078          110
#> 44  2018-05-01 03:30:00 1.079          110
#> 45  2018-05-01 03:45:00 1.080          110
#> 46  2018-05-01 04:00:00 1.081          110
#> 47  2018-05-01 04:15:00 1.081          110
#> 48  2018-05-01 04:30:00 1.081          110
#> 49  2018-05-01 04:45:00 1.082          110
#> 50  2018-05-01 05:00:00 1.082          110
#> 51  2018-05-01 05:15:00 1.082          110
#> 52  2018-05-01 05:30:00 1.082          110
#> 53  2018-05-01 05:45:00 1.082          110
#> 54  2018-05-01 06:00:00 1.082          110
#> 55  2018-05-01 06:15:00 1.081          110
#> 56  2018-05-01 06:30:00 1.081          110
#> 57  2018-05-01 06:45:00 1.081          110
#> 58  2018-05-01 07:00:00 1.081          110
#> 59  2018-05-01 07:15:00 1.081          110
#> 60  2018-05-01 07:30:00 1.081          110
#> 61  2018-05-01 07:45:00 1.082          110
#> 62  2018-05-01 08:00:00 1.082          110
#> 63  2018-05-01 08:15:00 1.082          110
#> 64  2018-05-01 08:30:00 1.082          110
#> 65  2018-05-01 08:45:00 1.081          110
#> 66  2018-05-01 09:00:00 1.082          110
#> 67  2018-05-01 09:15:00 1.082          110
#> 68  2018-05-01 09:30:00 1.082          110
#> 69  2018-05-01 09:45:00 1.082          110
#> 70  2018-05-01 10:00:00 1.082          110
#> 71  2018-05-01 10:15:00 1.082          110
#> 72  2018-05-01 10:30:00 1.081          110
#> 73  2018-05-01 10:45:00 1.081          110
#> 74  2018-05-01 11:00:00 1.081          110
#> 75  2018-05-01 11:15:00 1.081          110
#> 76  2018-05-01 11:30:00 1.081          110
#> 77  2018-05-01 11:45:00 1.081          110
#> 78  2018-05-01 12:00:00 1.082          110
#> 79  2018-05-01 12:15:00 1.080          110
#> 80  2018-05-01 12:30:00 1.080          110
#> 81  2018-05-01 12:45:00 1.079          110
#> 82  2018-05-01 13:00:00 1.079          110
#> 83  2018-05-01 13:15:00 1.079          110
#> 84  2018-05-01 13:30:00 1.079          110
#> 85  2018-05-01 13:45:00 1.079          110
#> 86  2018-05-01 14:00:00 1.078          110
#> 87  2018-05-01 14:15:00 1.079          110
#> 88  2018-05-01 14:30:00 1.079          110
#> 89  2018-05-01 14:45:00 1.079          110
#> 90  2018-05-01 15:00:00 1.080          110
#> 91  2018-05-01 15:15:00 1.081          110
#> 92  2018-05-01 15:30:00 1.081          110
#> 93  2018-05-01 15:45:00 1.081          110
#> 94  2018-05-01 16:00:00 1.081          110
#> 95  2018-05-01 16:15:00 1.081          110
#> 96  2018-05-01 16:30:00 1.080          110
#> 97  2018-05-01 16:45:00 1.080          110
#> 98  2018-05-01 17:00:00 1.079          110
#> 99  2018-05-01 17:15:00 1.078          110
#> 100 2018-05-01 17:30:00 1.077          110
#> 101 2018-05-01 17:45:00 1.076          110
#> 102 2018-05-01 18:00:00 1.076          110
#> 103 2018-05-01 18:15:00 1.074          110
#> 104 2018-05-01 18:30:00 1.074          110
#> 105 2018-05-01 18:45:00 1.073          110
#> 106 2018-05-01 19:00:00 1.072          110
#> 107 2018-05-01 19:15:00 1.071          110
#> 108 2018-05-01 19:30:00 1.070          110
#> 109 2018-05-01 19:45:00 1.070          110
#> 110 2018-05-01 20:00:00 1.068          110
#> 111 2018-05-01 20:15:00 1.067          110
#> 112 2018-05-01 20:30:00 1.067          110
#> 113 2018-05-01 20:45:00 1.066          110
#> 114 2018-05-01 21:00:00 1.065          110
#> 115 2018-05-01 21:15:00 1.064          110
#> 116 2018-05-01 21:30:00 1.063          110
#> 117 2018-05-01 21:45:00 1.061          110
#> 118 2018-05-01 22:00:00 1.061          110
#> 119 2018-05-01 22:15:00 1.059          110
#> 120 2018-05-01 22:30:00 1.058          110
#> 121 2018-05-01 22:45:00 1.057          110
#> 122 2018-05-01 23:00:00 1.056          110
#> 123 2018-05-01 23:15:00 1.055          110
#> 124 2018-05-01 23:30:00 1.053          110
#> 125 2018-05-01 23:45:00 1.052          110
#> 126 2018-05-02 00:00:00 1.051          110
#> 127 2018-05-02 00:15:00 1.049          110
#> 128 2018-05-02 00:30:00 1.048          110
#> 129 2018-05-02 00:45:00 1.047          110
#> 130 2018-05-02 01:00:00 1.045          110
#> 131 2018-05-02 01:15:00 1.043          110
#> 132 2018-05-02 01:30:00 1.042          110
#> 133 2018-05-02 01:45:00 1.041          110
#> 134 2018-05-02 02:00:00 1.039          110
#> 135 2018-05-02 02:15:00 1.037          110
#> 136 2018-05-02 02:30:00 1.036          110
#> 137 2018-05-02 02:45:00 1.034          110
#> 138 2018-05-02 03:00:00 1.032          110
#> 139 2018-05-02 03:15:00 1.030          110
#> 140 2018-05-02 03:30:00 1.028          110
#> 141 2018-05-02 03:45:00 1.026          110
#> 142 2018-05-02 04:00:00 1.024          110
#> 143 2018-05-02 04:15:00 1.023          110
#> 144 2018-05-02 04:30:00 1.021          110
#> 145 2018-05-02 04:45:00 1.019          110
#> 146 2018-05-02 05:00:00 1.016          110
#> 147 2018-05-02 05:15:00 1.014          110
#> 148 2018-05-02 05:30:00 1.011          110
#> 149 2018-05-02 05:45:00 1.007          110
#> 150 2018-05-02 06:00:00 1.004          110
#> 151 2018-05-02 06:15:00 1.001          110
#> 152 2018-05-02 06:30:00 0.997          110
#> 153 2018-05-02 06:45:00 0.993          110
#> 154 2018-05-02 07:00:00 0.990          110
#> 155 2018-05-02 07:15:00 0.987          110
#> 156 2018-05-02 07:30:00 0.984          110
#> 157 2018-05-02 07:45:00 0.981          110
#> 158 2018-05-02 08:00:00 0.978          110
#> 159 2018-05-02 08:15:00 0.976          110
#> 160 2018-05-02 08:30:00 0.973          110
#> 161 2018-05-02 08:45:00 0.971          110
#> 162 2018-05-02 09:00:00 0.969          110
#> 163 2018-05-02 09:15:00 0.965          110
#> 164 2018-05-02 09:30:00 0.963          110
#> 165 2018-05-02 09:45:00 0.960          110
#> 166 2018-05-02 10:00:00 0.958          110
#> 167 2018-05-02 10:15:00 0.955          110
#> 168 2018-05-02 10:30:00 0.952          110
#> 169 2018-05-02 10:45:00 0.949          110
#> 170 2018-05-02 11:00:00 0.947          110
#> 171 2018-05-02 11:15:00 0.943          110
#> 172 2018-05-02 11:30:00 0.941          110
#> 173 2018-05-02 11:45:00 0.937          110
#> 174 2018-05-02 12:00:00 0.935          110
#> 175 2018-05-02 12:15:00 0.932          110
#> 176 2018-05-02 12:30:00 0.929          110
#> 177 2018-05-02 12:45:00 0.926          110
#> 178 2018-05-02 13:00:00 0.923          110
#> 179 2018-05-02 13:15:00 0.920          110
#> 180 2018-05-02 13:30:00 0.918          110
#> 181 2018-05-02 13:45:00 0.915          110
#> 182 2018-05-02 14:00:00 0.913          110
#> 183 2018-05-02 14:15:00 0.910          110
#> 184 2018-05-02 14:30:00 0.907          110
#> 185 2018-05-02 14:45:00 0.904          110
#> 186 2018-05-02 15:00:00 0.902          110
#> 187 2018-05-02 15:15:00 0.899          110
#> 188 2018-05-02 15:30:00 0.896          110
#> 189 2018-05-02 15:45:00 0.894          110
#> 190 2018-05-02 16:00:00 0.890          110
#> 191 2018-05-02 16:15:00 0.888          110
#> 192 2018-05-02 16:30:00 0.885          110
#> 193 2018-05-02 16:45:00 0.883          110
#> 194 2018-05-02 17:00:00 0.881          110
#> 195 2018-05-02 17:15:00 0.879          110
#> 196 2018-05-02 17:30:00 0.876          110
#> 197 2018-05-02 17:45:00 0.873          110
#> 198 2018-05-02 18:00:00 0.871          110
#> 199 2018-05-02 18:15:00 0.869          110
#> 200 2018-05-02 18:30:00 0.866          110
#> 201 2018-05-02 18:45:00 0.864          110
#> 202 2018-05-02 19:00:00 0.861          110
#> 203 2018-05-02 19:15:00 0.858          110
#> 204 2018-05-02 19:30:00 0.855          110
#> 205 2018-05-02 19:45:00 0.853          110
#> 206 2018-05-02 20:00:00 0.850          110
#> 207 2018-05-02 20:15:00 0.847          110
#> 208 2018-05-02 20:30:00 0.844          110
#> 209 2018-05-02 20:45:00 0.841          110
#> 210 2018-05-02 21:00:00 0.838          110
#> 211 2018-05-02 21:15:00 0.834          110
#> 212 2018-05-02 21:30:00 0.832          110
#> 213 2018-05-02 21:45:00 0.829          110
#> 214 2018-05-02 22:00:00 0.826          110
#> 215 2018-05-02 22:15:00 0.824          110
#> 216 2018-05-02 22:30:00 0.821          110
#> 217 2018-05-02 22:45:00 0.818          110
#> 218 2018-05-02 23:00:00 0.815          110
#> 219 2018-05-02 23:15:00 0.813          110
#> 220 2018-05-02 23:30:00 0.810          110
#> 221 2018-05-02 23:45:00 0.807          110
#> 222 2018-05-03 00:00:00 0.804          110
#> 223 2018-05-03 00:15:00 0.802          110
#> 224 2018-05-03 00:30:00 0.799          110
#> 225 2018-05-03 00:45:00 0.796          110
#> 226 2018-05-03 01:00:00 0.794          110
#> 227 2018-05-03 01:15:00 0.792          110
#> 228 2018-05-03 01:30:00 0.789          110
#> 229 2018-05-03 01:45:00 0.786          110
#> 230 2018-05-03 02:00:00 0.784          110
#> 231 2018-05-03 02:15:00 0.781          110
#> 232 2018-05-03 02:30:00 0.779          110
#> 233 2018-05-03 02:45:00 0.777          110
#> 234 2018-05-03 03:00:00 0.774          110
#> 235 2018-05-03 03:15:00 0.771          110
#> 236 2018-05-03 03:30:00 0.769          110
#> 237 2018-05-03 03:45:00 0.767          110
#> 238 2018-05-03 04:00:00 0.765          110
#> 239 2018-05-03 04:15:00 0.762          110
#> 240 2018-05-03 04:30:00 0.760          110
#> 241 2018-05-03 04:45:00 0.758          110
#> 242 2018-05-03 05:00:00 0.755          110
#> 243 2018-05-03 05:15:00 0.753          110
#> 244 2018-05-03 05:30:00 0.751          110
#> 245 2018-05-03 05:45:00 0.749          110
#> 246 2018-05-03 06:00:00 0.746          110
#> 247 2018-05-03 06:15:00 0.744          110
#> 248 2018-05-03 06:30:00 0.742          110
#> 249 2018-05-03 06:45:00 0.740          110
#> 250 2018-05-03 07:00:00 0.737          110
#> 251 2018-05-03 07:15:00 0.736          110
#> 252 2018-05-03 07:30:00 0.734          110
#> 253 2018-05-03 07:45:00 0.731          110
#> 254 2018-05-03 08:00:00 0.729          110
#> 255 2018-05-03 08:15:00 0.726          110
#> 256 2018-05-03 08:30:00 0.724          110
#> 257 2018-05-03 08:45:00 0.721          110
#> 258 2018-05-03 09:00:00 0.719          110
#> 259 2018-05-03 09:15:00 0.716          110
#> 260 2018-05-03 09:30:00 0.712          110
#> 261 2018-05-03 09:45:00 0.709          110
#> 262 2018-05-03 10:00:00 0.706          110
#> 263 2018-05-03 10:15:00 0.704          110
#> 264 2018-05-03 10:30:00 0.701          110
#> 265 2018-05-03 10:45:00 0.698          110
#> 266 2018-05-03 11:00:00 0.696          110
#> 267 2018-05-03 11:15:00 0.693          110
#> 268 2018-05-03 11:30:00 0.692          110
#> 269 2018-05-03 11:45:00 0.690          110
#> 270 2018-05-03 12:00:00 0.688          110
#> 271 2018-05-03 12:15:00 0.686          110
#> 272 2018-05-03 12:30:00 0.684          110
#> 273 2018-05-03 12:45:00 0.682          110
#> 274 2018-05-03 13:00:00 0.680          110

get_timeseries_tsid("2813562", period = "P1D", datasource = 2)

#>                Timestamp Value Quality Code
#> 1    2018-05-02 16:35:00  4.19          200
#> 2    2018-05-02 16:36:00  4.23          200
#> 3    2018-05-02 16:37:00  4.25          200
#> 4    2018-05-02 16:38:00  4.28          200
#> 5    2018-05-02 16:39:00  4.31          200
#> 6    2018-05-02 16:40:00  4.34          200
#> 7    2018-05-02 16:41:00  4.36          200
#> 8    2018-05-02 16:42:00  4.39          200
#> 9    2018-05-02 16:43:00  4.42          200
#> 10   2018-05-02 16:44:00  4.44          200
#> 11   2018-05-02 16:45:00  4.47          200
#> 12   2018-05-02 16:46:00  4.50          200
#> 13   2018-05-02 16:47:00  4.53          200
#> 14   2018-05-02 16:48:00  4.56          200
#> 15   2018-05-02 16:49:00  4.58          200
#> 16   2018-05-02 16:50:00  4.61          200
#> 17   2018-05-02 16:51:00  4.64          200
#> 18   2018-05-02 16:52:00  4.66          200
#> 19   2018-05-02 16:53:00  4.68          200
#> 20   2018-05-02 16:54:00  4.71          200
#> 21   2018-05-02 16:55:00  4.73          200
#> 22   2018-05-02 16:56:00  4.76          200
#> 23   2018-05-02 16:57:00  4.78          200
#> 24   2018-05-02 16:58:00  4.79          200
#> 25   2018-05-02 16:59:00  4.82          200
#> 26   2018-05-02 17:00:00  4.83          200
#> 27   2018-05-02 17:01:00  4.86          200
#> 28   2018-05-02 17:02:00  4.88          200
#> 29   2018-05-02 17:03:00  4.90          200
#> 30   2018-05-02 17:04:00  4.91          200
#> 31   2018-05-02 17:05:00  4.94          200
#> 32   2018-05-02 17:06:00  4.95          200
#> 33   2018-05-02 17:07:00  4.98          200
#> 34   2018-05-02 17:08:00  5.00          200
#> 35   2018-05-02 17:09:00  5.01          200
#> 36   2018-05-02 17:10:00  5.03          200
#> 37   2018-05-02 17:11:00  5.04          200
#> 38   2018-05-02 17:12:00  5.06          200
#> 39   2018-05-02 17:13:00  5.07          200
#> 40   2018-05-02 17:14:00  5.09          200
#> 41   2018-05-02 17:15:00  5.10          200
#> 42   2018-05-02 17:16:00  5.12          200
#> 43   2018-05-02 17:17:00  5.14          200
#> 44   2018-05-02 17:18:00  5.15          200
#> 45   2018-05-02 17:19:00  5.17          200
#> 46   2018-05-02 17:20:00  5.18          200
#> 47   2018-05-02 17:21:00  5.19          200
#> 48   2018-05-02 17:22:00  5.20          200
#> 49   2018-05-02 17:23:00  5.22          200
#> 50   2018-05-02 17:24:00  5.23          200
#> 51   2018-05-02 17:25:00  5.24          200
#> 52   2018-05-02 17:26:00  5.26          200
#> 53   2018-05-02 17:27:00  5.27          200
#> 54   2018-05-02 17:28:00  5.28          200
#> 55   2018-05-02 17:29:00  5.29          200
#> 56   2018-05-02 17:30:00  5.30          200
#> 57   2018-05-02 17:31:00  5.31          200
#> 58   2018-05-02 17:32:00  5.32          200
#> 59   2018-05-02 17:33:00  5.33          200
#> 60   2018-05-02 17:34:00  5.34          200
#> 61   2018-05-02 17:35:00  5.35          200
#> 62   2018-05-02 17:36:00  5.36          200
#> 63   2018-05-02 17:37:00  5.37          200
#> 64   2018-05-02 17:38:00  5.37          200
#> 65   2018-05-02 17:39:00  5.38          200
#> 66   2018-05-02 17:40:00  5.39          200
#> 67   2018-05-02 17:41:00  5.40          200
#> 68   2018-05-02 17:42:00  5.40          200
#> 69   2018-05-02 17:43:00  5.41          200
#> 70   2018-05-02 17:44:00  5.41          200
#> 71   2018-05-02 17:45:00  5.42          200
#> 72   2018-05-02 17:46:00  5.42          200
#> 73   2018-05-02 17:47:00  5.42          200
#> 74   2018-05-02 17:48:00  5.43          200
#> 75   2018-05-02 17:49:00  5.43          200
#> 76   2018-05-02 17:50:00  5.44          200
#> 77   2018-05-02 17:51:00  5.44          200
#> 78   2018-05-02 17:52:00  5.44          200
#> 79   2018-05-02 17:53:00  5.44          200
#> 80   2018-05-02 17:54:00  5.45          200
#> 81   2018-05-02 17:55:00  5.45          200
#> 82   2018-05-02 17:56:00  5.45          200
#> 83   2018-05-02 17:57:00  5.45          200
#> 84   2018-05-02 17:58:00  5.45          200
#> 85   2018-05-02 17:59:00  5.46          200
#> 86   2018-05-02 18:00:00  5.46          200
#> 87   2018-05-02 18:01:00  5.45          200
#> 88   2018-05-02 18:02:00  5.45          200
#> 89   2018-05-02 18:03:00  5.46          200
#> 90   2018-05-02 18:04:00  5.45          200
#> 91   2018-05-02 18:05:00  5.45          200
#> 92   2018-05-02 18:06:00  5.45          200
#> 93   2018-05-02 18:07:00  5.45          200
#> 94   2018-05-02 18:08:00  5.45          200
#> 95   2018-05-02 18:09:00  5.44          200
#> 96   2018-05-02 18:10:00  5.44          200
#> 97   2018-05-02 18:11:00  5.43          200
#> 98   2018-05-02 18:12:00  5.43          200
#> 99   2018-05-02 18:13:00  5.42          200
#> 100  2018-05-02 18:14:00  5.42          200
#> 101  2018-05-02 18:15:00  5.41          200
#> 102  2018-05-02 18:16:00  5.41          200
#> 103  2018-05-02 18:17:00  5.40          200
#> 104  2018-05-02 18:18:00  5.39          200
#> 105  2018-05-02 18:19:00  5.38          200
#> 106  2018-05-02 18:20:00  5.37          200
#> 107  2018-05-02 18:21:00  5.36          200
#> 108  2018-05-02 18:22:00  5.35          200
#> 109  2018-05-02 18:23:00  5.34          200
#> 110  2018-05-02 18:24:00  5.33          200
#> 111  2018-05-02 18:25:00  5.32          200
#> 112  2018-05-02 18:26:00  5.31          200
#> 113  2018-05-02 18:27:00  5.30          200
#> 114  2018-05-02 18:28:00  5.28          200
#> 115  2018-05-02 18:29:00  5.27          200
#> 116  2018-05-02 18:30:00  5.26          200
#> 117  2018-05-02 18:31:00  5.24          200
#> 118  2018-05-02 18:32:00  5.23          200
#> 119  2018-05-02 18:33:00  5.21          200
#> 120  2018-05-02 18:34:00  5.20          200
#> 121  2018-05-02 18:35:00  5.18          200
#> 122  2018-05-02 18:36:00  5.16          200
#> 123  2018-05-02 18:37:00  5.14          200
#> 124  2018-05-02 18:38:00  5.12          200
#> 125  2018-05-02 18:39:00  5.10          200
#> 126  2018-05-02 18:40:00  5.08          200
#> 127  2018-05-02 18:41:00  5.06          200
#> 128  2018-05-02 18:42:00  5.04          200
#> 129  2018-05-02 18:43:00  5.02          200
#> 130  2018-05-02 18:44:00  5.00          200
#> 131  2018-05-02 18:45:00  4.98          200
#> 132  2018-05-02 18:46:00  4.96          200
#> 133  2018-05-02 18:47:00  4.94          200
#> 134  2018-05-02 18:48:00  4.92          200
#> 135  2018-05-02 18:49:00  4.90          200
#> 136  2018-05-02 18:50:00  4.87          200
#> 137  2018-05-02 18:51:00  4.86          200
#> 138  2018-05-02 18:52:00  4.84          200
#> 139  2018-05-02 18:53:00  4.82          200
#> 140  2018-05-02 18:54:00  4.80          200
#> 141  2018-05-02 18:55:00  4.78          200
#> 142  2018-05-02 18:56:00  4.76          200
#> 143  2018-05-02 18:57:00  4.74          200
#> 144  2018-05-02 18:58:00  4.72          200
#> 145  2018-05-02 18:59:00  4.70          200
#> 146  2018-05-02 19:00:00  4.68          200
#> 147  2018-05-02 19:01:00  4.66          200
#> 148  2018-05-02 19:02:00  4.64          200
#> 149  2018-05-02 19:03:00  4.62          200
#> 150  2018-05-02 19:04:00  4.60          200
#> 151  2018-05-02 19:05:00  4.58          200
#> 152  2018-05-02 19:06:00  4.56          200
#> 153  2018-05-02 19:07:00  4.54          200
#> 154  2018-05-02 19:08:00  4.52          200
#> 155  2018-05-02 19:09:00  4.50          200
#> 156  2018-05-02 19:10:00  4.48          200
#> 157  2018-05-02 19:11:00  4.46          200
#> 158  2018-05-02 19:12:00  4.44          200
#> 159  2018-05-02 19:13:00  4.43          200
#> 160  2018-05-02 19:14:00  4.41          200
#> 161  2018-05-02 19:15:00  4.39          200
#> 162  2018-05-02 19:16:00  4.37          200
#> 163  2018-05-02 19:17:00  4.36          200
#> 164  2018-05-02 19:18:00  4.34          200
#> 165  2018-05-02 19:19:00  4.32          200
#> 166  2018-05-02 19:20:00  4.30          200
#> 167  2018-05-02 19:21:00  4.29          200
#> 168  2018-05-02 19:22:00  4.27          200
#> 169  2018-05-02 19:23:00  4.26          200
#> 170  2018-05-02 19:24:00  4.24          200
#> 171  2018-05-02 19:25:00  4.22          200
#> 172  2018-05-02 19:26:00  4.21          200
#> 173  2018-05-02 19:27:00  4.20          200
#> 174  2018-05-02 19:28:00  4.18          200
#> 175  2018-05-02 19:29:00  4.16          200
#> 176  2018-05-02 19:30:00  4.15          200
#> 177  2018-05-02 19:31:00  4.14          200
#> 178  2018-05-02 19:32:00  4.12          200
#> 179  2018-05-02 19:33:00  4.11          200
#> 180  2018-05-02 19:34:00  4.09          200
#> 181  2018-05-02 19:35:00  4.08          200
#> 182  2018-05-02 19:36:00  4.04          200
#> 183  2018-05-02 19:37:00  4.05          200
#> 184  2018-05-02 19:38:00  4.04          200
#> 185  2018-05-02 19:39:00  4.02          200
#> 186  2018-05-02 19:40:00  4.01          200
#> 187  2018-05-02 19:41:00  4.00          200
#> 188  2018-05-02 19:42:00  3.98          200
#> 189  2018-05-02 19:43:00  3.97          200
#> 190  2018-05-02 19:44:00  3.95          200
#> 191  2018-05-02 19:45:00  3.94          200
#> 192  2018-05-02 19:46:00  3.93          200
#> 193  2018-05-02 19:47:00  3.92          200
#> 194  2018-05-02 19:48:00  3.90          200
#> 195  2018-05-02 19:49:00  3.90          200
#> 196  2018-05-02 19:50:00  3.88          200
#> 197  2018-05-02 19:51:00  3.86          200
#> 198  2018-05-02 19:52:00  3.86          200
#> 199  2018-05-02 19:53:00  3.85          200
#> 200  2018-05-02 19:54:00  3.83          200
#> 201  2018-05-02 19:55:00  3.82          200
#> 202  2018-05-02 19:56:00  3.81          200
#> 203  2018-05-02 19:57:00  3.80          200
#> 204  2018-05-02 19:58:00  3.79          200
#> 205  2018-05-02 19:59:00  3.78          200
#> 206  2018-05-02 20:00:00  3.76          200
#> 207  2018-05-02 20:01:00  3.75          200
#> 208  2018-05-02 20:02:00  3.74          200
#> 209  2018-05-02 20:03:00  3.73          200
#> 210  2018-05-02 20:04:00  3.72          200
#> 211  2018-05-02 20:05:00  3.71          200
#> 212  2018-05-02 20:06:00  3.69          200
#> 213  2018-05-02 20:07:00  3.68          200
#> 214  2018-05-02 20:08:00  3.67          200
#> 215  2018-05-02 20:09:00  3.66          200
#> 216  2018-05-02 20:10:00  3.64          200
#> 217  2018-05-02 20:11:00  3.63          200
#> 218  2018-05-02 20:12:00  3.63          200
#> 219  2018-05-02 20:13:00  3.62          200
#> 220  2018-05-02 20:14:00  3.60          200
#> 221  2018-05-02 20:15:00  3.59          200
#> 222  2018-05-02 20:16:00  3.58          200
#> 223  2018-05-02 20:17:00  3.57          200
#> 224  2018-05-02 20:18:00  3.56          200
#> 225  2018-05-02 20:19:00  3.55          200
#> 226  2018-05-02 20:20:00  3.53          200
#> 227  2018-05-02 20:21:00  3.52          200
#> 228  2018-05-02 20:22:00  3.51          200
#> 229  2018-05-02 20:23:00  3.50          200
#> 230  2018-05-02 20:24:00  3.49          200
#> 231  2018-05-02 20:25:00  3.47          200
#> 232  2018-05-02 20:26:00  3.46          200
#> 233  2018-05-02 20:27:00  3.45          200
#> 234  2018-05-02 20:28:00  3.44          200
#> 235  2018-05-02 20:29:00  3.43          200
#> 236  2018-05-02 20:30:00  3.42          200
#> 237  2018-05-02 20:31:00  3.41          200
#> 238  2018-05-02 20:32:00  3.40          200
#> 239  2018-05-02 20:33:00  3.39          200
#> 240  2018-05-02 20:34:00  3.38          200
#> 241  2018-05-02 20:35:00  3.37          200
#> 242  2018-05-02 20:36:00  3.35          200
#> 243  2018-05-02 20:37:00  3.34          200
#> 244  2018-05-02 20:38:00  3.33          200
#> 245  2018-05-02 20:39:00  3.32          200
#> 246  2018-05-02 20:40:00  3.31          200
#> 247  2018-05-02 20:41:00  3.30          200
#> 248  2018-05-02 20:42:00  3.29          200
#> 249  2018-05-02 20:43:00  3.28          200
#> 250  2018-05-02 20:44:00  3.27          200
#> 251  2018-05-02 20:45:00  3.26          200
#> 252  2018-05-02 20:46:00  3.24          200
#> 253  2018-05-02 20:47:00  3.23          200
#> 254  2018-05-02 20:48:00  3.22          200
#> 255  2018-05-02 20:49:00  3.22          200
#> 256  2018-05-02 20:50:00  3.21          200
#> 257  2018-05-02 20:51:00  3.20          200
#> 258  2018-05-02 20:52:00  3.18          200
#> 259  2018-05-02 20:53:00  3.17          200
#> 260  2018-05-02 20:54:00  3.16          200
#> 261  2018-05-02 20:55:00  3.15          200
#> 262  2018-05-02 20:56:00  3.14          200
#> 263  2018-05-02 20:57:00  3.13          200
#> 264  2018-05-02 20:58:00  3.12          200
#> 265  2018-05-02 20:59:00  3.12          200
#> 266  2018-05-02 21:00:00  3.10          200
#> 267  2018-05-02 21:01:00  3.09          200
#> 268  2018-05-02 21:02:00  3.08          200
#> 269  2018-05-02 21:03:00  3.07          200
#> 270  2018-05-02 21:04:00  3.06          200
#> 271  2018-05-02 21:05:00  3.05          200
#> 272  2018-05-02 21:06:00  3.04          200
#> 273  2018-05-02 21:07:00  3.03          200
#> 274  2018-05-02 21:08:00  3.02          200
#> 275  2018-05-02 21:09:00  3.01          200
#> 276  2018-05-02 21:10:00  3.00          200
#> 277  2018-05-02 21:11:00  3.00          200
#> 278  2018-05-02 21:12:00  2.98          200
#> 279  2018-05-02 21:13:00  2.97          200
#> 280  2018-05-02 21:14:00  2.97          200
#> 281  2018-05-02 21:15:00  2.96          200
#> 282  2018-05-02 21:16:00  2.95          200
#> 283  2018-05-02 21:17:00  2.94          200
#> 284  2018-05-02 21:18:00  2.92          200
#> 285  2018-05-02 21:19:00  2.92          200
#> 286  2018-05-02 21:20:00  2.90          200
#> 287  2018-05-02 21:21:00  2.90          200
#> 288  2018-05-02 21:22:00  2.89          200
#> 289  2018-05-02 21:23:00  2.88          200
#> 290  2018-05-02 21:24:00  2.87          200
#> 291  2018-05-02 21:25:00  2.86          200
#> 292  2018-05-02 21:26:00  2.85          200
#> 293  2018-05-02 21:27:00  2.84          200
#> 294  2018-05-02 21:28:00  2.83          200
#> 295  2018-05-02 21:29:00  2.82          200
#> 296  2018-05-02 21:30:00  2.82          200
#> 297  2018-05-02 21:31:00  2.80          200
#> 298  2018-05-02 21:32:00  2.80          200
#> 299  2018-05-02 21:33:00  2.79          200
#> 300  2018-05-02 21:34:00  2.78          200
#> 301  2018-05-02 21:35:00  2.77          200
#> 302  2018-05-02 21:36:00  2.76          200
#> 303  2018-05-02 21:37:00  2.75          200
#> 304  2018-05-02 21:38:00  2.74          200
#> 305  2018-05-02 21:39:00  2.73          200
#> 306  2018-05-02 21:40:00  2.72          200
#> 307  2018-05-02 21:41:00  2.71          200
#> 308  2018-05-02 21:42:00  2.70          200
#> 309  2018-05-02 21:43:00  2.69          200
#> 310  2018-05-02 21:44:00  2.68          200
#> 311  2018-05-02 21:45:00  2.68          200
#> 312  2018-05-02 21:46:00  2.66          200
#> 313  2018-05-02 21:47:00  2.66          200
#> 314  2018-05-02 21:48:00  2.65          200
#> 315  2018-05-02 21:49:00  2.64          200
#> 316  2018-05-02 21:50:00  2.63          200
#> 317  2018-05-02 21:51:00  2.62          200
#> 318  2018-05-02 21:52:00  2.61          200
#> 319  2018-05-02 21:53:00  2.60          200
#> 320  2018-05-02 21:54:00  2.59          200
#> 321  2018-05-02 21:55:00  2.58          200
#> 322  2018-05-02 21:56:00  2.57          200
#> 323  2018-05-02 21:57:00  2.56          200
#> 324  2018-05-02 21:58:00  2.55          200
#> 325  2018-05-02 21:59:00  2.54          200
#> 326  2018-05-02 22:00:00  2.53          200
#> 327  2018-05-02 22:01:00  2.52          200
#> 328  2018-05-02 22:02:00  2.51          200
#> 329  2018-05-02 22:03:00  2.50          200
#> 330  2018-05-02 22:04:00  2.49          200
#> 331  2018-05-02 22:05:00  2.48          200
#> 332  2018-05-02 22:06:00  2.47          200
#> 333  2018-05-02 22:07:00  2.46          200
#>  [ reached getOption("max.print") -- omitted 1068 rows ]