Download timeseries data from waterinfo.be

Using the ts_id codes and by providing a given date period, download the corresponding time series from the waterinfo.be website

get_timeseries_tsid(ts_id, period = NULL, from = NULL, to = NULL,
  datasource = 1, token = NULL)

Arguments

ts_id
waterinfo.be database ts_id, defining a timeserie variable and frequency it is defined.

period
input string according to format required by waterinfo: De period string is provided as P#Y#M#DT#H#M#S, with P defines `Period`, each # is an integer value and the codes define the number of... Y - years M - months D - days T required if information about sub-day resolution is present H - hours D - days M - minutes S - seconds Instead of D (days), the usage of W - weeks is possible as well Examples of valid period strings: P3D, P1Y, P1DT12H, PT6H, P1Y6M3DT4H20M30S.

from
date of datestring as start of the time series

to
date of datestring as end of the time series

datasource
int [0-3] defines the `meetnet` of which the measurement station is part of. VMM based stations are net '1', MOW-HIC is net '2'

token
token to use with the call (optional, can be retrieved via get_token)

Format

A data.frame with 3 variables:

Timestamp

Datetime of the measurement.

Value

Measured value.

Quality Code

Quality code of the measurement, dependent on the data source used:

VMM Quality Code Interpretation (datasource 1)

10/110 - Excellent

30/100/130 - Good

50/150 - Moderate

70/170 - Poor

80/180 - Estimated

90/190 - Suspect

220 - Default

-1 - Missing

HIC Quality Code Interpretation (datasource 2)

40 - Good

80 - Estimated

120 - Suspect

200 - Unchecked

60 - Complete

160 - Incomplete

-1 - Missing

Aggregated timeseries

40 - Good

100 - Estimated

120 - Suspect

200 - Unchecked

-1 - Missing

Value

data.frame with the timestamps, values and quality code

Examples

get_timeseries_tsid("35055042", from = "2017-01-01", to = "2017-01-02")

#>              Timestamp Value Quality Code
#> 1  2017-01-01 00:00:00  0.00          130
#> 2  2017-01-01 00:15:00  0.00          130
#> 3  2017-01-01 00:30:00  0.00          130
#> 4  2017-01-01 00:45:00  0.00          130
#> 5  2017-01-01 01:00:00  0.00          130
#> 6  2017-01-01 01:15:00  0.00          130
#> 7  2017-01-01 01:30:00  0.00          130
#> 8  2017-01-01 01:45:00  0.00          130
#> 9  2017-01-01 02:00:00  0.00          130
#> 10 2017-01-01 02:15:00  0.00          130
#> 11 2017-01-01 02:30:00  0.00          130
#> 12 2017-01-01 02:45:00  0.00          130
#> 13 2017-01-01 03:00:00  0.00          130
#> 14 2017-01-01 03:15:00  0.00          130
#> 15 2017-01-01 03:30:00  0.00          130
#> 16 2017-01-01 03:45:00  0.00          130
#> 17 2017-01-01 04:00:00  0.00          130
#> 18 2017-01-01 04:15:00  0.00          130
#> 19 2017-01-01 04:30:00  0.00          130
#> 20 2017-01-01 04:45:00  0.00          130
#> 21 2017-01-01 05:00:00  0.00          130
#> 22 2017-01-01 05:15:00  0.00          130
#> 23 2017-01-01 05:30:00  0.00          130
#> 24 2017-01-01 05:45:00  0.00          130
#> 25 2017-01-01 06:00:00  0.00          130
#> 26 2017-01-01 06:15:00  0.00          130
#> 27 2017-01-01 06:30:00  0.00          130
#> 28 2017-01-01 06:45:00  0.00          130
#> 29 2017-01-01 07:00:00  0.00          130
#> 30 2017-01-01 07:15:00  0.00          130
#> 31 2017-01-01 07:30:00  0.00          130
#> 32 2017-01-01 07:45:00  0.00          130
#> 33 2017-01-01 08:00:00  0.00          130
#> 34 2017-01-01 08:15:00  0.00          130
#> 35 2017-01-01 08:30:00  0.00          130
#> 36 2017-01-01 08:45:00  0.00          130
#> 37 2017-01-01 09:00:00  0.00          130
#> 38 2017-01-01 09:15:00  0.00          130
#> 39 2017-01-01 09:30:00  0.00          130
#> 40 2017-01-01 09:45:00  0.00          130
#> 41 2017-01-01 10:00:00  0.00          130
#> 42 2017-01-01 10:15:00  0.00          130
#> 43 2017-01-01 10:30:00  0.00          130
#> 44 2017-01-01 10:45:00  0.00          130
#> 45 2017-01-01 11:00:00  0.00          130
#> 46 2017-01-01 11:15:00  0.00          130
#> 47 2017-01-01 11:30:00  0.00          130
#> 48 2017-01-01 11:45:00  0.00          130
#> 49 2017-01-01 12:00:00  0.00          130
#> 50 2017-01-01 12:15:00  0.00          130
#> 51 2017-01-01 12:30:00  0.00          130
#> 52 2017-01-01 12:45:00  0.00          130
#> 53 2017-01-01 13:00:00  0.00          130
#> 54 2017-01-01 13:15:00  0.00          130
#> 55 2017-01-01 13:30:00  0.00          130
#> 56 2017-01-01 13:45:00  0.00          130
#> 57 2017-01-01 14:00:00  0.00          130
#> 58 2017-01-01 14:15:00  0.00          130
#> 59 2017-01-01 14:30:00  0.00          130
#> 60 2017-01-01 14:45:00  0.00          130
#> 61 2017-01-01 15:00:00  0.00          130
#> 62 2017-01-01 15:15:00  0.00          130
#> 63 2017-01-01 15:30:00  0.00          130
#> 64 2017-01-01 15:45:00  0.00          130
#> 65 2017-01-01 16:00:00  0.00          130
#> 66 2017-01-01 16:15:00  0.00          130
#> 67 2017-01-01 16:30:00  0.00          130
#> 68 2017-01-01 16:45:00  0.00          130
#> 69 2017-01-01 17:00:00  0.00          130
#> 70 2017-01-01 17:15:00  0.00          130
#> 71 2017-01-01 17:30:00  0.00          130
#> 72 2017-01-01 17:45:00  0.03          130
#> 73 2017-01-01 18:00:00  0.00          130
#> 74 2017-01-01 18:15:00  0.04          130
#> 75 2017-01-01 18:30:00  0.00          130
#> 76 2017-01-01 18:45:00  0.03          130
#> 77 2017-01-01 19:00:00  0.00          130
#> 78 2017-01-01 19:15:00  0.03          130
#> 79 2017-01-01 19:30:00  0.03          130
#> 80 2017-01-01 19:45:00  0.00          130
#> 81 2017-01-01 20:00:00  0.03          130
#> 82 2017-01-01 20:15:00  0.00          130
#> 83 2017-01-01 20:30:00  0.00          130
#> 84 2017-01-01 20:45:00  0.00          130
#> 85 2017-01-01 21:00:00  0.00          130
#> 86 2017-01-01 21:15:00  0.03          130
#> 87 2017-01-01 21:30:00  0.00          130
#> 88 2017-01-01 21:45:00  0.00          130
#> 89 2017-01-01 22:00:00  0.00          130
#> 90 2017-01-01 22:15:00  0.11          130
#> 91 2017-01-01 22:30:00  0.82          130
#> 92 2017-01-01 22:45:00  0.69          130
#> 93 2017-01-01 23:00:00  0.34          130
#> 94 2017-01-01 23:15:00  0.13          130
#> 95 2017-01-01 23:30:00  0.03          130
#> 96 2017-01-01 23:45:00  0.03          130
#> 97 2017-01-02 00:00:00  0.05          130

get_timeseries_tsid("5156042", period = "P3D")

#>               Timestamp Value Quality Code
#> 1   2018-11-04 14:30:00 0.257          110
#> 2   2018-11-04 14:45:00 0.257          110
#> 3   2018-11-04 15:00:00 0.256          110
#> 4   2018-11-04 15:15:00 0.256          110
#> 5   2018-11-04 15:30:00 0.256          110
#> 6   2018-11-04 15:45:00 0.255          110
#> 7   2018-11-04 16:00:00 0.255          110
#> 8   2018-11-04 16:15:00 0.255          110
#> 9   2018-11-04 16:30:00 0.254          110
#> 10  2018-11-04 16:45:00 0.254          110
#> 11  2018-11-04 17:00:00 0.254          110
#> 12  2018-11-04 17:15:00 0.254          110
#> 13  2018-11-04 17:30:00 0.254          110
#> 14  2018-11-04 17:45:00 0.253          110
#> 15  2018-11-04 18:00:00 0.253          110
#> 16  2018-11-04 18:15:00 0.253          110
#> 17  2018-11-04 18:30:00 0.253          110
#> 18  2018-11-04 18:45:00 0.253          110
#> 19  2018-11-04 19:00:00 0.253          110
#> 20  2018-11-04 19:15:00 0.252          110
#> 21  2018-11-04 19:30:00 0.252          110
#> 22  2018-11-04 19:45:00 0.252          110
#> 23  2018-11-04 20:00:00 0.252          110
#> 24  2018-11-04 20:15:00 0.252          110
#> 25  2018-11-04 20:30:00 0.252          110
#> 26  2018-11-04 20:45:00 0.252          110
#> 27  2018-11-04 21:00:00 0.252          110
#> 28  2018-11-04 21:15:00 0.252          110
#> 29  2018-11-04 21:30:00 0.252          110
#> 30  2018-11-04 21:45:00 0.252          110
#> 31  2018-11-04 22:00:00 0.252          110
#> 32  2018-11-04 22:15:00 0.252          110
#> 33  2018-11-04 22:30:00 0.252          110
#> 34  2018-11-04 22:45:00 0.252          110
#> 35  2018-11-04 23:00:00 0.252          110
#> 36  2018-11-04 23:15:00 0.252          110
#> 37  2018-11-04 23:30:00 0.252          110
#> 38  2018-11-04 23:45:00 0.252          110
#> 39  2018-11-05 00:00:00 0.252          110
#> 40  2018-11-05 00:15:00 0.252          110
#> 41  2018-11-05 00:30:00 0.252          110
#> 42  2018-11-05 00:45:00 0.252          110
#> 43  2018-11-05 01:00:00 0.252          110
#> 44  2018-11-05 01:15:00 0.252          110
#> 45  2018-11-05 01:30:00 0.252          110
#> 46  2018-11-05 01:45:00 0.252          110
#> 47  2018-11-05 02:00:00 0.252          110
#> 48  2018-11-05 02:15:00 0.252          110
#> 49  2018-11-05 02:30:00 0.252          110
#> 50  2018-11-05 02:45:00 0.252          110
#> 51  2018-11-05 03:00:00 0.252          110
#> 52  2018-11-05 03:15:00 0.252          110
#> 53  2018-11-05 03:30:00 0.252          110
#> 54  2018-11-05 03:45:00 0.252          110
#> 55  2018-11-05 04:00:00 0.252          110
#> 56  2018-11-05 04:15:00 0.251          110
#> 57  2018-11-05 04:30:00 0.251          110
#> 58  2018-11-05 04:45:00 0.251          110
#> 59  2018-11-05 05:00:00 0.251          110
#> 60  2018-11-05 05:15:00 0.251          110
#> 61  2018-11-05 05:30:00 0.251          110
#> 62  2018-11-05 05:45:00 0.251          110
#> 63  2018-11-05 06:00:00 0.250          110
#> 64  2018-11-05 06:15:00 0.250          110
#> 65  2018-11-05 06:30:00 0.250          110
#> 66  2018-11-05 06:45:00 0.250          110
#> 67  2018-11-05 07:00:00 0.250          110
#> 68  2018-11-05 07:15:00 0.249          110
#> 69  2018-11-05 07:30:00 0.249          110
#> 70  2018-11-05 07:45:00 0.249          110
#> 71  2018-11-05 08:00:00 0.249          110
#> 72  2018-11-05 08:15:00 0.248          110
#> 73  2018-11-05 08:30:00 0.248          110
#> 74  2018-11-05 08:45:00 0.247          110
#> 75  2018-11-05 09:00:00 0.247          110
#> 76  2018-11-05 09:15:00 0.247          110
#> 77  2018-11-05 09:30:00 0.247          110
#> 78  2018-11-05 09:45:00 0.246          110
#> 79  2018-11-05 10:00:00 0.246          110
#> 80  2018-11-05 10:15:00 0.246          110
#> 81  2018-11-05 10:30:00 0.246          110
#> 82  2018-11-05 10:45:00 0.246          110
#> 83  2018-11-05 11:00:00 0.246          110
#> 84  2018-11-05 11:15:00 0.246          110
#> 85  2018-11-05 11:30:00 0.246          110
#> 86  2018-11-05 11:45:00 0.246          110
#> 87  2018-11-05 12:00:00 0.245          110
#> 88  2018-11-05 12:15:00 0.245          110
#> 89  2018-11-05 12:30:00 0.244          110
#> 90  2018-11-05 12:45:00 0.243          110
#> 91  2018-11-05 13:00:00 0.242          110
#> 92  2018-11-05 13:15:00 0.242          110
#> 93  2018-11-05 13:30:00 0.241          110
#> 94  2018-11-05 13:45:00 0.240          110
#> 95  2018-11-05 14:00:00 0.239          110
#> 96  2018-11-05 14:15:00 0.239          110
#> 97  2018-11-05 14:30:00 0.238          110
#> 98  2018-11-05 14:45:00 0.238          110
#> 99  2018-11-05 15:00:00 0.237          110
#> 100 2018-11-05 15:15:00 0.237          110
#> 101 2018-11-05 15:30:00 0.236          110
#> 102 2018-11-05 15:45:00 0.236          110
#> 103 2018-11-05 16:00:00 0.236          110
#> 104 2018-11-05 16:15:00 0.235          110
#> 105 2018-11-05 16:30:00 0.236          110
#> 106 2018-11-05 16:45:00 0.236          110
#> 107 2018-11-05 17:00:00 0.236          110
#> 108 2018-11-05 17:15:00 0.236          110
#> 109 2018-11-05 17:30:00 0.236          110
#> 110 2018-11-05 17:45:00 0.235          110
#> 111 2018-11-05 18:00:00 0.236          110
#> 112 2018-11-05 18:15:00 0.236          110
#> 113 2018-11-05 18:30:00 0.236          110
#> 114 2018-11-05 18:45:00 0.236          110
#> 115 2018-11-05 19:00:00 0.236          110
#> 116 2018-11-05 19:15:00 0.236          110
#> 117 2018-11-05 19:30:00 0.236          110
#> 118 2018-11-05 19:45:00 0.236          110
#> 119 2018-11-05 20:00:00 0.236          110
#> 120 2018-11-05 20:15:00 0.236          110
#> 121 2018-11-05 20:30:00 0.236          110
#> 122 2018-11-05 20:45:00 0.237          110
#> 123 2018-11-05 21:00:00 0.237          110
#> 124 2018-11-05 21:15:00 0.237          110
#> 125 2018-11-05 21:30:00 0.237          110
#> 126 2018-11-05 21:45:00 0.237          110
#> 127 2018-11-05 22:00:00 0.237          110
#> 128 2018-11-05 22:15:00 0.237          110
#> 129 2018-11-05 22:30:00 0.237          110
#> 130 2018-11-05 22:45:00 0.237          110
#> 131 2018-11-05 23:00:00 0.237          110
#> 132 2018-11-05 23:15:00 0.237          110
#> 133 2018-11-05 23:30:00 0.237          110
#> 134 2018-11-05 23:45:00 0.237          110
#> 135 2018-11-06 00:00:00 0.237          110
#> 136 2018-11-06 00:15:00 0.237          110
#> 137 2018-11-06 00:30:00 0.236          110
#> 138 2018-11-06 00:45:00 0.236          110
#> 139 2018-11-06 01:00:00 0.236          110
#> 140 2018-11-06 01:15:00 0.236          110
#> 141 2018-11-06 01:30:00 0.236          110
#> 142 2018-11-06 01:45:00 0.235          110
#> 143 2018-11-06 02:00:00 0.235          110
#> 144 2018-11-06 02:15:00 0.235          110
#> 145 2018-11-06 02:30:00 0.234          110
#> 146 2018-11-06 02:45:00 0.234          110
#> 147 2018-11-06 03:00:00 0.234          110
#> 148 2018-11-06 03:15:00 0.234          110
#> 149 2018-11-06 03:30:00 0.234          110
#> 150 2018-11-06 03:45:00 0.233          110
#> 151 2018-11-06 04:00:00 0.233          110
#> 152 2018-11-06 04:15:00 0.233          110
#> 153 2018-11-06 04:30:00 0.233          110
#> 154 2018-11-06 04:45:00 0.233          110
#> 155 2018-11-06 05:00:00 0.233          110
#> 156 2018-11-06 05:15:00 0.233          110
#> 157 2018-11-06 05:30:00 0.232          110
#> 158 2018-11-06 05:45:00 0.232          110
#> 159 2018-11-06 06:00:00 0.232          110
#> 160 2018-11-06 06:15:00 0.232          110
#> 161 2018-11-06 06:30:00 0.231          110
#> 162 2018-11-06 06:45:00 0.232          110
#> 163 2018-11-06 07:00:00 0.231          110
#> 164 2018-11-06 07:15:00 0.231          110
#> 165 2018-11-06 07:30:00 0.231          110
#> 166 2018-11-06 07:45:00 0.231          110
#> 167 2018-11-06 08:00:00 0.230          110
#> 168 2018-11-06 08:15:00 0.230          110
#> 169 2018-11-06 08:30:00 0.229          110
#> 170 2018-11-06 08:45:00 0.229          110
#> 171 2018-11-06 09:00:00 0.229          110
#> 172 2018-11-06 09:15:00 0.228          110
#> 173 2018-11-06 09:30:00 0.228          110
#> 174 2018-11-06 09:45:00 0.227          110
#> 175 2018-11-06 10:00:00 0.227          110
#> 176 2018-11-06 10:15:00 0.227          110
#> 177 2018-11-06 10:30:00 0.226          110
#> 178 2018-11-06 10:45:00 0.226          110
#> 179 2018-11-06 11:00:00 0.225          110
#> 180 2018-11-06 11:15:00 0.224          110
#> 181 2018-11-06 11:30:00 0.223          110
#> 182 2018-11-06 11:45:00 0.223          110
#> 183 2018-11-06 12:00:00 0.223          110
#> 184 2018-11-06 12:15:00 0.223          110
#> 185 2018-11-06 12:30:00 0.222          110
#> 186 2018-11-06 12:45:00 0.222          110
#> 187 2018-11-06 13:00:00 0.222          110
#> 188 2018-11-06 13:15:00 0.221          110
#> 189 2018-11-06 13:30:00 0.221          110
#> 190 2018-11-06 13:45:00 0.221          110
#> 191 2018-11-06 14:00:00 0.220          110
#> 192 2018-11-06 14:15:00 0.220          110
#> 193 2018-11-06 14:30:00 0.219          110
#> 194 2018-11-06 14:45:00 0.219          110
#> 195 2018-11-06 15:00:00 0.218          110
#> 196 2018-11-06 15:15:00 0.218          110
#> 197 2018-11-06 15:30:00 0.218          110
#> 198 2018-11-06 15:45:00 0.218          110
#> 199 2018-11-06 16:00:00 0.218          110
#> 200 2018-11-06 16:15:00 0.218          110
#> 201 2018-11-06 16:30:00 0.218          110
#> 202 2018-11-06 16:45:00 0.218          110
#> 203 2018-11-06 17:00:00 0.218          110
#> 204 2018-11-06 17:15:00 0.219          110
#> 205 2018-11-06 17:30:00 0.219          110
#> 206 2018-11-06 17:45:00 0.219          110
#> 207 2018-11-06 18:00:00 0.219          110
#> 208 2018-11-06 18:15:00 0.219          110
#> 209 2018-11-06 18:30:00 0.220          110
#> 210 2018-11-06 18:45:00 0.219          110
#> 211 2018-11-06 19:00:00 0.220          110
#> 212 2018-11-06 19:15:00 0.219          110
#> 213 2018-11-06 19:30:00 0.219          110
#> 214 2018-11-06 19:45:00 0.219          110
#> 215 2018-11-06 20:00:00 0.219          110
#> 216 2018-11-06 20:15:00 0.218          110
#> 217 2018-11-06 20:30:00 0.218          110
#> 218 2018-11-06 20:45:00 0.218          110
#> 219 2018-11-06 21:00:00 0.217          110
#> 220 2018-11-06 21:15:00 0.217          110
#> 221 2018-11-06 21:30:00 0.216          110
#> 222 2018-11-06 21:45:00 0.216          110
#> 223 2018-11-06 22:00:00 0.215          110
#> 224 2018-11-06 22:15:00 0.215          110
#> 225 2018-11-06 22:30:00 0.215          110
#> 226 2018-11-06 22:45:00 0.214          110
#> 227 2018-11-06 23:00:00 0.213          110
#> 228 2018-11-06 23:15:00 0.213          110
#> 229 2018-11-06 23:30:00 0.212          110
#> 230 2018-11-06 23:45:00 0.211          110
#> 231 2018-11-07 00:00:00 0.211          110
#> 232 2018-11-07 00:15:00 0.210          110
#> 233 2018-11-07 00:30:00 0.210          110
#> 234 2018-11-07 00:45:00 0.210          110
#> 235 2018-11-07 01:00:00 0.210          110
#> 236 2018-11-07 01:15:00 0.210          110
#> 237 2018-11-07 01:30:00 0.210          110
#> 238 2018-11-07 01:45:00 0.209          110
#> 239 2018-11-07 02:00:00 0.209          110
#> 240 2018-11-07 02:15:00 0.209          110
#> 241 2018-11-07 02:30:00 0.209          110
#> 242 2018-11-07 02:45:00 0.210          110
#> 243 2018-11-07 03:00:00 0.210          110
#> 244 2018-11-07 03:15:00 0.210          110
#> 245 2018-11-07 03:30:00 0.210          110
#> 246 2018-11-07 03:45:00 0.210          110
#> 247 2018-11-07 04:00:00 0.210          110
#> 248 2018-11-07 04:15:00 0.211          110
#> 249 2018-11-07 04:30:00 0.211          110
#> 250 2018-11-07 04:45:00 0.211          110
#> 251 2018-11-07 05:00:00 0.211          110
#> 252 2018-11-07 05:15:00 0.212          110
#> 253 2018-11-07 05:30:00 0.212          110
#> 254 2018-11-07 05:45:00 0.212          110
#> 255 2018-11-07 06:00:00 0.212          110
#> 256 2018-11-07 06:15:00 0.213          110
#> 257 2018-11-07 06:30:00 0.213          110
#> 258 2018-11-07 06:45:00 0.213          110
#> 259 2018-11-07 07:00:00 0.213          110
#> 260 2018-11-07 07:15:00 0.213          110
#> 261 2018-11-07 07:30:00 0.213          110
#> 262 2018-11-07 07:45:00 0.213          110
#> 263 2018-11-07 08:00:00 0.213          110
#> 264 2018-11-07 08:15:00 0.213          110
#> 265 2018-11-07 08:30:00 0.213          110
#> 266 2018-11-07 08:45:00 0.213          110
#> 267 2018-11-07 09:00:00 0.213          110
#> 268 2018-11-07 09:15:00 0.213          110
#> 269 2018-11-07 09:30:00 0.213          110
#> 270 2018-11-07 09:45:00 0.213          110
#> 271 2018-11-07 10:00:00 0.213          110
#> 272 2018-11-07 10:15:00 0.213          110
#> 273 2018-11-07 10:30:00 0.213          110
#> 274 2018-11-07 10:45:00 0.213          110
#> 275 2018-11-07 11:00:00 0.213          110
#> 276 2018-11-07 11:15:00 0.213          110
#> 277 2018-11-07 11:30:00 0.213          110
#> 278 2018-11-07 11:45:00 0.213          110
#> 279 2018-11-07 12:00:00 0.213          110
#> 280 2018-11-07 12:15:00 0.213          110
#> 281 2018-11-07 12:30:00 0.213          110
#> 282 2018-11-07 12:45:00 0.213          110
#> 283 2018-11-07 13:00:00 0.213          110
#> 284 2018-11-07 13:15:00 0.213          110

get_timeseries_tsid("55419010", from = "2017-06-01", to = "2017-06-03",
                    datasource = 4)

#>               Timestamp Value Quality Code
#> 1   2017-06-01 00:00:00  2.27           11
#> 2   2017-06-01 00:10:00  2.15           11
#> 3   2017-06-01 00:20:00  2.03           11
#> 4   2017-06-01 00:30:00  1.91           11
#> 5   2017-06-01 00:40:00  1.80           11
#> 6   2017-06-01 00:50:00  1.69           11
#> 7   2017-06-01 01:00:00  1.58           11
#> 8   2017-06-01 01:10:00  1.46           11
#> 9   2017-06-01 01:20:00  1.35           11
#> 10  2017-06-01 01:30:00  1.25           11
#> 11  2017-06-01 01:40:00  1.14           11
#> 12  2017-06-01 01:50:00  1.04           11
#> 13  2017-06-01 02:00:00  0.95           11
#> 14  2017-06-01 02:10:00  0.85           11
#> 15  2017-06-01 02:20:00  0.76           11
#> 16  2017-06-01 02:30:00  0.67           11
#> 17  2017-06-01 02:40:00  0.58           11
#> 18  2017-06-01 02:50:00  0.50           11
#> 19  2017-06-01 03:00:00  0.42           11
#> 20  2017-06-01 03:10:00  0.34           11
#> 21  2017-06-01 03:20:00  0.26           11
#> 22  2017-06-01 03:30:00  0.19           11
#> 23  2017-06-01 03:40:00  0.12           11
#> 24  2017-06-01 03:50:00  0.07           11
#> 25  2017-06-01 04:00:00  0.04           11
#> 26  2017-06-01 04:10:00  0.13           11
#> 27  2017-06-01 04:20:00  0.43           11
#> 28  2017-06-01 04:30:00  0.75           11
#> 29  2017-06-01 04:40:00  0.99           11
#> 30  2017-06-01 04:50:00  1.19           11
#> 31  2017-06-01 05:00:00  1.36           11
#> 32  2017-06-01 05:10:00  1.52           11
#> 33  2017-06-01 05:20:00  1.67           11
#> 34  2017-06-01 05:30:00  1.84           11
#> 35  2017-06-01 05:40:00  1.99           11
#> 36  2017-06-01 05:50:00  2.16           11
#> 37  2017-06-01 06:00:00  2.31           11
#> 38  2017-06-01 06:10:00  2.46           11
#> 39  2017-06-01 06:20:00  2.62           11
#> 40  2017-06-01 06:30:00  2.79           11
#> 41  2017-06-01 06:40:00  2.96           11
#> 42  2017-06-01 06:50:00  3.17           11
#> 43  2017-06-01 07:00:00  3.39           11
#> 44  2017-06-01 07:10:00  3.65           11
#> 45  2017-06-01 07:20:00  3.94           11
#> 46  2017-06-01 07:30:00  4.22           11
#> 47  2017-06-01 07:40:00  4.50           11
#> 48  2017-06-01 07:50:00  4.73           11
#> 49  2017-06-01 08:00:00  4.94           11
#> 50  2017-06-01 08:10:00  5.11           11
#> 51  2017-06-01 08:20:00  5.23           11
#> 52  2017-06-01 08:30:00  5.33           11
#> 53  2017-06-01 08:40:00  5.39           11
#> 54  2017-06-01 08:50:00  5.42           11
#> 55  2017-06-01 09:00:00  5.42           11
#> 56  2017-06-01 09:10:00  5.38           11
#> 57  2017-06-01 09:20:00  5.29           11
#> 58  2017-06-01 09:30:00  5.16           11
#> 59  2017-06-01 09:40:00  4.98           11
#> 60  2017-06-01 09:50:00  4.78           11
#> 61  2017-06-01 10:00:00  4.58           11
#> 62  2017-06-01 10:10:00  4.40           11
#> 63  2017-06-01 10:20:00  4.21           11
#> 64  2017-06-01 10:30:00  4.04           11
#> 65  2017-06-01 10:40:00  3.87           11
#> 66  2017-06-01 10:50:00  3.70           11
#> 67  2017-06-01 11:00:00  3.53           11
#> 68  2017-06-01 11:10:00  3.37           11
#> 69  2017-06-01 11:20:00  3.21           11
#> 70  2017-06-01 11:30:00  3.06           11
#> 71  2017-06-01 11:40:00  2.91           11
#> 72  2017-06-01 11:50:00  2.77           11
#> 73  2017-06-01 12:00:00  2.64           11
#> 74  2017-06-01 12:10:00  2.50           11
#> 75  2017-06-01 12:20:00  2.37           11
#> 76  2017-06-01 12:30:00  2.24           11
#> 77  2017-06-01 12:40:00  2.12           11
#> 78  2017-06-01 12:50:00  2.00           11
#> 79  2017-06-01 13:00:00  1.88           11
#> 80  2017-06-01 13:10:00  1.77           11
#> 81  2017-06-01 13:20:00  1.66           11
#> 82  2017-06-01 13:30:00  1.55           11
#> 83  2017-06-01 13:40:00  1.44           11
#> 84  2017-06-01 13:50:00  1.34           11
#> 85  2017-06-01 14:00:00  1.24           11
#> 86  2017-06-01 14:10:00  1.14           11
#> 87  2017-06-01 14:20:00  1.04           11
#> 88  2017-06-01 14:30:00  0.94           11
#> 89  2017-06-01 14:40:00  0.85           11
#> 90  2017-06-01 14:50:00  0.77           11
#> 91  2017-06-01 15:00:00  0.68           11
#> 92  2017-06-01 15:10:00  0.60           11
#> 93  2017-06-01 15:20:00  0.53           11
#> 94  2017-06-01 15:30:00  0.45           11
#> 95  2017-06-01 15:40:00  0.39           11
#> 96  2017-06-01 15:50:00  0.33           11
#> 97  2017-06-01 16:00:00  0.29           11
#> 98  2017-06-01 16:10:00  0.27           11
#> 99  2017-06-01 16:20:00  0.39           11
#> 100 2017-06-01 16:30:00  0.71           11
#> 101 2017-06-01 16:40:00  0.98           11
#> 102 2017-06-01 16:50:00  1.21           11
#> 103 2017-06-01 17:00:00  1.40           11
#> 104 2017-06-01 17:10:00  1.55           11
#> 105 2017-06-01 17:20:00  1.70           11
#> 106 2017-06-01 17:30:00  1.84           11
#> 107 2017-06-01 17:40:00  1.99           11
#> 108 2017-06-01 17:50:00  2.14           11
#> 109 2017-06-01 18:00:00  2.28           11
#> 110 2017-06-01 18:10:00  2.42           11
#> 111 2017-06-01 18:20:00  2.55           11
#> 112 2017-06-01 18:30:00  2.68           11
#> 113 2017-06-01 18:40:00  2.81           11
#> 114 2017-06-01 18:50:00  2.93           11
#> 115 2017-06-01 19:00:00  3.07           11
#> 116 2017-06-01 19:10:00  3.22           11
#> 117 2017-06-01 19:20:00  3.36           11
#> 118 2017-06-01 19:30:00  3.52           11
#> 119 2017-06-01 19:40:00  3.69           11
#> 120 2017-06-01 19:50:00  3.90           11
#> 121 2017-06-01 20:00:00  4.11           11
#> 122 2017-06-01 20:10:00  4.34           11
#> 123 2017-06-01 20:20:00  4.55           11
#> 124 2017-06-01 20:30:00  4.75           11
#> 125 2017-06-01 20:40:00  4.91           11
#> 126 2017-06-01 20:50:00  5.04           11
#> 127 2017-06-01 21:00:00  5.14           11
#> 128 2017-06-01 21:10:00  5.20           11
#> 129 2017-06-01 21:20:00  5.23           11
#> 130 2017-06-01 21:30:00  5.21           11
#> 131 2017-06-01 21:40:00  5.15           11
#> 132 2017-06-01 21:50:00  5.02           11
#> 133 2017-06-01 22:00:00  4.85           11
#> 134 2017-06-01 22:10:00  4.65           11
#> 135 2017-06-01 22:20:00  4.43           11
#> 136 2017-06-01 22:30:00  4.24           11
#> 137 2017-06-01 22:40:00  4.06           11
#> 138 2017-06-01 22:50:00  3.89           11
#> 139 2017-06-01 23:00:00  3.73           15
#> 140 2017-06-01 23:10:00  3.57           11
#> 141 2017-06-01 23:20:00  3.42           11
#> 142 2017-06-01 23:30:00  3.26           11
#> 143 2017-06-01 23:40:00  3.11           11
#> 144 2017-06-01 23:50:00  2.97           11
#> 145 2017-06-02 00:00:00  2.83           11
#> 146 2017-06-02 00:10:00  2.69           11
#> 147 2017-06-02 00:20:00  2.56           11
#> 148 2017-06-02 00:30:00  2.43           11
#> 149 2017-06-02 00:40:00  2.31           11
#> 150 2017-06-02 00:50:00  2.18           11
#> 151 2017-06-02 01:00:00  2.06           11
#> 152 2017-06-02 01:10:00  1.95           11
#> 153 2017-06-02 01:20:00  1.84           11
#> 154 2017-06-02 01:30:00  1.72           11
#> 155 2017-06-02 01:40:00  1.61           11
#> 156 2017-06-02 01:50:00  1.51           11
#> 157 2017-06-02 02:00:00  1.40           11
#> 158 2017-06-02 02:10:00  1.30           11
#> 159 2017-06-02 02:20:00  1.20           11
#> 160 2017-06-02 02:30:00  1.10           11
#> 161 2017-06-02 02:40:00  1.00           11
#> 162 2017-06-02 02:50:00  0.91           11
#> 163 2017-06-02 03:00:00  0.82           11
#> 164 2017-06-02 03:10:00  0.73           11
#> 165 2017-06-02 03:20:00  0.65           11
#> 166 2017-06-02 03:30:00  0.56           11
#> 167 2017-06-02 03:40:00  0.48           11
#> 168 2017-06-02 03:50:00  0.41           11
#> 169 2017-06-02 04:00:00  0.34           11
#> 170 2017-06-02 04:10:00  0.27           11
#> 171 2017-06-02 04:20:00  0.20           11
#> 172 2017-06-02 04:30:00  0.15           11
#> 173 2017-06-02 04:40:00  0.10           11
#> 174 2017-06-02 04:50:00  0.07           11
#> 175 2017-06-02 05:00:00  0.16           11
#> 176 2017-06-02 05:10:00  0.43           11
#> 177 2017-06-02 05:20:00  0.71           11
#> 178 2017-06-02 05:30:00  0.95           11
#> 179 2017-06-02 05:40:00  1.15           11
#> 180 2017-06-02 05:50:00  1.32           11
#> 181 2017-06-02 06:00:00  1.48           11
#> 182 2017-06-02 06:10:00  1.64           11
#> 183 2017-06-02 06:20:00  1.81           11
#> 184 2017-06-02 06:30:00  1.98           11
#> 185 2017-06-02 06:40:00  2.14           11
#> 186 2017-06-02 06:50:00  2.30           11
#> 187 2017-06-02 07:00:00  2.47           11
#> 188 2017-06-02 07:10:00  2.63           11
#> 189 2017-06-02 07:20:00  2.80           11
#> 190 2017-06-02 07:30:00  2.99           11
#> 191 2017-06-02 07:40:00  3.18           11
#> 192 2017-06-02 07:50:00  3.39           11
#> 193 2017-06-02 08:00:00  3.62           11
#> 194 2017-06-02 08:10:00  3.86           11
#> 195 2017-06-02 08:20:00  4.11           11
#> 196 2017-06-02 08:30:00  4.36           11
#> 197 2017-06-02 08:40:00  4.60           11
#> 198 2017-06-02 08:50:00  4.81           11
#> 199 2017-06-02 09:00:00  4.99           11
#> 200 2017-06-02 09:10:00  5.14           11
#> 201 2017-06-02 09:20:00  5.25           11
#> 202 2017-06-02 09:30:00  5.34           11
#> 203 2017-06-02 09:40:00  5.40           11
#> 204 2017-06-02 09:50:00  5.43           11
#> 205 2017-06-02 10:00:00  5.44           11
#> 206 2017-06-02 10:10:00  5.41           11
#> 207 2017-06-02 10:20:00  5.33           11
#> 208 2017-06-02 10:30:00  5.21           11
#> 209 2017-06-02 10:40:00  5.05           11
#> 210 2017-06-02 10:50:00  4.87           11
#> 211 2017-06-02 11:00:00  4.68           11
#> 212 2017-06-02 11:10:00  4.50           11
#> 213 2017-06-02 11:20:00  4.33           11
#> 214 2017-06-02 11:30:00  4.14           11
#> 215 2017-06-02 11:40:00  3.98           11
#> 216 2017-06-02 11:50:00  3.82           11
#> 217 2017-06-02 12:00:00  3.66           11
#> 218 2017-06-02 12:10:00  3.50           11
#> 219 2017-06-02 12:20:00  3.33           11
#> 220 2017-06-02 12:30:00  3.19           11
#> 221 2017-06-02 12:40:00  3.04           11
#> 222 2017-06-02 12:50:00  2.89           11
#> 223 2017-06-02 13:00:00  2.76           11
#> 224 2017-06-02 13:10:00  2.63           11
#> 225 2017-06-02 13:20:00  2.50           11
#> 226 2017-06-02 13:30:00  2.37           11
#> 227 2017-06-02 13:40:00  2.24           11
#> 228 2017-06-02 13:50:00  2.13           11
#> 229 2017-06-02 14:00:00  2.02           11
#> 230 2017-06-02 14:10:00  1.90           11
#> 231 2017-06-02 14:20:00  1.79           11
#> 232 2017-06-02 14:30:00  1.69           11
#> 233 2017-06-02 14:40:00  1.59           11
#> 234 2017-06-02 14:50:00  1.48           11
#> 235 2017-06-02 15:00:00  1.38           11
#> 236 2017-06-02 15:10:00  1.28           11
#> 237 2017-06-02 15:20:00  1.19           11
#> 238 2017-06-02 15:30:00  1.09           11
#> 239 2017-06-02 15:40:00  1.00           11
#> 240 2017-06-02 15:50:00  0.92           11
#> 241 2017-06-02 16:00:00  0.83           11
#> 242 2017-06-02 16:10:00  0.76           11
#> 243 2017-06-02 16:20:00  0.68           11
#> 244 2017-06-02 16:30:00  0.62           11
#> 245 2017-06-02 16:40:00  0.56           11
#> 246 2017-06-02 16:50:00  0.52           11
#> 247 2017-06-02 17:00:00  0.51           11
#> 248 2017-06-02 17:10:00  0.58           11
#> 249 2017-06-02 17:20:00  0.80           11
#> 250 2017-06-02 17:30:00  1.09           11
#> 251 2017-06-02 17:40:00  1.31           11
#> 252 2017-06-02 17:50:00  1.51           11
#> 253 2017-06-02 18:00:00  1.67           11
#> 254 2017-06-02 18:10:00  1.82           11
#> 255 2017-06-02 18:20:00  1.95           11
#> 256 2017-06-02 18:30:00  2.11           11
#> 257 2017-06-02 18:40:00  2.25           11
#> 258 2017-06-02 18:50:00  2.39           11
#> 259 2017-06-02 19:00:00  2.52           11
#> 260 2017-06-02 19:10:00  2.65           11
#> 261 2017-06-02 19:20:00  2.78           11
#> 262 2017-06-02 19:30:00  2.90           11
#> 263 2017-06-02 19:40:00  3.02           11
#> 264 2017-06-02 19:50:00  3.15           11
#> 265 2017-06-02 20:00:00  3.28           11
#> 266 2017-06-02 20:10:00  3.41           11
#> 267 2017-06-02 20:20:00  3.55           11
#> 268 2017-06-02 20:30:00  3.70           11
#> 269 2017-06-02 20:40:00  3.87           11
#> 270 2017-06-02 20:50:00  4.04           11
#> 271 2017-06-02 21:00:00  4.23           11
#> 272 2017-06-02 21:10:00  4.41           11
#> 273 2017-06-02 21:20:00  4.60           11
#> 274 2017-06-02 21:30:00  4.78           11
#> 275 2017-06-02 21:40:00  4.93           11
#> 276 2017-06-02 21:50:00  5.05           11
#> 277 2017-06-02 22:00:00  5.15           11
#> 278 2017-06-02 22:10:00  5.22           11
#> 279 2017-06-02 22:20:00  5.24           11
#> 280 2017-06-02 22:30:00  5.24           11
#> 281 2017-06-02 22:40:00  5.19           11
#> 282 2017-06-02 22:50:00  5.09           11
#> 283 2017-06-02 23:00:00  4.95           11
#> 284 2017-06-02 23:10:00  4.77           11
#> 285 2017-06-02 23:20:00  4.57           11
#> 286 2017-06-02 23:30:00  4.37           11
#> 287 2017-06-02 23:40:00  4.20           11
#> 288 2017-06-02 23:50:00  4.04           11
#> 289 2017-06-03 00:00:00  3.88           11