Download timeseries data from waterinfo.be

Using the ts_id codes and by providing a given date period, download the corresponding time series from the waterinfo.be website

get_timeseries_tsid(ts_id, period = NULL, from = NULL, to = NULL,
  datasource = 1)

Arguments

ts_id
waterinfo.be database ts_id, defining a timeserie variable and frequency it is defined.

period
input string according to format required by waterinfo: De period string is provided as P#Y#M#DT#H#M#S, with P defines `Period`, each # is an integer value and the codes define the number of... Y - years M - months D - days T required if information about sub-day resolution is present H - hours D - days M - minutes S - seconds Instead of D (days), the usage of W - weeks is possible as well Examples of valid period strings: P3D, P1Y, P1DT12H, PT6H, P1Y6M3DT4H20M30S.

from
date of datestring as start of the time series

to
date of datestring as end of the time series

datasource
int [0-3] defines the `meetnet` of which the measurement station is part of. VMM based stations are net '1', MOW-HIC is net '2'

Value

data.frame with the timestamps, values and quality code

Examples

get_timeseries_tsid("35055042", from="2017-01-01", to="2017-01-02")

#>              Timestamp Value Quality Code
#> 1  2017-01-01 00:00:00  0.00          130
#> 2  2017-01-01 00:15:00  0.00          130
#> 3  2017-01-01 00:30:00  0.00          130
#> 4  2017-01-01 00:45:00  0.00          130
#> 5  2017-01-01 01:00:00  0.00          130
#> 6  2017-01-01 01:15:00  0.00          130
#> 7  2017-01-01 01:30:00  0.00          130
#> 8  2017-01-01 01:45:00  0.00          130
#> 9  2017-01-01 02:00:00  0.00          130
#> 10 2017-01-01 02:15:00  0.00          130
#> 11 2017-01-01 02:30:00  0.00          130
#> 12 2017-01-01 02:45:00  0.00          130
#> 13 2017-01-01 03:00:00  0.00          130
#> 14 2017-01-01 03:15:00  0.00          130
#> 15 2017-01-01 03:30:00  0.00          130
#> 16 2017-01-01 03:45:00  0.00          130
#> 17 2017-01-01 04:00:00  0.00          130
#> 18 2017-01-01 04:15:00  0.00          130
#> 19 2017-01-01 04:30:00  0.00          130
#> 20 2017-01-01 04:45:00  0.00          130
#> 21 2017-01-01 05:00:00  0.00          130
#> 22 2017-01-01 05:15:00  0.00          130
#> 23 2017-01-01 05:30:00  0.00          130
#> 24 2017-01-01 05:45:00  0.00          130
#> 25 2017-01-01 06:00:00  0.00          130
#> 26 2017-01-01 06:15:00  0.00          130
#> 27 2017-01-01 06:30:00  0.00          130
#> 28 2017-01-01 06:45:00  0.00          130
#> 29 2017-01-01 07:00:00  0.00          130
#> 30 2017-01-01 07:15:00  0.00          130
#> 31 2017-01-01 07:30:00  0.00          130
#> 32 2017-01-01 07:45:00  0.00          130
#> 33 2017-01-01 08:00:00  0.00          130
#> 34 2017-01-01 08:15:00  0.00          130
#> 35 2017-01-01 08:30:00  0.00          130
#> 36 2017-01-01 08:45:00  0.00          130
#> 37 2017-01-01 09:00:00  0.00          130
#> 38 2017-01-01 09:15:00  0.00          130
#> 39 2017-01-01 09:30:00  0.00          130
#> 40 2017-01-01 09:45:00  0.00          130
#> 41 2017-01-01 10:00:00  0.00          130
#> 42 2017-01-01 10:15:00  0.00          130
#> 43 2017-01-01 10:30:00  0.00          130
#> 44 2017-01-01 10:45:00  0.00          130
#> 45 2017-01-01 11:00:00  0.00          130
#> 46 2017-01-01 11:15:00  0.00          130
#> 47 2017-01-01 11:30:00  0.00          130
#> 48 2017-01-01 11:45:00  0.00          130
#> 49 2017-01-01 12:00:00  0.00          130
#> 50 2017-01-01 12:15:00  0.00          130
#> 51 2017-01-01 12:30:00  0.00          130
#> 52 2017-01-01 12:45:00  0.00          130
#> 53 2017-01-01 13:00:00  0.00          130
#> 54 2017-01-01 13:15:00  0.00          130
#> 55 2017-01-01 13:30:00  0.00          130
#> 56 2017-01-01 13:45:00  0.00          130
#> 57 2017-01-01 14:00:00  0.00          130
#> 58 2017-01-01 14:15:00  0.00          130
#> 59 2017-01-01 14:30:00  0.00          130
#> 60 2017-01-01 14:45:00  0.00          130
#> 61 2017-01-01 15:00:00  0.00          130
#> 62 2017-01-01 15:15:00  0.00          130
#> 63 2017-01-01 15:30:00  0.00          130
#> 64 2017-01-01 15:45:00  0.00          130
#> 65 2017-01-01 16:00:00  0.00          130
#> 66 2017-01-01 16:15:00  0.00          130
#> 67 2017-01-01 16:30:00  0.00          130
#> 68 2017-01-01 16:45:00  0.00          130
#> 69 2017-01-01 17:00:00  0.00          130
#> 70 2017-01-01 17:15:00  0.00          130
#> 71 2017-01-01 17:30:00  0.00          130
#> 72 2017-01-01 17:45:00  0.03          130
#> 73 2017-01-01 18:00:00  0.00          130
#> 74 2017-01-01 18:15:00  0.04          130
#> 75 2017-01-01 18:30:00  0.00          130
#> 76 2017-01-01 18:45:00  0.03          130
#> 77 2017-01-01 19:00:00  0.00          130
#> 78 2017-01-01 19:15:00  0.03          130
#> 79 2017-01-01 19:30:00  0.03          130
#> 80 2017-01-01 19:45:00  0.00          130
#> 81 2017-01-01 20:00:00  0.03          130
#> 82 2017-01-01 20:15:00  0.00          130
#> 83 2017-01-01 20:30:00  0.00          130
#> 84 2017-01-01 20:45:00  0.00          130
#> 85 2017-01-01 21:00:00  0.00          130
#> 86 2017-01-01 21:15:00  0.03          130
#> 87 2017-01-01 21:30:00  0.00          130
#> 88 2017-01-01 21:45:00  0.00          130
#> 89 2017-01-01 22:00:00  0.00          130
#> 90 2017-01-01 22:15:00  0.11          130
#> 91 2017-01-01 22:30:00  0.82          130
#> 92 2017-01-01 22:45:00  0.69          130
#> 93 2017-01-01 23:00:00  0.34          130
#> 94 2017-01-01 23:15:00  0.13          130
#> 95 2017-01-01 23:30:00  0.03          130
#> 96 2017-01-01 23:45:00  0.03          130
#> 97 2017-01-02 00:00:00  0.05          130

get_timeseries_tsid("5156042", period = "P3D")

#>               Timestamp Value Quality Code
#> 1   2017-10-18 12:00:00 0.233          110
#> 2   2017-10-18 12:15:00 0.233          110
#> 3   2017-10-18 12:30:00 0.233          110
#> 4   2017-10-18 12:45:00 0.233          110
#> 5   2017-10-18 13:00:00 0.234          110
#> 6   2017-10-18 13:15:00 0.234          110
#> 7   2017-10-18 13:30:00 0.234          110
#> 8   2017-10-18 13:45:00 0.235          110
#> 9   2017-10-18 14:00:00 0.235          110
#> 10  2017-10-18 14:15:00 0.235          110
#> 11  2017-10-18 14:30:00 0.235          110
#> 12  2017-10-18 14:45:00 0.235          110
#> 13  2017-10-18 15:00:00 0.235          110
#> 14  2017-10-18 15:15:00 0.234          110
#> 15  2017-10-18 15:30:00 0.234          110
#> 16  2017-10-18 15:45:00 0.233          110
#> 17  2017-10-18 16:00:00 0.233          110
#> 18  2017-10-18 16:15:00 0.233          110
#> 19  2017-10-18 16:30:00 0.232          110
#> 20  2017-10-18 16:45:00 0.231          110
#> 21  2017-10-18 17:00:00 0.230          110
#> 22  2017-10-18 17:15:00 0.231          110
#> 23  2017-10-18 17:30:00 0.230          110
#> 24  2017-10-18 17:45:00 0.230          110
#> 25  2017-10-18 18:00:00 0.230          110
#> 26  2017-10-18 18:15:00 0.230          110
#> 27  2017-10-18 18:30:00 0.230          110
#> 28  2017-10-18 18:45:00 0.231          110
#> 29  2017-10-18 19:00:00 0.231          110
#> 30  2017-10-18 19:15:00 0.231          110
#> 31  2017-10-18 19:30:00 0.231          110
#> 32  2017-10-18 19:45:00 0.232          110
#> 33  2017-10-18 20:00:00 0.232          110
#> 34  2017-10-18 20:15:00 0.232          110
#> 35  2017-10-18 20:30:00 0.233          110
#> 36  2017-10-18 20:45:00 0.234          110
#> 37  2017-10-18 21:00:00 0.234          110
#> 38  2017-10-18 21:15:00 0.235          110
#> 39  2017-10-18 21:30:00 0.236          110
#> 40  2017-10-18 21:45:00 0.236          110
#> 41  2017-10-18 22:00:00 0.237          110
#> 42  2017-10-18 22:15:00 0.238          110
#> 43  2017-10-18 22:30:00 0.238          110
#> 44  2017-10-18 22:45:00 0.239          110
#> 45  2017-10-18 23:00:00 0.240          110
#> 46  2017-10-18 23:15:00 0.240          110
#> 47  2017-10-18 23:30:00 0.241          110
#> 48  2017-10-18 23:45:00 0.241          110
#> 49  2017-10-19 00:00:00 0.242          110
#> 50  2017-10-19 00:15:00 0.242          110
#> 51  2017-10-19 00:30:00 0.243          110
#> 52  2017-10-19 00:45:00 0.243          110
#> 53  2017-10-19 01:00:00 0.243          110
#> 54  2017-10-19 01:15:00 0.243          110
#> 55  2017-10-19 01:30:00 0.243          110
#> 56  2017-10-19 01:45:00 0.243          110
#> 57  2017-10-19 02:00:00 0.243          110
#> 58  2017-10-19 02:15:00 0.243          110
#> 59  2017-10-19 02:30:00 0.243          110
#> 60  2017-10-19 02:45:00 0.243          110
#> 61  2017-10-19 03:00:00 0.242          110
#> 62  2017-10-19 03:15:00 0.242          110
#> 63  2017-10-19 03:30:00 0.243          110
#> 64  2017-10-19 03:45:00 0.242          110
#> 65  2017-10-19 04:00:00 0.242          110
#> 66  2017-10-19 04:15:00 0.242          110
#> 67  2017-10-19 04:30:00 0.242          110
#> 68  2017-10-19 04:45:00 0.242          110
#> 69  2017-10-19 05:00:00 0.242          110
#> 70  2017-10-19 05:15:00 0.242          110
#> 71  2017-10-19 05:30:00 0.242          110
#> 72  2017-10-19 05:45:00 0.242          110
#> 73  2017-10-19 06:00:00 0.242          110
#> 74  2017-10-19 06:15:00 0.242          110
#> 75  2017-10-19 06:30:00 0.241          110
#> 76  2017-10-19 06:45:00 0.241          110
#> 77  2017-10-19 07:00:00 0.241          110
#> 78  2017-10-19 07:15:00 0.241          110
#> 79  2017-10-19 07:30:00 0.240          110
#> 80  2017-10-19 07:45:00 0.240          110
#> 81  2017-10-19 08:00:00 0.240          110
#> 82  2017-10-19 08:15:00 0.239          110
#> 83  2017-10-19 08:30:00 0.240          110
#> 84  2017-10-19 08:45:00 0.240          110
#> 85  2017-10-19 09:00:00 0.239          110
#> 86  2017-10-19 09:15:00 0.239          110
#> 87  2017-10-19 09:30:00 0.239          110
#> 88  2017-10-19 09:45:00 0.239          110
#> 89  2017-10-19 10:00:00 0.238          110
#> 90  2017-10-19 10:15:00 0.238          110
#> 91  2017-10-19 10:30:00 0.238          110
#> 92  2017-10-19 10:45:00 0.237          110
#> 93  2017-10-19 11:00:00 0.237          110
#> 94  2017-10-19 11:15:00 0.237          110
#> 95  2017-10-19 11:30:00 0.237          110
#> 96  2017-10-19 11:45:00 0.236          110
#> 97  2017-10-19 12:00:00 0.237          110
#> 98  2017-10-19 12:15:00 0.237          110
#> 99  2017-10-19 12:30:00 0.237          110
#> 100 2017-10-19 12:45:00 0.237          110
#> 101 2017-10-19 13:00:00 0.237          110
#> 102 2017-10-19 13:15:00 0.237          110
#> 103 2017-10-19 13:30:00 0.237          110
#> 104 2017-10-19 13:45:00 0.237          110
#> 105 2017-10-19 14:00:00 0.237          110
#> 106 2017-10-19 14:15:00 0.237          110
#> 107 2017-10-19 14:30:00 0.237          110
#> 108 2017-10-19 14:45:00 0.237          110
#> 109 2017-10-19 15:00:00 0.237          110
#> 110 2017-10-19 15:15:00 0.237          110
#> 111 2017-10-19 15:30:00 0.237          110
#> 112 2017-10-19 15:45:00 0.237          110
#> 113 2017-10-19 16:00:00 0.237          110
#> 114 2017-10-19 16:15:00 0.237          110
#> 115 2017-10-19 16:30:00 0.236          110
#> 116 2017-10-19 16:45:00 0.236          110
#> 117 2017-10-19 17:00:00 0.235          110
#> 118 2017-10-19 17:15:00 0.235          110
#> 119 2017-10-19 17:30:00 0.235          110
#> 120 2017-10-19 17:45:00 0.235          110
#> 121 2017-10-19 18:00:00 0.235          110
#> 122 2017-10-19 18:15:00 0.235          110
#> 123 2017-10-19 18:30:00 0.234          110
#> 124 2017-10-19 18:45:00 0.234          110
#> 125 2017-10-19 19:00:00 0.234          110
#> 126 2017-10-19 19:15:00 0.234          110
#> 127 2017-10-19 19:30:00 0.234          110
#> 128 2017-10-19 19:45:00 0.234          110
#> 129 2017-10-19 20:00:00 0.234          110
#> 130 2017-10-19 20:15:00 0.234          110
#> 131 2017-10-19 20:30:00 0.234          110
#> 132 2017-10-19 20:45:00 0.234          110
#> 133 2017-10-19 21:00:00 0.234          110
#> 134 2017-10-19 21:15:00 0.234          110
#> 135 2017-10-19 21:30:00 0.234          110
#> 136 2017-10-19 21:45:00 0.234          110
#> 137 2017-10-19 22:00:00 0.234          110
#> 138 2017-10-19 22:15:00 0.234          110
#> 139 2017-10-19 22:30:00 0.234          110
#> 140 2017-10-19 22:45:00 0.234          110
#> 141 2017-10-19 23:00:00 0.234          110
#> 142 2017-10-19 23:15:00 0.234          110
#> 143 2017-10-19 23:30:00 0.234          110
#> 144 2017-10-19 23:45:00 0.234          110
#> 145 2017-10-20 00:00:00 0.234          110
#> 146 2017-10-20 00:15:00 0.234          110
#> 147 2017-10-20 00:30:00 0.234          110
#> 148 2017-10-20 00:45:00 0.234          110
#> 149 2017-10-20 01:00:00 0.234          110
#> 150 2017-10-20 01:15:00 0.234          110
#> 151 2017-10-20 01:30:00 0.234          110
#> 152 2017-10-20 01:45:00 0.234          110
#> 153 2017-10-20 02:00:00 0.234          110
#> 154 2017-10-20 02:15:00 0.234          110
#> 155 2017-10-20 02:30:00 0.234          110
#> 156 2017-10-20 02:45:00 0.234          110
#> 157 2017-10-20 03:00:00 0.235          110
#> 158 2017-10-20 03:15:00 0.235          110
#> 159 2017-10-20 03:30:00 0.235          110
#> 160 2017-10-20 03:45:00 0.235          110
#> 161 2017-10-20 04:00:00 0.235          110
#> 162 2017-10-20 04:15:00 0.235          110
#> 163 2017-10-20 04:30:00 0.235          110
#> 164 2017-10-20 04:45:00 0.235          110
#> 165 2017-10-20 05:00:00 0.235          110
#> 166 2017-10-20 05:15:00 0.236          110
#> 167 2017-10-20 05:30:00 0.236          110
#> 168 2017-10-20 05:45:00 0.236          110
#> 169 2017-10-20 06:00:00 0.236          110
#> 170 2017-10-20 06:15:00 0.237          110
#> 171 2017-10-20 06:30:00 0.236          110
#> 172 2017-10-20 06:45:00 0.236          110
#> 173 2017-10-20 07:00:00 0.236          110
#> 174 2017-10-20 07:15:00 0.236          110
#> 175 2017-10-20 07:30:00 0.236          110
#> 176 2017-10-20 07:45:00 0.235          110
#> 177 2017-10-20 08:00:00 0.235          110
#> 178 2017-10-20 08:15:00 0.235          110
#> 179 2017-10-20 08:30:00 0.235          110
#> 180 2017-10-20 08:45:00 0.236          110
#> 181 2017-10-20 09:00:00 0.236          110
#> 182 2017-10-20 09:15:00 0.235          110
#> 183 2017-10-20 09:30:00 0.235          110
#> 184 2017-10-20 09:45:00 0.234          110
#> 185 2017-10-20 10:00:00 0.234          110
#> 186 2017-10-20 10:15:00 0.234          110
#> 187 2017-10-20 10:30:00 0.233          110
#> 188 2017-10-20 10:45:00 0.233          110
#> 189 2017-10-20 11:00:00 0.233          110
#> 190 2017-10-20 11:15:00 0.232          110
#> 191 2017-10-20 11:30:00 0.232          110
#> 192 2017-10-20 11:45:00 0.231          110
#> 193 2017-10-20 12:00:00 0.231          110
#> 194 2017-10-20 12:15:00 0.231          110
#> 195 2017-10-20 12:30:00 0.230          110
#> 196 2017-10-20 12:45:00 0.230          110
#> 197 2017-10-20 13:00:00 0.230          110
#> 198 2017-10-20 13:15:00 0.230          110
#> 199 2017-10-20 13:30:00 0.229          110
#> 200 2017-10-20 13:45:00 0.229          110
#> 201 2017-10-20 14:00:00 0.229          110
#> 202 2017-10-20 14:15:00 0.229          110
#> 203 2017-10-20 14:30:00 0.229          110
#> 204 2017-10-20 14:45:00 0.230          110
#> 205 2017-10-20 15:00:00 0.230          110
#> 206 2017-10-20 15:15:00 0.230          110
#> 207 2017-10-20 15:30:00 0.230          110
#> 208 2017-10-20 15:45:00 0.230          110
#> 209 2017-10-20 16:00:00 0.230          110
#> 210 2017-10-20 16:15:00 0.230          110
#> 211 2017-10-20 16:30:00 0.231          110
#> 212 2017-10-20 16:45:00 0.231          110
#> 213 2017-10-20 17:00:00 0.231          110
#> 214 2017-10-20 17:15:00 0.231          110
#> 215 2017-10-20 17:30:00 0.231          110
#> 216 2017-10-20 17:45:00 0.231          110
#> 217 2017-10-20 18:00:00 0.231          110
#> 218 2017-10-20 18:15:00 0.231          110
#> 219 2017-10-20 18:30:00 0.231          110
#> 220 2017-10-20 18:45:00 0.231          110
#> 221 2017-10-20 19:00:00 0.232          110
#> 222 2017-10-20 19:15:00 0.232          110
#> 223 2017-10-20 19:30:00 0.232          110
#> 224 2017-10-20 19:45:00 0.232          110
#> 225 2017-10-20 20:00:00 0.233          110
#> 226 2017-10-20 20:15:00 0.233          110
#> 227 2017-10-20 20:30:00 0.233          110
#> 228 2017-10-20 20:45:00 0.233          110
#> 229 2017-10-20 21:00:00 0.233          110
#> 230 2017-10-20 21:15:00 0.234          110
#> 231 2017-10-20 21:30:00 0.234          110
#> 232 2017-10-20 21:45:00 0.234          110
#> 233 2017-10-20 22:00:00 0.235          110
#> 234 2017-10-20 22:15:00 0.234          110
#> 235 2017-10-20 22:30:00 0.235          110
#> 236 2017-10-20 22:45:00 0.235          110
#> 237 2017-10-20 23:00:00 0.235          110
#> 238 2017-10-20 23:15:00 0.235          110
#> 239 2017-10-20 23:30:00 0.235          110
#> 240 2017-10-20 23:45:00 0.235          110
#> 241 2017-10-21 00:00:00 0.235          110
#> 242 2017-10-21 00:15:00 0.235          110
#> 243 2017-10-21 00:30:00 0.235          110
#> 244 2017-10-21 00:45:00 0.235          110
#> 245 2017-10-21 01:00:00 0.235          110
#> 246 2017-10-21 01:15:00 0.235          110
#> 247 2017-10-21 01:30:00 0.235          110
#> 248 2017-10-21 01:45:00 0.235          110
#> 249 2017-10-21 02:00:00 0.235          110
#> 250 2017-10-21 02:15:00 0.236          110
#> 251 2017-10-21 02:30:00 0.236          110
#> 252 2017-10-21 02:45:00 0.235          110
#> 253 2017-10-21 03:00:00 0.235          110
#> 254 2017-10-21 03:15:00 0.236          110
#> 255 2017-10-21 03:30:00 0.235          110
#> 256 2017-10-21 03:45:00 0.235          110
#> 257 2017-10-21 04:00:00 0.235          110
#> 258 2017-10-21 04:15:00 0.235          110
#> 259 2017-10-21 04:30:00 0.235          110
#> 260 2017-10-21 04:45:00 0.235          110
#> 261 2017-10-21 05:00:00 0.235          110
#> 262 2017-10-21 05:15:00 0.235          110
#> 263 2017-10-21 05:30:00 0.234          110
#> 264 2017-10-21 05:45:00 0.234          110
#> 265 2017-10-21 06:00:00 0.234          110
#> 266 2017-10-21 06:15:00 0.235          110
#> 267 2017-10-21 06:30:00 0.235          110
#> 268 2017-10-21 06:45:00 0.235          110
#> 269 2017-10-21 07:00:00 0.235          110
#> 270 2017-10-21 07:15:00 0.235          110
#> 271 2017-10-21 07:30:00 0.235          110
#> 272 2017-10-21 07:45:00 0.235          110
#> 273 2017-10-21 08:00:00 0.234          110
#> 274 2017-10-21 08:15:00 0.234          110
#> 275 2017-10-21 08:30:00 0.234          110
#> 276 2017-10-21 08:45:00 0.233          110
#> 277 2017-10-21 09:00:00 0.233          110
#> 278 2017-10-21 09:15:00 0.233          110
#> 279 2017-10-21 09:30:00 0.233          110
#> 280 2017-10-21 09:45:00 0.233          110
#> 281 2017-10-21 10:00:00 0.233          110

get_timeseries_tsid("2813562", period = "P1D", datasource = 2)

#>                Timestamp Value Quality Code
#> 1    2017-10-20 11:53:00  0.99          200
#> 2    2017-10-20 11:54:00  0.98          200
#> 3    2017-10-20 11:55:00  0.99          200
#> 4    2017-10-20 11:56:00  0.98          200
#> 5    2017-10-20 11:57:00  0.98          200
#> 6    2017-10-20 11:58:00  0.97          200
#> 7    2017-10-20 11:59:00  0.97          200
#> 8    2017-10-20 12:00:00  0.97          200
#> 9    2017-10-20 12:01:00  0.98          200
#> 10   2017-10-20 12:02:00  1.00          200
#> 11   2017-10-20 12:03:00  1.03          200
#> 12   2017-10-20 12:04:00  1.07          200
#> 13   2017-10-20 12:05:00  1.11          200
#> 14   2017-10-20 12:06:00  1.17          200
#> 15   2017-10-20 12:07:00  1.21          200
#> 16   2017-10-20 12:08:00  1.27          200
#> 17   2017-10-20 12:09:00  1.29          200
#> 18   2017-10-20 12:10:00  1.33          200
#> 19   2017-10-20 12:11:00  1.36          200
#> 20   2017-10-20 12:12:00  1.40          200
#> 21   2017-10-20 12:13:00  1.43          200
#> 22   2017-10-20 12:14:00  1.46          200
#> 23   2017-10-20 12:15:00  1.48          200
#> 24   2017-10-20 12:16:00  1.50          200
#> 25   2017-10-20 12:17:00  1.55          200
#> 26   2017-10-20 12:18:00  1.57          200
#> 27   2017-10-20 12:19:00  1.60          200
#> 28   2017-10-20 12:20:00  1.63          200
#> 29   2017-10-20 12:21:00  1.64          200
#> 30   2017-10-20 12:22:00  1.65          200
#> 31   2017-10-20 12:23:00  1.69          200
#> 32   2017-10-20 12:24:00  1.70          200
#> 33   2017-10-20 12:25:00  1.72          200
#> 34   2017-10-20 12:26:00  1.74          200
#> 35   2017-10-20 12:27:00  1.76          200
#> 36   2017-10-20 12:28:00  1.78          200
#> 37   2017-10-20 12:29:00  1.81          200
#> 38   2017-10-20 12:30:00  1.82          200
#> 39   2017-10-20 12:31:00  1.82          200
#> 40   2017-10-20 12:32:00  1.85          200
#> 41   2017-10-20 12:33:00  1.86          200
#> 42   2017-10-20 12:34:00  1.88          200
#> 43   2017-10-20 12:35:00  1.90          200
#> 44   2017-10-20 12:36:00  1.91          200
#> 45   2017-10-20 12:37:00  1.92          200
#> 46   2017-10-20 12:38:00  1.94          200
#> 47   2017-10-20 12:39:00  1.95          200
#> 48   2017-10-20 12:40:00  1.97          200
#> 49   2017-10-20 12:41:00  1.97          200
#> 50   2017-10-20 12:42:00  2.00          200
#> 51   2017-10-20 12:43:00  2.00          200
#> 52   2017-10-20 12:44:00  2.02          200
#> 53   2017-10-20 12:45:00  2.04          200
#> 54   2017-10-20 12:46:00  2.05          200
#> 55   2017-10-20 12:47:00  2.06          200
#> 56   2017-10-20 12:48:00  2.08          200
#> 57   2017-10-20 12:49:00  2.09          200
#> 58   2017-10-20 12:50:00  2.10          200
#> 59   2017-10-20 12:51:00  2.12          200
#> 60   2017-10-20 12:52:00  2.13          200
#> 61   2017-10-20 12:53:00  2.15          200
#> 62   2017-10-20 12:54:00  2.15          200
#> 63   2017-10-20 12:55:00  2.16          200
#> 64   2017-10-20 12:56:00  2.17          200
#> 65   2017-10-20 12:57:00  2.19          200
#> 66   2017-10-20 12:58:00  2.19          200
#> 67   2017-10-20 12:59:00  2.21          200
#> 68   2017-10-20 13:00:00  2.22          200
#> 69   2017-10-20 13:01:00  2.24          200
#> 70   2017-10-20 13:02:00  2.24          200
#> 71   2017-10-20 13:03:00  2.25          200
#> 72   2017-10-20 13:04:00  2.26          200
#> 73   2017-10-20 13:05:00  2.27          200
#> 74   2017-10-20 13:06:00  2.31          200
#> 75   2017-10-20 13:07:00  2.31          200
#> 76   2017-10-20 13:08:00  2.32          200
#> 77   2017-10-20 13:09:00  2.34          200
#> 78   2017-10-20 13:10:00  2.36          200
#> 79   2017-10-20 13:11:00  2.37          200
#> 80   2017-10-20 13:12:00  2.37          200
#> 81   2017-10-20 13:13:00  2.40          200
#> 82   2017-10-20 13:14:00  2.40          200
#> 83   2017-10-20 13:15:00  2.43          200
#> 84   2017-10-20 13:16:00  2.44          200
#> 85   2017-10-20 13:17:00  2.46          200
#> 86   2017-10-20 13:18:00  2.45          200
#> 87   2017-10-20 13:19:00  2.48          200
#> 88   2017-10-20 13:20:00  2.48          200
#> 89   2017-10-20 13:21:00  2.50          200
#> 90   2017-10-20 13:22:00  2.51          200
#> 91   2017-10-20 13:23:00  2.53          200
#> 92   2017-10-20 13:24:00  2.53          200
#> 93   2017-10-20 13:25:00  2.56          200
#> 94   2017-10-20 13:26:00  2.57          200
#> 95   2017-10-20 13:27:00  2.58          200
#> 96   2017-10-20 13:28:00  2.58          200
#> 97   2017-10-20 13:29:00  2.60          200
#> 98   2017-10-20 13:30:00  2.63          200
#> 99   2017-10-20 13:31:00  2.64          200
#> 100  2017-10-20 13:32:00  2.65          200
#> 101  2017-10-20 13:33:00  2.67          200
#> 102  2017-10-20 13:34:00  2.71          200
#> 103  2017-10-20 13:35:00  2.72          200
#> 104  2017-10-20 13:36:00  2.72          200
#> 105  2017-10-20 13:37:00  2.72          200
#> 106  2017-10-20 13:38:00  2.75          200
#> 107  2017-10-20 13:39:00  2.75          200
#> 108  2017-10-20 13:40:00  2.76          200
#> 109  2017-10-20 13:41:00  2.78          200
#> 110  2017-10-20 13:42:00  2.80          200
#> 111  2017-10-20 13:43:00  2.82          200
#> 112  2017-10-20 13:44:00  2.82          200
#> 113  2017-10-20 13:45:00  2.84          200
#> 114  2017-10-20 13:46:00  2.86          200
#> 115  2017-10-20 13:47:00  2.87          200
#> 116  2017-10-20 13:48:00  2.89          200
#> 117  2017-10-20 13:49:00  2.90          200
#> 118  2017-10-20 13:50:00  2.92          200
#> 119  2017-10-20 13:51:00  2.94          200
#> 120  2017-10-20 13:52:00  2.95          200
#> 121  2017-10-20 13:53:00  2.97          200
#> 122  2017-10-20 13:54:00  2.98          200
#> 123  2017-10-20 13:55:00  3.00          200
#> 124  2017-10-20 13:56:00  3.02          200
#> 125  2017-10-20 13:57:00  3.03          200
#> 126  2017-10-20 13:58:00  3.04          200
#> 127  2017-10-20 13:59:00  3.06          200
#> 128  2017-10-20 14:00:00  3.06          200
#> 129  2017-10-20 14:01:00  3.09          200
#> 130  2017-10-20 14:02:00  3.10          200
#> 131  2017-10-20 14:03:00  3.11          200
#> 132  2017-10-20 14:04:00  3.13          200
#> 133  2017-10-20 14:05:00  3.14          200
#> 134  2017-10-20 14:06:00  3.16          200
#> 135  2017-10-20 14:07:00  3.18          200
#> 136  2017-10-20 14:08:00  3.19          200
#> 137  2017-10-20 14:09:00  3.21          200
#> 138  2017-10-20 14:10:00  3.22          200
#> 139  2017-10-20 14:11:00  3.24          200
#> 140  2017-10-20 14:12:00  3.25          200
#> 141  2017-10-20 14:13:00  3.27          200
#> 142  2017-10-20 14:14:00  3.29          200
#> 143  2017-10-20 14:15:00  3.31          200
#> 144  2017-10-20 14:16:00  3.33          200
#> 145  2017-10-20 14:17:00  3.35          200
#> 146  2017-10-20 14:18:00  3.36          200
#> 147  2017-10-20 14:19:00  3.38          200
#> 148  2017-10-20 14:20:00  3.40          200
#> 149  2017-10-20 14:21:00  3.42          200
#> 150  2017-10-20 14:22:00  3.42          200
#> 151  2017-10-20 14:23:00  3.44          200
#> 152  2017-10-20 14:24:00  3.46          200
#> 153  2017-10-20 14:25:00  3.48          200
#> 154  2017-10-20 14:26:00  3.49          200
#> 155  2017-10-20 14:27:00  3.51          200
#> 156  2017-10-20 14:28:00  3.53          200
#> 157  2017-10-20 14:29:00  3.54          200
#> 158  2017-10-20 14:30:00  3.57          200
#> 159  2017-10-20 14:31:00  3.59          200
#> 160  2017-10-20 14:32:00  3.61          200
#> 161  2017-10-20 14:33:00  3.62          200
#> 162  2017-10-20 14:34:00  3.65          200
#> 163  2017-10-20 14:35:00  3.67          200
#> 164  2017-10-20 14:36:00  3.69          200
#> 165  2017-10-20 14:37:00  3.71          200
#> 166  2017-10-20 14:38:00  3.73          200
#> 167  2017-10-20 14:39:00  3.75          200
#> 168  2017-10-20 14:40:00  3.77          200
#> 169  2017-10-20 14:41:00  3.80          200
#> 170  2017-10-20 14:42:00  3.82          200
#> 171  2017-10-20 14:43:00  3.84          200
#> 172  2017-10-20 14:44:00  3.86          200
#> 173  2017-10-20 14:45:00  3.89          200
#> 174  2017-10-20 14:46:00  3.91          200
#> 175  2017-10-20 14:47:00  3.94          200
#> 176  2017-10-20 14:48:00  3.96          200
#> 177  2017-10-20 14:49:00  3.98          200
#> 178  2017-10-20 14:50:00  4.01          200
#> 179  2017-10-20 14:51:00  4.03          200
#> 180  2017-10-20 14:52:00  4.05          200
#> 181  2017-10-20 14:53:00  4.07          200
#> 182  2017-10-20 14:54:00  4.10          200
#> 183  2017-10-20 14:55:00  4.12          200
#> 184  2017-10-20 14:56:00  4.15          200
#> 185  2017-10-20 14:57:00  4.17          200
#> 186  2017-10-20 14:58:00  4.19          200
#> 187  2017-10-20 14:59:00  4.22          200
#> 188  2017-10-20 15:00:00  4.24          200
#> 189  2017-10-20 15:01:00  4.27          200
#> 190  2017-10-20 15:02:00  4.30          200
#> 191  2017-10-20 15:03:00  4.32          200
#> 192  2017-10-20 15:04:00  4.35          200
#> 193  2017-10-20 15:05:00  4.37          200
#> 194  2017-10-20 15:06:00  4.40          200
#> 195  2017-10-20 15:07:00  4.42          200
#> 196  2017-10-20 15:08:00  4.45          200
#> 197  2017-10-20 15:09:00  4.48          200
#> 198  2017-10-20 15:10:00  4.50          200
#> 199  2017-10-20 15:11:00  4.52          200
#> 200  2017-10-20 15:12:00  4.55          200
#> 201  2017-10-20 15:13:00  4.58          200
#> 202  2017-10-20 15:14:00  4.60          200
#> 203  2017-10-20 15:15:00  4.63          200
#> 204  2017-10-20 15:16:00  4.66          200
#> 205  2017-10-20 15:17:00  4.69          200
#> 206  2017-10-20 15:18:00  4.72          200
#> 207  2017-10-20 15:19:00  4.75          200
#> 208  2017-10-20 15:20:00  4.78          200
#> 209  2017-10-20 15:21:00  4.80          200
#> 210  2017-10-20 15:22:00  4.84          200
#> 211  2017-10-20 15:23:00  4.86          200
#> 212  2017-10-20 15:24:00  4.88          200
#> 213  2017-10-20 15:25:00  4.91          200
#> 214  2017-10-20 15:26:00  4.92          200
#> 215  2017-10-20 15:27:00  4.95          200
#> 216  2017-10-20 15:28:00  4.98          200
#> 217  2017-10-20 15:29:00  5.00          200
#> 218  2017-10-20 15:30:00  5.02          200
#> 219  2017-10-20 15:31:00  5.05          200
#> 220  2017-10-20 15:32:00  5.07          200
#> 221  2017-10-20 15:33:00  5.09          200
#> 222  2017-10-20 15:34:00  5.11          200
#> 223  2017-10-20 15:35:00  5.13          200
#> 224  2017-10-20 15:36:00  5.14          200
#> 225  2017-10-20 15:37:00  5.17          200
#> 226  2017-10-20 15:38:00  5.19          200
#> 227  2017-10-20 15:39:00  5.21          200
#> 228  2017-10-20 15:40:00  5.23          200
#> 229  2017-10-20 15:41:00  5.24          200
#> 230  2017-10-20 15:42:00  5.26          200
#> 231  2017-10-20 15:43:00  5.28          200
#> 232  2017-10-20 15:44:00  5.30          200
#> 233  2017-10-20 15:45:00  5.32          200
#> 234  2017-10-20 15:46:00  5.33          200
#> 235  2017-10-20 15:47:00  5.35          200
#> 236  2017-10-20 15:48:00  5.37          200
#> 237  2017-10-20 15:49:00  5.38          200
#> 238  2017-10-20 15:50:00  5.39          200
#> 239  2017-10-20 15:51:00  5.41          200
#> 240  2017-10-20 15:52:00  5.42          200
#> 241  2017-10-20 15:53:00  5.44          200
#> 242  2017-10-20 15:54:00  5.45          200
#> 243  2017-10-20 15:55:00  5.46          200
#> 244  2017-10-20 15:56:00  5.47          200
#> 245  2017-10-20 15:57:00  5.48          200
#> 246  2017-10-20 15:58:00  5.49          200
#> 247  2017-10-20 15:59:00  5.50          200
#> 248  2017-10-20 16:00:00  5.51          200
#> 249  2017-10-20 16:01:00  5.52          200
#> 250  2017-10-20 16:02:00  5.53          200
#> 251  2017-10-20 16:03:00  5.54          200
#> 252  2017-10-20 16:04:00  5.55          200
#> 253  2017-10-20 16:05:00  5.56          200
#> 254  2017-10-20 16:06:00  5.57          200
#> 255  2017-10-20 16:07:00  5.58          200
#> 256  2017-10-20 16:08:00  5.59          200
#> 257  2017-10-20 16:09:00  5.60          200
#> 258  2017-10-20 16:10:00  5.61          200
#> 259  2017-10-20 16:11:00  5.62          200
#> 260  2017-10-20 16:12:00  5.62          200
#> 261  2017-10-20 16:13:00  5.63          200
#> 262  2017-10-20 16:14:00  5.64          200
#> 263  2017-10-20 16:15:00  5.65          200
#> 264  2017-10-20 16:16:00  5.66          200
#> 265  2017-10-20 16:17:00  5.66          200
#> 266  2017-10-20 16:18:00  5.66          200
#> 267  2017-10-20 16:19:00  5.67          200
#> 268  2017-10-20 16:20:00  5.68          200
#> 269  2017-10-20 16:21:00  5.69          200
#> 270  2017-10-20 16:22:00  5.69          200
#> 271  2017-10-20 16:23:00  5.69          200
#> 272  2017-10-20 16:24:00  5.70          200
#> 273  2017-10-20 16:25:00  5.70          200
#> 274  2017-10-20 16:26:00  5.70          200
#> 275  2017-10-20 16:27:00  5.71          200
#> 276  2017-10-20 16:28:00  5.71          200
#> 277  2017-10-20 16:29:00  5.72          200
#> 278  2017-10-20 16:30:00  5.72          200
#> 279  2017-10-20 16:31:00  5.72          200
#> 280  2017-10-20 16:32:00  5.72          200
#> 281  2017-10-20 16:33:00  5.73          200
#> 282  2017-10-20 16:34:00  5.73          200
#> 283  2017-10-20 16:35:00  5.73          200
#> 284  2017-10-20 16:36:00  5.74          200
#> 285  2017-10-20 16:37:00  5.74          200
#> 286  2017-10-20 16:38:00  5.74          200
#> 287  2017-10-20 16:39:00  5.74          200
#> 288  2017-10-20 16:40:00  5.74          200
#> 289  2017-10-20 16:41:00  5.74          200
#> 290  2017-10-20 16:42:00  5.75          200
#> 291  2017-10-20 16:43:00  5.75          200
#> 292  2017-10-20 16:44:00  5.75          200
#> 293  2017-10-20 16:45:00  5.75          200
#> 294  2017-10-20 16:46:00  5.75          200
#> 295  2017-10-20 16:47:00  5.74          200
#> 296  2017-10-20 16:48:00  5.74          200
#> 297  2017-10-20 16:49:00  5.74          200
#> 298  2017-10-20 16:50:00  5.74          200
#> 299  2017-10-20 16:51:00  5.74          200
#> 300  2017-10-20 16:52:00  5.73          200
#> 301  2017-10-20 16:53:00  5.73          200
#> 302  2017-10-20 16:54:00  5.72          200
#> 303  2017-10-20 16:55:00  5.72          200
#> 304  2017-10-20 16:56:00  5.71          200
#> 305  2017-10-20 16:57:00  5.71          200
#> 306  2017-10-20 16:58:00  5.70          200
#> 307  2017-10-20 16:59:00  5.69          200
#> 308  2017-10-20 17:00:00  5.69          200
#> 309  2017-10-20 17:01:00  5.68          200
#> 310  2017-10-20 17:02:00  5.67          200
#> 311  2017-10-20 17:03:00  5.66          200
#> 312  2017-10-20 17:04:00  5.65          200
#> 313  2017-10-20 17:05:00  5.64          200
#> 314  2017-10-20 17:06:00  5.63          200
#> 315  2017-10-20 17:07:00  5.62          200
#> 316  2017-10-20 17:08:00  5.61          200
#> 317  2017-10-20 17:09:00  5.60          200
#> 318  2017-10-20 17:10:00  5.58          200
#> 319  2017-10-20 17:11:00  5.57          200
#> 320  2017-10-20 17:12:00  5.56          200
#> 321  2017-10-20 17:13:00  5.54          200
#> 322  2017-10-20 17:14:00  5.53          200
#> 323  2017-10-20 17:15:00  5.51          200
#> 324  2017-10-20 17:16:00  5.50          200
#> 325  2017-10-20 17:17:00  5.48          200
#> 326  2017-10-20 17:18:00  5.47          200
#> 327  2017-10-20 17:19:00  5.45          200
#> 328  2017-10-20 17:20:00  5.43          200
#> 329  2017-10-20 17:21:00  5.42          200
#> 330  2017-10-20 17:22:00  5.40          200
#> 331  2017-10-20 17:23:00  5.38          200
#> 332  2017-10-20 17:24:00  5.36          200
#> 333  2017-10-20 17:25:00  5.34          200
#>  [ reached getOption("max.print") -- omitted 1070 rows ]