Download timeseries data from waterinfo.be
Source: R/get_timeseries.R

Using the ts_id codes and by providing a given date period, download the corresponding time series from the waterinfo.be website

get_timeseries_tsid(ts_id, period = NULL, from = NULL, to = NULL,
  datasource = 1, token = NULL)

Arguments

ts_id
waterinfo.be database ts_id, defining a timeserie variable and frequency it is defined.

period
input string according to format required by waterinfo: De period string is provided as P#Y#M#DT#H#M#S, with P defines `Period`, each # is an integer value and the codes define the number of... Y - years M - months D - days T required if information about sub-day resolution is present H - hours D - days M - minutes S - seconds Instead of D (days), the usage of W - weeks is possible as well Examples of valid period strings: P3D, P1Y, P1DT12H, PT6H, P1Y6M3DT4H20M30S.

from
date of datestring as start of the time series

to
date of datestring as end of the time series

datasource
int [0-3] defines the `meetnet` of which the measurement station is part of. VMM based stations are net '1', MOW-HIC is net '2'

token
token to use with the call (optional, can be retrieved via get_token)

Value

data.frame with the timestamps, values and quality code

Examples

get_timeseries_tsid("35055042", from="2017-01-01", to="2017-01-02")

#>              Timestamp Value Quality Code
#> 1  2017-01-01 00:00:00  0.00          130
#> 2  2017-01-01 00:15:00  0.00          130
#> 3  2017-01-01 00:30:00  0.00          130
#> 4  2017-01-01 00:45:00  0.00          130
#> 5  2017-01-01 01:00:00  0.00          130
#> 6  2017-01-01 01:15:00  0.00          130
#> 7  2017-01-01 01:30:00  0.00          130
#> 8  2017-01-01 01:45:00  0.00          130
#> 9  2017-01-01 02:00:00  0.00          130
#> 10 2017-01-01 02:15:00  0.00          130
#> 11 2017-01-01 02:30:00  0.00          130
#> 12 2017-01-01 02:45:00  0.00          130
#> 13 2017-01-01 03:00:00  0.00          130
#> 14 2017-01-01 03:15:00  0.00          130
#> 15 2017-01-01 03:30:00  0.00          130
#> 16 2017-01-01 03:45:00  0.00          130
#> 17 2017-01-01 04:00:00  0.00          130
#> 18 2017-01-01 04:15:00  0.00          130
#> 19 2017-01-01 04:30:00  0.00          130
#> 20 2017-01-01 04:45:00  0.00          130
#> 21 2017-01-01 05:00:00  0.00          130
#> 22 2017-01-01 05:15:00  0.00          130
#> 23 2017-01-01 05:30:00  0.00          130
#> 24 2017-01-01 05:45:00  0.00          130
#> 25 2017-01-01 06:00:00  0.00          130
#> 26 2017-01-01 06:15:00  0.00          130
#> 27 2017-01-01 06:30:00  0.00          130
#> 28 2017-01-01 06:45:00  0.00          130
#> 29 2017-01-01 07:00:00  0.00          130
#> 30 2017-01-01 07:15:00  0.00          130
#> 31 2017-01-01 07:30:00  0.00          130
#> 32 2017-01-01 07:45:00  0.00          130
#> 33 2017-01-01 08:00:00  0.00          130
#> 34 2017-01-01 08:15:00  0.00          130
#> 35 2017-01-01 08:30:00  0.00          130
#> 36 2017-01-01 08:45:00  0.00          130
#> 37 2017-01-01 09:00:00  0.00          130
#> 38 2017-01-01 09:15:00  0.00          130
#> 39 2017-01-01 09:30:00  0.00          130
#> 40 2017-01-01 09:45:00  0.00          130
#> 41 2017-01-01 10:00:00  0.00          130
#> 42 2017-01-01 10:15:00  0.00          130
#> 43 2017-01-01 10:30:00  0.00          130
#> 44 2017-01-01 10:45:00  0.00          130
#> 45 2017-01-01 11:00:00  0.00          130
#> 46 2017-01-01 11:15:00  0.00          130
#> 47 2017-01-01 11:30:00  0.00          130
#> 48 2017-01-01 11:45:00  0.00          130
#> 49 2017-01-01 12:00:00  0.00          130
#> 50 2017-01-01 12:15:00  0.00          130
#> 51 2017-01-01 12:30:00  0.00          130
#> 52 2017-01-01 12:45:00  0.00          130
#> 53 2017-01-01 13:00:00  0.00          130
#> 54 2017-01-01 13:15:00  0.00          130
#> 55 2017-01-01 13:30:00  0.00          130
#> 56 2017-01-01 13:45:00  0.00          130
#> 57 2017-01-01 14:00:00  0.00          130
#> 58 2017-01-01 14:15:00  0.00          130
#> 59 2017-01-01 14:30:00  0.00          130
#> 60 2017-01-01 14:45:00  0.00          130
#> 61 2017-01-01 15:00:00  0.00          130
#> 62 2017-01-01 15:15:00  0.00          130
#> 63 2017-01-01 15:30:00  0.00          130
#> 64 2017-01-01 15:45:00  0.00          130
#> 65 2017-01-01 16:00:00  0.00          130
#> 66 2017-01-01 16:15:00  0.00          130
#> 67 2017-01-01 16:30:00  0.00          130
#> 68 2017-01-01 16:45:00  0.00          130
#> 69 2017-01-01 17:00:00  0.00          130
#> 70 2017-01-01 17:15:00  0.00          130
#> 71 2017-01-01 17:30:00  0.00          130
#> 72 2017-01-01 17:45:00  0.03          130
#> 73 2017-01-01 18:00:00  0.00          130
#> 74 2017-01-01 18:15:00  0.04          130
#> 75 2017-01-01 18:30:00  0.00          130
#> 76 2017-01-01 18:45:00  0.03          130
#> 77 2017-01-01 19:00:00  0.00          130
#> 78 2017-01-01 19:15:00  0.03          130
#> 79 2017-01-01 19:30:00  0.03          130
#> 80 2017-01-01 19:45:00  0.00          130
#> 81 2017-01-01 20:00:00  0.03          130
#> 82 2017-01-01 20:15:00  0.00          130
#> 83 2017-01-01 20:30:00  0.00          130
#> 84 2017-01-01 20:45:00  0.00          130
#> 85 2017-01-01 21:00:00  0.00          130
#> 86 2017-01-01 21:15:00  0.03          130
#> 87 2017-01-01 21:30:00  0.00          130
#> 88 2017-01-01 21:45:00  0.00          130
#> 89 2017-01-01 22:00:00  0.00          130
#> 90 2017-01-01 22:15:00  0.11          130
#> 91 2017-01-01 22:30:00  0.82          130
#> 92 2017-01-01 22:45:00  0.69          130
#> 93 2017-01-01 23:00:00  0.34          130
#> 94 2017-01-01 23:15:00  0.13          130
#> 95 2017-01-01 23:30:00  0.03          130
#> 96 2017-01-01 23:45:00  0.03          130
#> 97 2017-01-02 00:00:00  0.05          130

get_timeseries_tsid("5156042", period = "P3D")

#>               Timestamp Value Quality Code
#> 1   2018-04-30 09:15:00 1.004          110
#> 2   2018-04-30 09:30:00 1.005          110
#> 3   2018-04-30 09:45:00 1.006          110
#> 4   2018-04-30 10:00:00 1.006          110
#> 5   2018-04-30 10:15:00 1.007          110
#> 6   2018-04-30 10:30:00 1.007          110
#> 7   2018-04-30 10:45:00 1.009          110
#> 8   2018-04-30 11:00:00 1.009          110
#> 9   2018-04-30 11:15:00 1.010          110
#> 10  2018-04-30 11:30:00 1.012          110
#> 11  2018-04-30 11:45:00 1.012          110
#> 12  2018-04-30 12:00:00 1.013          110
#> 13  2018-04-30 12:15:00 1.014          110
#> 14  2018-04-30 12:30:00 1.014          110
#> 15  2018-04-30 12:45:00 1.014          110
#> 16  2018-04-30 13:00:00 1.014          110
#> 17  2018-04-30 13:15:00 1.015          110
#> 18  2018-04-30 13:30:00 1.015          110
#> 19  2018-04-30 13:45:00 1.015          110
#> 20  2018-04-30 14:00:00 1.015          110
#> 21  2018-04-30 14:15:00 1.015          110
#> 22  2018-04-30 14:30:00 1.014          110
#> 23  2018-04-30 14:45:00 1.013          110
#> 24  2018-04-30 15:00:00 1.013          110
#> 25  2018-04-30 15:15:00 1.013          110
#> 26  2018-04-30 15:30:00 1.012          110
#> 27  2018-04-30 15:45:00 1.012          110
#> 28  2018-04-30 16:00:00 1.012          110
#> 29  2018-04-30 16:15:00 1.012          110
#> 30  2018-04-30 16:30:00 1.012          110
#> 31  2018-04-30 16:45:00 1.012          110
#> 32  2018-04-30 17:00:00 1.011          110
#> 33  2018-04-30 17:15:00 1.010          110
#> 34  2018-04-30 17:30:00 1.009          110
#> 35  2018-04-30 17:45:00 1.008          110
#> 36  2018-04-30 18:00:00 1.008          110
#> 37  2018-04-30 18:15:00 1.008          110
#> 38  2018-04-30 18:30:00 1.009          110
#> 39  2018-04-30 18:45:00 1.010          110
#> 40  2018-04-30 19:00:00 1.011          110
#> 41  2018-04-30 19:15:00 1.012          110
#> 42  2018-04-30 19:30:00 1.013          110
#> 43  2018-04-30 19:45:00 1.014          110
#> 44  2018-04-30 20:00:00 1.015          110
#> 45  2018-04-30 20:15:00 1.016          110
#> 46  2018-04-30 20:30:00 1.016          110
#> 47  2018-04-30 20:45:00 1.017          110
#> 48  2018-04-30 21:00:00 1.019          110
#> 49  2018-04-30 21:15:00 1.020          110
#> 50  2018-04-30 21:30:00 1.021          110
#> 51  2018-04-30 21:45:00 1.022          110
#> 52  2018-04-30 22:00:00 1.023          110
#> 53  2018-04-30 22:15:00 1.024          110
#> 54  2018-04-30 22:30:00 1.026          110
#> 55  2018-04-30 22:45:00 1.028          110
#> 56  2018-04-30 23:00:00 1.030          110
#> 57  2018-04-30 23:15:00 1.033          110
#> 58  2018-04-30 23:30:00 1.036          110
#> 59  2018-04-30 23:45:00 1.040          110
#> 60  2018-05-01 00:00:00 1.043          110
#> 61  2018-05-01 00:15:00 1.047          110
#> 62  2018-05-01 00:30:00 1.051          110
#> 63  2018-05-01 00:45:00 1.054          110
#> 64  2018-05-01 01:00:00 1.057          110
#> 65  2018-05-01 01:15:00 1.061          110
#> 66  2018-05-01 01:30:00 1.064          110
#> 67  2018-05-01 01:45:00 1.066          110
#> 68  2018-05-01 02:00:00 1.069          110
#> 69  2018-05-01 02:15:00 1.072          110
#> 70  2018-05-01 02:30:00 1.074          110
#> 71  2018-05-01 02:45:00 1.075          110
#> 72  2018-05-01 03:00:00 1.077          110
#> 73  2018-05-01 03:15:00 1.078          110
#> 74  2018-05-01 03:30:00 1.079          110
#> 75  2018-05-01 03:45:00 1.080          110
#> 76  2018-05-01 04:00:00 1.081          110
#> 77  2018-05-01 04:15:00 1.081          110
#> 78  2018-05-01 04:30:00 1.081          110
#> 79  2018-05-01 04:45:00 1.082          110
#> 80  2018-05-01 05:00:00 1.082          110
#> 81  2018-05-01 05:15:00 1.082          110
#> 82  2018-05-01 05:30:00 1.082          110
#> 83  2018-05-01 05:45:00 1.082          110
#> 84  2018-05-01 06:00:00 1.082          110
#> 85  2018-05-01 06:15:00 1.081          110
#> 86  2018-05-01 06:30:00 1.081          110
#> 87  2018-05-01 06:45:00 1.081          110
#> 88  2018-05-01 07:00:00 1.081          110
#> 89  2018-05-01 07:15:00 1.081          110
#> 90  2018-05-01 07:30:00 1.081          110
#> 91  2018-05-01 07:45:00 1.082          110
#> 92  2018-05-01 08:00:00 1.082          110
#> 93  2018-05-01 08:15:00 1.082          110
#> 94  2018-05-01 08:30:00 1.082          110
#> 95  2018-05-01 08:45:00 1.081          110
#> 96  2018-05-01 09:00:00 1.082          110
#> 97  2018-05-01 09:15:00 1.082          110
#> 98  2018-05-01 09:30:00 1.082          110
#> 99  2018-05-01 09:45:00 1.082          110
#> 100 2018-05-01 10:00:00 1.082          110
#> 101 2018-05-01 10:15:00 1.082          110
#> 102 2018-05-01 10:30:00 1.081          110
#> 103 2018-05-01 10:45:00 1.081          110
#> 104 2018-05-01 11:00:00 1.081          110
#> 105 2018-05-01 11:15:00 1.081          110
#> 106 2018-05-01 11:30:00 1.081          110
#> 107 2018-05-01 11:45:00 1.081          110
#> 108 2018-05-01 12:00:00 1.082          110
#> 109 2018-05-01 12:15:00 1.080          110
#> 110 2018-05-01 12:30:00 1.080          110
#> 111 2018-05-01 12:45:00 1.079          110
#> 112 2018-05-01 13:00:00 1.079          110
#> 113 2018-05-01 13:15:00 1.079          110
#> 114 2018-05-01 13:30:00 1.079          110
#> 115 2018-05-01 13:45:00 1.079          110
#> 116 2018-05-01 14:00:00 1.078          110
#> 117 2018-05-01 14:15:00 1.079          110
#> 118 2018-05-01 14:30:00 1.079          110
#> 119 2018-05-01 14:45:00 1.079          110
#> 120 2018-05-01 15:00:00 1.080          110
#> 121 2018-05-01 15:15:00 1.081          110
#> 122 2018-05-01 15:30:00 1.081          110
#> 123 2018-05-01 15:45:00 1.081          110
#> 124 2018-05-01 16:00:00 1.081          110
#> 125 2018-05-01 16:15:00 1.081          110
#> 126 2018-05-01 16:30:00 1.080          110
#> 127 2018-05-01 16:45:00 1.080          110
#> 128 2018-05-01 17:00:00 1.079          110
#> 129 2018-05-01 17:15:00 1.078          110
#> 130 2018-05-01 17:30:00 1.077          110
#> 131 2018-05-01 17:45:00 1.076          110
#> 132 2018-05-01 18:00:00 1.076          110
#> 133 2018-05-01 18:15:00 1.074          110
#> 134 2018-05-01 18:30:00 1.074          110
#> 135 2018-05-01 18:45:00 1.073          110
#> 136 2018-05-01 19:00:00 1.072          110
#> 137 2018-05-01 19:15:00 1.071          110
#> 138 2018-05-01 19:30:00 1.070          110
#> 139 2018-05-01 19:45:00 1.070          110
#> 140 2018-05-01 20:00:00 1.068          110
#> 141 2018-05-01 20:15:00 1.067          110
#> 142 2018-05-01 20:30:00 1.067          110
#> 143 2018-05-01 20:45:00 1.066          110
#> 144 2018-05-01 21:00:00 1.065          110
#> 145 2018-05-01 21:15:00 1.064          110
#> 146 2018-05-01 21:30:00 1.063          110
#> 147 2018-05-01 21:45:00 1.061          110
#> 148 2018-05-01 22:00:00 1.061          110
#> 149 2018-05-01 22:15:00 1.059          110
#> 150 2018-05-01 22:30:00 1.058          110
#> 151 2018-05-01 22:45:00 1.057          110
#> 152 2018-05-01 23:00:00 1.056          110
#> 153 2018-05-01 23:15:00 1.055          110
#> 154 2018-05-01 23:30:00 1.053          110
#> 155 2018-05-01 23:45:00 1.052          110
#> 156 2018-05-02 00:00:00 1.051          110
#> 157 2018-05-02 00:15:00 1.049          110
#> 158 2018-05-02 00:30:00 1.048          110
#> 159 2018-05-02 00:45:00 1.047          110
#> 160 2018-05-02 01:00:00 1.045          110
#> 161 2018-05-02 01:15:00 1.043          110
#> 162 2018-05-02 01:30:00 1.042          110
#> 163 2018-05-02 01:45:00 1.041          110
#> 164 2018-05-02 02:00:00 1.039          110
#> 165 2018-05-02 02:15:00 1.037          110
#> 166 2018-05-02 02:30:00 1.036          110
#> 167 2018-05-02 02:45:00 1.034          110
#> 168 2018-05-02 03:00:00 1.032          110
#> 169 2018-05-02 03:15:00 1.030          110
#> 170 2018-05-02 03:30:00 1.028          110
#> 171 2018-05-02 03:45:00 1.026          110
#> 172 2018-05-02 04:00:00 1.024          110
#> 173 2018-05-02 04:15:00 1.023          110
#> 174 2018-05-02 04:30:00 1.021          110
#> 175 2018-05-02 04:45:00 1.019          110
#> 176 2018-05-02 05:00:00 1.016          110
#> 177 2018-05-02 05:15:00 1.014          110
#> 178 2018-05-02 05:30:00 1.011          110
#> 179 2018-05-02 05:45:00 1.007          110
#> 180 2018-05-02 06:00:00 1.004          110
#> 181 2018-05-02 06:15:00 1.001          110
#> 182 2018-05-02 06:30:00 0.997          110
#> 183 2018-05-02 06:45:00 0.993          110
#> 184 2018-05-02 07:00:00 0.990          110
#> 185 2018-05-02 07:15:00 0.987          110
#> 186 2018-05-02 07:30:00 0.984          110
#> 187 2018-05-02 07:45:00 0.981          110
#> 188 2018-05-02 08:00:00 0.978          110
#> 189 2018-05-02 08:15:00 0.976          110
#> 190 2018-05-02 08:30:00 0.973          110
#> 191 2018-05-02 08:45:00 0.971          110
#> 192 2018-05-02 09:00:00 0.969          110
#> 193 2018-05-02 09:15:00 0.965          110
#> 194 2018-05-02 09:30:00 0.963          110
#> 195 2018-05-02 09:45:00 0.960          110
#> 196 2018-05-02 10:00:00 0.958          110
#> 197 2018-05-02 10:15:00 0.955          110
#> 198 2018-05-02 10:30:00 0.952          110
#> 199 2018-05-02 10:45:00 0.949          110
#> 200 2018-05-02 11:00:00 0.947          110
#> 201 2018-05-02 11:15:00 0.943          110
#> 202 2018-05-02 11:30:00 0.941          110
#> 203 2018-05-02 11:45:00 0.937          110
#> 204 2018-05-02 12:00:00 0.935          110
#> 205 2018-05-02 12:15:00 0.932          110
#> 206 2018-05-02 12:30:00 0.929          110
#> 207 2018-05-02 12:45:00 0.926          110
#> 208 2018-05-02 13:00:00 0.923          110
#> 209 2018-05-02 13:15:00 0.920          110
#> 210 2018-05-02 13:30:00 0.918          110
#> 211 2018-05-02 13:45:00 0.915          110
#> 212 2018-05-02 14:00:00 0.913          110
#> 213 2018-05-02 14:15:00 0.910          110
#> 214 2018-05-02 14:30:00 0.907          110
#> 215 2018-05-02 14:45:00 0.904          110
#> 216 2018-05-02 15:00:00 0.902          110
#> 217 2018-05-02 15:15:00 0.899          110
#> 218 2018-05-02 15:30:00 0.896          110
#> 219 2018-05-02 15:45:00 0.894          110
#> 220 2018-05-02 16:00:00 0.890          110
#> 221 2018-05-02 16:15:00 0.888          110
#> 222 2018-05-02 16:30:00 0.885          110
#> 223 2018-05-02 16:45:00 0.883          110
#> 224 2018-05-02 17:00:00 0.881          110
#> 225 2018-05-02 17:15:00 0.879          110
#> 226 2018-05-02 17:30:00 0.876          110
#> 227 2018-05-02 17:45:00 0.873          110
#> 228 2018-05-02 18:00:00 0.871          110
#> 229 2018-05-02 18:15:00 0.869          110
#> 230 2018-05-02 18:30:00 0.866          110
#> 231 2018-05-02 18:45:00 0.864          110
#> 232 2018-05-02 19:00:00 0.861          110
#> 233 2018-05-02 19:15:00 0.858          110
#> 234 2018-05-02 19:30:00 0.855          110
#> 235 2018-05-02 19:45:00 0.853          110
#> 236 2018-05-02 20:00:00 0.850          110
#> 237 2018-05-02 20:15:00 0.847          110
#> 238 2018-05-02 20:30:00 0.844          110
#> 239 2018-05-02 20:45:00 0.841          110
#> 240 2018-05-02 21:00:00 0.838          110
#> 241 2018-05-02 21:15:00 0.834          110
#> 242 2018-05-02 21:30:00 0.832          110
#> 243 2018-05-02 21:45:00 0.829          110
#> 244 2018-05-02 22:00:00 0.826          110
#> 245 2018-05-02 22:15:00 0.824          110
#> 246 2018-05-02 22:30:00 0.821          110
#> 247 2018-05-02 22:45:00 0.818          110
#> 248 2018-05-02 23:00:00 0.815          110
#> 249 2018-05-02 23:15:00 0.813          110
#> 250 2018-05-02 23:30:00 0.810          110
#> 251 2018-05-02 23:45:00 0.807          110
#> 252 2018-05-03 00:00:00 0.804          110
#> 253 2018-05-03 00:15:00 0.802          110
#> 254 2018-05-03 00:30:00 0.799          110
#> 255 2018-05-03 00:45:00 0.796          110
#> 256 2018-05-03 01:00:00 0.794          110
#> 257 2018-05-03 01:15:00 0.792          110
#> 258 2018-05-03 01:30:00 0.789          110
#> 259 2018-05-03 01:45:00 0.786          110
#> 260 2018-05-03 02:00:00 0.784          110
#> 261 2018-05-03 02:15:00 0.781          110
#> 262 2018-05-03 02:30:00 0.779          110
#> 263 2018-05-03 02:45:00 0.777          110
#> 264 2018-05-03 03:00:00 0.774          110
#> 265 2018-05-03 03:15:00 0.771          110
#> 266 2018-05-03 03:30:00 0.769          110
#> 267 2018-05-03 03:45:00 0.767          110
#> 268 2018-05-03 04:00:00 0.765          110
#> 269 2018-05-03 04:15:00 0.762          110
#> 270 2018-05-03 04:30:00 0.760          110
#> 271 2018-05-03 04:45:00 0.758          110
#> 272 2018-05-03 05:00:00 0.755          110
#> 273 2018-05-03 05:15:00 0.753          110
#> 274 2018-05-03 05:30:00 0.751          110
#> 275 2018-05-03 05:45:00 0.749          110
#> 276 2018-05-03 06:00:00 0.746          110
#> 277 2018-05-03 06:15:00 0.744          110
#> 278 2018-05-03 06:30:00 0.742          110
#> 279 2018-05-03 06:45:00 0.740          110
#> 280 2018-05-03 07:00:00 0.737          110

get_timeseries_tsid("2813562", period = "P1D", datasource = 2)

#>                Timestamp Value Quality Code
#> 1    2018-05-02 09:06:00  3.03          200
#> 2    2018-05-02 09:07:00  3.02          200
#> 3    2018-05-02 09:08:00  3.01          200
#> 4    2018-05-02 09:09:00  3.00          200
#> 5    2018-05-02 09:10:00  2.99          200
#> 6    2018-05-02 09:11:00  2.99          200
#> 7    2018-05-02 09:12:00  2.98          200
#> 8    2018-05-02 09:13:00  2.96          200
#> 9    2018-05-02 09:14:00  2.94          200
#> 10   2018-05-02 09:15:00  2.94          200
#> 11   2018-05-02 09:16:00  2.93          200
#> 12   2018-05-02 09:17:00  2.92          200
#> 13   2018-05-02 09:18:00  2.91          200
#> 14   2018-05-02 09:19:00  2.90          200
#> 15   2018-05-02 09:20:00  2.89          200
#> 16   2018-05-02 09:21:00  2.88          200
#> 17   2018-05-02 09:22:00  2.88          200
#> 18   2018-05-02 09:23:00  2.86          200
#> 19   2018-05-02 09:24:00  2.86          200
#> 20   2018-05-02 09:25:00  2.84          200
#> 21   2018-05-02 09:26:00  2.83          200
#> 22   2018-05-02 09:27:00  2.82          200
#> 23   2018-05-02 09:28:00  2.81          200
#> 24   2018-05-02 09:29:00  2.80          200
#> 25   2018-05-02 09:30:00  2.78          200
#> 26   2018-05-02 09:31:00  2.78          200
#> 27   2018-05-02 09:32:00  2.77          200
#> 28   2018-05-02 09:33:00  2.75          200
#> 29   2018-05-02 09:34:00  2.74          200
#> 30   2018-05-02 09:35:00  2.74          200
#> 31   2018-05-02 09:36:00  2.72          200
#> 32   2018-05-02 09:37:00  2.71          200
#> 33   2018-05-02 09:38:00  2.70          200
#> 34   2018-05-02 09:39:00  2.69          200
#> 35   2018-05-02 09:40:00  2.68          200
#> 36   2018-05-02 09:41:00  2.68          200
#> 37   2018-05-02 09:42:00  2.67          200
#> 38   2018-05-02 09:43:00  2.65          200
#> 39   2018-05-02 09:44:00  2.64          200
#> 40   2018-05-02 09:45:00  2.64          200
#> 41   2018-05-02 09:46:00  2.63          200
#> 42   2018-05-02 09:47:00  2.62          200
#> 43   2018-05-02 09:48:00  2.61          200
#> 44   2018-05-02 09:49:00  2.59          200
#> 45   2018-05-02 09:50:00  2.58          200
#> 46   2018-05-02 09:51:00  2.57          200
#> 47   2018-05-02 09:52:00  2.56          200
#> 48   2018-05-02 09:53:00  2.55          200
#> 49   2018-05-02 09:54:00  2.54          200
#> 50   2018-05-02 09:55:00  2.53          200
#> 51   2018-05-02 09:56:00  2.52          200
#> 52   2018-05-02 09:57:00  2.52          200
#> 53   2018-05-02 09:58:00  2.52          200
#> 54   2018-05-02 09:59:00  2.51          200
#> 55   2018-05-02 10:00:00  2.49          200
#> 56   2018-05-02 10:01:00  2.47          200
#> 57   2018-05-02 10:02:00  2.47          200
#> 58   2018-05-02 10:03:00  2.46          200
#> 59   2018-05-02 10:04:00  2.44          200
#> 60   2018-05-02 10:05:00  2.43          200
#> 61   2018-05-02 10:06:00  2.42          200
#> 62   2018-05-02 10:07:00  2.41          200
#> 63   2018-05-02 10:08:00  2.40          200
#> 64   2018-05-02 10:09:00  2.39          200
#> 65   2018-05-02 10:10:00  2.38          200
#> 66   2018-05-02 10:11:00  2.37          200
#> 67   2018-05-02 10:12:00  2.36          200
#> 68   2018-05-02 10:13:00  2.35          200
#> 69   2018-05-02 10:14:00  2.34          200
#> 70   2018-05-02 10:15:00  2.33          200
#> 71   2018-05-02 10:16:00  2.32          200
#> 72   2018-05-02 10:17:00  2.31          200
#> 73   2018-05-02 10:18:00  2.30          200
#> 74   2018-05-02 10:19:00  2.29          200
#> 75   2018-05-02 10:20:00  2.28          200
#> 76   2018-05-02 10:21:00  2.28          200
#> 77   2018-05-02 10:22:00  2.26          200
#> 78   2018-05-02 10:23:00  2.26          200
#> 79   2018-05-02 10:24:00  2.24          200
#> 80   2018-05-02 10:25:00  2.23          200
#> 81   2018-05-02 10:26:00  2.23          200
#> 82   2018-05-02 10:27:00  2.21          200
#> 83   2018-05-02 10:28:00  2.21          200
#> 84   2018-05-02 10:29:00  2.20          200
#> 85   2018-05-02 10:30:00  2.19          200
#> 86   2018-05-02 10:31:00  2.18          200
#> 87   2018-05-02 10:32:00  2.17          200
#> 88   2018-05-02 10:33:00  2.16          200
#> 89   2018-05-02 10:34:00  2.15          200
#> 90   2018-05-02 10:35:00  2.14          200
#> 91   2018-05-02 10:36:00  2.13          200
#> 92   2018-05-02 10:37:00  2.12          200
#> 93   2018-05-02 10:38:00  2.11          200
#> 94   2018-05-02 10:39:00  2.10          200
#> 95   2018-05-02 10:40:00  2.10          200
#> 96   2018-05-02 10:41:00  2.09          200
#> 97   2018-05-02 10:42:00  2.07          200
#> 98   2018-05-02 10:43:00  2.06          200
#> 99   2018-05-02 10:44:00  2.05          200
#> 100  2018-05-02 10:45:00  2.05          200
#> 101  2018-05-02 10:46:00  2.03          200
#> 102  2018-05-02 10:47:00  2.03          200
#> 103  2018-05-02 10:48:00  2.02          200
#> 104  2018-05-02 10:49:00  2.01          200
#> 105  2018-05-02 10:50:00  2.00          200
#> 106  2018-05-02 10:51:00  2.00          200
#> 107  2018-05-02 10:52:00  1.98          200
#> 108  2018-05-02 10:53:00  1.97          200
#> 109  2018-05-02 10:54:00  1.96          200
#> 110  2018-05-02 10:55:00  1.96          200
#> 111  2018-05-02 10:56:00  1.94          200
#> 112  2018-05-02 10:57:00  1.93          200
#> 113  2018-05-02 10:58:00  1.93          200
#> 114  2018-05-02 10:59:00  1.92          200
#> 115  2018-05-02 11:00:00  1.91          200
#> 116  2018-05-02 11:01:00  1.90          200
#> 117  2018-05-02 11:02:00  1.89          200
#> 118  2018-05-02 11:03:00  1.89          200
#> 119  2018-05-02 11:04:00  1.88          200
#> 120  2018-05-02 11:05:00  1.86          200
#> 121  2018-05-02 11:06:00  1.85          200
#> 122  2018-05-02 11:07:00  1.84          200
#> 123  2018-05-02 11:08:00  1.83          200
#> 124  2018-05-02 11:09:00  1.82          200
#> 125  2018-05-02 11:10:00  1.82          200
#> 126  2018-05-02 11:11:00  1.81          200
#> 127  2018-05-02 11:12:00  1.80          200
#> 128  2018-05-02 11:13:00  1.79          200
#> 129  2018-05-02 11:14:00  1.78          200
#> 130  2018-05-02 11:15:00  1.77          200
#> 131  2018-05-02 11:16:00  1.76          200
#> 132  2018-05-02 11:17:00  1.76          200
#> 133  2018-05-02 11:18:00  1.75          200
#> 134  2018-05-02 11:19:00  1.74          200
#> 135  2018-05-02 11:20:00  1.73          200
#> 136  2018-05-02 11:21:00  1.72          200
#> 137  2018-05-02 11:22:00  1.71          200
#> 138  2018-05-02 11:23:00  1.70          200
#> 139  2018-05-02 11:24:00  1.70          200
#> 140  2018-05-02 11:25:00  1.69          200
#> 141  2018-05-02 11:26:00  1.68          200
#> 142  2018-05-02 11:27:00  1.67          200
#> 143  2018-05-02 11:28:00  1.66          200
#> 144  2018-05-02 11:29:00  1.65          200
#> 145  2018-05-02 11:30:00  1.64          200
#> 146  2018-05-02 11:31:00  1.63          200
#> 147  2018-05-02 11:32:00  1.63          200
#> 148  2018-05-02 11:33:00  1.62          200
#> 149  2018-05-02 11:34:00  1.61          200
#> 150  2018-05-02 11:35:00  1.60          200
#> 151  2018-05-02 11:36:00  1.60          200
#> 152  2018-05-02 11:37:00  1.59          200
#> 153  2018-05-02 11:38:00  1.58          200
#> 154  2018-05-02 11:39:00  1.57          200
#> 155  2018-05-02 11:40:00  1.56          200
#> 156  2018-05-02 11:41:00  1.55          200
#> 157  2018-05-02 11:42:00  1.54          200
#> 158  2018-05-02 11:43:00  1.54          200
#> 159  2018-05-02 11:44:00  1.52          200
#> 160  2018-05-02 11:45:00  1.52          200
#> 161  2018-05-02 11:46:00  1.51          200
#> 162  2018-05-02 11:47:00  1.50          200
#> 163  2018-05-02 11:48:00  1.49          200
#> 164  2018-05-02 11:49:00  1.48          200
#> 165  2018-05-02 11:50:00  1.48          200
#> 166  2018-05-02 11:51:00  1.46          200
#> 167  2018-05-02 11:52:00  1.46          200
#> 168  2018-05-02 11:53:00  1.45          200
#> 169  2018-05-02 11:54:00  1.44          200
#> 170  2018-05-02 11:55:00  1.43          200
#> 171  2018-05-02 11:56:00  1.42          200
#> 172  2018-05-02 11:57:00  1.42          200
#> 173  2018-05-02 11:58:00  1.41          200
#> 174  2018-05-02 11:59:00  1.40          200
#> 175  2018-05-02 12:00:00  1.39          200
#> 176  2018-05-02 12:01:00  1.38          200
#> 177  2018-05-02 12:02:00  1.38          200
#> 178  2018-05-02 12:03:00  1.37          200
#> 179  2018-05-02 12:04:00  1.36          200
#> 180  2018-05-02 12:05:00  1.35          200
#> 181  2018-05-02 12:06:00  1.34          200
#> 182  2018-05-02 12:07:00  1.34          200
#> 183  2018-05-02 12:08:00  1.33          200
#> 184  2018-05-02 12:09:00  1.32          200
#> 185  2018-05-02 12:10:00  1.31          200
#> 186  2018-05-02 12:11:00  1.30          200
#> 187  2018-05-02 12:12:00  1.29          200
#> 188  2018-05-02 12:13:00  1.28          200
#> 189  2018-05-02 12:14:00  1.28          200
#> 190  2018-05-02 12:15:00  1.27          200
#> 191  2018-05-02 12:16:00  1.26          200
#> 192  2018-05-02 12:17:00  1.25          200
#> 193  2018-05-02 12:18:00  1.24          200
#> 194  2018-05-02 12:19:00  1.24          200
#> 195  2018-05-02 12:20:00  1.23          200
#> 196  2018-05-02 12:21:00  1.22          200
#> 197  2018-05-02 12:22:00  1.21          200
#> 198  2018-05-02 12:23:00  1.21          200
#> 199  2018-05-02 12:24:00  1.20          200
#> 200  2018-05-02 12:25:00  1.19          200
#> 201  2018-05-02 12:26:00  1.18          200
#> 202  2018-05-02 12:27:00  1.18          200
#> 203  2018-05-02 12:28:00  1.17          200
#> 204  2018-05-02 12:29:00  1.16          200
#> 205  2018-05-02 12:30:00  1.15          200
#> 206  2018-05-02 12:31:00  1.14          200
#> 207  2018-05-02 12:32:00  1.14          200
#> 208  2018-05-02 12:33:00  1.13          200
#> 209  2018-05-02 12:34:00  1.12          200
#> 210  2018-05-02 12:35:00  1.11          200
#> 211  2018-05-02 12:36:00  1.10          200
#> 212  2018-05-02 12:37:00  1.09          200
#> 213  2018-05-02 12:38:00  1.08          200
#> 214  2018-05-02 12:39:00  1.08          200
#> 215  2018-05-02 12:40:00  1.07          200
#> 216  2018-05-02 12:41:00  1.06          200
#> 217  2018-05-02 12:42:00  1.06          200
#> 218  2018-05-02 12:43:00  1.06          200
#> 219  2018-05-02 12:44:00  1.04          200
#> 220  2018-05-02 12:45:00  1.04          200
#> 221  2018-05-02 12:46:00  1.03          200
#> 222  2018-05-02 12:47:00  1.03          200
#> 223  2018-05-02 12:48:00  1.02          200
#> 224  2018-05-02 12:49:00  1.01          200
#> 225  2018-05-02 12:50:00  1.00          200
#> 226  2018-05-02 12:51:00  1.00          200
#> 227  2018-05-02 12:52:00  0.99          200
#> 228  2018-05-02 12:53:00  0.98          200
#> 229  2018-05-02 12:54:00  0.97          200
#> 230  2018-05-02 12:55:00  0.97          200
#> 231  2018-05-02 12:56:00  0.96          200
#> 232  2018-05-02 12:57:00  0.96          200
#> 233  2018-05-02 12:58:00  0.95          200
#> 234  2018-05-02 12:59:00  0.94          200
#> 235  2018-05-02 13:00:00  0.93          200
#> 236  2018-05-02 13:01:00  0.93          200
#> 237  2018-05-02 13:02:00  0.92          200
#> 238  2018-05-02 13:03:00  0.92          200
#> 239  2018-05-02 13:04:00  0.90          200
#> 240  2018-05-02 13:05:00  0.90          200
#> 241  2018-05-02 13:06:00  0.90          200
#> 242  2018-05-02 13:07:00  0.89          200
#> 243  2018-05-02 13:08:00  0.88          200
#> 244  2018-05-02 13:09:00  0.88          200
#> 245  2018-05-02 13:10:00  0.86          200
#> 246  2018-05-02 13:11:00  0.86          200
#> 247  2018-05-02 13:12:00  0.85          200
#> 248  2018-05-02 13:13:00  0.84          200
#> 249  2018-05-02 13:14:00  0.84          200
#> 250  2018-05-02 13:15:00  0.83          200
#> 251  2018-05-02 13:16:00  0.83          200
#> 252  2018-05-02 13:17:00  0.81          200
#> 253  2018-05-02 13:18:00  0.82          200
#> 254  2018-05-02 13:19:00  0.81          200
#> 255  2018-05-02 13:20:00  0.80          200
#> 256  2018-05-02 13:21:00  0.80          200
#> 257  2018-05-02 13:22:00  0.79          200
#> 258  2018-05-02 13:23:00  0.78          200
#> 259  2018-05-02 13:24:00  0.77          200
#> 260  2018-05-02 13:25:00  0.76          200
#> 261  2018-05-02 13:26:00  0.76          200
#> 262  2018-05-02 13:27:00  0.76          200
#> 263  2018-05-02 13:28:00  0.74          200
#> 264  2018-05-02 13:29:00  0.75          200
#> 265  2018-05-02 13:30:00  0.74          200
#> 266  2018-05-02 13:31:00  0.74          200
#> 267  2018-05-02 13:32:00  0.74          200
#> 268  2018-05-02 13:33:00  0.75          200
#> 269  2018-05-02 13:34:00  0.76          200
#> 270  2018-05-02 13:35:00  0.78          200
#> 271  2018-05-02 13:36:00  0.82          200
#> 272  2018-05-02 13:37:00  0.86          200
#> 273  2018-05-02 13:38:00  0.89          200
#> 274  2018-05-02 13:39:00  0.94          200
#> 275  2018-05-02 13:40:00  0.98          200
#> 276  2018-05-02 13:41:00  1.01          200
#> 277  2018-05-02 13:42:00  1.04          200
#> 278  2018-05-02 13:43:00  1.07          200
#> 279  2018-05-02 13:44:00  1.11          200
#> 280  2018-05-02 13:45:00  1.14          200
#> 281  2018-05-02 13:46:00  1.15          200
#> 282  2018-05-02 13:47:00  1.19          200
#> 283  2018-05-02 13:48:00  1.22          200
#> 284  2018-05-02 13:49:00  1.25          200
#> 285  2018-05-02 13:50:00  1.28          200
#> 286  2018-05-02 13:51:00  1.30          200
#> 287  2018-05-02 13:52:00  1.32          200
#> 288  2018-05-02 13:53:00  1.33          200
#> 289  2018-05-02 13:54:00  1.35          200
#> 290  2018-05-02 13:55:00  1.38          200
#> 291  2018-05-02 13:56:00  1.39          200
#> 292  2018-05-02 13:57:00  1.42          200
#> 293  2018-05-02 13:58:00  1.44          200
#> 294  2018-05-02 13:59:00  1.44          200
#> 295  2018-05-02 14:00:00  1.46          200
#> 296  2018-05-02 14:01:00  1.48          200
#> 297  2018-05-02 14:02:00  1.51          200
#> 298  2018-05-02 14:03:00  1.52          200
#> 299  2018-05-02 14:04:00  1.53          200
#> 300  2018-05-02 14:05:00  1.55          200
#> 301  2018-05-02 14:06:00  1.57          200
#> 302  2018-05-02 14:07:00  1.59          200
#> 303  2018-05-02 14:08:00  1.60          200
#> 304  2018-05-02 14:09:00  1.62          200
#> 305  2018-05-02 14:10:00  1.64          200
#> 306  2018-05-02 14:11:00  1.64          200
#> 307  2018-05-02 14:12:00  1.64          200
#> 308  2018-05-02 14:13:00  1.67          200
#> 309  2018-05-02 14:14:00  1.70          200
#> 310  2018-05-02 14:15:00  1.68          200
#> 311  2018-05-02 14:16:00  1.69          200
#> 312  2018-05-02 14:17:00  1.70          200
#> 313  2018-05-02 14:18:00  1.73          200
#> 314  2018-05-02 14:19:00  1.76          200
#> 315  2018-05-02 14:20:00  1.75          200
#> 316  2018-05-02 14:21:00  1.76          200
#> 317  2018-05-02 14:22:00  1.78          200
#> 318  2018-05-02 14:23:00  1.81          200
#> 319  2018-05-02 14:24:00  1.80          200
#> 320  2018-05-02 14:25:00  1.83          200
#> 321  2018-05-02 14:26:00  1.84          200
#> 322  2018-05-02 14:27:00  1.84          200
#> 323  2018-05-02 14:28:00  1.90          200
#> 324  2018-05-02 14:29:00  1.86          200
#> 325  2018-05-02 14:30:00  1.89          200
#> 326  2018-05-02 14:31:00  1.92          200
#> 327  2018-05-02 14:32:00  1.90          200
#> 328  2018-05-02 14:33:00  1.91          200
#> 329  2018-05-02 14:34:00  1.94          200
#> 330  2018-05-02 14:35:00  1.97          200
#> 331  2018-05-02 14:36:00  1.96          200
#> 332  2018-05-02 14:37:00  1.97          200
#> 333  2018-05-02 14:38:00  1.99          200
#>  [ reached getOption("max.print") -- omitted 1067 rows ]