Download timeseries data from waterinfo.be

Using the ts_id codes and by providing a given date period, download the corresponding time series from the waterinfo.be website

get_timeseries_tsid(ts_id, period = NULL, from = NULL, to = NULL,
  datasource = 1)

Arguments

ts_id
waterinfo.be database ts_id, defining a timeserie variable and frequency it is defined.

period
input string according to format required by waterinfo: De period string is provided as P#Y#M#DT#H#M#S, with P defines `Period`, each # is an integer value and the codes define the number of... Y - years M - months D - days T required if information about sub-day resolution is present H - hours D - days M - minutes S - seconds Instead of D (days), the usage of W - weeks is possible as well Examples of valid period strings: P3D, P1Y, P1DT12H, PT6H, P1Y6M3DT4H20M30S.

from
date of datestring as start of the time series

to
date of datestring as end of the time series

datasource
int [0-3] defines the `meetnet` of which the measurement station is part of. VMM based stations are net '1', MOW-HIC is net '2'

Value

data.frame with the timestamps, values and quality code

Examples

get_timeseries_tsid("35055042", from="2017-01-01", to="2017-01-02")

#>              Timestamp Value Quality Code
#> 1  2017-01-01 00:00:00  0.00          130
#> 2  2017-01-01 00:15:00  0.00          130
#> 3  2017-01-01 00:30:00  0.00          130
#> 4  2017-01-01 00:45:00  0.00          130
#> 5  2017-01-01 01:00:00  0.00          130
#> 6  2017-01-01 01:15:00  0.00          130
#> 7  2017-01-01 01:30:00  0.00          130
#> 8  2017-01-01 01:45:00  0.00          130
#> 9  2017-01-01 02:00:00  0.00          130
#> 10 2017-01-01 02:15:00  0.00          130
#> 11 2017-01-01 02:30:00  0.00          130
#> 12 2017-01-01 02:45:00  0.00          130
#> 13 2017-01-01 03:00:00  0.00          130
#> 14 2017-01-01 03:15:00  0.00          130
#> 15 2017-01-01 03:30:00  0.00          130
#> 16 2017-01-01 03:45:00  0.00          130
#> 17 2017-01-01 04:00:00  0.00          130
#> 18 2017-01-01 04:15:00  0.00          130
#> 19 2017-01-01 04:30:00  0.00          130
#> 20 2017-01-01 04:45:00  0.00          130
#> 21 2017-01-01 05:00:00  0.00          130
#> 22 2017-01-01 05:15:00  0.00          130
#> 23 2017-01-01 05:30:00  0.00          130
#> 24 2017-01-01 05:45:00  0.00          130
#> 25 2017-01-01 06:00:00  0.00          130
#> 26 2017-01-01 06:15:00  0.00          130
#> 27 2017-01-01 06:30:00  0.00          130
#> 28 2017-01-01 06:45:00  0.00          130
#> 29 2017-01-01 07:00:00  0.00          130
#> 30 2017-01-01 07:15:00  0.00          130
#> 31 2017-01-01 07:30:00  0.00          130
#> 32 2017-01-01 07:45:00  0.00          130
#> 33 2017-01-01 08:00:00  0.00          130
#> 34 2017-01-01 08:15:00  0.00          130
#> 35 2017-01-01 08:30:00  0.00          130
#> 36 2017-01-01 08:45:00  0.00          130
#> 37 2017-01-01 09:00:00  0.00          130
#> 38 2017-01-01 09:15:00  0.00          130
#> 39 2017-01-01 09:30:00  0.00          130
#> 40 2017-01-01 09:45:00  0.00          130
#> 41 2017-01-01 10:00:00  0.00          130
#> 42 2017-01-01 10:15:00  0.00          130
#> 43 2017-01-01 10:30:00  0.00          130
#> 44 2017-01-01 10:45:00  0.00          130
#> 45 2017-01-01 11:00:00  0.00          130
#> 46 2017-01-01 11:15:00  0.00          130
#> 47 2017-01-01 11:30:00  0.00          130
#> 48 2017-01-01 11:45:00  0.00          130
#> 49 2017-01-01 12:00:00  0.00          130
#> 50 2017-01-01 12:15:00  0.00          130
#> 51 2017-01-01 12:30:00  0.00          130
#> 52 2017-01-01 12:45:00  0.00          130
#> 53 2017-01-01 13:00:00  0.00          130
#> 54 2017-01-01 13:15:00  0.00          130
#> 55 2017-01-01 13:30:00  0.00          130
#> 56 2017-01-01 13:45:00  0.00          130
#> 57 2017-01-01 14:00:00  0.00          130
#> 58 2017-01-01 14:15:00  0.00          130
#> 59 2017-01-01 14:30:00  0.00          130
#> 60 2017-01-01 14:45:00  0.00          130
#> 61 2017-01-01 15:00:00  0.00          130
#> 62 2017-01-01 15:15:00  0.00          130
#> 63 2017-01-01 15:30:00  0.00          130
#> 64 2017-01-01 15:45:00  0.00          130
#> 65 2017-01-01 16:00:00  0.00          130
#> 66 2017-01-01 16:15:00  0.00          130
#> 67 2017-01-01 16:30:00  0.00          130
#> 68 2017-01-01 16:45:00  0.00          130
#> 69 2017-01-01 17:00:00  0.00          130
#> 70 2017-01-01 17:15:00  0.00          130
#> 71 2017-01-01 17:30:00  0.00          130
#> 72 2017-01-01 17:45:00  0.03          130
#> 73 2017-01-01 18:00:00  0.00          130
#> 74 2017-01-01 18:15:00  0.04          130
#> 75 2017-01-01 18:30:00  0.00          130
#> 76 2017-01-01 18:45:00  0.03          130
#> 77 2017-01-01 19:00:00  0.00          130
#> 78 2017-01-01 19:15:00  0.03          130
#> 79 2017-01-01 19:30:00  0.03          130
#> 80 2017-01-01 19:45:00  0.00          130
#> 81 2017-01-01 20:00:00  0.03          130
#> 82 2017-01-01 20:15:00  0.00          130
#> 83 2017-01-01 20:30:00  0.00          130
#> 84 2017-01-01 20:45:00  0.00          130
#> 85 2017-01-01 21:00:00  0.00          130
#> 86 2017-01-01 21:15:00  0.03          130
#> 87 2017-01-01 21:30:00  0.00          130
#> 88 2017-01-01 21:45:00  0.00          130
#> 89 2017-01-01 22:00:00  0.00          130
#> 90 2017-01-01 22:15:00  0.11          130
#> 91 2017-01-01 22:30:00  0.82          130
#> 92 2017-01-01 22:45:00  0.69          130
#> 93 2017-01-01 23:00:00  0.34          130
#> 94 2017-01-01 23:15:00  0.13          130
#> 95 2017-01-01 23:30:00  0.03          130
#> 96 2017-01-01 23:45:00  0.03          130
#> 97 2017-01-02 00:00:00  0.05          130

get_timeseries_tsid("5156042", period = "P3D")

#>               Timestamp Value Quality Code
#> 1   2017-11-14 11:00:00 0.264          110
#> 2   2017-11-14 11:15:00 0.264          110
#> 3   2017-11-14 11:30:00 0.263          110
#> 4   2017-11-14 11:45:00 0.262          110
#> 5   2017-11-14 12:00:00 0.262          110
#> 6   2017-11-14 12:15:00 0.261          110
#> 7   2017-11-14 12:30:00 0.260          110
#> 8   2017-11-14 12:45:00 0.260          110
#> 9   2017-11-14 13:00:00 0.260          110
#> 10  2017-11-14 13:15:00 0.259          110
#> 11  2017-11-14 13:30:00 0.259          110
#> 12  2017-11-14 13:45:00 0.259          110
#> 13  2017-11-14 14:00:00 0.258          110
#> 14  2017-11-14 14:15:00 0.258          110
#> 15  2017-11-14 14:30:00 0.258          110
#> 16  2017-11-14 14:45:00 0.257          110
#> 17  2017-11-14 15:00:00 0.257          110
#> 18  2017-11-14 15:15:00 0.257          110
#> 19  2017-11-14 15:30:00 0.257          110
#> 20  2017-11-14 15:45:00 0.257          110
#> 21  2017-11-14 16:00:00 0.256          110
#> 22  2017-11-14 16:15:00 0.256          110
#> 23  2017-11-14 16:30:00 0.256          110
#> 24  2017-11-14 16:45:00 0.256          110
#> 25  2017-11-14 17:00:00 0.256          110
#> 26  2017-11-14 17:15:00 0.256          110
#> 27  2017-11-14 17:30:00 0.255          110
#> 28  2017-11-14 17:45:00 0.255          110
#> 29  2017-11-14 18:00:00 0.255          110
#> 30  2017-11-14 18:15:00 0.255          110
#> 31  2017-11-14 18:30:00 0.255          110
#> 32  2017-11-14 18:45:00 0.255          110
#> 33  2017-11-14 19:00:00 0.254          110
#> 34  2017-11-14 19:15:00 0.254          110
#> 35  2017-11-14 19:30:00 0.254          110
#> 36  2017-11-14 19:45:00 0.254          110
#> 37  2017-11-14 20:00:00 0.254          110
#> 38  2017-11-14 20:15:00 0.254          110
#> 39  2017-11-14 20:30:00 0.254          110
#> 40  2017-11-14 20:45:00 0.253          110
#> 41  2017-11-14 21:00:00 0.253          110
#> 42  2017-11-14 21:15:00 0.253          110
#> 43  2017-11-14 21:30:00 0.253          110
#> 44  2017-11-14 21:45:00 0.253          110
#> 45  2017-11-14 22:00:00 0.253          110
#> 46  2017-11-14 22:15:00 0.253          110
#> 47  2017-11-14 22:30:00 0.252          110
#> 48  2017-11-14 22:45:00 0.252          110
#> 49  2017-11-14 23:00:00 0.253          110
#> 50  2017-11-14 23:15:00 0.252          110
#> 51  2017-11-14 23:30:00 0.252          110
#> 52  2017-11-14 23:45:00 0.252          110
#> 53  2017-11-15 00:00:00 0.252          110
#> 54  2017-11-15 00:15:00 0.252          110
#> 55  2017-11-15 00:30:00 0.252          110
#> 56  2017-11-15 00:45:00 0.252          110
#> 57  2017-11-15 01:00:00 0.252          110
#> 58  2017-11-15 01:15:00 0.252          110
#> 59  2017-11-15 01:30:00 0.252          110
#> 60  2017-11-15 01:45:00 0.252          110
#> 61  2017-11-15 02:00:00 0.252          110
#> 62  2017-11-15 02:15:00 0.252          110
#> 63  2017-11-15 02:30:00 0.252          110
#> 64  2017-11-15 02:45:00 0.252          110
#> 65  2017-11-15 03:00:00 0.252          110
#> 66  2017-11-15 03:15:00 0.252          110
#> 67  2017-11-15 03:30:00 0.252          110
#> 68  2017-11-15 03:45:00 0.252          110
#> 69  2017-11-15 04:00:00 0.252          110
#> 70  2017-11-15 04:15:00 0.252          110
#> 71  2017-11-15 04:30:00 0.252          110
#> 72  2017-11-15 04:45:00 0.252          110
#> 73  2017-11-15 05:00:00 0.252          110
#> 74  2017-11-15 05:15:00 0.252          110
#> 75  2017-11-15 05:30:00 0.251          110
#> 76  2017-11-15 05:45:00 0.251          110
#> 77  2017-11-15 06:00:00 0.251          110
#> 78  2017-11-15 06:15:00 0.251          110
#> 79  2017-11-15 06:30:00 0.250          110
#> 80  2017-11-15 06:45:00 0.250          110
#> 81  2017-11-15 07:00:00 0.250          110
#> 82  2017-11-15 07:15:00 0.250          110
#> 83  2017-11-15 07:30:00 0.250          110
#> 84  2017-11-15 07:45:00 0.249          110
#> 85  2017-11-15 08:00:00 0.249          110
#> 86  2017-11-15 08:15:00 0.249          110
#> 87  2017-11-15 08:30:00 0.249          110
#> 88  2017-11-15 08:45:00 0.249          110
#> 89  2017-11-15 09:00:00 0.249          110
#> 90  2017-11-15 09:15:00 0.249          110
#> 91  2017-11-15 09:30:00 0.249          110
#> 92  2017-11-15 09:45:00 0.249          110
#> 93  2017-11-15 10:00:00 0.248          110
#> 94  2017-11-15 10:15:00 0.249          110
#> 95  2017-11-15 10:30:00 0.248          110
#> 96  2017-11-15 10:45:00 0.248          110
#> 97  2017-11-15 11:00:00 0.248          110
#> 98  2017-11-15 11:15:00 0.248          110
#> 99  2017-11-15 11:30:00 0.248          110
#> 100 2017-11-15 11:45:00 0.248          110
#> 101 2017-11-15 12:00:00 0.249          110
#> 102 2017-11-15 12:15:00 0.248          110
#> 103 2017-11-15 12:30:00 0.249          110
#> 104 2017-11-15 12:45:00 0.249          110
#> 105 2017-11-15 13:00:00 0.249          110
#> 106 2017-11-15 13:15:00 0.249          110
#> 107 2017-11-15 13:30:00 0.249          110
#> 108 2017-11-15 13:45:00 0.250          110
#> 109 2017-11-15 14:00:00 0.250          110
#> 110 2017-11-15 14:15:00 0.251          110
#> 111 2017-11-15 14:30:00 0.252          110
#> 112 2017-11-15 14:45:00 0.253          110
#> 113 2017-11-15 15:00:00 0.254          110
#> 114 2017-11-15 15:15:00 0.256          110
#> 115 2017-11-15 15:30:00 0.257          110
#> 116 2017-11-15 15:45:00 0.259          110
#> 117 2017-11-15 16:00:00 0.261          110
#> 118 2017-11-15 16:15:00 0.262          110
#> 119 2017-11-15 16:30:00 0.263          110
#> 120 2017-11-15 16:45:00 0.264          110
#> 121 2017-11-15 17:00:00 0.264          110
#> 122 2017-11-15 17:15:00 0.264          110
#> 123 2017-11-15 17:30:00 0.264          110
#> 124 2017-11-15 17:45:00 0.263          110
#> 125 2017-11-15 18:00:00 0.263          110
#> 126 2017-11-15 18:15:00 0.262          110
#> 127 2017-11-15 18:30:00 0.262          110
#> 128 2017-11-15 18:45:00 0.261          110
#> 129 2017-11-15 19:00:00 0.261          110
#> 130 2017-11-15 19:15:00 0.260          110
#> 131 2017-11-15 19:30:00 0.259          110
#> 132 2017-11-15 19:45:00 0.259          110
#> 133 2017-11-15 20:00:00 0.258          110
#> 134 2017-11-15 20:15:00 0.257          110
#> 135 2017-11-15 20:30:00 0.257          110
#> 136 2017-11-15 20:45:00 0.256          110
#> 137 2017-11-15 21:00:00 0.255          110
#> 138 2017-11-15 21:15:00 0.256          110
#> 139 2017-11-15 21:30:00 0.256          110
#> 140 2017-11-15 21:45:00 0.256          110
#> 141 2017-11-15 22:00:00 0.256          110
#> 142 2017-11-15 22:15:00 0.257          110
#> 143 2017-11-15 22:30:00 0.258          110
#> 144 2017-11-15 22:45:00 0.258          110
#> 145 2017-11-15 23:00:00 0.260          110
#> 146 2017-11-15 23:15:00 0.261          110
#> 147 2017-11-15 23:30:00 0.262          110
#> 148 2017-11-15 23:45:00 0.263          110
#> 149 2017-11-16 00:00:00 0.264          110
#> 150 2017-11-16 00:15:00 0.265          110
#> 151 2017-11-16 00:30:00 0.266          110
#> 152 2017-11-16 00:45:00 0.267          110
#> 153 2017-11-16 01:00:00 0.268          110
#> 154 2017-11-16 01:15:00 0.269          110
#> 155 2017-11-16 01:30:00 0.270          110
#> 156 2017-11-16 01:45:00 0.270          110
#> 157 2017-11-16 02:00:00 0.271          110
#> 158 2017-11-16 02:15:00 0.271          110
#> 159 2017-11-16 02:30:00 0.272          110
#> 160 2017-11-16 02:45:00 0.272          110
#> 161 2017-11-16 03:00:00 0.272          110
#> 162 2017-11-16 03:15:00 0.272          110
#> 163 2017-11-16 03:30:00 0.273          110
#> 164 2017-11-16 03:45:00 0.273          110
#> 165 2017-11-16 04:00:00 0.273          110
#> 166 2017-11-16 04:15:00 0.273          110
#> 167 2017-11-16 04:30:00 0.274          110
#> 168 2017-11-16 04:45:00 0.274          110
#> 169 2017-11-16 05:00:00 0.274          110
#> 170 2017-11-16 05:15:00 0.274          110
#> 171 2017-11-16 05:30:00 0.274          110
#> 172 2017-11-16 05:45:00 0.275          110
#> 173 2017-11-16 06:00:00 0.275          110
#> 174 2017-11-16 06:15:00 0.275          110
#> 175 2017-11-16 06:30:00 0.275          110
#> 176 2017-11-16 06:45:00 0.275          110
#> 177 2017-11-16 07:00:00 0.275          110
#> 178 2017-11-16 07:15:00 0.275          110
#> 179 2017-11-16 07:30:00 0.275          110
#> 180 2017-11-16 07:45:00 0.276          110
#> 181 2017-11-16 08:00:00 0.275          110
#> 182 2017-11-16 08:15:00 0.276          110
#> 183 2017-11-16 08:30:00 0.276          110
#> 184 2017-11-16 08:45:00 0.276          110
#> 185 2017-11-16 09:00:00 0.276          110
#> 186 2017-11-16 09:15:00 0.277          110
#> 187 2017-11-16 09:30:00 0.277          110
#> 188 2017-11-16 09:45:00 0.277          110
#> 189 2017-11-16 10:00:00 0.277          110
#> 190 2017-11-16 10:15:00 0.277          110
#> 191 2017-11-16 10:30:00 0.277          110
#> 192 2017-11-16 10:45:00 0.278          110
#> 193 2017-11-16 11:00:00 0.278          110
#> 194 2017-11-16 11:15:00 0.279          110
#> 195 2017-11-16 11:30:00 0.279          110
#> 196 2017-11-16 11:45:00 0.279          110
#> 197 2017-11-16 12:00:00 0.279          110
#> 198 2017-11-16 12:15:00 0.279          110
#> 199 2017-11-16 12:30:00 0.279          110
#> 200 2017-11-16 12:45:00 0.279          110
#> 201 2017-11-16 13:00:00 0.279          110
#> 202 2017-11-16 13:15:00 0.279          110
#> 203 2017-11-16 13:30:00 0.280          110
#> 204 2017-11-16 13:45:00 0.280          110
#> 205 2017-11-16 14:00:00 0.280          110
#> 206 2017-11-16 14:15:00 0.280          110
#> 207 2017-11-16 14:30:00 0.280          110
#> 208 2017-11-16 14:45:00 0.280          110
#> 209 2017-11-16 15:00:00 0.280          110
#> 210 2017-11-16 15:15:00 0.280          110
#> 211 2017-11-16 15:30:00 0.280          110
#> 212 2017-11-16 15:45:00 0.280          110
#> 213 2017-11-16 16:00:00 0.280          110
#> 214 2017-11-16 16:15:00 0.280          110
#> 215 2017-11-16 16:30:00 0.280          110
#> 216 2017-11-16 16:45:00 0.280          110
#> 217 2017-11-16 17:00:00 0.280          110
#> 218 2017-11-16 17:15:00 0.280          110
#> 219 2017-11-16 17:30:00 0.280          110
#> 220 2017-11-16 17:45:00 0.280          110
#> 221 2017-11-16 18:00:00 0.280          110
#> 222 2017-11-16 18:15:00 0.280          110
#> 223 2017-11-16 18:30:00 0.280          110
#> 224 2017-11-16 18:45:00 0.280          110
#> 225 2017-11-16 19:00:00 0.280          110
#> 226 2017-11-16 19:15:00 0.280          110
#> 227 2017-11-16 19:30:00 0.280          110
#> 228 2017-11-16 19:45:00 0.280          110
#> 229 2017-11-16 20:00:00 0.279          110
#> 230 2017-11-16 20:15:00 0.279          110
#> 231 2017-11-16 20:30:00 0.279          110
#> 232 2017-11-16 20:45:00 0.278          110
#> 233 2017-11-16 21:00:00 0.278          110
#> 234 2017-11-16 21:15:00 0.277          110
#> 235 2017-11-16 21:30:00 0.277          110
#> 236 2017-11-16 21:45:00 0.277          110
#> 237 2017-11-16 22:00:00 0.277          110
#> 238 2017-11-16 22:15:00 0.277          110
#> 239 2017-11-16 22:30:00 0.278          110
#> 240 2017-11-16 22:45:00 0.278          110
#> 241 2017-11-16 23:00:00 0.278          110
#> 242 2017-11-16 23:15:00 0.278          110
#> 243 2017-11-16 23:30:00 0.279          110
#> 244 2017-11-16 23:45:00 0.278          110
#> 245 2017-11-17 00:00:00 0.279          110
#> 246 2017-11-17 00:15:00 0.279          110
#> 247 2017-11-17 00:30:00 0.279          110
#> 248 2017-11-17 00:45:00 0.279          110
#> 249 2017-11-17 01:00:00 0.279          110
#> 250 2017-11-17 01:15:00 0.279          110
#> 251 2017-11-17 01:30:00 0.279          110
#> 252 2017-11-17 01:45:00 0.279          110
#> 253 2017-11-17 02:00:00 0.280          110
#> 254 2017-11-17 02:15:00 0.280          110
#> 255 2017-11-17 02:30:00 0.280          110
#> 256 2017-11-17 02:45:00 0.280          110
#> 257 2017-11-17 03:00:00 0.281          110
#> 258 2017-11-17 03:15:00 0.280          110
#> 259 2017-11-17 03:30:00 0.281          110
#> 260 2017-11-17 03:45:00 0.281          110
#> 261 2017-11-17 04:00:00 0.281          110
#> 262 2017-11-17 04:15:00 0.281          110
#> 263 2017-11-17 04:30:00 0.281          110
#> 264 2017-11-17 04:45:00 0.281          110
#> 265 2017-11-17 05:00:00 0.282          110
#> 266 2017-11-17 05:15:00 0.282          110
#> 267 2017-11-17 05:30:00 0.282          110
#> 268 2017-11-17 05:45:00 0.282          110
#> 269 2017-11-17 06:00:00 0.282          110
#> 270 2017-11-17 06:15:00 0.282          110
#> 271 2017-11-17 06:30:00 0.282          110
#> 272 2017-11-17 06:45:00 0.282          110
#> 273 2017-11-17 07:00:00 0.282          110
#> 274 2017-11-17 07:15:00 0.282          110
#> 275 2017-11-17 07:30:00 0.282          110
#> 276 2017-11-17 07:45:00 0.282          110
#> 277 2017-11-17 08:00:00 0.282          110
#> 278 2017-11-17 08:15:00 0.282          110
#> 279 2017-11-17 08:30:00 0.282          110
#> 280 2017-11-17 08:45:00 0.282          110
#> 281 2017-11-17 09:00:00 0.282          110
#> 282 2017-11-17 09:15:00 0.281          110
#> 283 2017-11-17 09:30:00 0.281          110
#> 284 2017-11-17 09:45:00 0.281          110
#> 285 2017-11-17 10:00:00 0.281          110

get_timeseries_tsid("2813562", period = "P1D", datasource = 2)

#>                Timestamp Value Quality Code
#> 1    2017-11-16 10:50:00  1.33          200
#> 2    2017-11-16 10:51:00  1.35          200
#> 3    2017-11-16 10:52:00  1.37          200
#> 4    2017-11-16 10:53:00  1.39          200
#> 5    2017-11-16 10:54:00  1.41          200
#> 6    2017-11-16 10:55:00  1.43          200
#> 7    2017-11-16 10:56:00  1.45          200
#> 8    2017-11-16 10:57:00  1.46          200
#> 9    2017-11-16 10:58:00  1.48          200
#> 10   2017-11-16 10:59:00  1.50          200
#> 11   2017-11-16 11:00:00  1.52          200
#> 12   2017-11-16 11:01:00  1.54          200
#> 13   2017-11-16 11:02:00  1.56          200
#> 14   2017-11-16 11:03:00  1.57          200
#> 15   2017-11-16 11:04:00  1.59          200
#> 16   2017-11-16 11:05:00  1.60          200
#> 17   2017-11-16 11:06:00  1.61          200
#> 18   2017-11-16 11:07:00  1.63          200
#> 19   2017-11-16 11:08:00  1.64          200
#> 20   2017-11-16 11:09:00  1.66          200
#> 21   2017-11-16 11:10:00  1.67          200
#> 22   2017-11-16 11:11:00  1.69          200
#> 23   2017-11-16 11:12:00  1.70          200
#> 24   2017-11-16 11:13:00  1.71          200
#> 25   2017-11-16 11:14:00  1.72          200
#> 26   2017-11-16 11:15:00  1.73          200
#> 27   2017-11-16 11:16:00  1.73          200
#> 28   2017-11-16 11:17:00  1.76          200
#> 29   2017-11-16 11:18:00  1.76          200
#> 30   2017-11-16 11:19:00  1.78          200
#> 31   2017-11-16 11:20:00  1.78          200
#> 32   2017-11-16 11:21:00  1.80          200
#> 33   2017-11-16 11:22:00  1.82          200
#> 34   2017-11-16 11:23:00  1.83          200
#> 35   2017-11-16 11:24:00  1.84          200
#> 36   2017-11-16 11:25:00  1.85          200
#> 37   2017-11-16 11:26:00  1.88          200
#> 38   2017-11-16 11:27:00  1.88          200
#> 39   2017-11-16 11:28:00  1.89          200
#> 40   2017-11-16 11:29:00  1.91          200
#> 41   2017-11-16 11:30:00  1.92          200
#> 42   2017-11-16 11:31:00  1.94          200
#> 43   2017-11-16 11:32:00  1.95          200
#> 44   2017-11-16 11:33:00  1.95          200
#> 45   2017-11-16 11:34:00  1.96          200
#> 46   2017-11-16 11:35:00  1.97          200
#> 47   2017-11-16 11:36:00  1.98          200
#> 48   2017-11-16 11:37:00  1.99          200
#> 49   2017-11-16 11:38:00  2.00          200
#> 50   2017-11-16 11:39:00  2.01          200
#> 51   2017-11-16 11:40:00  2.03          200
#> 52   2017-11-16 11:41:00  2.04          200
#> 53   2017-11-16 11:42:00  2.06          200
#> 54   2017-11-16 11:43:00  2.08          200
#> 55   2017-11-16 11:44:00  2.07          200
#> 56   2017-11-16 11:45:00  2.08          200
#> 57   2017-11-16 11:46:00  2.10          200
#> 58   2017-11-16 11:47:00  2.12          200
#> 59   2017-11-16 11:48:00  2.13          200
#> 60   2017-11-16 11:49:00  2.14          200
#> 61   2017-11-16 11:50:00  2.16          200
#> 62   2017-11-16 11:51:00  2.16          200
#> 63   2017-11-16 11:52:00  2.17          200
#> 64   2017-11-16 11:53:00  2.17          200
#> 65   2017-11-16 11:54:00  2.19          200
#> 66   2017-11-16 11:55:00  2.20          200
#> 67   2017-11-16 11:56:00  2.21          200
#> 68   2017-11-16 11:57:00  2.23          200
#> 69   2017-11-16 11:58:00  2.27          200
#> 70   2017-11-16 11:59:00  2.28          200
#> 71   2017-11-16 12:00:00  2.28          200
#> 72   2017-11-16 12:01:00  2.28          200
#> 73   2017-11-16 12:02:00  2.27          200
#> 74   2017-11-16 12:03:00  2.31          200
#> 75   2017-11-16 12:04:00  2.31          200
#> 76   2017-11-16 12:05:00  2.31          200
#> 77   2017-11-16 12:06:00  2.34          200
#> 78   2017-11-16 12:07:00  2.35          200
#> 79   2017-11-16 12:08:00  2.36          200
#> 80   2017-11-16 12:09:00  2.37          200
#> 81   2017-11-16 12:10:00  2.39          200
#> 82   2017-11-16 12:11:00  2.40          200
#> 83   2017-11-16 12:12:00  2.41          200
#> 84   2017-11-16 12:13:00  2.42          200
#> 85   2017-11-16 12:14:00  2.43          200
#> 86   2017-11-16 12:15:00  2.45          200
#> 87   2017-11-16 12:16:00  2.44          200
#> 88   2017-11-16 12:17:00  2.47          200
#> 89   2017-11-16 12:18:00  2.48          200
#> 90   2017-11-16 12:19:00  2.49          200
#> 91   2017-11-16 12:20:00  2.51          200
#> 92   2017-11-16 12:21:00  2.54          200
#> 93   2017-11-16 12:22:00  2.54          200
#> 94   2017-11-16 12:23:00  2.54          200
#> 95   2017-11-16 12:24:00  2.54          200
#> 96   2017-11-16 12:25:00  2.56          200
#> 97   2017-11-16 12:26:00  2.57          200
#> 98   2017-11-16 12:27:00  2.58          200
#> 99   2017-11-16 12:28:00  2.60          200
#> 100  2017-11-16 12:29:00  2.62          200
#> 101  2017-11-16 12:30:00  2.63          200
#> 102  2017-11-16 12:31:00  2.64          200
#> 103  2017-11-16 12:32:00  2.65          200
#> 104  2017-11-16 12:33:00  2.66          200
#> 105  2017-11-16 12:34:00  2.68          200
#> 106  2017-11-16 12:35:00  2.68          200
#> 107  2017-11-16 12:36:00  2.69          200
#> 108  2017-11-16 12:37:00  2.70          200
#> 109  2017-11-16 12:38:00  2.72          200
#> 110  2017-11-16 12:39:00  2.73          200
#> 111  2017-11-16 12:40:00  2.74          200
#> 112  2017-11-16 12:41:00  2.75          200
#> 113  2017-11-16 12:42:00  2.76          200
#> 114  2017-11-16 12:43:00  2.78          200
#> 115  2017-11-16 12:44:00  2.79          200
#> 116  2017-11-16 12:45:00  2.81          200
#> 117  2017-11-16 12:46:00  2.81          200
#> 118  2017-11-16 12:47:00  2.83          200
#> 119  2017-11-16 12:48:00  2.85          200
#> 120  2017-11-16 12:49:00  2.85          200
#> 121  2017-11-16 12:50:00  2.87          200
#> 122  2017-11-16 12:51:00  2.87          200
#> 123  2017-11-16 12:52:00  2.90          200
#> 124  2017-11-16 12:53:00  2.91          200
#> 125  2017-11-16 12:54:00  2.93          200
#> 126  2017-11-16 12:55:00  2.93          200
#> 127  2017-11-16 12:56:00  2.94          200
#> 128  2017-11-16 12:57:00  2.96          200
#> 129  2017-11-16 12:58:00  2.98          200
#> 130  2017-11-16 12:59:00  2.99          200
#> 131  2017-11-16 13:00:00  3.00          200
#> 132  2017-11-16 13:01:00  3.02          200
#> 133  2017-11-16 13:02:00  3.04          200
#> 134  2017-11-16 13:03:00  3.06          200
#> 135  2017-11-16 13:04:00  3.08          200
#> 136  2017-11-16 13:05:00  3.09          200
#> 137  2017-11-16 13:06:00  3.09          200
#> 138  2017-11-16 13:07:00  3.09          200
#> 139  2017-11-16 13:08:00  3.12          200
#> 140  2017-11-16 13:09:00  3.14          200
#> 141  2017-11-16 13:10:00  3.16          200
#> 142  2017-11-16 13:11:00  3.18          200
#> 143  2017-11-16 13:12:00  3.18          200
#> 144  2017-11-16 13:13:00  3.16          200
#> 145  2017-11-16 13:14:00  3.21          200
#> 146  2017-11-16 13:15:00  3.24          200
#> 147  2017-11-16 13:16:00  3.26          200
#> 148  2017-11-16 13:17:00  3.28          200
#> 149  2017-11-16 13:18:00  3.30          200
#> 150  2017-11-16 13:19:00  3.32          200
#> 151  2017-11-16 13:20:00  3.33          200
#> 152  2017-11-16 13:21:00  3.37          200
#> 153  2017-11-16 13:22:00  3.38          200
#> 154  2017-11-16 13:23:00  3.41          200
#> 155  2017-11-16 13:24:00  3.42          200
#> 156  2017-11-16 13:25:00  3.44          200
#> 157  2017-11-16 13:26:00  3.46          200
#> 158  2017-11-16 13:27:00  3.48          200
#> 159  2017-11-16 13:28:00  3.51          200
#> 160  2017-11-16 13:29:00  3.53          200
#> 161  2017-11-16 13:30:00  3.56          200
#> 162  2017-11-16 13:31:00  3.58          200
#> 163  2017-11-16 13:32:00  3.60          200
#> 164  2017-11-16 13:33:00  3.63          200
#> 165  2017-11-16 13:34:00  3.64          200
#> 166  2017-11-16 13:35:00  3.67          200
#> 167  2017-11-16 13:36:00  3.69          200
#> 168  2017-11-16 13:37:00  3.72          200
#> 169  2017-11-16 13:38:00  3.74          200
#> 170  2017-11-16 13:39:00  3.78          200
#> 171  2017-11-16 13:40:00  3.80          200
#> 172  2017-11-16 13:41:00  3.83          200
#> 173  2017-11-16 13:42:00  3.85          200
#> 174  2017-11-16 13:43:00  3.88          200
#> 175  2017-11-16 13:44:00  3.89          200
#> 176  2017-11-16 13:45:00  3.92          200
#> 177  2017-11-16 13:46:00  3.94          200
#> 178  2017-11-16 13:47:00  3.97          200
#> 179  2017-11-16 13:48:00  3.99          200
#> 180  2017-11-16 13:49:00  4.02          200
#> 181  2017-11-16 13:50:00  4.05          200
#> 182  2017-11-16 13:51:00  4.07          200
#> 183  2017-11-16 13:52:00  4.10          200
#> 184  2017-11-16 13:53:00  4.12          200
#> 185  2017-11-16 13:54:00  4.15          200
#> 186  2017-11-16 13:55:00  4.17          200
#> 187  2017-11-16 13:56:00  4.19          200
#> 188  2017-11-16 13:57:00  4.21          200
#> 189  2017-11-16 13:58:00  4.24          200
#> 190  2017-11-16 13:59:00  4.26          200
#> 191  2017-11-16 14:00:00  4.29          200
#> 192  2017-11-16 14:01:00  4.31          200
#> 193  2017-11-16 14:02:00  4.34          200
#> 194  2017-11-16 14:03:00  4.36          200
#> 195  2017-11-16 14:04:00  4.38          200
#> 196  2017-11-16 14:05:00  4.40          200
#> 197  2017-11-16 14:06:00  4.43          200
#> 198  2017-11-16 14:07:00  4.45          200
#> 199  2017-11-16 14:08:00  4.47          200
#> 200  2017-11-16 14:09:00  4.49          200
#> 201  2017-11-16 14:10:00  4.51          200
#> 202  2017-11-16 14:11:00  4.53          200
#> 203  2017-11-16 14:12:00  4.55          200
#> 204  2017-11-16 14:13:00  4.57          200
#> 205  2017-11-16 14:14:00  4.60          200
#> 206  2017-11-16 14:15:00  4.61          200
#> 207  2017-11-16 14:16:00  4.63          200
#> 208  2017-11-16 14:17:00  4.65          200
#> 209  2017-11-16 14:18:00  4.67          200
#> 210  2017-11-16 14:19:00  4.68          200
#> 211  2017-11-16 14:20:00  4.69          200
#> 212  2017-11-16 14:21:00  4.70          200
#> 213  2017-11-16 14:22:00  4.72          200
#> 214  2017-11-16 14:23:00  4.74          200
#> 215  2017-11-16 14:24:00  4.76          200
#> 216  2017-11-16 14:25:00  4.77          200
#> 217  2017-11-16 14:26:00  4.79          200
#> 218  2017-11-16 14:27:00  4.80          200
#> 219  2017-11-16 14:28:00  4.82          200
#> 220  2017-11-16 14:29:00  4.83          200
#> 221  2017-11-16 14:30:00  4.84          200
#> 222  2017-11-16 14:31:00  4.86          200
#> 223  2017-11-16 14:32:00  4.87          200
#> 224  2017-11-16 14:33:00  4.88          200
#> 225  2017-11-16 14:34:00  4.89          200
#> 226  2017-11-16 14:35:00  4.90          200
#> 227  2017-11-16 14:36:00  4.91          200
#> 228  2017-11-16 14:37:00  4.93          200
#> 229  2017-11-16 14:38:00  4.94          200
#> 230  2017-11-16 14:39:00  4.95          200
#> 231  2017-11-16 14:40:00  4.96          200
#> 232  2017-11-16 14:41:00  4.97          200
#> 233  2017-11-16 14:42:00  4.98          200
#> 234  2017-11-16 14:43:00  4.98          200
#> 235  2017-11-16 14:44:00  4.99          200
#> 236  2017-11-16 14:45:00  5.00          200
#> 237  2017-11-16 14:46:00  5.01          200
#> 238  2017-11-16 14:47:00  5.02          200
#> 239  2017-11-16 14:48:00  5.03          200
#> 240  2017-11-16 14:49:00  5.03          200
#> 241  2017-11-16 14:50:00  5.04          200
#> 242  2017-11-16 14:51:00  5.05          200
#> 243  2017-11-16 14:52:00  5.05          200
#> 244  2017-11-16 14:53:00  5.06          200
#> 245  2017-11-16 14:54:00  5.06          200
#> 246  2017-11-16 14:55:00  5.06          200
#> 247  2017-11-16 14:56:00  5.07          200
#> 248  2017-11-16 14:57:00  5.08          200
#> 249  2017-11-16 14:58:00  5.08          200
#> 250  2017-11-16 14:59:00  5.09          200
#> 251  2017-11-16 15:00:00  5.09          200
#> 252  2017-11-16 15:01:00  5.10          200
#> 253  2017-11-16 15:02:00  5.10          200
#> 254  2017-11-16 15:03:00  5.10          200
#> 255  2017-11-16 15:04:00  5.10          200
#> 256  2017-11-16 15:05:00  5.11          200
#> 257  2017-11-16 15:06:00  5.12          200
#> 258  2017-11-16 15:07:00  5.12          200
#> 259  2017-11-16 15:08:00  5.12          200
#> 260  2017-11-16 15:09:00  5.12          200
#> 261  2017-11-16 15:10:00  5.12          200
#> 262  2017-11-16 15:11:00  5.12          200
#> 263  2017-11-16 15:12:00  5.12          200
#> 264  2017-11-16 15:13:00  5.12          200
#> 265  2017-11-16 15:14:00  5.12          200
#> 266  2017-11-16 15:15:00  5.12          200
#> 267  2017-11-16 15:16:00  5.12          200
#> 268  2017-11-16 15:17:00  5.12          200
#> 269  2017-11-16 15:18:00  5.12          200
#> 270  2017-11-16 15:19:00  5.11          200
#> 271  2017-11-16 15:20:00  5.11          200
#> 272  2017-11-16 15:21:00  5.10          200
#> 273  2017-11-16 15:22:00  5.10          200
#> 274  2017-11-16 15:23:00  5.09          200
#> 275  2017-11-16 15:24:00  5.08          200
#> 276  2017-11-16 15:25:00  5.08          200
#> 277  2017-11-16 15:26:00  5.07          200
#> 278  2017-11-16 15:27:00  5.06          200
#> 279  2017-11-16 15:28:00  5.06          200
#> 280  2017-11-16 15:29:00  5.05          200
#> 281  2017-11-16 15:30:00  5.04          200
#> 282  2017-11-16 15:31:00  5.03          200
#> 283  2017-11-16 15:32:00  5.02          200
#> 284  2017-11-16 15:33:00  5.01          200
#> 285  2017-11-16 15:34:00  5.00          200
#> 286  2017-11-16 15:35:00  4.98          200
#> 287  2017-11-16 15:36:00  4.97          200
#> 288  2017-11-16 15:37:00  4.96          200
#> 289  2017-11-16 15:38:00  4.94          200
#> 290  2017-11-16 15:39:00  4.93          200
#> 291  2017-11-16 15:40:00  4.91          200
#> 292  2017-11-16 15:41:00  4.90          200
#> 293  2017-11-16 15:42:00  4.88          200
#> 294  2017-11-16 15:43:00  4.87          200
#> 295  2017-11-16 15:44:00  4.85          200
#> 296  2017-11-16 15:45:00  4.84          200
#> 297  2017-11-16 15:46:00  4.82          200
#> 298  2017-11-16 15:47:00  4.80          200
#> 299  2017-11-16 15:48:00  4.79          200
#> 300  2017-11-16 15:49:00  4.77          200
#> 301  2017-11-16 15:50:00  4.75          200
#> 302  2017-11-16 15:51:00  4.74          200
#> 303  2017-11-16 15:52:00  4.72          200
#> 304  2017-11-16 15:53:00  4.71          200
#> 305  2017-11-16 15:54:00  4.69          200
#> 306  2017-11-16 15:55:00  4.66          200
#> 307  2017-11-16 15:56:00  4.64          200
#> 308  2017-11-16 15:57:00  4.62          200
#> 309  2017-11-16 15:58:00  4.60          200
#> 310  2017-11-16 15:59:00  4.58          200
#> 311  2017-11-16 16:00:00  4.55          200
#> 312  2017-11-16 16:01:00  4.54          200
#> 313  2017-11-16 16:02:00  4.52          200
#> 314  2017-11-16 16:03:00  4.50          200
#> 315  2017-11-16 16:04:00  4.48          200
#> 316  2017-11-16 16:05:00  4.46          200
#> 317  2017-11-16 16:06:00  4.44          200
#> 318  2017-11-16 16:07:00  4.42          200
#> 319  2017-11-16 16:08:00  4.40          200
#> 320  2017-11-16 16:09:00  4.38          200
#> 321  2017-11-16 16:10:00  4.35          200
#> 322  2017-11-16 16:11:00  4.34          200
#> 323  2017-11-16 16:12:00  4.33          200
#> 324  2017-11-16 16:13:00  4.31          200
#> 325  2017-11-16 16:14:00  4.30          200
#> 326  2017-11-16 16:15:00  4.29          200
#> 327  2017-11-16 16:16:00  4.28          200
#> 328  2017-11-16 16:17:00  4.26          200
#> 329  2017-11-16 16:18:00  4.24          200
#> 330  2017-11-16 16:19:00  4.22          200
#> 331  2017-11-16 16:20:00  4.20          200
#> 332  2017-11-16 16:21:00  4.18          200
#> 333  2017-11-16 16:22:00  4.17          200
#>  [ reached getOption("max.print") -- omitted 1053 rows ]