Download timeseries data from waterinfo.be

Using the ts_id codes and by providing a given date period, download the corresponding time series from the waterinfo.be website

get_timeseries_tsid(ts_id, period = NULL, from = NULL, to = NULL,
  datasource = 1)

Arguments

ts_id
waterinfo.be database ts_id, defining a timeserie variable and frequency it is defined.

period
input string according to format required by waterinfo: De period string is provided as P#Y#M#DT#H#M#S, with P defines `Period`, each # is an integer value and the codes define the number of... Y - years M - months D - days T required if information about sub-day resolution is present H - hours D - days M - minutes S - seconds Instead of D (days), the usage of W - weeks is possible as well Examples of valid period strings: P3D, P1Y, P1DT12H, PT6H, P1Y6M3DT4H20M30S.

from
date of datestring as start of the time series

to
date of datestring as end of the time series

datasource
int [0-3] defines the `meetnet` of which the measurement station is part of. VMM based stations are net '1', MOW-HIC is net '2'

Value

data.frame with the timestamps, values and quality code

Examples

get_timeseries_tsid("35055042", from="2017-01-01", to="2017-01-02")

#>              Timestamp Value Quality Code
#> 1  2017-01-01 00:00:00  0.00          130
#> 2  2017-01-01 00:15:00  0.00          130
#> 3  2017-01-01 00:30:00  0.00          130
#> 4  2017-01-01 00:45:00  0.00          130
#> 5  2017-01-01 01:00:00  0.00          130
#> 6  2017-01-01 01:15:00  0.00          130
#> 7  2017-01-01 01:30:00  0.00          130
#> 8  2017-01-01 01:45:00  0.00          130
#> 9  2017-01-01 02:00:00  0.00          130
#> 10 2017-01-01 02:15:00  0.00          130
#> 11 2017-01-01 02:30:00  0.00          130
#> 12 2017-01-01 02:45:00  0.00          130
#> 13 2017-01-01 03:00:00  0.00          130
#> 14 2017-01-01 03:15:00  0.00          130
#> 15 2017-01-01 03:30:00  0.00          130
#> 16 2017-01-01 03:45:00  0.00          130
#> 17 2017-01-01 04:00:00  0.00          130
#> 18 2017-01-01 04:15:00  0.00          130
#> 19 2017-01-01 04:30:00  0.00          130
#> 20 2017-01-01 04:45:00  0.00          130
#> 21 2017-01-01 05:00:00  0.00          130
#> 22 2017-01-01 05:15:00  0.00          130
#> 23 2017-01-01 05:30:00  0.00          130
#> 24 2017-01-01 05:45:00  0.00          130
#> 25 2017-01-01 06:00:00  0.00          130
#> 26 2017-01-01 06:15:00  0.00          130
#> 27 2017-01-01 06:30:00  0.00          130
#> 28 2017-01-01 06:45:00  0.00          130
#> 29 2017-01-01 07:00:00  0.00          130
#> 30 2017-01-01 07:15:00  0.00          130
#> 31 2017-01-01 07:30:00  0.00          130
#> 32 2017-01-01 07:45:00  0.00          130
#> 33 2017-01-01 08:00:00  0.00          130
#> 34 2017-01-01 08:15:00  0.00          130
#> 35 2017-01-01 08:30:00  0.00          130
#> 36 2017-01-01 08:45:00  0.00          130
#> 37 2017-01-01 09:00:00  0.00          130
#> 38 2017-01-01 09:15:00  0.00          130
#> 39 2017-01-01 09:30:00  0.00          130
#> 40 2017-01-01 09:45:00  0.00          130
#> 41 2017-01-01 10:00:00  0.00          130
#> 42 2017-01-01 10:15:00  0.00          130
#> 43 2017-01-01 10:30:00  0.00          130
#> 44 2017-01-01 10:45:00  0.00          130
#> 45 2017-01-01 11:00:00  0.00          130
#> 46 2017-01-01 11:15:00  0.00          130
#> 47 2017-01-01 11:30:00  0.00          130
#> 48 2017-01-01 11:45:00  0.00          130
#> 49 2017-01-01 12:00:00  0.00          130
#> 50 2017-01-01 12:15:00  0.00          130
#> 51 2017-01-01 12:30:00  0.00          130
#> 52 2017-01-01 12:45:00  0.00          130
#> 53 2017-01-01 13:00:00  0.00          130
#> 54 2017-01-01 13:15:00  0.00          130
#> 55 2017-01-01 13:30:00  0.00          130
#> 56 2017-01-01 13:45:00  0.00          130
#> 57 2017-01-01 14:00:00  0.00          130
#> 58 2017-01-01 14:15:00  0.00          130
#> 59 2017-01-01 14:30:00  0.00          130
#> 60 2017-01-01 14:45:00  0.00          130
#> 61 2017-01-01 15:00:00  0.00          130
#> 62 2017-01-01 15:15:00  0.00          130
#> 63 2017-01-01 15:30:00  0.00          130
#> 64 2017-01-01 15:45:00  0.00          130
#> 65 2017-01-01 16:00:00  0.00          130
#> 66 2017-01-01 16:15:00  0.00          130
#> 67 2017-01-01 16:30:00  0.00          130
#> 68 2017-01-01 16:45:00  0.00          130
#> 69 2017-01-01 17:00:00  0.00          130
#> 70 2017-01-01 17:15:00  0.00          130
#> 71 2017-01-01 17:30:00  0.00          130
#> 72 2017-01-01 17:45:00  0.03          130
#> 73 2017-01-01 18:00:00  0.00          130
#> 74 2017-01-01 18:15:00  0.04          130
#> 75 2017-01-01 18:30:00  0.00          130
#> 76 2017-01-01 18:45:00  0.03          130
#> 77 2017-01-01 19:00:00  0.00          130
#> 78 2017-01-01 19:15:00  0.03          130
#> 79 2017-01-01 19:30:00  0.03          130
#> 80 2017-01-01 19:45:00  0.00          130
#> 81 2017-01-01 20:00:00  0.03          130
#> 82 2017-01-01 20:15:00  0.00          130
#> 83 2017-01-01 20:30:00  0.00          130
#> 84 2017-01-01 20:45:00  0.00          130
#> 85 2017-01-01 21:00:00  0.00          130
#> 86 2017-01-01 21:15:00  0.03          130
#> 87 2017-01-01 21:30:00  0.00          130
#> 88 2017-01-01 21:45:00  0.00          130
#> 89 2017-01-01 22:00:00  0.00          130
#> 90 2017-01-01 22:15:00  0.11          130
#> 91 2017-01-01 22:30:00  0.82          130
#> 92 2017-01-01 22:45:00  0.69          130
#> 93 2017-01-01 23:00:00  0.34          130
#> 94 2017-01-01 23:15:00  0.13          130
#> 95 2017-01-01 23:30:00  0.03          130
#> 96 2017-01-01 23:45:00  0.03          130
#> 97 2017-01-02 00:00:00  0.05          130

get_timeseries_tsid("5156042", period = "P3D")

#>               Timestamp Value Quality Code
#> 1   2017-10-18 01:00:00 0.233          110
#> 2   2017-10-18 01:15:00 0.233          110
#> 3   2017-10-18 01:30:00 0.233          110
#> 4   2017-10-18 01:45:00 0.233          110
#> 5   2017-10-18 02:00:00 0.233          110
#> 6   2017-10-18 02:15:00 0.233          110
#> 7   2017-10-18 02:30:00 0.234          110
#> 8   2017-10-18 02:45:00 0.234          110
#> 9   2017-10-18 03:00:00 0.234          110
#> 10  2017-10-18 03:15:00 0.234          110
#> 11  2017-10-18 03:30:00 0.234          110
#> 12  2017-10-18 03:45:00 0.234          110
#> 13  2017-10-18 04:00:00 0.234          110
#> 14  2017-10-18 04:15:00 0.235          110
#> 15  2017-10-18 04:30:00 0.234          110
#> 16  2017-10-18 04:45:00 0.234          110
#> 17  2017-10-18 05:00:00 0.235          110
#> 18  2017-10-18 05:15:00 0.235          110
#> 19  2017-10-18 05:30:00 0.235          110
#> 20  2017-10-18 05:45:00 0.235          110
#> 21  2017-10-18 06:00:00 0.236          110
#> 22  2017-10-18 06:15:00 0.236          110
#> 23  2017-10-18 06:30:00 0.235          110
#> 24  2017-10-18 06:45:00 0.236          110
#> 25  2017-10-18 07:00:00 0.236          110
#> 26  2017-10-18 07:15:00 0.236          110
#> 27  2017-10-18 07:30:00 0.236          110
#> 28  2017-10-18 07:45:00 0.236          110
#> 29  2017-10-18 08:00:00 0.236          110
#> 30  2017-10-18 08:15:00 0.236          110
#> 31  2017-10-18 08:30:00 0.236          110
#> 32  2017-10-18 08:45:00 0.235          110
#> 33  2017-10-18 09:00:00 0.235          110
#> 34  2017-10-18 09:15:00 0.235          110
#> 35  2017-10-18 09:30:00 0.235          110
#> 36  2017-10-18 09:45:00 0.234          110
#> 37  2017-10-18 10:00:00 0.234          110
#> 38  2017-10-18 10:15:00 0.234          110
#> 39  2017-10-18 10:30:00 0.234          110
#> 40  2017-10-18 10:45:00 0.234          110
#> 41  2017-10-18 11:00:00 0.234          110
#> 42  2017-10-18 11:15:00 0.233          110
#> 43  2017-10-18 11:30:00 0.233          110
#> 44  2017-10-18 11:45:00 0.233          110
#> 45  2017-10-18 12:00:00 0.233          110
#> 46  2017-10-18 12:15:00 0.233          110
#> 47  2017-10-18 12:30:00 0.233          110
#> 48  2017-10-18 12:45:00 0.233          110
#> 49  2017-10-18 13:00:00 0.234          110
#> 50  2017-10-18 13:15:00 0.234          110
#> 51  2017-10-18 13:30:00 0.234          110
#> 52  2017-10-18 13:45:00 0.235          110
#> 53  2017-10-18 14:00:00 0.235          110
#> 54  2017-10-18 14:15:00 0.235          110
#> 55  2017-10-18 14:30:00 0.235          110
#> 56  2017-10-18 14:45:00 0.235          110
#> 57  2017-10-18 15:00:00 0.235          110
#> 58  2017-10-18 15:15:00 0.234          110
#> 59  2017-10-18 15:30:00 0.234          110
#> 60  2017-10-18 15:45:00 0.233          110
#> 61  2017-10-18 16:00:00 0.233          110
#> 62  2017-10-18 16:15:00 0.233          110
#> 63  2017-10-18 16:30:00 0.232          110
#> 64  2017-10-18 16:45:00 0.231          110
#> 65  2017-10-18 17:00:00 0.230          110
#> 66  2017-10-18 17:15:00 0.231          110
#> 67  2017-10-18 17:30:00 0.230          110
#> 68  2017-10-18 17:45:00 0.230          110
#> 69  2017-10-18 18:00:00 0.230          110
#> 70  2017-10-18 18:15:00 0.230          110
#> 71  2017-10-18 18:30:00 0.230          110
#> 72  2017-10-18 18:45:00 0.231          110
#> 73  2017-10-18 19:00:00 0.231          110
#> 74  2017-10-18 19:15:00 0.231          110
#> 75  2017-10-18 19:30:00 0.231          110
#> 76  2017-10-18 19:45:00 0.232          110
#> 77  2017-10-18 20:00:00 0.232          110
#> 78  2017-10-18 20:15:00 0.232          110
#> 79  2017-10-18 20:30:00 0.233          110
#> 80  2017-10-18 20:45:00 0.234          110
#> 81  2017-10-18 21:00:00 0.234          110
#> 82  2017-10-18 21:15:00 0.235          110
#> 83  2017-10-18 21:30:00 0.236          110
#> 84  2017-10-18 21:45:00 0.236          110
#> 85  2017-10-18 22:00:00 0.237          110
#> 86  2017-10-18 22:15:00 0.238          110
#> 87  2017-10-18 22:30:00 0.238          110
#> 88  2017-10-18 22:45:00 0.239          110
#> 89  2017-10-18 23:00:00 0.240          110
#> 90  2017-10-18 23:15:00 0.240          110
#> 91  2017-10-18 23:30:00 0.241          110
#> 92  2017-10-18 23:45:00 0.241          110
#> 93  2017-10-19 00:00:00 0.242          110
#> 94  2017-10-19 00:15:00 0.242          110
#> 95  2017-10-19 00:30:00 0.243          110
#> 96  2017-10-19 00:45:00 0.243          110
#> 97  2017-10-19 01:00:00 0.243          110
#> 98  2017-10-19 01:15:00 0.243          110
#> 99  2017-10-19 01:30:00 0.243          110
#> 100 2017-10-19 01:45:00 0.243          110
#> 101 2017-10-19 02:00:00 0.243          110
#> 102 2017-10-19 02:15:00 0.243          110
#> 103 2017-10-19 02:30:00 0.243          110
#> 104 2017-10-19 02:45:00 0.243          110
#> 105 2017-10-19 03:00:00 0.242          110
#> 106 2017-10-19 03:15:00 0.242          110
#> 107 2017-10-19 03:30:00 0.243          110
#> 108 2017-10-19 03:45:00 0.242          110
#> 109 2017-10-19 04:00:00 0.242          110
#> 110 2017-10-19 04:15:00 0.242          110
#> 111 2017-10-19 04:30:00 0.242          110
#> 112 2017-10-19 04:45:00 0.242          110
#> 113 2017-10-19 05:00:00 0.242          110
#> 114 2017-10-19 05:15:00 0.242          110
#> 115 2017-10-19 05:30:00 0.242          110
#> 116 2017-10-19 05:45:00 0.242          110
#> 117 2017-10-19 06:00:00 0.242          110
#> 118 2017-10-19 06:15:00 0.242          110
#> 119 2017-10-19 06:30:00 0.241          110
#> 120 2017-10-19 06:45:00 0.241          110
#> 121 2017-10-19 07:00:00 0.241          110
#> 122 2017-10-19 07:15:00 0.241          110
#> 123 2017-10-19 07:30:00 0.240          110
#> 124 2017-10-19 07:45:00 0.240          110
#> 125 2017-10-19 08:00:00 0.240          110
#> 126 2017-10-19 08:15:00 0.239          110
#> 127 2017-10-19 08:30:00 0.240          110
#> 128 2017-10-19 08:45:00 0.240          110
#> 129 2017-10-19 09:00:00 0.239          110
#> 130 2017-10-19 09:15:00 0.239          110
#> 131 2017-10-19 09:30:00 0.239          110
#> 132 2017-10-19 09:45:00 0.239          110
#> 133 2017-10-19 10:00:00 0.238          110
#> 134 2017-10-19 10:15:00 0.238          110
#> 135 2017-10-19 10:30:00 0.238          110
#> 136 2017-10-19 10:45:00 0.237          110
#> 137 2017-10-19 11:00:00 0.237          110
#> 138 2017-10-19 11:15:00 0.237          110
#> 139 2017-10-19 11:30:00 0.237          110
#> 140 2017-10-19 11:45:00 0.236          110
#> 141 2017-10-19 12:00:00 0.237          110
#> 142 2017-10-19 12:15:00 0.237          110
#> 143 2017-10-19 12:30:00 0.237          110
#> 144 2017-10-19 12:45:00 0.237          110
#> 145 2017-10-19 13:00:00 0.237          110
#> 146 2017-10-19 13:15:00 0.237          110
#> 147 2017-10-19 13:30:00 0.237          110
#> 148 2017-10-19 13:45:00 0.237          110
#> 149 2017-10-19 14:00:00 0.237          110
#> 150 2017-10-19 14:15:00 0.237          110
#> 151 2017-10-19 14:30:00 0.237          110
#> 152 2017-10-19 14:45:00 0.237          110
#> 153 2017-10-19 15:00:00 0.237          110
#> 154 2017-10-19 15:15:00 0.237          110
#> 155 2017-10-19 15:30:00 0.237          110
#> 156 2017-10-19 15:45:00 0.237          110
#> 157 2017-10-19 16:00:00 0.237          110
#> 158 2017-10-19 16:15:00 0.237          110
#> 159 2017-10-19 16:30:00 0.236          110
#> 160 2017-10-19 16:45:00 0.236          110
#> 161 2017-10-19 17:00:00 0.235          110
#> 162 2017-10-19 17:15:00 0.235          110
#> 163 2017-10-19 17:30:00 0.235          110
#> 164 2017-10-19 17:45:00 0.235          110
#> 165 2017-10-19 18:00:00 0.235          110
#> 166 2017-10-19 18:15:00 0.235          110
#> 167 2017-10-19 18:30:00 0.234          110
#> 168 2017-10-19 18:45:00 0.234          110
#> 169 2017-10-19 19:00:00 0.234          110
#> 170 2017-10-19 19:15:00 0.234          110
#> 171 2017-10-19 19:30:00 0.234          110
#> 172 2017-10-19 19:45:00 0.234          110
#> 173 2017-10-19 20:00:00 0.234          110
#> 174 2017-10-19 20:15:00 0.234          110
#> 175 2017-10-19 20:30:00 0.234          110
#> 176 2017-10-19 20:45:00 0.234          110
#> 177 2017-10-19 21:00:00 0.234          110
#> 178 2017-10-19 21:15:00 0.234          110
#> 179 2017-10-19 21:30:00 0.234          110
#> 180 2017-10-19 21:45:00 0.234          110
#> 181 2017-10-19 22:00:00 0.234          110
#> 182 2017-10-19 22:15:00 0.234          110
#> 183 2017-10-19 22:30:00 0.234          110
#> 184 2017-10-19 22:45:00 0.234          110
#> 185 2017-10-19 23:00:00 0.234          110
#> 186 2017-10-19 23:15:00 0.234          110
#> 187 2017-10-19 23:30:00 0.234          110
#> 188 2017-10-19 23:45:00 0.234          110
#> 189 2017-10-20 00:00:00 0.234          110
#> 190 2017-10-20 00:15:00 0.234          110
#> 191 2017-10-20 00:30:00 0.234          110
#> 192 2017-10-20 00:45:00 0.234          110
#> 193 2017-10-20 01:00:00 0.234          110
#> 194 2017-10-20 01:15:00 0.234          110
#> 195 2017-10-20 01:30:00 0.234          110
#> 196 2017-10-20 01:45:00 0.234          110
#> 197 2017-10-20 02:00:00 0.234          110
#> 198 2017-10-20 02:15:00 0.234          110
#> 199 2017-10-20 02:30:00 0.234          110
#> 200 2017-10-20 02:45:00 0.234          110
#> 201 2017-10-20 03:00:00 0.235          110
#> 202 2017-10-20 03:15:00 0.235          110
#> 203 2017-10-20 03:30:00 0.235          110
#> 204 2017-10-20 03:45:00 0.235          110
#> 205 2017-10-20 04:00:00 0.235          110
#> 206 2017-10-20 04:15:00 0.235          110
#> 207 2017-10-20 04:30:00 0.235          110
#> 208 2017-10-20 04:45:00 0.235          110
#> 209 2017-10-20 05:00:00 0.235          110
#> 210 2017-10-20 05:15:00 0.236          110
#> 211 2017-10-20 05:30:00 0.236          110
#> 212 2017-10-20 05:45:00 0.236          110
#> 213 2017-10-20 06:00:00 0.236          110
#> 214 2017-10-20 06:15:00 0.237          110
#> 215 2017-10-20 06:30:00 0.236          110
#> 216 2017-10-20 06:45:00 0.236          110
#> 217 2017-10-20 07:00:00 0.236          110
#> 218 2017-10-20 07:15:00 0.236          110
#> 219 2017-10-20 07:30:00 0.236          110
#> 220 2017-10-20 07:45:00 0.235          110
#> 221 2017-10-20 08:00:00 0.235          110
#> 222 2017-10-20 08:15:00 0.235          110
#> 223 2017-10-20 08:30:00 0.235          110
#> 224 2017-10-20 08:45:00 0.236          110
#> 225 2017-10-20 09:00:00 0.236          110
#> 226 2017-10-20 09:15:00 0.235          110
#> 227 2017-10-20 09:30:00 0.235          110
#> 228 2017-10-20 09:45:00 0.234          110
#> 229 2017-10-20 10:00:00 0.234          110
#> 230 2017-10-20 10:15:00 0.234          110
#> 231 2017-10-20 10:30:00 0.233          110
#> 232 2017-10-20 10:45:00 0.233          110
#> 233 2017-10-20 11:00:00 0.233          110
#> 234 2017-10-20 11:15:00 0.232          110
#> 235 2017-10-20 11:30:00 0.232          110
#> 236 2017-10-20 11:45:00 0.231          110
#> 237 2017-10-20 12:00:00 0.231          110
#> 238 2017-10-20 12:15:00 0.231          110
#> 239 2017-10-20 12:30:00 0.230          110
#> 240 2017-10-20 12:45:00 0.230          110
#> 241 2017-10-20 13:00:00 0.230          110
#> 242 2017-10-20 13:15:00 0.230          110
#> 243 2017-10-20 13:30:00 0.229          110
#> 244 2017-10-20 13:45:00 0.229          110
#> 245 2017-10-20 14:00:00 0.229          110
#> 246 2017-10-20 14:15:00 0.229          110
#> 247 2017-10-20 14:30:00 0.229          110
#> 248 2017-10-20 14:45:00 0.230          110
#> 249 2017-10-20 15:00:00 0.230          110
#> 250 2017-10-20 15:15:00 0.230          110
#> 251 2017-10-20 15:30:00 0.230          110
#> 252 2017-10-20 15:45:00 0.230          110
#> 253 2017-10-20 16:00:00 0.230          110
#> 254 2017-10-20 16:15:00 0.230          110
#> 255 2017-10-20 16:30:00 0.231          110
#> 256 2017-10-20 16:45:00 0.231          110
#> 257 2017-10-20 17:00:00 0.231          110
#> 258 2017-10-20 17:15:00 0.231          110
#> 259 2017-10-20 17:30:00 0.231          110
#> 260 2017-10-20 17:45:00 0.231          110
#> 261 2017-10-20 18:00:00 0.231          110
#> 262 2017-10-20 18:15:00 0.231          110
#> 263 2017-10-20 18:30:00 0.231          110
#> 264 2017-10-20 18:45:00 0.231          110
#> 265 2017-10-20 19:00:00 0.232          110
#> 266 2017-10-20 19:15:00 0.232          110
#> 267 2017-10-20 19:30:00 0.232          110
#> 268 2017-10-20 19:45:00 0.232          110
#> 269 2017-10-20 20:00:00 0.233          110
#> 270 2017-10-20 20:15:00 0.233          110
#> 271 2017-10-20 20:30:00 0.233          110
#> 272 2017-10-20 20:45:00 0.233          110
#> 273 2017-10-20 21:00:00 0.233          110
#> 274 2017-10-20 21:15:00 0.234          110
#> 275 2017-10-20 21:30:00 0.234          110
#> 276 2017-10-20 21:45:00 0.234          110
#> 277 2017-10-20 22:00:00 0.235          110

get_timeseries_tsid("2813562", period = "P1D", datasource = 2)

#>                Timestamp Value Quality Code
#> 1    2017-10-20 00:47:00  1.84          200
#> 2    2017-10-20 00:48:00  1.85          200
#> 3    2017-10-20 00:49:00  1.86          200
#> 4    2017-10-20 00:50:00  1.87          200
#> 5    2017-10-20 00:51:00  1.88          200
#> 6    2017-10-20 00:52:00  1.90          200
#> 7    2017-10-20 00:53:00  1.91          200
#> 8    2017-10-20 00:54:00  1.93          200
#> 9    2017-10-20 00:55:00  1.94          200
#> 10   2017-10-20 00:56:00  1.96          200
#> 11   2017-10-20 00:57:00  1.97          200
#> 12   2017-10-20 00:58:00  1.98          200
#> 13   2017-10-20 00:59:00  2.00          200
#> 14   2017-10-20 01:00:00  2.01          200
#> 15   2017-10-20 01:01:00  2.02          200
#> 16   2017-10-20 01:02:00  2.03          200
#> 17   2017-10-20 01:03:00  2.04          200
#> 18   2017-10-20 01:04:00  2.05          200
#> 19   2017-10-20 01:05:00  2.06          200
#> 20   2017-10-20 01:06:00  2.06          200
#> 21   2017-10-20 01:07:00  2.08          200
#> 22   2017-10-20 01:08:00  2.09          200
#> 23   2017-10-20 01:09:00  2.10          200
#> 24   2017-10-20 01:10:00  2.11          200
#> 25   2017-10-20 01:11:00  2.13          200
#> 26   2017-10-20 01:12:00  2.14          200
#> 27   2017-10-20 01:13:00  2.16          200
#> 28   2017-10-20 01:14:00  2.17          200
#> 29   2017-10-20 01:15:00  2.18          200
#> 30   2017-10-20 01:16:00  2.20          200
#> 31   2017-10-20 01:17:00  2.21          200
#> 32   2017-10-20 01:18:00  2.22          200
#> 33   2017-10-20 01:19:00  2.23          200
#> 34   2017-10-20 01:20:00  2.24          200
#> 35   2017-10-20 01:21:00  2.26          200
#> 36   2017-10-20 01:22:00  2.27          200
#> 37   2017-10-20 01:23:00  2.28          200
#> 38   2017-10-20 01:24:00  2.29          200
#> 39   2017-10-20 01:25:00  2.30          200
#> 40   2017-10-20 01:26:00  2.31          200
#> 41   2017-10-20 01:27:00  2.32          200
#> 42   2017-10-20 01:28:00  2.33          200
#> 43   2017-10-20 01:29:00  2.35          200
#> 44   2017-10-20 01:30:00  2.36          200
#> 45   2017-10-20 01:31:00  2.37          200
#> 46   2017-10-20 01:32:00  2.38          200
#> 47   2017-10-20 01:33:00  2.39          200
#> 48   2017-10-20 01:34:00  2.41          200
#> 49   2017-10-20 01:35:00  2.42          200
#> 50   2017-10-20 01:36:00  2.44          200
#> 51   2017-10-20 01:37:00  2.45          200
#> 52   2017-10-20 01:38:00  2.46          200
#> 53   2017-10-20 01:39:00  2.48          200
#> 54   2017-10-20 01:40:00  2.49          200
#> 55   2017-10-20 01:41:00  2.50          200
#> 56   2017-10-20 01:42:00  2.51          200
#> 57   2017-10-20 01:43:00  2.53          200
#> 58   2017-10-20 01:44:00  2.54          200
#> 59   2017-10-20 01:45:00  2.55          200
#> 60   2017-10-20 01:46:00  2.57          200
#> 61   2017-10-20 01:47:00  2.58          200
#> 62   2017-10-20 01:48:00  2.60          200
#> 63   2017-10-20 01:49:00  2.61          200
#> 64   2017-10-20 01:50:00  2.62          200
#> 65   2017-10-20 01:51:00  2.63          200
#> 66   2017-10-20 01:52:00  2.65          200
#> 67   2017-10-20 01:53:00  2.66          200
#> 68   2017-10-20 01:54:00  2.67          200
#> 69   2017-10-20 01:55:00  2.68          200
#> 70   2017-10-20 01:56:00  2.70          200
#> 71   2017-10-20 01:57:00  2.72          200
#> 72   2017-10-20 01:58:00  2.73          200
#> 73   2017-10-20 01:59:00  2.75          200
#> 74   2017-10-20 02:00:00  2.76          200
#> 75   2017-10-20 02:01:00  2.77          200
#> 76   2017-10-20 02:02:00  2.78          200
#> 77   2017-10-20 02:03:00  2.80          200
#> 78   2017-10-20 02:04:00  2.82          200
#> 79   2017-10-20 02:05:00  2.83          200
#> 80   2017-10-20 02:06:00  2.84          200
#> 81   2017-10-20 02:07:00  2.86          200
#> 82   2017-10-20 02:08:00  2.88          200
#> 83   2017-10-20 02:09:00  2.90          200
#> 84   2017-10-20 02:10:00  2.91          200
#> 85   2017-10-20 02:11:00  2.92          200
#> 86   2017-10-20 02:12:00  2.94          200
#> 87   2017-10-20 02:13:00  2.96          200
#> 88   2017-10-20 02:14:00  2.98          200
#> 89   2017-10-20 02:15:00  3.00          200
#> 90   2017-10-20 02:16:00  3.02          200
#> 91   2017-10-20 02:17:00  3.03          200
#> 92   2017-10-20 02:18:00  3.05          200
#> 93   2017-10-20 02:19:00  3.07          200
#> 94   2017-10-20 02:20:00  3.09          200
#> 95   2017-10-20 02:21:00  3.11          200
#> 96   2017-10-20 02:22:00  3.13          200
#> 97   2017-10-20 02:23:00  3.15          200
#> 98   2017-10-20 02:24:00  3.17          200
#> 99   2017-10-20 02:25:00  3.19          200
#> 100  2017-10-20 02:26:00  3.21          200
#> 101  2017-10-20 02:27:00  3.23          200
#> 102  2017-10-20 02:28:00  3.26          200
#> 103  2017-10-20 02:29:00  3.28          200
#> 104  2017-10-20 02:30:00  3.30          200
#> 105  2017-10-20 02:31:00  3.32          200
#> 106  2017-10-20 02:32:00  3.35          200
#> 107  2017-10-20 02:33:00  3.37          200
#> 108  2017-10-20 02:34:00  3.40          200
#> 109  2017-10-20 02:35:00  3.42          200
#> 110  2017-10-20 02:36:00  3.45          200
#> 111  2017-10-20 02:37:00  3.48          200
#> 112  2017-10-20 02:38:00  3.50          200
#> 113  2017-10-20 02:39:00  3.52          200
#> 114  2017-10-20 02:40:00  3.55          200
#> 115  2017-10-20 02:41:00  3.58          200
#> 116  2017-10-20 02:42:00  3.60          200
#> 117  2017-10-20 02:43:00  3.63          200
#> 118  2017-10-20 02:44:00  3.66          200
#> 119  2017-10-20 02:45:00  3.69          200
#> 120  2017-10-20 02:46:00  3.73          200
#> 121  2017-10-20 02:47:00  3.77          200
#> 122  2017-10-20 02:48:00  3.80          200
#> 123  2017-10-20 02:49:00  3.83          200
#> 124  2017-10-20 02:50:00  3.86          200
#> 125  2017-10-20 02:51:00  3.89          200
#> 126  2017-10-20 02:52:00  3.92          200
#> 127  2017-10-20 02:53:00  3.95          200
#> 128  2017-10-20 02:54:00  3.98          200
#> 129  2017-10-20 02:55:00  4.02          200
#> 130  2017-10-20 02:56:00  4.05          200
#> 131  2017-10-20 02:57:00  4.09          200
#> 132  2017-10-20 02:58:00  4.13          200
#> 133  2017-10-20 02:59:00  4.15          200
#> 134  2017-10-20 03:00:00  4.19          200
#> 135  2017-10-20 03:01:00  4.21          200
#> 136  2017-10-20 03:02:00  4.24          200
#> 137  2017-10-20 03:03:00  4.27          200
#> 138  2017-10-20 03:04:00  4.31          200
#> 139  2017-10-20 03:05:00  4.34          200
#> 140  2017-10-20 03:06:00  4.37          200
#> 141  2017-10-20 03:07:00  4.40          200
#> 142  2017-10-20 03:08:00  4.44          200
#> 143  2017-10-20 03:09:00  4.47          200
#> 144  2017-10-20 03:10:00  4.49          200
#> 145  2017-10-20 03:11:00  4.52          200
#> 146  2017-10-20 03:12:00  4.56          200
#> 147  2017-10-20 03:13:00  4.58          200
#> 148  2017-10-20 03:14:00  4.61          200
#> 149  2017-10-20 03:15:00  4.63          200
#> 150  2017-10-20 03:16:00  4.66          200
#> 151  2017-10-20 03:17:00  4.70          200
#> 152  2017-10-20 03:18:00  4.72          200
#> 153  2017-10-20 03:19:00  4.75          200
#> 154  2017-10-20 03:20:00  4.77          200
#> 155  2017-10-20 03:21:00  4.80          200
#> 156  2017-10-20 03:22:00  4.83          200
#> 157  2017-10-20 03:23:00  4.85          200
#> 158  2017-10-20 03:24:00  4.87          200
#> 159  2017-10-20 03:25:00  4.89          200
#> 160  2017-10-20 03:26:00  4.91          200
#> 161  2017-10-20 03:27:00  4.94          200
#> 162  2017-10-20 03:28:00  4.96          200
#> 163  2017-10-20 03:29:00  4.98          200
#> 164  2017-10-20 03:30:00  5.00          200
#> 165  2017-10-20 03:31:00  5.02          200
#> 166  2017-10-20 03:32:00  5.04          200
#> 167  2017-10-20 03:33:00  5.07          200
#> 168  2017-10-20 03:34:00  5.08          200
#> 169  2017-10-20 03:35:00  5.10          200
#> 170  2017-10-20 03:36:00  5.12          200
#> 171  2017-10-20 03:37:00  5.14          200
#> 172  2017-10-20 03:38:00  5.16          200
#> 173  2017-10-20 03:39:00  5.17          200
#> 174  2017-10-20 03:40:00  5.19          200
#> 175  2017-10-20 03:41:00  5.21          200
#> 176  2017-10-20 03:42:00  5.23          200
#> 177  2017-10-20 03:43:00  5.24          200
#> 178  2017-10-20 03:44:00  5.25          200
#> 179  2017-10-20 03:45:00  5.26          200
#> 180  2017-10-20 03:46:00  5.28          200
#> 181  2017-10-20 03:47:00  5.29          200
#> 182  2017-10-20 03:48:00  5.30          200
#> 183  2017-10-20 03:49:00  5.31          200
#> 184  2017-10-20 03:50:00  5.33          200
#> 185  2017-10-20 03:51:00  5.34          200
#> 186  2017-10-20 03:52:00  5.35          200
#> 187  2017-10-20 03:53:00  5.36          200
#> 188  2017-10-20 03:54:00  5.38          200
#> 189  2017-10-20 03:55:00  5.39          200
#> 190  2017-10-20 03:56:00  5.40          200
#> 191  2017-10-20 03:57:00  5.40          200
#> 192  2017-10-20 03:58:00  5.41          200
#> 193  2017-10-20 03:59:00  5.42          200
#> 194  2017-10-20 04:00:00  5.43          200
#> 195  2017-10-20 04:01:00  5.44          200
#> 196  2017-10-20 04:02:00  5.44          200
#> 197  2017-10-20 04:03:00  5.45          200
#> 198  2017-10-20 04:04:00  5.46          200
#> 199  2017-10-20 04:05:00  5.47          200
#> 200  2017-10-20 04:06:00  5.48          200
#> 201  2017-10-20 04:07:00  5.48          200
#> 202  2017-10-20 04:08:00  5.49          200
#> 203  2017-10-20 04:09:00  5.49          200
#> 204  2017-10-20 04:10:00  5.50          200
#> 205  2017-10-20 04:11:00  5.51          200
#> 206  2017-10-20 04:12:00  5.50          200
#> 207  2017-10-20 04:13:00  5.50          200
#> 208  2017-10-20 04:14:00  5.51          200
#> 209  2017-10-20 04:15:00  5.51          200
#> 210  2017-10-20 04:16:00  5.51          200
#> 211  2017-10-20 04:17:00  5.52          200
#> 212  2017-10-20 04:18:00  5.52          200
#> 213  2017-10-20 04:19:00  5.53          200
#> 214  2017-10-20 04:20:00  5.53          200
#> 215  2017-10-20 04:21:00  5.53          200
#> 216  2017-10-20 04:22:00  5.54          200
#> 217  2017-10-20 04:23:00  5.54          200
#> 218  2017-10-20 04:24:00  5.54          200
#> 219  2017-10-20 04:25:00  5.55          200
#> 220  2017-10-20 04:26:00  5.55          200
#> 221  2017-10-20 04:27:00  5.54          200
#> 222  2017-10-20 04:28:00  5.54          200
#> 223  2017-10-20 04:29:00  5.55          200
#> 224  2017-10-20 04:30:00  5.55          200
#> 225  2017-10-20 04:31:00  5.55          200
#> 226  2017-10-20 04:32:00  5.55          200
#> 227  2017-10-20 04:33:00  5.55          200
#> 228  2017-10-20 04:34:00  5.55          200
#> 229  2017-10-20 04:35:00  5.55          200
#> 230  2017-10-20 04:36:00  5.55          200
#> 231  2017-10-20 04:37:00  5.55          200
#> 232  2017-10-20 04:38:00  5.55          200
#> 233  2017-10-20 04:39:00  5.54          200
#> 234  2017-10-20 04:40:00  5.54          200
#> 235  2017-10-20 04:41:00  5.54          200
#> 236  2017-10-20 04:42:00  5.54          200
#> 237  2017-10-20 04:43:00  5.53          200
#> 238  2017-10-20 04:44:00  5.52          200
#> 239  2017-10-20 04:45:00  5.52          200
#> 240  2017-10-20 04:46:00  5.51          200
#> 241  2017-10-20 04:47:00  5.51          200
#> 242  2017-10-20 04:48:00  5.50          200
#> 243  2017-10-20 04:49:00  5.49          200
#> 244  2017-10-20 04:50:00  5.48          200
#> 245  2017-10-20 04:51:00  5.48          200
#> 246  2017-10-20 04:52:00  5.47          200
#> 247  2017-10-20 04:53:00  5.46          200
#> 248  2017-10-20 04:54:00  5.45          200
#> 249  2017-10-20 04:55:00  5.44          200
#> 250  2017-10-20 04:56:00  5.42          200
#> 251  2017-10-20 04:57:00  5.41          200
#> 252  2017-10-20 04:58:00  5.39          200
#> 253  2017-10-20 04:59:00  5.38          200
#> 254  2017-10-20 05:00:00  5.37          200
#> 255  2017-10-20 05:01:00  5.36          200
#> 256  2017-10-20 05:02:00  5.34          200
#> 257  2017-10-20 05:03:00  5.33          200
#> 258  2017-10-20 05:04:00  5.31          200
#> 259  2017-10-20 05:05:00  5.30          200
#> 260  2017-10-20 05:06:00  5.29          200
#> 261  2017-10-20 05:07:00  5.27          200
#> 262  2017-10-20 05:08:00  5.25          200
#> 263  2017-10-20 05:09:00  5.24          200
#> 264  2017-10-20 05:10:00  5.22          200
#> 265  2017-10-20 05:11:00  5.20          200
#> 266  2017-10-20 05:12:00  5.19          200
#> 267  2017-10-20 05:13:00  5.17          200
#> 268  2017-10-20 05:14:00  5.15          200
#> 269  2017-10-20 05:15:00  5.13          200
#> 270  2017-10-20 05:16:00  5.11          200
#> 271  2017-10-20 05:17:00  5.09          200
#> 272  2017-10-20 05:18:00  5.07          200
#> 273  2017-10-20 05:19:00  5.05          200
#> 274  2017-10-20 05:20:00  5.03          200
#> 275  2017-10-20 05:21:00  5.02          200
#> 276  2017-10-20 05:22:00  5.00          200
#> 277  2017-10-20 05:23:00  4.98          200
#> 278  2017-10-20 05:24:00  4.96          200
#> 279  2017-10-20 05:25:00  4.95          200
#> 280  2017-10-20 05:26:00  4.93          200
#> 281  2017-10-20 05:27:00  4.91          200
#> 282  2017-10-20 05:28:00  4.89          200
#> 283  2017-10-20 05:29:00  4.87          200
#> 284  2017-10-20 05:30:00  4.86          200
#> 285  2017-10-20 05:31:00  4.84          200
#> 286  2017-10-20 05:32:00  4.82          200
#> 287  2017-10-20 05:33:00  4.81          200
#> 288  2017-10-20 05:34:00  4.79          200
#> 289  2017-10-20 05:35:00  4.77          200
#> 290  2017-10-20 05:36:00  4.75          200
#> 291  2017-10-20 05:37:00  4.74          200
#> 292  2017-10-20 05:38:00  4.72          200
#> 293  2017-10-20 05:39:00  4.70          200
#> 294  2017-10-20 05:40:00  4.68          200
#> 295  2017-10-20 05:41:00  4.66          200
#> 296  2017-10-20 05:42:00  4.64          200
#> 297  2017-10-20 05:43:00  4.62          200
#> 298  2017-10-20 05:44:00  4.60          200
#> 299  2017-10-20 05:45:00  4.59          200
#> 300  2017-10-20 05:46:00  4.57          200
#> 301  2017-10-20 05:47:00  4.56          200
#> 302  2017-10-20 05:48:00  4.54          200
#> 303  2017-10-20 05:49:00  4.52          200
#> 304  2017-10-20 05:50:00  4.50          200
#> 305  2017-10-20 05:51:00  4.49          200
#> 306  2017-10-20 05:52:00  4.47          200
#> 307  2017-10-20 05:53:00  4.46          200
#> 308  2017-10-20 05:54:00  4.44          200
#> 309  2017-10-20 05:55:00  4.42          200
#> 310  2017-10-20 05:56:00  4.40          200
#> 311  2017-10-20 05:57:00  4.39          200
#> 312  2017-10-20 05:58:00  4.38          200
#> 313  2017-10-20 05:59:00  4.36          200
#> 314  2017-10-20 06:00:00  4.34          200
#> 315  2017-10-20 06:01:00  4.33          200
#> 316  2017-10-20 06:02:00  4.31          200
#> 317  2017-10-20 06:03:00  4.30          200
#> 318  2017-10-20 06:04:00  4.28          200
#> 319  2017-10-20 06:05:00  4.26          200
#> 320  2017-10-20 06:06:00  4.24          200
#> 321  2017-10-20 06:07:00  4.24          200
#> 322  2017-10-20 06:08:00  4.22          200
#> 323  2017-10-20 06:09:00  4.20          200
#> 324  2017-10-20 06:10:00  4.19          200
#> 325  2017-10-20 06:11:00  4.17          200
#> 326  2017-10-20 06:12:00  4.16          200
#> 327  2017-10-20 06:13:00  4.15          200
#> 328  2017-10-20 06:14:00  4.13          200
#> 329  2017-10-20 06:15:00  4.12          200
#> 330  2017-10-20 06:16:00  4.11          200
#> 331  2017-10-20 06:17:00  4.10          200
#> 332  2017-10-20 06:18:00  4.08          200
#> 333  2017-10-20 06:19:00  4.07          200
#>  [ reached getOption("max.print") -- omitted 1066 rows ]