Get waterinfo Token

Retrieve a fresh waterinfo token.

get_token(client = NULL, client_id = NULL, client_secret = NULL,
  token_url = "http://download.waterinfo.be/kiwis-auth/token")

is.expired(token)

expires.in(token)

Arguments

client
base64 encoded client containing the client id and client secret, seperated by :

client_id
client id string

client_secret
client secret string

token_url
url to get the token from

token
a token object

Value

An object of class token containing the token string with the token_url, token type and moment of expiration as attributes.

Details

Either client or client_id and client_secret need to be passed as arguments. If provided, client is always used. Tokens remain valid for 24 hours, after which a fresh one must be acquired. To limit load on the server, token objects should be reused as much as possible until expiration in stead of creating fresh ones for each call. The client_id and client_secret provided in the examples are for test purposes, get your very own client via hydrometrie@waterinfo.be!

Examples

# Get token via client_id and client_secret
client_id <- '32dceece-826c-4b98-9c2d-b16978f0ca6a'
client_secret <- '4ada871a-9528-4b4d-bfd5-b570ce8f4d2d'
my_token <- get_token(client_id = client_id,client_secret = client_secret)
print(my_token)

#> Token:
#> eyJhbGciOiJIUzI1NiJ9.eyJqdGkiOiIwNzBiZGVkMC1iYzFjLTQ2OWEtYTc3MS1kMDU4Nzc5Nzc3NzEiLCJpYXQiOjE1MzgzOTU3NjcsImlzcyI6Imh0dHA6Ly9sb2NhbGhvc3Q6ODA4MC9LaVdlYlBvcnRhbC9hdXRoIiwiYXVkIjoiMzJkY2VlY2UtODI2Yy00Yjk4LTljMmQtYjE2OTc4ZjBjYTZhIiwiZXhwIjoxNTM4NDgyMTY3fQ.KtxpHKxd7XnPoXog8rR8L6XguWHXCMN_nmOUWT8kqMw
#> 
#> Attributes:
#>  url: http://download.waterinfo.be/kiwis-auth/token
#>  type: Bearer
#>  expires: 2018-10-02 14:09:26 CEST

# get token via client
client <- paste0('MzJkY2VlY2UtODI2Yy00Yjk4LTljMmQtYjE2OTc4ZjBjYTZhOjRhZGE4',
                'NzFhLTk1MjgtNGI0ZC1iZmQ1LWI1NzBjZThmNGQyZA==')
my_token <- get_token(client = client)
print(my_token)

#> Token:
#> eyJhbGciOiJIUzI1NiJ9.eyJqdGkiOiI3M2I1MTYzYy04OTI2LTRhOTItOWIwNi0zN2U1ZmFjOGZjMTYiLCJpYXQiOjE1MzgzOTU3NjcsImlzcyI6Imh0dHA6Ly9sb2NhbGhvc3Q6ODA4MC9LaVdlYlBvcnRhbC9hdXRoIiwiYXVkIjoiMzJkY2VlY2UtODI2Yy00Yjk4LTljMmQtYjE2OTc4ZjBjYTZhIiwiZXhwIjoxNTM4NDgyMTY3fQ.g87ITrXZJ4oYg4PIXu-3L0xTm3XTnMcWjCDNnbxY5VM
#> 
#> Attributes:
#>  url: http://download.waterinfo.be/kiwis-auth/token
#>  type: Bearer
#>  expires: 2018-10-02 14:09:26 CEST

is.expired(my_token)

#> [1] FALSE

expires.in(my_token)

#> Time difference of 24 hours